### 2025-10-07 #9152 @HarrisonKramer
Hi @kyleniemeyer,

Thanks for getting this started.

On possible reviewers (from JOSS reviewer search for "optics" + Python + Physics/Engineering):
- [dihm](https://github.com/dihm)
- [Jashcraf](https://github.com/Jashcraf)
- [ldes89150](https://github.com/ldes89150)
- [szabo137](https://github.com/szabo137)
- [caldarolamartin](https://github.com/caldarolamartin)

> The paper is currently a bit long

The included code examples are distinct from the README and illustrate specific use cases. I'm happy to condense if you feel strongly, but I believe they add value at the current length.
