+++ optiland/analysis/spot_diagram.py
@@ -8,7 +8,7 @@ Kramer Harrison, 2024
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Literal
+from typing import TYPE_CHECKING, Literal
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -19,6 +19,13 @@ from optiland.utils import resolve_fields
 from optiland.visualization.system.utils import transform
 
 from .base import BaseAnalysis
+
+if TYPE_CHECKING:
+    from matplotlib.axes import Axes
+    from matplotlib.figure import Figure
+    from numpy.typing import NDArray
+
+    from optiland._types import BEArray, DistributionType
 
 
 @dataclass
@@ -60,7 +67,7 @@ class SpotDiagram(BaseAnalysis):
         fields: str | list = "all",
         wavelengths: str | list = "all",
         num_rings: int = 6,
-        distribution: str = "hexapolar",
+        distribution: DistributionType = "hexapolar",
         coordinates: Literal["global", "local"] = "local",
     ):
         """Initializes the SpotDiagram analysis.
@@ -91,7 +98,7 @@ class SpotDiagram(BaseAnalysis):
         self.coordinates = coordinates
 
         self.num_rings = num_rings
-        self.distribution = distribution
+        self.distribution: DistributionType = distribution
 
         super().__init__(optic, wavelengths)
         primary_wl_value = self.optic.primary_wavelength
@@ -106,10 +113,10 @@ class SpotDiagram(BaseAnalysis):
 
     def view(
         self,
-        fig_to_plot_on: plt.Figure = None,
+        fig_to_plot_on: Figure | None = None,
         figsize: tuple[float, float] = (12, 4),
         add_airy_disk: bool = False,
-    ) -> tuple[plt.Figure, list[plt.Axes]]:
+    ) -> tuple[Figure, list[Axes]]:
         """Displays the spot diagram plot.
 
         Args:
@@ -149,7 +156,7 @@ class SpotDiagram(BaseAnalysis):
 
     # --- Calculation Methods ---
 
-    def angle_from_cosine(self, a: be.array, b: be.array) -> float:
+    def angle_from_cosine(self, a: BEArray, b: BEArray) -> float:
         """Calculates the angle in radians between two direction cosine vectors.
 
         Args:
@@ -231,7 +238,7 @@ class SpotDiagram(BaseAnalysis):
         rays = self.generate_marginal_rays(H_x, H_y, wavelength)
         return tuple(be.array([ray.L, ray.M, ray.N]).ravel() for ray in rays)
 
-    def generate_chief_rays_cosines(self, wavelength: float) -> be.array:
+    def generate_chief_rays_cosines(self, wavelength: float) -> BEArray:
         """Generates direction cosines for the chief ray of each field.
 
         Args:
@@ -251,7 +258,7 @@ class SpotDiagram(BaseAnalysis):
         ]
         return be.stack(cosines, axis=0)
 
-    def generate_chief_rays_centers(self, wavelength: float) -> be.array:
+    def generate_chief_rays_centers(self, wavelength: float) -> BEArray:
         """Generates the (x, y) intersection points for the chief ray of each field.
 
         Args:
@@ -307,7 +314,7 @@ class SpotDiagram(BaseAnalysis):
 
         return airy_rad_x_list, airy_rad_y_list
 
-    def centroid(self) -> list[tuple[be.array, be.array]]:
+    def centroid(self) -> list[tuple[BEArray, BEArray]]:
         """Calculates the geometric centroid of each spot for the reference wavelength.
 
         Returns:
@@ -319,7 +326,7 @@ class SpotDiagram(BaseAnalysis):
[…]
