### 2025-10-07 cc3d110
feat: Add thickness attribute to Surface serialization and deserialization
