### 2025-10-07 7a2feca
Fix/rel cs loading (#359)
* feat: Add thickness attribute to Surface serialization and deserialization

* test: Add test for removing surface after loading from file
