### 2025-10-12 a50bd46
fix: Fix thickness updates in SurfaceGroup after flipping surfaces
