### 2025-10-12 c4ea918
Update optiland/surfaces/surface_group.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
