### 2025-10-12 e502dc6
fix: Fix thickness updates in SurfaceGroup after flipping surfaces (#366)
* fix: Fix thickness updates in SurfaceGroup after flipping surfaces

* Update optiland/surfaces/surface_group.py

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
