### 2025-10-12 058496e
fix: Ensure material linkage updates correctly when removing a surface
