### 2025-10-12 f967462
fix: Update material properties during surface removal and add corresponding test
