### 2025-10-12 d96bbb1
Fix/surface removal material setting (#367)
* fix: Ensure material linkage updates correctly when removing a surface

* fix: Update material properties during surface removal and add corresponding test
