### 2025-10-12 ef7edd8
remove multiple copies of materials across objects
# Conflicts:
#	optiland/surfaces/surface_group.py
