+++ optiland/surfaces/surface_group.py
@@ -11,6 +11,8 @@ Kramer Harrison, 2024
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
+from contextlib import suppress
+from functools import cached_property
 
 import optiland.backend as be
 from optiland.coatings import BaseCoatingPolarized
@@ -31,7 +33,7 @@ class SurfaceGroup:
 
     """
 
-    def __init__(self, surfaces: list[Surface] = None):
+    def __init__(self, surfaces: list[Surface] | None = None):
         """Initializes a new instance of the SurfaceGroup class.
 
         Args:
@@ -40,11 +42,23 @@ class SurfaceGroup:
 
         """
         if surfaces is None:
-            self.surfaces = []
+            self._surfaces = []
         else:
-            self.surfaces = surfaces
+            self._surfaces = surfaces
+            self._update_surface_links()
 
         self.surface_factory = SurfaceFactory(self)
+
+    def _update_surface_links(self):
+        with suppress(KeyError):
+            self.__dict__.pop("surfaces")
+        surfaces = self._surfaces
+        if surfaces:
+            surfaces[0].previous_surface = None
+
+            if len(surfaces) > 1:
+                for idx, surface in enumerate(surfaces[1:]):
+                    surface.previous_surface = surfaces[idx]
 
     def __add__(self, other):
         """Add two SurfaceGroup objects together.
@@ -53,7 +67,7 @@ class SurfaceGroup:
         surface of the other group.
         """
         # add the offset of the last surface in self to each surface in other
-        offset = self.surfaces[-1].geometry.cs.z
+        offset = self.surfaces[-1].geometry.cs.z if self.surfaces else 0.0
 
         # add object surface distance if finite
         object_distance = other.surfaces[0].geometry.cs.z
@@ -67,7 +81,11 @@ class SurfaceGroup:
         for surface in other.surfaces:
             surface.is_stop = False
 
-        return SurfaceGroup(self.surfaces[:-1] + other.surfaces[1:])
+        return SurfaceGroup(self._surfaces[:-1] + other._surfaces[1:])
+
+    @cached_property
+    def surfaces(self):
+        return tuple(item for item in self._surfaces)
 
     @property
     def x(self):
@@ -263,34 +281,31 @@ class SurfaceGroup:
         new_surface.thickness = kwargs.get("thickness", 0.0)
         self.surface_factory.material_factory.last_material = new_surface.material_post
 
-        if new_surface.is_stop:
-            for surface in self.surfaces:
-                surface.is_stop = False
-
         if index is None:
-            self.surfaces.append(new_surface)
+            self._surfaces.append(new_surface)
+            self._update_surface_links()
+            index = len(self._surfaces) - 1
         else:
             if index < 0:
                 raise IndexError(f"Index {index} cannot be negative.")
-            if index > len(self.surfaces):
+            if index > len(self._surfaces):
                 raise IndexError(
                     f"Index {index} is out of bounds for insertion. "
-                    f"Max index for insertion is {len(self.surfaces)} (to append)."
+                    f"Max index for insertion is {len(self._surfaces)} (to append)."
                 )
 
-            self.surfaces.insert(index, new_surface)
-
-            # If a surface was inserted (not appended) and there's a surface after it
-            if index < len(self.surfaces) - 1:
-                surface_after_inserted = self.surfaces[index + 1]
-                new_surface = self.surfaces[index]
-                surface_after_inserted.material_pre = new_surface.material_post
+            self._surfaces.insert(index, new_surface)
+            self._update_surface_links()
 
             # Update coordinate systems if surface was inserted
             if not self.surface_factory.use_absolute_cs and index < (
-                len(self.surfaces) - 1
+                len(self._surfaces) - 1
             ):
[…]
