+++ tests/test_thin_lens_interaction_model.py
@@ -23,23 +23,23 @@ def surface(set_test_backend):
     coating = SimpleCoating(0.5, 0.5)
     bsdf = None
     interaction_model = ThinLensInteractionModel(
+        parent_surface=None,
         focal_length=focal_length,
-        geometry=geometry,
-        material_pre=material_pre,
-        material_post=material_post,
         is_reflective=True,
         coating=coating,
         bsdf=bsdf,
     )
-    return Surface(
+    surf = Surface(
+        previous_surface=None,
         geometry=geometry,
-        material_pre=material_pre,
         material_post=material_post,
         is_stop=True,
         aperture=aperture,
         interaction_model=interaction_model,
         surface_type="paraxial",
     )
+    interaction_model.parent_surface = surf
+    return surf
 
 
 class TestThinLensInteractionModel:
