### 2025-10-13 8d5f3b6
Fix type annotations and improve docstrings (#361)
* Fix type annotations and improve docstrings (#360)

This commit resolves static type checking errors and extends type
annotations
in various files. Minor syntax changes lead to code reduction while a
few
None-type checks are added to help static analysis. No logic is changed
and
all tests are passing.

Changes:
- separate BEArray and ScalarOrArray TypeVar (BEArrayT, ScalarOrArrayT)
    from type alias (BEArray, ScalarOrArray)
- define BEArray in TYPE_CHECKING block to keep torch optional
- add and update Google-style docstrings in various fi
[…]
