+++ optiland/rays/real_rays.py
@@ -14,33 +14,63 @@ import optiland.backend as be
 from optiland.rays.base import BaseRays
 
 if TYPE_CHECKING:
+    from numpy.typing import ArrayLike
+
+    from optiland._types import BEArray, ScalarOrArray
     from optiland.materials import BaseMaterial
 
 
 class RealRays(BaseRays):
     """Represents a collection of real rays in 3D space.
 
+    This class stores ray positions, directions, and properties as arrays,
+    supporting both NumPy and PyTorch backends for efficient computation.
+
     Attributes:
-        x (ndarray): The x-coordinates of the rays.
-        y (ndarray): The y-coordinates of the rays.
-        z (ndarray): The z-coordinates of the rays.
-        L (ndarray): The x-components of the direction vectors of the rays.
-        M (ndarray): The y-components of the direction vectors of the rays.
-        N (ndarray): The z-components of the direction vectors of the rays.
-        i (ndarray): The intensity of the rays.
-        w (ndarray): The wavelength of the rays.
-        opd (ndarray): The optical path length of the rays.
+        x: x-coordinates of ray positions.
+        y: y-coordinates of ray positions.
+        z: z-coordinates of ray positions.
+        L: x-components of ray direction cosines.
+        M: y-components of ray direction cosines.
+        N: z-components of ray direction cosines.
+        i: Intensity values of the rays.
+        w: Wavelength values of the rays.
+        opd: Optical path difference values.
+        L0: Pre-surface x-direction cosines (None until surface interaction).
+        M0: Pre-surface y-direction cosines (None until surface interaction).
+        N0: Pre-surface z-direction cosines (None until surface interaction).
 
-    Methods:
-        rotate_x(rx: float): Rotate the rays about the x-axis.
-        rotate_y(ry: float): Rotate the rays about the y-axis.
-        rotate_z(rz: float): Rotate the rays about the z-axis.
-        propagate(t: float): Propagate the rays a distance t.
-        clip(condition): Clip the rays based on a condition.
-
+    Note:
+        Direction cosines should be normalized: L² + M² + N² = 1.
     """
 
-    def __init__(self, x, y, z, L, M, N, intensity, wavelength):
+    def __init__(
+        self,
+        x: ArrayLike,
+        y: ArrayLike,
+        z: ArrayLike,
+        L: ArrayLike,
+        M: ArrayLike,
+        N: ArrayLike,
+        intensity: ArrayLike,
+        wavelength: ArrayLike,
+    ):
+        """Initialize a collection of real rays in 3D space.
+
+        Args:
+            x: The x-coordinates of the ray starting positions.
+            y: The y-coordinates of the ray starting positions.
+            z: The z-coordinates of the ray starting positions.
+            L: The x-components of the ray direction cosines.
+            M: The y-components of the ray direction cosines.
+            N: The z-components of the ray direction cosines.
+            intensity: The intensity values of the rays.
+            wavelength: The wavelength values of the rays.
+
+        Note:
+            All input arrays are converted to 1D arrays. Direction cosines
+            (L, M, N) should be normalized such that L² + M² + N² = 1.
+        """
         self.x = be.as_array_1d(x)
         self.y = be.as_array_1d(y)
         self.z = be.as_array_1d(z)
@@ -52,49 +82,55 @@ class RealRays(BaseRays):
         self.opd = be.zeros_like(self.x)
 
         # variables to hold pre-surface direction cosines
-        self.L0 = None
-        self.M0 = None
[…]
