### 2025-10-13 ea9c347
docs: Add Littie28 to the list of contributors
