### 2025-10-13 978c2ea
This change finalizes the NURBS geometry implementation.
The NURBS code in `optiland/geometries/nurbs/` and the `NurbsPointsVariable` in `optiland/optimization/variable/nurbs.py` have been cleaned up to pass ruff formatting and linting checks. Google-style docstrings and inline comments have been added to improve readability, and old code and comments have been removed.

Documentation has been added to the `.rst` files in `docs/` to include references to the new NURBS functionality.

Tests have been added for the NURBS functionality, including geometries, variables, and the geometry factory. T
[…]
