+++ optiland/geometries/nurbs/nurbs_fitting.py
@@ -1,8 +1,10 @@
-"""
-Define functions used to find the NURBS parameters that approximate a given set of points.
-Most of the code is derived from nurbs-geomdl package written by Onur R. Bingol <contact@onurbingol.net>
+"""This module defines functions for NURBS surface fitting.
 
-Matteo Taccola,2025
+This module defines functions used to find the NURBS parameters that
+approximate a given set of points. Most of the code is derived from the
+nurbs-geomdl package written by Onur R. Bingol <contact@onurbingol.net>.
+
+Matteo Taccola, 2025
 """
 
 from __future__ import annotations
@@ -15,73 +17,53 @@ from .nurbs_basis_functions import basis_function, basis_function_one
 
 
 def approximate_surface(points, size_u, size_v, degree_u, degree_v, **kwargs):
-    """Surface approximation using least squares method with fixed number of control points.
+    """Approximates a surface using a least-squares method.
 
-    This algorithm interpolates the corner control points and approximates the remaining control points. Please refer to
-    Algorithm A9.7 of The NURBS Book (2nd Edition), pp.422-423 for details.
+    This algorithm interpolates the corner control points and approximates the
+    remaining control points. Please refer to Algorithm A9.7 of The NURBS Book
+    (2nd Edition), pp.422-423 for details.
 
-    Keyword Arguments:
-        * ``centripetal``: activates centripetal parametrization method. *Default: False*
-        * ``ctrlpts_size_u``: number of control points on the u-direction. *Default: size_u - 1*
-        * ``ctrlpts_size_v``: number of control points on the v-direction. *Default: size_v - 1*
+    Args:
+        points: The data points.
+        size_u: The number of data points on the u-direction, :math:`r`.
+        size_v: The number of data points on the v-direction, :math:`s`.
+        degree_u: The degree of the output surface for the u-direction.
+        degree_v: The degree of the output surface for the v-direction.
+        **kwargs: Keyword arguments.
 
-    :param points: data points
-    :type points: list, tuple
-    :param size_u: number of data points on the u-direction, :math:`r`
-    :type size_u: int
-    :param size_v: number of data points on the v-direction, :math:`s`
-    :type size_v: int
-    :param degree_u: degree of the output surface for the u-direction
-    :type degree_u: int
-    :param degree_v: degree of the output surface for the v-direction
-    :type degree_v: int
-    :return: approximated B-Spline surface
-    :rtype: BSpline.Surface
+    Returns:
+        The approximated B-Spline surface.
     """
-    # Keyword arguments
     use_centripetal = kwargs.get("centripetal", False)
-    num_cpts_u = kwargs.get(
-        "ctrlpts_size_u", size_u - 1
-    )  # number of datapts, r + 1 > number of ctrlpts, n + 1
-    num_cpts_v = kwargs.get(
-        "ctrlpts_size_v", size_v - 1
-    )  # number of datapts, s + 1 > number of ctrlpts, m + 1
+    num_cpts_u = kwargs.get("ctrlpts_size_u", size_u - 1)
+    num_cpts_v = kwargs.get("ctrlpts_size_v", size_v - 1)
 
-    # Dimension
     dim = len(points[0])
 
-    # Get uk and vl
     uk, vl = compute_params_surface(points, size_u, size_v, use_centripetal)
 
-    # Compute knot vectors
     kv_u = compute_knot_vector(degree_u, size_u, num_cpts_u, uk)
[…]
