### 2025-10-13 3ef92c3
refactor: Simplify expressions and improve readability in NURBS functions
