### 2025-10-13 0eb6c60
test: Remove skip decorator and update test_nurbs_factory to use set_test_backend
