### 2025-10-14 982861b
`Surface:_from_dict`: fix obsolete geometry param
