+++ optiland/rays/real_rays.py
@@ -130,21 +130,6 @@ class RealRays(BaseRays):
             self.L * be.sin(rz) + self.M * be.cos(rz),
         )
 
-    def propagate(self, t: float, material: BaseMaterial | None = None):
-        """Propagate the rays a distance t."""
-        self.x = self.x + t * self.L
-        self.y = self.y + t * self.M
-        self.z = self.z + t * self.N
-
-        if material is not None:
-            k = material.k(self.w)
-            alpha = 4 * be.pi * k / self.w
-            self.i = self.i * be.exp(-alpha * t * 1e3)  # mm to microns
-
-        # normalize, if required
-        if not self.is_normalized:
-            self.normalize()
-
     def clip(self, condition: BEArray):
         """Clip the rays based on a condition."""
         cond = be.array(condition)
