+++ optiland/materials/material_file.py
@@ -31,6 +31,9 @@ class MaterialFile(BaseMaterial):
 
     Args:
         filename (str): The path to the material file.
+        propagation_model (BasePropagationModel, optional): The propagation
+            model to use. Defaults to None, which creates a
+            HomogeneousPropagation model.
 
     Attributes:
         filename (str): The filename of the material file.
@@ -45,8 +48,8 @@ class MaterialFile(BaseMaterial):
 
     """
 
-    def __init__(self, filename):
-        super().__init__()
+    def __init__(self, filename, propagation_model=None):
+        super().__init__(propagation_model)
         self.filename = filename
         self._k_warning_printed = False
         self.coefficients = []
@@ -555,4 +558,14 @@ class MaterialFile(BaseMaterial):
             raise ValueError("Material file data missing filename.")
 
         material = cls(data["filename"])
+
+        propagation_model_data = data.get("propagation_model")
+        if (
+            propagation_model_data
+            and propagation_model_data.get("class") == "GrinPropagation"
+        ):
+            from optiland.propagation.grin import GrinPropagation
+
+            material.propagation_model = GrinPropagation()
+
         return material

[…]
