### 2025-10-14 efdd4b6
fix: Update type annotations for propagation model parameters
