+++ optiland/propagation/base.py
@@ -1,6 +1,9 @@
 """Abstract base class for propagation models."""
+
+from __future__ import annotations
+
 import abc
-from typing import TYPE_CHECKING, Dict, Any
+from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
     from optiland.rays.real_rays import RealRays
@@ -10,7 +13,7 @@ class BasePropagationModel(abc.ABC):
     """Abstract base class for all propagation models."""
 
     @abc.abstractmethod
-    def propagate(self, rays: 'RealRays', t: float) -> None:
+    def propagate(self, rays: RealRays, t: float) -> None:
         """Propagates rays a distance t through the medium.
 
         This method modifies the ray state in-place.
@@ -21,17 +24,17 @@ class BasePropagationModel(abc.ABC):
         """
         raise NotImplementedError
 
-    def to_dict(self) -> Dict[str, Any]:
+    def to_dict(self) -> dict[str, Any]:
         """Serializes the propagation model to a dictionary.
 
         By default, this only stores the class name, as the model is
         reconstructed and linked by the parent material during its own
         deserialization.
         """
-        return {'class': self.__class__.__name__}
+        return {"class": self.__class__.__name__}
 
     @classmethod
-    def from_dict(cls, d: Dict[str, Any]) -> 'BasePropagationModel':
+    def from_dict(cls, d: dict[str, Any]) -> BasePropagationModel:
         """Deserializes a propagation model from a dictionary.
 
         Note: This is part of the common serialization interface but should
