+++ optiland/propagation/grin.py
@@ -1,4 +1,7 @@
 """Placeholder for Graded-Index (GRIN) propagation model."""
+
+from __future__ import annotations
+
 from typing import TYPE_CHECKING
 
 from optiland.propagation.base import BasePropagationModel
@@ -13,13 +16,11 @@ class GrinPropagation(BasePropagationModel):
     This model is not yet implemented and will raise an error if used.
     """
 
-    def propagate(self, rays: 'RealRays', t: float) -> None:
+    def propagate(self, rays: RealRays, t: float) -> None:
         """Raises NotImplementedError.
 
         Args:
             rays: The rays object to be propagated.
             t: The distance to propagate.
         """
-        raise NotImplementedError(
-            "GRIN propagation is not yet implemented."
-        )
+        raise NotImplementedError("GRIN propagation is not yet implemented.")
