### 2025-10-14 aab5de3
refactor: Update docstrings to include author information and improve clarity
