### 2025-10-14 a3c0366
refactor: Rename GrinPropagation to GRINPropagation for consistency
