### 2025-10-14 cd8fdf1
refactor: Improve material propagation model handling and deserialization
