+++ optiland/materials/ideal.py
@@ -95,22 +95,4 @@ class IdealMaterial(BaseMaterial):
             IdealMaterial: The material.
 
         """
-        propagation_model_data = data.get("propagation_model")
-        # This is a temporary solution. A more robust factory would be needed
-        # for multiple propagation models.
-        if (
-            propagation_model_data
-            and propagation_model_data.get("class") == "GrinPropagation"
-        ):
-            # The material needs to be passed to the propagation model, but
-            # it hasn't been created yet. We create the material first, and
-            # the BaseMaterial __init__ handles creating the default model.
-            # Then we can replace it. This is a bit clunky.
-            # A better solution would be a true factory for materials.
-            from optiland.propagation.grin import GrinPropagation
-
-            material = cls(data["index"], data.get("absorp", 0))
-            material.propagation_model = GrinPropagation()
-            return material
-
         return cls(data["index"], data.get("absorp", 0))
