+++ tests/propagation/test_homogeneous.py
@@ -6,29 +6,26 @@ from optiland import backend as be
 from optiland.materials.ideal import IdealMaterial
 from optiland.propagation.homogeneous import HomogeneousPropagation
 from optiland.rays.real_rays import RealRays
+from ..utils import assert_allclose
 
 
-@pytest.fixture
-def basic_rays():
-    """A simple RealRays fixture for testing."""
-    return RealRays(x=[0], y=[0], z=[0], L=[0], M=[0], N=[1], intensity=[1.0], wavelength=[0.5])
-
-
-def test_homogeneous_propagation_position_update(basic_rays):
+def test_homogeneous_propagation_position_update(set_test_backend):
     """Verify that ray coordinates are updated correctly."""
+    basic_rays = RealRays(x=[0], y=[0], z=[0], L=[0], M=[0], N=[1], intensity=[1.0], wavelength=[0.5])
     material = IdealMaterial(n=1.0)  # k=0 by default
     model = HomogeneousPropagation(material)
     t = 10.0
 
     model.propagate(basic_rays, t)
 
-    assert be.allclose(basic_rays.x, be.array([0.0]))
-    assert be.allclose(basic_rays.y, be.array([0.0]))
-    assert be.allclose(basic_rays.z, be.array([10.0]))
+    assert_allclose(basic_rays.x, be.array([0.0]))
+    assert_allclose(basic_rays.y, be.array([0.0]))
+    assert_allclose(basic_rays.z, be.array([10.0]))
 
 
-def test_homogeneous_propagation_no_attenuation_with_k0(basic_rays):
+def test_homogeneous_propagation_no_attenuation_with_k0(set_test_backend):
     """Verify ray intensity is unchanged when k=0."""
+    basic_rays = RealRays(x=[0], y=[0], z=[0], L=[0], M=[0], N=[1], intensity=[1.0], wavelength=[0.5])
     material = IdealMaterial(n=1.0, k=0.0)
     model = HomogeneousPropagation(material)
     t = 10.0
@@ -36,11 +33,12 @@ def test_homogeneous_propagation_no_attenuation_with_k0(basic_rays):
     initial_intensity = be.copy(basic_rays.i)
     model.propagate(basic_rays, t)
 
-    assert be.allclose(basic_rays.i, initial_intensity)
+    assert_allclose(basic_rays.i, initial_intensity)
 
 
-def test_homogeneous_propagation_attenuation_with_k_gt_0(basic_rays):
+def test_homogeneous_propagation_attenuation_with_k_gt_0(set_test_backend):
     """Verify ray intensity is correctly attenuated when k > 0."""
+    basic_rays = RealRays(x=[0], y=[0], z=[0], L=[0], M=[0], N=[1], intensity=[1.0], wavelength=[0.5])
     k_val = 0.1
     wavelength = be.to_numpy(basic_rays.w)[0]
     t = 10.0  # distance in mm
@@ -58,11 +56,12 @@ def test_homogeneous_propagation_attenuation_with_k_gt_0(basic_rays):
     alpha = 4 * be.pi * k_val / wavelength
     expected_intensity = initial_intensity * be.exp(-alpha * t * 1e3)
 
-    assert be.allclose(basic_rays.i, be.array([expected_intensity]))
+    assert_allclose(basic_rays.i, be.array([expected_intensity]))
 
 
-def test_homogeneous_propagation_normalizes_rays(basic_rays):
+def test_homogeneous_propagation_normalizes_rays(set_test_backend):
     """Verify that unnormalized rays are normalized after propagation."""
[…]
