### 2025-10-14 322afbe
refactor: Enhance serialization tests and improve factory pattern for propagation models
