### 2025-10-14 29befae
Feature/propagation (#371)
* This refactoring decouples the ray propagation logic from the `RealRays` class into a flexible, pluggable `PropagationModel` system.

Key changes:
- A new `optiland/propagation` module is introduced with a `BasePropagationModel` ABC.
- `HomogeneousPropagation` is implemented for straight-line propagation and absorption.
- A placeholder `GrinPropagation` model is added for future development.
- `BaseMaterial` and all its subclasses now own a `propagation_model`, defaulting to `HomogeneousPropagation`.
- The core ray tracing engine in `Surface.trace` now uses the pro
[…]
