### 2025-10-15 0701ce6
Docs: Comprehensive update of Developer's Guide
This commit provides a comprehensive update to the developer's guide to align it with the current codebase, add missing information, and improve clarity.

Key changes include:

- Corrected outdated architectural descriptions in `architecture.rst`, `surface_overview.rst`, `ray_overview.rst`, and `geometry_overview.rst`.
- Clarified the backend's dynamic dispatch mechanism in `configurable_backend.rst`.
- Updated `gui.rst` to include the `ActionManager` and `PanelManager` and clarify the undo/redo process.

New documentation added:

- Created `code
[…]
