### 2025-10-15 f1e456b
docs: Update code structure and geometry overview for clarity; fix GRIN capitalization in propagation models
