+++ docs/developers_guide/geometry_overview.rst
@@ -19,9 +19,8 @@ A geometry is defined by:
 
 - **Geometry-Specific Parameters**: Additional attributes tailored to the surface type. For example:
 
-  - **Spherical Geometry**: Radius of curvature.
-  - **Conic Geometry**: Conic constant and radius of curvature.
-  - **Aspheric Geometry**: Coefficients for higher-order terms (e.g., polynomial aspheric terms).
+- **StandardGeometry**: This class in `standard.py` handles spherical and conic surfaces.
+- **Aspheric Geometries**: Implemented in `even_asphere.py` and `odd_asphere.py`.
   - **Custom Geometries**: User-defined parameters. This flexibility allows for a wide range of surface shapes, including freeforms.
 
 .. note::
@@ -34,9 +33,9 @@ Geometries provide methods for:
 
 1. **Finding the Intersection Point**:
 
-   - For simple shapes (e.g., planes, spheres, conics), this is computed using closed-form equations.
-   - For complex shapes (e.g., polynomial XY, Chebyshev surfaces, or aspheres), iterative methods like Newton-Raphson are used to find the intersection.
-   - A **Newton-Raphson Base Class** is included for geometries that require iterative solutions. Simply subclass this base class and implement the required methods for your geometry.
+   - For simple shapes, this is computed using closed-form equations.
+   - For complex shapes, iterative methods like Newton-Raphson are used to find the intersection.
+   - The `NewtonRaphsonGeometry` class serves as the base for geometries requiring iterative solutions.
 
 2. **Computing Surface Normals**:
 
@@ -47,12 +46,14 @@ Supported Geometry Types
 
 Optiland includes a wide range of built-in geometries:
 
+- **Standard**: Handles spherical and conic surfaces.
 - **Planes**: Flat surfaces with infinite or finite extent.
-- **Spheres and Conics**: Defined by radius of curvature and conic constant.
-- **Even Aspheres**: Described by polynomial terms for deviations from a sphere.
-- **Odd Aspheres**: Similar to even aspheres but with additional terms for odd powers.
-- **Polynomial XY and Chebyshev Surfaces**: Useful for advanced freeform optical systems.
-- **Toridal Surfaces**: Defined by two radii of curvature, allowing for toroidal shapes.
+- **EvenAsphere**: Described by polynomial terms for deviations from a sphere.
+- **OddAsphere**: Similar to even aspheres but with additional terms for odd powers.
+- **Biconic**: A surface with different radii and conic constants in x and y.
+- **Toroidal**: Defined by two radii of curvature, allowing for toroidal shapes.
+- **PlaneGrating and StandardGrating**: Surfaces with diffraction gratings.
+- **Polynomial and Chebyshev**: Useful for advanced freeform optical systems.
 - **Zernike Surfaces**: Represented by Zernike polynomials. For a detailed mathematical description of the Zernike geometry, see the `Zernike Geometry Mathematics Reference <https://github.com/HarrisonKramer/optiland/blob/master/docs/references/zernike_description.md>`_.
 - **Forbes Surfaces**: As described in the corresponding `Forbes Surface` gallery example, these surfaces are defined following the convention by the papers: [1] G. W. Forbes, “Manufacturability estimates for optical aspheres,” Opt. Express 19(10), 9923–9941 (2011) and [2] G. W. Forbes, "Characterizing the shape of freeform optics," Opt. Express 20, 2483-2499 (2012). The `qpoly.py` module in `optiland\geometries\forbes` was adapted from the implementation in the `prysm <https://github.com/brandondube/prysm?tab=readme-ov-file>` package.
 - **Custom Geometries**: Users can easily extend the framework by subclassing the `BaseGeometry` (analytical geometries) or `NewtonRaphsonGeometry` (iterative geometries) classes.
