+++ docs/developers_guide/architecture.rst
@@ -27,15 +27,15 @@ coordinates interactions between key components. It encapsulates:
 - **Field Group**: A collection of defined field points for ray tracing and analysis. These define the extent of the object.
 - **Wavelength Group**: The set of wavelengths used in calculations and ray tracing.
 - **Surface Group**: A collection of surfaces that define the optical system.
-- **Ray Generator**: A module for generating rays based on the system's parameters.
-- **Paraxial Helper**: A utility for performing paraxial calculations.
-- **Aberration Helper**: A utility for computing various aberration metrics.
+- **RealRayTracer**: A module for tracing real rays through the system.
+- **Paraxial**: A helper class for performing paraxial calculations.
+- **Aberrations**: A helper class for computing various aberration metrics.
 - **Pickup and Solve Managers**: Tools for linking parameters between surfaces and solving specific system constraints.
 
 Overview of Key Modules
 -----------------------
 
-- **Optic Class**: The central container for an optical system. It manages fields, wavelengths, surfaces, ray generation, and utility tools like paraxial calculations and aberration helpers. The `Optic` class serves as the interface for operations such as optimization, analysis, and tolerancing.
+- **Optic Class**: The central container for an optical system. It manages fields, wavelengths, surfaces, and utility tools like paraxial calculations and aberration helpers. The `Optic` class serves as the interface for operations such as optimization, analysis, and tolerancing.
 - **Surface Group**: A collection of surfaces within the `Optic` class. Each surface defines an optical interface with attributes like geometry, materials, coatings, and optional apertures.
 - **Ray Tracing Framework**: Manages ray generation and propagation through the optical system, supporting real, paraxial, and polarized rays.
 - **Analysis Framework**: Provides tools to evaluate system performance, such as spot diagrams, ray fans, and wavefront aberrations.
@@ -47,7 +47,7 @@ Component Interactions
 
 The components of Optiland are designed to work seamlessly together, with the `Optic` class orchestrating most interactions:
 
-- Rays are generated via the ray generator in the `Optic` class and traced through the system via the surface group.
+- Rays are generated via the `RealRayTracer` and traced through the system via the surface group.
 - Optimization relies on surface attributes (e.g., radius of curvature) as variables and ray tracing results for evaluating operands.
 - Analysis tools, such as spot diagrams and ray fans, operate on data generated by the ray tracing framework.
 - Visualization modules query the state of the `Optic` class, rendering components and simulation results.
