+++ docs/developers_guide/ray_overview.rst
@@ -14,8 +14,8 @@ Rays are the fundamental elements of the ray tracing process. A ray consists of 
 - **Direction (L, M, N)**: The ray's unit vector indicating its direction.
 - **Wavelength**: The wavelength of the ray in the system.
 - **Intensity**: The ray's relative intensity.
-- **Optical Path Length (OPL)**: The accumulated path the ray has traveled, weighted by refractive index.
-- **Polarization Matrix (optional)**: A 3x3 matrix representing the transformation of a ray's initial electric field into its final state.
+- **Optical Path Difference (OPD)**: The accumulated path difference of the ray relative to the chief ray.
+- **Polarization Matrix (optional)**: A 3x3 matrix representing the transformation of a ray's initial electric field into its final state. This is used by the `PolarizedRays` class.
 
 .. note::
   In Optiland, all ray attributes are defined as NumPy arrays for efficient computation.
@@ -39,18 +39,19 @@ The ray tracing framework uses a **Ray Generator** to produce rays for tracing. 
 - System properties (e.g., F/#, NA, telecentricity)
 - Apodization (intensity distribution within the pupil)
 
-Generated rays are passed to the **Surface Group** for tracing through the optical system. Each `Optic` instance has both a ray generator and a surface group
-specific to that system.
+Generated rays are passed to the **Surface Group** for tracing through the optical system. Each `Optic` instance has a `ray_tracer` attribute, which in turn contains the `RayGenerator`.
 
 Ray Tracing
 -----------
 For real rays, the ray tracing process is managed by the **RealRayTracer**. The ray tracer is responsible for:
 
-- Generating the appropriate rays based on the selected ray generator.
+- Generating the appropriate rays via its `RayGenerator`.
 - Ray tracing input validation.
 - Propagating rays through the surface group.
 
-For paraxial rays, the **ParaxialRayTracer** is used. This tracer is similar in form to the RealRayTracer but uses simplified equations for faster computations.
+For paraxial rays, there are two primary methods for tracing:
+- The **ParaxialRayTracer** class provides a dedicated tracer for paraxial rays.
+- The `Surface.trace` method can also directly handle `ParaxialRays`, providing an alternative way to trace them.
 
 Tracing Process
 ---------------
