### 2025-10-15 52fc2d7
docs: Update README and gallery to include NURBS examples; refactor NURBS variable classes to improve scaling handling
