### 2025-10-15 8b66678
feat: Finalize NURBS implementation and enhance test coverage
This change finalizes the NURBS implementation in Optiland.

It ensures that the NURBS functionalities are well-tested, documented, and adhere to the project's coding standards.

- All previously skipped NURBS-related tests have been fixed and enabled.
- New tests have been added to increase test coverage for the `NurbsGeometry` class.
- All tests are now parameterized to run against all available backends (NumPy and PyTorch).
- A segmentation fault in the NURBS derivative calculation has been fixed.
- Docstrings and code style have
[…]
