### 2025-10-15 0126ef1
Update optiland/geometries/nurbs/nurbs_basis_functions.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
