### 2025-10-15 2f5274e
Update optiland/surfaces/factories/surface_factory.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
