+++ optiland/surfaces/factories/surface_factory.py
@@ -123,7 +123,6 @@ class SurfaceFactory:
             zernike_type=kwargs.get("zernike_type", "fringe"),
             radial_terms=kwargs.get("radial_terms"),
             freeform_coeffs=kwargs.get("freeform_coeffs"),
-            # forbes_norm_radius=kwargs.get("forbes_norm_radius", 1.0),
             nurbs_norm_x=kwargs.get("nurbs_norm_x"),
             nurbs_norm_y=kwargs.get("nurbs_norm_y"),
             nurbs_x_center=kwargs.get("nurbs_x_center", 0.0),
