### 2025-10-15 4cdc1ce
Update optiland/geometries/nurbs/nurbs_basis_functions.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
