### 2025-10-15 338b807
fix: improve error messages in compute_basis_polynomials_derivatives function
