### 2025-10-15 e4a2a3a
feat: Finalize NURBS implementation and enhance test coverage
This change finalizes the NURBS implementation in Optiland.

It ensures that the NURBS functionalities are well-tested, documented, and adhere to the project's coding standards.

- All previously skipped NURBS-related tests have been fixed and enabled.
- New tests have been added to increase test coverage for the `NurbsGeometry` class.
- All tests are now parameterized to run against all available backends (NumPy and PyTorch).
- Fixed multiple backend compatibility issues in the NURBS implementation, including a segmentation fault i
[…]
