+++ tests/test_nurbs_factory.py
@@ -5,6 +5,7 @@ from optiland import backend as be
 from optiland.geometries.nurbs.nurbs_geometry import NurbsGeometry
 from optiland.materials.ideal import IdealMaterial
 from optiland.surfaces.factories.surface_factory import SurfaceFactory
+from tests.utils import assert_allclose
 
 
 class MockSurface:
@@ -18,7 +19,9 @@ class MockSurfaceGroup:
         self.surfaces = [MockSurface()]
 
 
-def test_nurbs_factory(set_test_backend):
+@pytest.mark.parametrize("backend", be.list_available_backends())
+def test_nurbs_factory(set_test_backend, backend):
+    be.set_backend(backend)
     config = {
         "surface_type": "NurbsGeometry",
         "radius": 100,
@@ -31,5 +34,5 @@ def test_nurbs_factory(set_test_backend):
     factory = SurfaceFactory(surface_group=MockSurfaceGroup(2))
     surf = factory.create_surface(**config)
     assert isinstance(surf.geometry, NurbsGeometry)
-    assert be.allclose(surf.geometry.radius, 100)
-    assert be.allclose(surf.geometry.k, -1)
+    assert_allclose(be.asarray(surf.geometry.radius), be.asarray(100))
+    assert_allclose(be.asarray(surf.geometry.k), be.asarray(-1))
