### 2025-10-15 c38f75f
fix: standardize surface type string in NURBS factory test and update assertions
