### 2025-10-15 a7d0f2e
fix: improve docstring formatting for random_uniform functions in numpy and torch backends refactor: remove unused import in nurbs_basis_functions.py
