### 2025-10-15 c3a40a9
fix: remove backend parameterization from NURBS geometry tests for consistency
