### 2025-10-15 c0f84ef
Fix NURBS tests for torch backend
This commit fixes several issues that caused the NURBS tests to fail when using the torch backend.

The main changes are:
- Conditionally disable Numba JIT compilation when the torch backend is active to prevent segmentation faults.
- Refactor NURBS geometry methods to use out-of-place tensor operations, avoiding errors with PyTorch's autograd.
- Ensure consistent float64 data types in tests to prevent mismatches.
- Add torch to the development dependencies in pyproject.toml.
