+++ optiland/geometries/nurbs/nurbs_geometry.py
@@ -13,6 +13,7 @@ Matteo Taccola, 2025
 
 from __future__ import annotations
 
+import numpy as np
 from scipy.special import binom
 
 import optiland.backend as be
@@ -357,10 +358,10 @@ class NurbsGeometry(BaseGeometry):
         m = mm - 1
 
         if be.get_backend() == "torch":
-            np_u = be.to_numpy(u)
-            np_U = be.to_numpy(U)
-            np_v = be.to_numpy(v)
-            np_V = be.to_numpy(V)
+            np_u = be.to_numpy(u).copy().astype(np.float64)
+            np_U = be.to_numpy(U).copy().astype(np.float64)
+            np_v = be.to_numpy(v).copy().astype(np.float64)
+            np_V = be.to_numpy(V).copy().astype(np.float64)
             N_basis_u_np = compute_basis_polynomials(n, p, np_U, np_u)
             N_basis_v_np = compute_basis_polynomials(m, q, np_V, np_v)
             N_basis_u = be.asarray(N_basis_u_np)
@@ -499,33 +500,32 @@ class NurbsGeometry(BaseGeometry):
         w_ders = bspline_derivatives[:, :, [-1], :]
 
         n_dim, N = be.shape(P)[0], be.size(u)
-        nurbs_derivatives = be.zeros((up_to_order_u + 1, up_to_order_v + 1, n_dim, N))
 
+        rows = []
         for k in range(up_to_order_u + 1):
+            cols = []
             for L in range(up_to_order_v + 1):
-                temp_numerator = A_ders[[k], [L], ...]
+                temp_numerator = A_ders[k, L, ...]
 
                 for i in range(1, k + 1):
-                    temp_numerator -= (
-                        binom(k, i)
-                        * w_ders[[i], [0], ...]
-                        * nurbs_derivatives[[k - i], [L], ...]
+                    temp_numerator = temp_numerator - (
+                        binom(k, i) * w_ders[i, 0, ...] * rows[k - i][L]
                     )
                 for j in range(1, L + 1):
-                    temp_numerator -= (
-                        binom(L, j)
-                        * w_ders[[0], [j], ...]
-                        * nurbs_derivatives[[k], [L - j], ...]
+                    temp_numerator = temp_numerator - (
+                        binom(L, j) * w_ders[0, j, ...] * cols[L - j]
                     )
                 for i in range(1, k + 1):
                     for j in range(1, L + 1):
-                        temp_numerator -= (
+                        temp_numerator = temp_numerator - (
                             binom(k, i)
                             * binom(L, j)
-                            * w_ders[[i], [j], ...]
-                            * nurbs_derivatives[[k - i], [L - j], ...]
+                            * w_ders[i, j, ...]
+                            * rows[k - i][L - j]
                         )
-                nurbs_derivatives[k, L, ...] = temp_numerator / w_ders[[0], [0], ...]
+                cols.append(temp_numerator / w_ders[0, 0, ...])
+            rows.append(be.stack(cols, axis=0))
+        nurbs_derivatives = be.stack(rows, axis=0)
         return nurbs_derivatives
 
     @staticmethod
@@ -554,18 +554,18 @@ class NurbsGeometry(BaseGeometry):
         u = be.asarray(u)
 
         n_dim, N = be.shape(P)[0], be.size(u)
[…]
