+++ pyproject.toml
@@ -35,8 +35,7 @@ Homepage = "https://github.com/HarrisonKramer/optiland"
 optiland = "optiland_gui.run_gui:main"
 
 [project.optional-dependencies]
-dev = ["codecov>=2.1.13", "pytest>=8.3.5"]
-torch = ["torch"]
+dev = ["codecov>=2.1.13", "pytest>=8.3.5", "torch"]
 gui = ["PySide6", "qtconsole"]
 
 [tool.hatch.metadata]
