### 2025-10-15 e1a9ff2
fix: rename variables for clarity in NurbsGeometry class
