### 2025-10-15 76982fb
fix: adjust tolerance levels in surface normal assertions for consistency
