### 2025-10-15 b7b2ff3
First implementation of NURBS surface (#298)
* Update material_file.py

* Update material_file.py

Updated to compute refractive index also as a function of temperature

* Update material_file.py

fixed 1 bug
sometimes the reference temperature exist in the yaml file but not the thermal dispersion coefficients (e.g. fused_silica). In that case there was a problem in the previous version

* chore: minor formatting changes

* docs: use Google-style doc strings

* refactor: add _apply_environmental_correction method, keeping mathematical implementation consistent with original

* chore: ensure en
[…]
