+++ optiland/surfaces/factories/geometry_factory.py
@@ -22,6 +22,7 @@ from optiland.geometries import (
     ForbesQbfsGeometry,
     ForbesSolverConfig,  # forbes
     ForbesSurfaceConfig,  # forbes
+    NurbsGeometry,
     OddAsphere,
     Plane,
     PlaneGrating,
@@ -90,6 +91,19 @@ class GeometryConfig:
     # Forbes parameters
     radial_terms: dict[int, float] = field(default_factory=dict)
     freeform_coeffs: dict[tuple[str, int, int], float] = field(default_factory=dict)
+    # NURBS parameters
+    control_points: list[list[list[float]]] = field(default_factory=list)
+    weights: list[float] = field(default_factory=list)
+    u_knots: list[float] = field(default_factory=list)
+    v_knots: list[float] = field(default_factory=list)
+    nurbs_norm_x: float = 0.0
+    nurbs_norm_y: float = 0.0
+    nurbs_x_center: float = 0.0
+    nurbs_y_center: float = 0.0
+    u_degree: int = 3
+    v_degree: int = 3
+    n_points_u: int = 5
+    n_points_v: int = 5
 
 
 def _create_plane(cs: CoordinateSystem, config: GeometryConfig):
@@ -350,6 +364,30 @@ def _create_forbes_q2d(cs: CoordinateSystem, config: GeometryConfig):
     )
 
 
+def _create_nurbs(cs: CoordinateSystem, config: GeometryConfig):
+    """Create a NURBS geometry."""
+
+    return NurbsGeometry(
+        cs,
+        config.radius,
+        config.conic,
+        config.nurbs_norm_x,
+        config.nurbs_norm_y,
+        config.nurbs_x_center,
+        config.nurbs_y_center,
+        config.control_points,
+        config.weights,
+        config.u_degree,
+        config.v_degree,
+        config.u_knots,
+        config.v_knots,
+        config.n_points_u,
+        config.n_points_v,
+        config.tol,
+        config.max_iter,
+    )
+
+
 def _create_paraxial(cs: CoordinateSystem, config: GeometryConfig):
     """
     Create a paraxial geometry, which is simply a planar surface.
@@ -377,6 +415,7 @@ geometry_mapper = {
     "zernike": _create_zernike,
     "forbes_qbfs": _create_forbes_qbfs,
     "forbes_q2d": _create_forbes_q2d,
+    "nurbs": _create_nurbs,
 }
 
 
