### 2025-10-15 2bff244
feat: Make Huygens PSF backend-agnostic
Refactors the `HuygensPSF` class to support both NumPy and PyTorch backends.

Implements the Strategy design pattern to encapsulate the core summation algorithm, separating the backend-specific logic from the main class.

Creates a new `huygens_fresnel_strategies.py` module to house the `NumbaSummation` and `TorchSummation` strategies. The `TorchSummation` strategy is vectorized and batched to manage memory and improve performance.

Updates the test suite to include comprehensive tests for the PyTorch backend, ensuring consistency with the NumPy implemen
[…]
