+++ optiland/backend/utils.py
@@ -71,3 +71,19 @@ def to_numpy(obj: ScalarOrArrayT) -> NDArray:
         except TypeError:
             continue
     raise TypeError(f"Unsupported object type: {type(obj)}")
+
+
+def is_torch_tensor(obj: any) -> bool:
+    """Checks if an object is a PyTorch tensor.
+
+    Args:
+        obj (any): The object to check.
+
+    Returns:
+        bool: True if the object is a PyTorch tensor, False otherwise.
+    """
+    if importlib.util.find_spec("torch"):
+        import torch
+
+        return isinstance(obj, torch.Tensor)
+    return False
