### 2025-10-15 f982025
feat: Add complex precision retrieval and improve backend restoration in tests
