### 2025-10-15 54f9c90
fix: Improve torch backend fixture to ensure numpy restoration after tests
