### 2025-10-15 a73fd52
refactor: Remove unnecessary restore_numpy_backend fixture from tests
