### 2025-10-15 ae32f03
fix: Enhance numpy backend fixture to restore original backend after tests
