### 2025-10-16 915d9e1
feat: Add PhaseInteractionModel for surfaces with phase profiles
This commit introduces a new `PhaseInteractionModel` to handle surfaces
with arbitrary phase profiles. This new model follows the Strategy pattern,
using a `BasePhase` object to calculate the effect of a phase function on
the rays.

This change is made in response to user feedback, which clarified that the
existing `DiffractiveInteractionModel` is for physical gratings and should
not be modified. The new `PhaseInteractionModel` is implemented as a
separate entity, and the `SurfaceFactory` has been updated to use it when
a `phase_
[…]
