+++ optiland/phase/radial.py
@@ -1,154 +1,146 @@
-"""Phase
+"""Radial Phase
 
-TPhaser that gets added to any surface defiend in Geometry:
-Write some kind of disccription
+This module defines the `RadialPhase` class, which represents a radially
+symmetric phase function.
 
+The `RadialPhase` class calculates the change in the direction of a ray and
+the optical path difference (OPD) introduced by a radially symmetric phase
+profile. The phase profile is defined by a series of coefficients.
+
+The implementation is based on the calculation of the phase gradient, which is
+used to determine the change in the ray's direction.
 
 Hhsoj, 2025
 """
 
 from __future__ import annotations
 
+from typing import TYPE_CHECKING, List
+
 import optiland.backend as be
-from optiland.coordinate_system import CoordinateSystem
 from optiland.phase.base import BasePhase
+
+if TYPE_CHECKING:
+    from optiland.rays import RealRays
+    from optiland._types import BEArray
 
 
 class RadialPhase(BasePhase):
-    """Represents a phase function that can be added to a Phase
-
-
-
+    """Represents a radially symmetric phase function.
 
     Args:
-        A (float): Period of the diffraction garting in lines/um
-        Order (integer): diffraction order for the grating
-
-    Methods:
-        phase_grating(self, rays): Calculates the phase for a diffraction grating
-
-
+        order (int, optional): The diffraction order. Defaults to 1.
+        coefficients (List[float], optional): A list of coefficients that
+            define the radial phase profile. Defaults to None.
 
     """
 
-    def __init__(self, order=1, coef=None, eff="ideal"):
-        if coef is None:
-            coef = []
+    def __init__(self, order: int = 1, coefficients: List[float] = None):
+        if coefficients is None:
+            coefficients = []
         self.order = be.array(order)
-        self.coef = be.array(coef)
-        self.eff = eff
+        self.coefficients = be.array(coefficients)
 
-    def __str__(self):
-        return "Radial"
-
-    def phasefunction(self, x=0, y=0):
-        """Calculate the surface sag of the geometry at the given coordinates.
+    def phase_calc(
+        self,
+        rays: RealRays,
+        nx: BEArray,
+        ny: BEArray,
+        nz: BEArray,
+        n1: BEArray,
+        n2: BEArray,
+    ) -> tuple[BEArray, BEArray, BEArray, BEArray]:
+        """Calculates the effect of the radial phase function on the rays.
 
         Args:
-            x (float or be.ndarray, optional): The x-coordinate(s). Defaults to 0.
-            y (float or be.ndarray, optional): The y-coordinate(s). Defaults to 0.
+            rays (RealRays): The rays incident on the surface.
+            nx (BEArray): The x-component of the surface normal.
+            ny (BEArray): The y-component of the surface normal.
+            nz (BEArray): The z-component of the surface normal.
+            n1 (BEArray): The refractive index of the medium before the
+                surface.
+            n2 (BEArray): The refractive index of the medium after the
+                surface.
 
         Returns:
-            be.ndarray or float: The sag value(s) at the given coordinates.
+            A tuple containing the new x, y, and z direction cosines (L, M, N)
+            and the optical path difference (OPD) to be added to the rays.
 
         """
-        m = self.order
-        r = be.sqrt(x**2 + y**2)
[…]
