### 2025-10-16 9293d0e
feat(docs): Add documentation for phase profiles
Adds comprehensive documentation for the new phase profile functionality.

- Creates two new Jupyter notebooks in the gallery:
  - `simple_grating.ipynb`: Demonstrates a simple grating using `PhaseInteractionModel`.
  - `complex_phase_profile.ipynb`: Showcases a custom, user-defined phase profile.
- Restructures the documentation gallery to include the new examples.
- Updates the developer's guide with information on the `PhaseInteractionModel`.
