+++ docs/gallery/diffractive_and_phase.rst
@@ -1,0 +1,25 @@
+.. _gallery_diffractive_and_phase:
+
+Diffractive and Phase Systems
+=============================
+
+This section contains examples of optical systems that utilize diffractive gratings and custom phase profiles.
+
+Diffractive Systems
+-------------------
+
+These examples demonstrate grating parameter specification and wavelength dispersion modeling in various spectrometer designs.
+
+.. nbgallery::
+    diffractive/Dyson_Spectrometer
+    diffractive/Littrow_Spectrometer
+    diffractive/Czerny_Turner_Spectrometer
+
+Phase Profile Systems
+---------------------
+
+These examples showcase the `PhaseInteractionModel` for creating surfaces with custom phase profiles, from simple gratings to more complex, user-defined functions.
+
+.. nbgallery::
+    notebooks/simple_grating
+    notebooks/complex_phase_profile
