### 2025-10-16 421acf6
refactor: Clean up imports and improve type hints in phase and interaction models
