+++ optiland/interactions/base.py
@@ -73,8 +73,8 @@ class BaseInteractionModel(ABC):
     def from_dict(cls, data, geometry, material_pre, material_post):
         """Creates an interaction model from a dictionary representation."""
         from optiland.coatings import BaseCoating
-        from optiland.scatter import BaseBSDF
         from optiland.phase.base import BasePhase
+        from optiland.scatter import BaseBSDF
 
         interaction_type = data["type"]
         subclass = cls._registry.get(interaction_type)
@@ -90,9 +90,7 @@ class BaseInteractionModel(ABC):
         if "bsdf" in init_data and init_data["bsdf"] is not None:
             init_data["bsdf"] = BaseBSDF.from_dict(init_data["bsdf"])
         if "phase_model" in init_data and init_data["phase_model"] is not None:
-            init_data["phase_model"] = BasePhase.from_dict(
-                init_data["phase_model"]
-            )
+            init_data["phase_model"] = BasePhase.from_dict(init_data["phase_model"])
 
         return subclass(
             geometry=geometry,
