+++ optiland/phase/__init__.py
@@ -1,9 +1,7 @@
-"""This package defines various optical surface geometries used in Optiland,
-ranging from simple planes and spheres to complex aspheres and polynomial
-surfaces."""
-
 from __future__ import annotations
 
 from .base import BasePhase
 from .grating import GratingPhase
 from .radial import RadialPhase
+
+__all__ = ["BasePhase", "GratingPhase", "RadialPhase"]
