### 2025-10-16 dfc21ba
refactor: Remove obsolete phase_test.py file
