+++ optiland/samples/phase_test.py
@@ -1,137 +1,0 @@
-# Defines simple sample optical systems.
-from __future__ import annotations
-
-import matplotlib.pyplot as plt
-import numpy as np
-
-import optiland.backend as be
-from optiland import optic
-from optiland.materials import Material
-from optiland.phase.grating import GratingPhase
-from optiland.phase.radial import RadialPhase
-
-
-class AsphericSingletMirror(optic.Optic):
-    """An aspheric singlet lens."""
-
-    def __init__(self):
-        super().__init__()
-        # add surfaces
-        a1 = -0.01
-        print(a1)
-        self.add_surface(index=0, radius=be.inf, thickness=be.inf)
-        # self.add_surface(index=1, radius=be.inf, thickness=1, material = "N-BK7")
-        self.add_surface(index=1, radius=be.inf, thickness=0, material="air")
-        self.add_surface(
-            index=2,
-            thickness=-50,
-            radius=-75,
-            is_stop=True,
-            material="mirror",
-            surface_type="standard",
-            phase_type=RadialPhase(coef=[a1], order=-1, eff="ideal"),
-        )
-        self.add_surface(index=3, thickness=-0)
-        self.add_surface(index=4)
-
-        # add aperture
-        self.set_aperture(aperture_type="EPD", value=20.0)
-
-        # add field
-        self.set_field_type(field_type="angle")
-        self.add_field(y=0)
-
-        # add wavelength
-        self.add_wavelength(value=0.530, is_primary=True)
-
-
-class AsphericSinglet(optic.Optic):
-    """An aspheric singlet lens."""
-
-    def __init__(self):
-        super().__init__()
-        # add surfaces
-        self.add_surface(index=0, radius=be.inf, thickness=be.inf)
-        self.add_surface(index=1, radius=be.inf, thickness=1, material="N-BK7")
-        # self.add_surface(index=1, radius=be.inf, thickness=1, material = "air")
-        self.add_surface(
-            index=2,
-            thickness=100,
-            radius=50,
-            is_stop=True,
-            material="air",
-            surface_type="standard",
-            phase_type=GratingPhase(A=1, order=-1, eff="ideal"),
-        )
-        self.add_surface(index=3, thickness=0)
-        self.add_surface(index=4)
-
-        # add aperture
-        self.set_aperture(aperture_type="EPD", value=20.0)
-
-        # add field
-        self.set_field_type(field_type="angle")
-        self.add_field(y=0)
-
-        # add wavelength
-        self.add_wavelength(value=0.530, is_primary=True)
-
-
-lens = AsphericSingletMirror()
-# You may try a different lens here
-# lens = AsphericSinglet()
-bk7 = Material("N-Bk7")
-
-rayData = lens.trace_generic(Hx=0, Hy=0, Px=0, Py=1, wavelength=0.53)
-num_surfaces = lens.surface_group.num_surfaces
-lens.draw(distribution="line_y", num_rays=11)
-plt.show()
-
-
-def generate_wavefront(z, num_points=32):
-    x, y = np.meshgrid(np.linspace(-1, 1, num_points), np.linspace(-1, 1, num_points))
-    radius = np.sqrt(x**2 + y**2)
-    phi = np.arctan2(y, x)
-    values = z.poly(radius, phi)
-
-    values[radius > 1] = 0.0
-
-    # Normalize the values between -1 and 1
[…]
