### 2025-10-16 df4f51a
refactor: Improve arithmetic operations and update test assertions for phase interactions
