+++ tests/test_phase.py
@@ -7,6 +7,7 @@ import optiland.backend as be
 from optiland.phase.grating import GratingPhase
 from optiland.phase.radial import RadialPhase
 from optiland.rays.real_rays import RealRays
+from .utils import assert_allclose
 
 
 @pytest.mark.parametrize("backend", be.list_available_backends())
@@ -25,10 +26,10 @@ def test_grating_phase(backend):
     )
     phase = GratingPhase(period=1.0, order=1)
     L, M, N, opd = phase.phase_calc(rays, 0, 0, -1, 1, 1)
-    assert be.allclose(L, 0.0)
-    assert be.allclose(M, 0.5)
-    assert be.allclose(N, be.sqrt(1 - 0.5**2))
-    assert be.allclose(opd, 0.5)
+    assert_allclose(L, 0.0)
+    assert_allclose(M, 0.5)
+    assert_allclose(N, be.sqrt(1 - 0.5**2))
+    assert_allclose(opd, 0.5)
 
 
 @pytest.mark.parametrize("backend", be.list_available_backends())
@@ -47,7 +48,7 @@ def test_radial_phase(backend):
     )
     phase = RadialPhase(order=1, coefficients=[0.1])
     L, M, N, opd = phase.phase_calc(rays, 0, 0, -1, 1, 1)
-    assert be.allclose(L, 0.1, atol=1e-6)
-    assert be.allclose(M, 0.0, atol=1e-6)
-    assert be.allclose(N, be.sqrt(1 - 0.1**2), atol=1e-6)
-    assert be.allclose(opd, 0.1)
+    assert_allclose(L, 0.1, atol=1e-6)
+    assert_allclose(M, 0.0, atol=1e-6)
+    assert_allclose(N, be.sqrt(1 - 0.1**2), atol=1e-6)
+    assert_allclose(opd, 0.1)
