### 2025-10-16 b498100
refactor: Enhance documentation and improve clarity in RealRays and DiffractiveInteractionModel classes
