+++ optiland/rays/real_rays.py
@@ -1,7 +1,8 @@
 """Real Rays
 
 This module contains the RealRays class, which represents a collection of real
-rays in 3D space.
+rays in 3D space. The complex diffraction logic has been removed to adhere to
+the Single Responsibility Principle.
 
 Kramer Harrison, 2024
 """
@@ -22,25 +23,9 @@ if TYPE_CHECKING:
 class RealRays(BaseRays):
     """Represents a collection of real rays in 3D space.
 
-    This class stores ray positions, directions, and properties as arrays,
-    supporting both NumPy and PyTorch backends for efficient computation.
-
-    Attributes:
-        x: x-coordinates of ray positions.
-        y: y-coordinates of ray positions.
-        z: z-coordinates of ray positions.
-        L: x-components of ray direction cosines.
-        M: y-components of ray direction cosines.
-        N: z-components of ray direction cosines.
-        i: Intensity values of the rays.
-        w: Wavelength values of the rays.
-        opd: Optical path difference values.
-        L0: Pre-surface x-direction cosines (None until surface interaction).
-        M0: Pre-surface y-direction cosines (None until surface interaction).
-        N0: Pre-surface z-direction cosines (None until surface interaction).
-
-    Note:
-        Direction cosines should be normalized: L² + M² + N² = 1.
+    This class stores ray positions, directions, and properties as arrays.
+    It is responsible for managing the state of the rays but does not contain
+    complex physics for surface interactions.
     """
 
     def __init__(
@@ -54,22 +39,7 @@ class RealRays(BaseRays):
         intensity: ArrayLike,
         wavelength: ArrayLike,
     ):
-        """Initialize a collection of real rays in 3D space.
-
-        Args:
-            x: The x-coordinates of the ray starting positions.
-            y: The y-coordinates of the ray starting positions.
-            z: The z-coordinates of the ray starting positions.
-            L: The x-components of the ray direction cosines.
-            M: The y-components of the ray direction cosines.
-            N: The z-components of the ray direction cosines.
-            intensity: The intensity values of the rays.
-            wavelength: The wavelength values of the rays.
-
-        Note:
-            All input arrays are converted to 1D arrays. Direction cosines
-            (L, M, N) should be normalized such that L² + M² + N² = 1.
-        """
+        """Initialize a collection of real rays in 3D space."""
         self.x = be.as_array_1d(x)
         self.y = be.as_array_1d(y)
         self.z = be.as_array_1d(z)
@@ -88,11 +58,7 @@ class RealRays(BaseRays):
         self.is_normalized = True
 
     def rotate_x(self, rx: ScalarOrArray):
-        """Rotate the rays about the x-axis.
-
-        Args:
-            rx: Rotation angle around x-axis in radians.
-        """
+        """Rotate the rays about the x-axis."""
         rx = be.array(rx)
         self.y, self.z, self.M, self.N = (
             self.y * be.cos(rx) - self.z * be.sin(rx),
@@ -102,11 +68,7 @@ class RealRays(BaseRays):
         )
 
     def rotate_y(self, ry: ScalarOrArray):
-        """Rotate the rays about the y-axis.
-
-        Args:
-            ry: Rotation angle around y-axis in radians.
-        """
+        """Rotate the rays about the y-axis."""
         ry = be.array(ry)
         self.x, self.z, self.L, self.N = (
             self.x * be.cos(ry) + self.z * be.sin(ry),
@@ -116,11 +78,7 @@ class RealRays(BaseRays):
         )
 
     def rotate_z(self, rz: ScalarOrArray):
-        """Rotate the rays about the z-axis.
-
-        Args:
-            rz: Rotation angle around z-axis in radians.
-        """
+        """Rotate the rays about the z-axis."""
[…]
