### 2025-10-17 fe2fc29
Refactor phase model handling in SurfaceFactory and update documentation
- Updated the documentation to reference the correct notebook for simple phase grating examples.
- Enhanced the phase model argument handling in SurfaceFactory to accept both BasePhase objects and dictionaries, improving flexibility and error handling.
