### 2025-10-18 4172b10
more fixes
