### 2025-10-18 b760e31
add listener for surface changes
