### 2025-10-18 b45c8b4
fix pytests
