### 2025-10-18 851dd0f
Add safeguards for paraxial surface MTF edge cases
- Handle u[-1] ≈ 0 in paraxial.f1() and f2() methods
- Return inf for focal length when output is collimated
- Handle inf/nan FNO in GeometricMTF and FFTMTF classes
- Use default max_freq with user warning for afocal systems
- Add comprehensive tests for paraxial surface MTF calculations

Co-authored-by: HarrisonKramer <29144587+HarrisonKramer@users.noreply.github.com>
