+++ tests/test_paraxial_mtf.py
@@ -1,0 +1,234 @@
+"""Test for paraxial surface MTF edge cases
+
+This test validates that the fixes for paraxial surface MTF calculations
+properly handle edge cases like collimated output and infinite focal lengths.
+"""
+import pytest
+import numpy as np
+
+import optiland.backend as be
+from optiland import optic, materials
+from optiland.mtf import GeometricMTF, FFTMTF
+
+
+def test_paraxial_surface_mtf_original_issue():
+    """Test the original issue case from GitHub issue report."""
+    n = 1.5
+    wavelength = 0.5
+    singlet = optic.Optic()
+
+    singlet.set_aperture(aperture_type="EPD", value=0.3*2)
+    singlet.add_surface(index=0, radius=np.inf, thickness=400)
+    singlet.add_surface(index=1, radius=np.inf, thickness=0.9, is_stop=True)
+    singlet.add_surface(index=2, radius=np.inf, thickness=0.5, material=materials.IdealMaterial(n=n))
+    singlet.add_surface(index=3,
+        f=0.500,
+        surface_type="paraxial",
+        thickness=0.5
+    )
+    singlet.add_surface(index=4)
+
+    singlet.set_field_type(field_type="angle")
+    singlet.add_field(y=0)
+    singlet.add_field(y=15)
+    singlet.add_field(y=30)
+
+    singlet.add_wavelength(value=wavelength, is_primary=True)
+
+    # Test paraxial properties
+    f2 = singlet.paraxial.f2()
+    assert not be.isnan(f2) and not be.isinf(f2), "f2 should be finite"
+    
+    fno = singlet.paraxial.FNO()
+    assert not be.isnan(fno) and not be.isinf(fno), "FNO should be finite"
+    
+    # Test GeometricMTF - should not raise warnings or errors
+    geo_mtf = GeometricMTF(singlet)
+    assert geo_mtf.max_freq > 0, "max_freq should be positive"
+    assert not be.isnan(geo_mtf.max_freq), "max_freq should not be nan"
+    assert not be.isinf(geo_mtf.max_freq), "max_freq should not be inf"
+    
+    # Test FFTMTF - should not raise warnings or errors
+    fft_mtf = FFTMTF(singlet)
+    assert fft_mtf.max_freq > 0, "max_freq should be positive"
+    assert not be.isnan(fft_mtf.max_freq), "max_freq should not be nan"
+    assert not be.isinf(fft_mtf.max_freq), "max_freq should not be inf"
+
+
+def test_paraxial_surface_collimated_output():
+    """Test that the safeguards handle edge cases without breaking normal operation.
+    
+    While it's difficult to create a real optical system with u=0 (collimated output),
+    we test that systems that might be close to that edge case are handled gracefully.
+    """
+    # Create a simple paraxial lens
+    lens = optic.Optic()
+    
+    lens.add_surface(index=0, radius=np.inf, thickness=np.inf)
+    lens.add_surface(index=1, surface_type="paraxial", f=100, thickness=100, is_stop=True)
+    lens.add_surface(index=2)
+    
+    lens.set_aperture(aperture_type="EPD", value=20)
+    lens.set_field_type(field_type="angle")
+    lens.add_field(y=0)
+    lens.add_wavelength(value=0.55, is_primary=True)
+    
+    # Test paraxial properties - should handle edge cases gracefully
+    f2 = lens.paraxial.f2()
+    assert not be.isnan(f2), "f2 should not be nan"
+    
+    fno = lens.paraxial.FNO()
+    assert not be.isnan(fno), "FNO should not be nan"
+    
[…]
