+++ optiland/mtf/fft.py
@@ -73,7 +73,20 @@ class FFTMTF(BaseMTF):
         self.FNO = self._get_fno()
 
         if max_freq == "cutoff":
-            self.max_freq = 1 / (self.resolved_wavelength * 1e-3 * self.FNO)
+            # Handle edge case where FNO is infinite (afocal or collimated system)
+            if be.isinf(self.FNO) or be.isnan(self.FNO):
+                import warnings
+
+                warnings.warn(
+                    "System appears to be afocal or produce collimated output. "
+                    "Using default cutoff frequency of 1000 cycles/mm. "
+                    "Consider specifying max_freq explicitly.",
+                    UserWarning,
+                    stacklevel=2,
+                )
+                self.max_freq = 1000.0  # Default reasonable value in cycles/mm
+            else:
+                self.max_freq = 1 / (self.resolved_wavelength * 1e-3 * self.FNO)
         else:
             self.max_freq = max_freq
 
@@ -158,6 +171,11 @@ class FFTMTF(BaseMTF):
             float: The MTF units calculated based on the grid size, number
                 of rays, wavelength (from BaseMTF), and FNO.
         """
+        # Handle edge case where FNO is infinite
+        if be.isinf(self.FNO) or be.isnan(self.FNO):
+            # Use a reasonable default when FNO is invalid
+            return 1.0 / self.grid_size
+
         dx = 1 / ((self.num_rays - 1) * self.resolved_wavelength * 1e-3 * self.FNO)
 
         return dx
