### 2025-10-18 f394085
Add FLOA aperture support for Zemax files
Co-authored-by: HarrisonKramer <29144587+HarrisonKramer@users.noreply.github.com>
