### 2025-10-18 c0d4211
Address code review feedback - use named constants
- Define NEAR_ZERO_THRESHOLD in paraxial.py for u[-1] checks
- Define DEFAULT_CUTOFF_FREQ in mtf modules for afocal systems
- Use consistent threshold values in production code and tests
- Improve code maintainability and consistency

Co-authored-by: HarrisonKramer <29144587+HarrisonKramer@users.noreply.github.com>
