### 2025-10-18 ee1bd03
Address code review feedback - make aperture handling more robust
Co-authored-by: HarrisonKramer <29144587+HarrisonKramer@users.noreply.github.com>
