### 2025-10-18 21d0d0c
Fix line length for warning messages
- Break long warning messages across multiple lines
- Ensure all lines are under 88 characters

Co-authored-by: HarrisonKramer <29144587+HarrisonKramer@users.noreply.github.com>
