### 2025-10-18 e4036b0
test for Fresnel updates
