### 2025-10-18 3a9746a
Fix FLOA (Float by Stop) aperture error when reading Zemax files (#375)
* Initial plan

* Add FLOA aperture support for Zemax files

Co-authored-by: HarrisonKramer <29144587+HarrisonKramer@users.noreply.github.com>

* Address code review feedback - make aperture handling more robust

Co-authored-by: HarrisonKramer <29144587+HarrisonKramer@users.noreply.github.com>

* Apply suggestion from @Copilot

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>
Co-authored-by: HarrisonKramer <2
[…]
