### 2025-10-18 fcba4c2
Don't allow more than one Surfaces at index = 0
