### 2025-10-18 69a7bcd
Allow setting SurfaceGroup.stop_index
