### 2025-10-18 87e7313
Disallow setting stop on image surface
