### 2025-10-19 4fcc61f
use weak references instead of __del__
