+++ tests/test_standard_surface.py
@@ -1,4 +1,5 @@
 import pytest
+import gc
 
 import optiland.backend as be
 from optiland.coatings import FresnelCoating, SimpleCoating
@@ -142,14 +143,37 @@ class TestSurface:
     def test_listener_registration(self):
         surf1 = self.create_surface()
         surf2 = self.create_surface()
-        surf2.previous_surface = surf1
-        assert surf1._listeners == [surf2._update_callback]
 
-        surf1.unregister_callback(surf2._update_callback)
+        # Test for explicit registering of the callback
+        surf1.register_callback(surf2._update_callback)
+        assert all(
+            [weakref() == surf2._update_callback for weakref in surf1._listeners]
+        )
+
+        # Test for explicit deregistering
+        surf1.deregister_callback(surf2._update_callback)
         assert surf1._listeners == []
 
-        surf1.register_callback(surf2._update_callback)
-        assert surf1._listeners == [surf2._update_callback]
+        # Test for implicit registering of the callback
+        surf2.previous_surface = surf1
+        assert all(
+            [weakref() == surf2._update_callback for weakref in surf1._listeners]
+        )
+
+        # Test for implicit deregistering
+        del surf2
+        gc.collect()
+        assert surf1._listeners == []
+
+        # Test for updating the callback
+        surf2 = self.create_surface()
+        surf3 = self.create_surface()
+        surf2.previous_surface = surf1
+        surf2.previous_surface = surf3
+        assert surf1._listeners == []
+        assert all(
+            [weakref() == surf3._update_callback for weakref in surf1._listeners]
+        )
 
     def test_fresnel_coating_material(self):
         surf1 = self.create_surface()
