### 2025-10-19 69c6d6e
fix deepcopy with weakref
