### 2025-10-19 9797a19
fix missing thickness update
