### 2025-10-19 10ecff1
`SurfaceGroup.add_surface`: Disallow second object surface (#376)
* Don't allow more than one Surfaces at index = 0

* add pytest

* update failing pytests
