### 2025-10-20 d65594a
fix `interactions.base.BaseInteraction.from_dict()`
and add pytest
