### 2025-10-20 f6ed59a
make (de)registering a callback private
