### 2025-10-20 ad94a6d
simplify `_update_coordinate_systems`
and test for start_index == 0
