### 2025-10-20 211ede7
allow testing for equality of materials
