### 2025-10-22 4beea5d
`SurfaceGroup`: Add `stop_index` setter (#377)
* Allow setting SurfaceGroup.stop_index

* add pytest

* Disallow setting stop on image surface

* update pytest

---------

Co-authored-by: Kramer <kdanielharrison@gmail.com>
