### 2025-10-22 49a4bb0
Avoid multiple copies of material objects (#369)
* remove multiple copies of materials across objects

# Conflicts:
#	optiland/surfaces/surface_group.py

* update pytests

* fixup! remove unused code

* ignore `material_pre`

* `Surface:_from_dict`: fix obsolete geometry param

* update Optic/SurfaceGroup `flip` method

* update new pytests

* more fixes

* add listener for surface changes

* fix pytests

* test for Fresnel updates

* update pytest

* use weak references instead of __del__

* fix deepcopy with weakref

* fix missing thickness update

* fix `interactions.base.BaseInteraction.fr
[…]
