+++ optiland/surfaces/standard_surface.py
@@ -12,6 +12,9 @@ Kramer Harrison, 2023
 
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+from weakref import WeakMethod
+
 import optiland.backend as be
 from optiland.coatings import BaseCoating, FresnelCoating
 from optiland.geometries import BaseGeometry
@@ -23,13 +26,16 @@ from optiland.physical_apertures.radial import configure_aperture
 from optiland.rays import BaseRays, ParaxialRays, RealRays
 from optiland.scatter import BaseBSDF
 
+if TYPE_CHECKING:
+    from collections.abc import Callable
+
 
 class Surface:
     """Represents a standard refractice surface in an optical system.
 
     Args:
         geometry (BaseGeometry): The geometry of the surface.
-        material_pre (BaseMaterial): The material before the surface.
+        previous_surface (Surface): The surface preceding this instance.
         material_post (BaseMaterial): The material after the surface.
         is_stop (bool, optional): Indicates if the surface is the aperture
             stop. Defaults to False.
@@ -47,18 +53,18 @@ class Surface:
 
     def __init__(
         self,
-        geometry: BaseGeometry,
-        material_pre: BaseMaterial,
+        previous_surface: Surface | None,
         material_post: BaseMaterial,
+        geometry: BaseGeometry,
         is_stop: bool = False,
-        aperture: BaseAperture = None,
-        surface_type: str = None,
+        aperture: BaseAperture | None = None,
+        surface_type: str | None = None,
         comment: str = "",
-        interaction_model: BaseInteractionModel = None,
+        interaction_model: BaseInteractionModel | None = None,
     ):
         self.geometry = geometry
-        self.material_pre = material_pre
-        self.material_post = material_post
+        self._previous_surface = previous_surface
+        self._material_post = material_post
         self.is_stop = is_stop
         self.aperture = configure_aperture(aperture)
         self.semi_aperture = None
@@ -67,31 +73,105 @@ class Surface:
 
         if interaction_model is None:
             self.interaction_model = RefractiveReflectiveModel(
-                geometry=self.geometry,
-                material_pre=self.material_pre,
-                material_post=self.material_post,
+                parent_surface=self,
                 is_reflective=False,
                 coating=None,
                 bsdf=None,
             )
         else:
             self.interaction_model = interaction_model
+            self.interaction_model.parent_surface = self
 
         self.thickness = 0.0  # used for surface positioning
-
+        self._listeners = []
         self.reset()
+
+    def __getstate__(self):
+        # Remove self._listeners when a deep copy is made
+        state = self.__dict__.copy()
+        del state["_listeners"]
+        return state
+
+    def __setstate__(self, state):
+        # Initialize self._listeners to an empty list after deepcopy
+        self.__dict__.update(state)
+        self._listeners = []
+
+    def _update_callback(self, caller: Surface) -> None:
+        # Called when a surface that we're related to changes
+        if caller != self.previous_surface:
+            raise RuntimeError("Unexpected Surface called _update_callback")
+
+        # Handle the changes. Right now, we're only interested in a change in refractive
+        # index. If this surface's interaction model has a Fresnel coating, update it:
+        if self.coating is not None:
+            self.set_fresnel_coating()
+
+    def _register_callback(self, callback: Callable):
+        if callback not in self._listeners:
+            self._listeners.append(
+                WeakMethod(callback, lambda obj: self._deregister_callback(obj))
+            )
+
+    def _deregister_callback(self, callback: Callable):
+        if isinstance(callback, WeakMethod) and callback in self._listeners:
+            self._listeners.remove(callback)
+            return
+        for weakref in self._listeners:
+            if weakref() == callback:
+                self._listeners.remove(weakref)
+
+    @property
[…]
