### 2025-10-22 59da7db
allow testing for equality of materials (#379)
