### 2025-10-22 7e3588a
re-enable negative image plane locations (#380)
