### 2025-10-25 2e1956a
`flip()`: don't flip sign of paraxial surfaces
