### 2025-10-25 45b6753
`flip()`: don't flip sign of paraxial surfaces (#381)
