### 2025-10-26 88c0da8
feat: Implement generalized phase surfaces (metasurfaces)
This commit introduces a flexible and extensible system for modeling generalized phase surfaces, such as metasurfaces and spatial light modulators.

The core of this feature is a new `PhaseInteractionModel` that uses the Strategy Pattern to delegate phase calculations to `BasePhaseProfile` objects. This allows users to define custom phase profiles without modifying the core ray tracing code.

Key changes:
- Created a new `optiland/phase` module with a `BasePhaseProfile` abstract base class and concrete implementations for constant, radi
[…]
