### 2025-10-26 6f0c576
`Paraxial`: handle internal foci and negative lens
