### 2025-10-26 14a63d2
chore: ruff formatting
