+++ optiland/phase/base.py
@@ -41,9 +41,7 @@ class BasePhaseProfile(abc.ABC):
         raise NotImplementedError
 
     @abc.abstractmethod
-    def get_gradient(
-        self, x: be.Array, y: be.Array
-    ) -> tuple[be.Array, be.Array]:
+    def get_gradient(self, x: be.Array, y: be.Array) -> tuple[be.Array, be.Array]:
         """Calculates the gradient of the phase at coordinates (x, y).
 
         Args:
