+++ optiland/phase/constant.py
@@ -36,9 +36,7 @@ class ConstantPhaseProfile(BasePhaseProfile):
         """
         return be.full_like(x, self.phase)
 
-    def get_gradient(
-        self, x: be.Array, y: be.Array
-    ) -> tuple[be.Array, be.Array]:
+    def get_gradient(self, x: be.Array, y: be.Array) -> tuple[be.Array, be.Array]:
         """Calculates the gradient of the phase at coordinates (x, y).
 
         Args:
