### 2025-10-26 1209a46
refactor: move phase profile imports to __init__.py and streamline geometry check in PhaseInteractionModel
