### 2025-10-26 33585ff
refactor: remove Plane geometry check and streamline ray interaction calculations in PhaseInteractionModel
