### 2025-10-26 0d02a39
feat: add phase interaction type handling in SurfaceFactory
