### 2025-10-26 2059478
feat: add LinearGratingPhaseProfile implementation and update __init__.py
