+++ optiland/phase/__init__.py
@@ -3,11 +3,13 @@ from __future__ import annotations
 from .base import BasePhaseProfile
 from .constant import ConstantPhaseProfile
 from .grid import GridPhaseProfile
+from .linear_grating import LinearGratingPhaseProfile
 from .radial import RadialPhaseProfile
 
 __all__ = [
     "BasePhaseProfile",
     "ConstantPhaseProfile",
     "GridPhaseProfile",
+    "LinearGratingPhaseProfile",
     "RadialPhaseProfile",
 ]
