### 2025-10-26 7854cae
feat: Finalize generalized phase grating feature
This commit finalizes the generalized phase grating feature.

It includes the following changes:

- Adds `order` and `efficiency` arguments to the `LinearGratingPhaseProfile`.
- Updates the `PhaseInteractionModel` to scale ray intensities by the efficiency.
- Adds comprehensive tests for the new features.
- Fixes several bugs in the `PhaseInteractionModel` and `RealRayTracer`.
- Updates the visualization to correctly render standalone surfaces with phase profiles.
- Adds three new example notebooks for the `LinearGrating`, `RadialGrating`, and `
[…]
