+++ optiland/surfaces/factories/surface_factory.py
@@ -158,7 +158,11 @@ class SurfaceFactory:
             interaction_type = "thin_lens"
         elif surface_type == "grating":
             interaction_type = "diffractive"
-        elif surface_type == "phase" or phase_profile is not None:
+        elif phase_profile is not None:
+            if surface_type != "plane":
+                raise ValueError(
+                    "Phase profiles can only be applied to plane surfaces."
+                )
             interaction_type = "phase"
 
         # Build interaction model
