+++ docs/api/api_phase.rst
@@ -1,0 +1,9 @@
+.. _api_phase:
+
+Phase
+=====
+
+.. automodule:: optiland.phase
+   :members:
+   :undoc-members:
+   :show-inheritance:

[…]
