### 2025-10-26 719ba84
refactor: clean up whitespace and improve readability in PhaseInteractionModel and LinearGratingPhaseProfile
