### 2025-10-26 a2d95fe
feat: Remove restriction on phase profiles for non-plane surfaces in SurfaceFactory
