### 2025-10-26 fd5a933
feat: Define interaction_type for diffractive, refractive_reflective, and thin_lens models
