### 2025-10-26 1a988b7
fix: Correct gradient calculation order in GridPhaseProfile and update test data dimensions
