### 2025-10-26 0466dd4
docs: Add Hsoj to the list of contributors in authors.rst
