### 2025-10-26 a5288dd
feat: Implement Zernike polynomial phase profile class
