+++ optiland/phase/zernike.py
@@ -1,0 +1,243 @@
+"""Zernike polynomial phase profile
+
+This module defines the `ZernikePhaseProfile` class, which applies a
+Zernike polynomial map as a phase profile over a circular aperture.
+
+Kramer Harrison, 2025
+"""
+
+from __future__ import annotations
+
+import typing
+
+from optiland.phase.base import BasePhaseProfile
+
+if typing.TYPE_CHECKING:
+    from optiland import backend as be
+    from optiland.zernike.base import BaseZernike
+
+
+class ZernikePhaseProfile(BasePhaseProfile):
+    """
+    Applies a Zernike polynomial map as a phase profile.
+
+    This class uses a provided Zernike polynomial object (conforming to
+    `BaseZernike`) to calculate the phase and its gradient over a
+    circular aperture defined by `norm_radius`.
+
+    Args:
+        zernike: An instance of a `BaseZernike` subclass (e.g.,
+            `ZernikeStandard`) that defines the polynomial.
+        norm_radius: The normalized radius of the aperture over which the Zernike
+            polynomial is defined.
+    """
+
+    phase_type = "zernike"
+
+    def __init__(self, zernike: BaseZernike, norm_radius: float):
+        super().__init__()
+        self.zernike = zernike
+        self.norm_radius = norm_radius
+
+    def _get_polar_coords(
+        self, x: be.Array, y: be.Array
+    ) -> tuple[be.Array, be.Array, be.Array]:
+        """
+        Converts Cartesian coordinates to normalized polar coordinates.
+
+        Args:
+            x: The x-coordinates.
+            y: The y-coordinates.
+
+        Returns:
+            A tuple of (r_cart, r_norm, phi):
+            - r_cart: The unnormalized radial coordinate (sqrt(x^2 + y^2)).
+            - r_norm: The radial coordinate normalized by `self.norm_radius`.
+            - phi: The azimuthal angle (arctan2(y, x)).
+        """
+        # Import backend here to be used in methods
+        from optiland import backend as be
+
+        r_cart = be.sqrt(x**2 + y**2)
+        r_norm = r_cart / self.norm_radius
+        phi = be.arctan2(y, x)
+        return r_cart, r_norm, phi
+
+    def get_phase(self, x: be.Array, y: be.Array) -> be.Array:
+        """
+        Calculates the phase added by the Zernike profile at (x, y).
+
+        Args:
+            x: The x-coordinates of the points of interest.
+            y: The y-coordinates of the points of interest.
+
+        Returns:
+            The phase at each (x, y) coordinate, calculated by summing all
+            Zernike terms.
+        """
+        _r_cart, r_norm, phi = self._get_polar_coords(x, y)
+        return self.zernike.poly(r_norm, phi)
+
+    def get_gradient(self, x: be.Array, y: be.Array) -> tuple[be.Array, be.Array]:
+        """
+        Calculates the gradient of the phase at coordinates (x, y).
+
+        This method computes the total gradient (d_phi/dx, d_phi/dy) by
+        summing the Cartesian gradients of each individual Zernike term.
+        It uses the chain rule to convert from polar derivatives
+        (dZ/dr_norm, dZ/dphi) to Cartesian derivatives (dZ/dx, dZ/dy).
+
+        dZ/dx = (dZ/dr_norm) * (dr_norm/dx) + (dZ/dphi) * (dphi/dx)
+        dZ/dy = (dZ/dr_norm) * (dr_norm/dy) + (dZ/dphi) * (dphi/dy)
+
+        where:
+        dr_norm/dx = x / (r_cart * R) = cos(phi) / R
+        dr_norm/dy = y / (r_cart * R) = sin(phi) / R
+        dphi/dx = -y / r_cart^2 = -sin(phi) / (r_norm * R)
[…]
