### 2025-10-26 0b9db92
feat(phase): Finalize ZernikePhaseProfile implementation
This commit finalizes the implementation of the `ZernikePhaseProfile` class.

- Refactors the `ZernikePhaseProfile` to accept a `zernike_type` string and a list of coefficients, following the existing pattern in `optiland/geometries/zernike.py`.
- Adds comprehensive unit tests for the `ZernikePhaseProfile` class, including tests for phase, gradient, and serialization.
- Updates the developer's guide to include a section on the `ZernikePhaseProfile` with a usage example.
- Creates a new Jupyter notebook example to demonstrate the use of t
[…]
