+++ optiland/phase/zernike.py
@@ -10,11 +10,20 @@ from __future__ import annotations
 
 import typing
 
+from optiland import backend as be
 from optiland.phase.base import BasePhaseProfile
+from optiland.zernike import ZernikeFringe, ZernikeNoll, ZernikeStandard
 
 if typing.TYPE_CHECKING:
-    from optiland import backend as be
+    from optiland._types import ZernikeType
     from optiland.zernike.base import BaseZernike
+
+
+_ZERNIKE_TYPES: dict[ZernikeType, type[BaseZernike]] = {
+    "standard": ZernikeStandard,
+    "noll": ZernikeNoll,
+    "fringe": ZernikeFringe,
+}
 
 
 class ZernikePhaseProfile(BasePhaseProfile):
@@ -26,18 +35,49 @@ class ZernikePhaseProfile(BasePhaseProfile):
     circular aperture defined by `norm_radius`.
 
     Args:
-        zernike: An instance of a `BaseZernike` subclass (e.g.,
-            `ZernikeStandard`) that defines the polynomial.
-        norm_radius: The normalized radius of the aperture over which the Zernike
-            polynomial is defined.
+        coefficients (list or be.ndarray, optional): The coefficients of the
+            Zernike polynomial surface. Defaults to an empty list, indicating
+            no Zernike polynomial coefficients are used.
+        zernike_type (str, optional): The type of Zernike polynomial to use.
+            Defaults to "standard". Options are "standard", "noll", or "fringe".
+        norm_radius (float, optional): The normalization radius for the
+            Zernike polynomial coordinates. Defaults to 1.
     """
 
     phase_type = "zernike"
 
-    def __init__(self, zernike: BaseZernike, norm_radius: float):
+    def __init__(
+        self,
+        coefficients: be.Array | None = None,
+        zernike_type: ZernikeType = "standard",
+        norm_radius: float = 1.0,
+    ):
         super().__init__()
-        self.zernike = zernike
+        if zernike_type not in _ZERNIKE_TYPES:
+            raise ValueError(
+                "Zernike type must be one of 'standard', 'noll', or 'fringe', got "
+                f"{zernike_type}",
+            )
+        if norm_radius <= 0:
+            raise ValueError(
+                f"Normalization radius must be positive, got {norm_radius}"
+            )
+
+        coefficients = be.atleast_1d(coefficients if coefficients is not None else [])
+
+        self.zernike = _ZERNIKE_TYPES[zernike_type](coeffs=coefficients)
+        self.zernike_type: ZernikeType = zernike_type
         self.norm_radius = norm_radius
+
+    @property
+    def coefficients(self) -> be.Array:
+        """Get the coefficients of the Zernike polynomial surface."""
+        return self.zernike.coeffs
+
+    @coefficients.setter
+    def coefficients(self, value: be.Array) -> None:
+        """Set the coefficients of the Zernike polynomial surface."""
+        self.zernike = _ZERNIKE_TYPES[self.zernike_type](coeffs=be.atleast_1d(value))
 
     def _get_polar_coords(
         self, x: be.Array, y: be.Array
@@ -55,9 +95,6 @@ class ZernikePhaseProfile(BasePhaseProfile):
             - r_norm: The radial coordinate normalized by `self.norm_radius`.
             - phi: The azimuthal angle (arctan2(y, x)).
         """
-        # Import backend here to be used in methods
-        from optiland import backend as be
-
         r_cart = be.sqrt(x**2 + y**2)
         r_norm = r_cart / self.norm_radius
         phi = be.arctan2(y, x)
@@ -109,9 +146,6 @@ class ZernikePhaseProfile(BasePhaseProfile):
             A tuple containing the x and y components of the phase gradient
             (d_phi/dx, d_phi/dy).
         """
[…]
