### 2025-10-26 df2aa43
feat(phase): Finalize ZernikePhaseProfile implementation
This commit finalizes the implementation of the `ZernikePhaseProfile` class.

- Refactors the `ZernikePhaseProfile` to accept a `zernike_type` string and a list of coefficients, following the existing pattern in `optiland/geometries/zernike.py`. This provides a more user-friendly API.
- Adds comprehensive unit tests for the `ZernikePhaseProfile` class, including tests for phase, gradient (including at the origin), and serialization/deserialization for all Zernike types.
- Updates the developer's guide to include a section on the `Zernike
[…]
