+++ optiland/zernike/base.py
@@ -62,6 +62,7 @@ class BaseZernike(ABC):
         """
         return (
             coeff
+            * self._norm_constant(n, m)
             * self._radial_term(n, m, r)
             * self._azimuthal_term(m, phi)
         )
@@ -98,14 +99,7 @@ class BaseZernike(ABC):
             float: The calculated value of the Zernike polynomial.
 
         """
-        val = []
-
-        for coeff, idx in zip(self.coeffs, self.indices, strict=True):
-            n, m = idx
-            norm = self._norm_constant(n, m)
-            val.append(norm * self.get_term(coeff, n, m, r, phi))
-
-        return sum(val)
+        return sum(self.terms(r, phi))
 
     def get_derivative(self, n=0, m=0, r=0, phi=0):
         """Calculate the derivative of the Zernike polynomial for the given
@@ -124,7 +118,6 @@ class BaseZernike(ABC):
             tuple[float, float]: The radial and azimuthal derivatives of the
                 Zernike polynomial.
         """
-        norm = self._norm_constant(n, m)
         radial_term = self._radial_term(n, abs(m), r)
         radial_derivative = self._radial_derivative(n, abs(m), r)
 
@@ -140,7 +133,7 @@ class BaseZernike(ABC):
                 be.abs(m) * radial_term * be.cos(be.abs(m) * phi)
             )
 
-        return norm * partial_radial_derivative, norm * partial_azimuthal_derivative
+        return partial_radial_derivative, partial_azimuthal_derivative
 
     @classmethod
     def _generate_indices(cls, n_indices: int) -> np.ndarray:
