### 2025-10-27 9b9eb34
Refactor GeometryConfig to be SOLID compliant
- Replaced the monolithic `GeometryConfig` class with a set of smaller, more focused dataclasses for each geometry type.
- Implemented a registry pattern to dynamically select the correct configuration class based on the surface type.
- Updated the `GeometryFactory` and `SurfaceFactory` to use the new configuration system.
- Updated the GUI Connector to work with the new system, ensuring no change to the user-facing API.
