+++ optiland/surfaces/factories/geometry_factory.py
@@ -10,7 +10,7 @@ Kramer Harrison, 2025
 
 from __future__ import annotations
 
-from dataclasses import dataclass, field
+from dataclasses import fields
 from typing import TYPE_CHECKING, Any
 
 import optiland.backend as be
@@ -32,87 +32,34 @@ from optiland.geometries import (
     ToroidalGeometry,
     ZernikePolynomialGeometry,
 )
+from optiland.surfaces.factories.geometry_configs import (
+    BiconicConfig,
+    ChebyshevConfig,
+    EvenAsphereConfig,
+    ForbesQ2dConfig,
+    ForbesQbfsConfig,
+    GratingConfig,
+    NurbsConfig,
+    OddAsphereConfig,
+    PlaneConfig,
+    PolynomialConfig,
+    StandardConfig,
+    ToroidalConfig,
+    ZernikeConfig,
+    config_registry,
+)
 
 if TYPE_CHECKING:
-    from optiland._types import ZernikeType
     from optiland.coordinate_system import CoordinateSystem
 
 
-@dataclass
-class GeometryConfig:
-    """Configuration parameters for creating a surface geometry.
-
-    Attributes:
-        radius (float): radius of curvature of the geometry (R = 1/c).
-                        Defaults to be.inf.
-        conic (float): conic constant (k) of the geometry. Defaults to 0.0.
-        grating_order (int): order of the grating. Defaults to 0.
-        grating_period (float): period of the grating. Defaults to be.inf.
-        groove_orientation_angle (float): angle of the groove orientation.
-                                        Defaults to 0.0.
-        coefficients (list): list of geometry coefficients. Defaults to empty list.
-        tol (float): tolerance to use for Newton-Raphson method. Defaults to 1e-6.
-        max_iter (int): maximum number of iterations to use for Newton-Raphson method.
-            Defaults to 100.
-        norm_x (float): normalization factor in x. Defaults to 1.0.
-        norm_y (float): normalization factor in y. Defaults to 1.0.
-        norm_radius (float): normalization radius. Defaults to 1.0.
-        radius_x (float): radius of curvature in x for biconic.
-                          Defaults to be.inf.
-        radius_y (float): radius of curvature in y for biconic or YZ radius for
-            toroidal. Defaults to be.inf.
-        conic_x (float): conic constant in x for biconic. Defaults to 0.0.
-        conic_y (float): conic constant in y for biconic. Defaults to 0.0.
-        toroidal_coeffs_poly_y (list): toroidal YZ polynomial coefficients.
-                                    Defaults to empty list.
-        zernike_type (str): type of Zernike polynomial to use. Defaults to "fringe".
-        radial_terms (dict): radial terms for Forbes Q-BFS surfaces.
-        freeform_coeffs (dict): freeform coefficients for Forbes Q-2D surfaces.
-    """
-
-    radius: float = be.inf
-    conic: float = 0.0
-    grating_order: int = 0
-    grating_period: float = be.inf
-    groove_orientation_angle: float = 0.0
-    coefficients: list[float] = field(default_factory=list)
-    tol: float = 1e-6
-    max_iter: int = 100
-    norm_x: float = 1.0
-    norm_y: float = 1.0
-    norm_radius: float = 1.0
-    # Biconic and Toroidal parameters
-    radius_x: float = be.inf  # Used by Biconic and Toroidal (as radius_x)
-    radius_y: float = be.inf  # Used by Biconic and Toroidal (as radius_yz)
-    conic_x: float = 0.0  # Used by Biconic
-    conic_y: float = 0.0  # Used by Biconic
-    toroidal_coeffs_poly_y: list[float] = field(default_factory=list)
[…]

[…]
