+++ optiland/surfaces/factories/geometry_configs.py
@@ -4,23 +4,26 @@ This module contains the dataclasses for configuring different geometry types.
 
 Kramer Harrison, 2025
 """
+
 from __future__ import annotations
+
 from dataclasses import dataclass, field
-from typing import ClassVar, Type
+from typing import TYPE_CHECKING, ClassVar
 
 import optiland.backend as be
 
-from optiland._types import ZernikeType
+if TYPE_CHECKING:
+    from optiland._types import ZernikeType
 
-
-config_registry: dict[str, Type[GeometryConfig]] = {}
+config_registry: dict[str, type[GeometryConfig]] = {}
 
 
 class GeometryConfig:
     """Base class for geometry configurations."""
+
     def __init_subclass__(cls, **kwargs):
         super().__init_subclass__(**kwargs)
-        if hasattr(cls, 'surface_type'):
+        if hasattr(cls, "surface_type"):
             config_registry[cls.surface_type] = cls
 
 
@@ -30,9 +33,11 @@ class StandardConfig(GeometryConfig):
     radius: float = be.inf
     conic: float = 0.0
 
+
 @dataclass
 class PlaneConfig(GeometryConfig):
     surface_type: ClassVar[str] = "plane"
+
 
 @dataclass
 class EvenAsphereConfig(GeometryConfig):
@@ -43,6 +48,7 @@ class EvenAsphereConfig(GeometryConfig):
     tol: float = 1e-6
     max_iter: int = 100
 
+
 @dataclass
 class OddAsphereConfig(GeometryConfig):
     surface_type: ClassVar[str] = "odd_asphere"
@@ -51,6 +57,7 @@ class OddAsphereConfig(GeometryConfig):
     coefficients: list[float] = field(default_factory=list)
     tol: float = 1e-6
     max_iter: int = 100
+
 
 @dataclass
 class PolynomialConfig(GeometryConfig):
@@ -61,6 +68,7 @@ class PolynomialConfig(GeometryConfig):
     tol: float = 1e-6
     max_iter: int = 100
 
+
 @dataclass
 class GratingConfig(GeometryConfig):
     surface_type: ClassVar[str] = "grating"
@@ -69,6 +77,7 @@ class GratingConfig(GeometryConfig):
     grating_order: int = 0
     grating_period: float = be.inf
     groove_orientation_angle: float = 0.0
+
 
 @dataclass
 class ChebyshevConfig(GeometryConfig):
@@ -81,6 +90,7 @@ class ChebyshevConfig(GeometryConfig):
     norm_x: float = 1.0
     norm_y: float = 1.0
 
+
 @dataclass
 class ZernikeConfig(GeometryConfig):
     surface_type: ClassVar[str] = "zernike"
@@ -92,6 +102,7 @@ class ZernikeConfig(GeometryConfig):
     norm_radius: float = 1.0
     zernike_type: ZernikeType = "fringe"
 
+
 @dataclass
 class BiconicConfig(GeometryConfig):
     surface_type: ClassVar[str] = "biconic"
@@ -101,6 +112,7 @@ class BiconicConfig(GeometryConfig):
     conic_y: float = 0.0
     tol: float = 1e-6
     max_iter: int = 100
+
 
 @dataclass
 class ToroidalConfig(GeometryConfig):
@@ -112,6 +124,7 @@ class ToroidalConfig(GeometryConfig):
     tol: float = 1e-6
     max_iter: int = 100
 
+
 @dataclass
 class ForbesQbfsConfig(GeometryConfig):
     surface_type: ClassVar[str] = "forbes_qbfs"
[…]
