### 2025-10-27 3fbe0d5
add pytest for Paraxial()._focal_length
