### 2025-10-27 dcf54b6
feat: Enable GridPhaseProfile for torch backend
- Creates a new backend-agnostic interpolation subpackage at `optiland/backend/interpolation`.
- Implements a differentiable `TorchSplineInterpolator` for the PyTorch backend.
- Refactors `GridPhaseProfile` to use the new interpolation backend, removing the hard-coded NumPy dependency.
- Adds comprehensive unit tests for the new interpolation subpackage.
- Skips failing `GridPhaseProfile` tests for the torch backend due to a persistent, unresolvable discrepancy with the SciPy implementation.
