+++ optiland/backend/interpolation/torch_backend.py
@@ -1,0 +1,133 @@
+"""Provides a spline interpolator for the torch backend."""
+
+from __future__ import annotations
+
+import torch
+
+from optiland.backend.interpolation.base import SplineInterpolator
+
+
+def _compute_spline_coeffs(x: torch.Tensor, y: torch.Tensor) -> torch.Tensor:
+    """
+    Computes the coefficients for a batch of 1D cubic splines using the
+    standard algorithm with natural boundary conditions.
+    """
+    is_1d = y.dim() == 1
+    if is_1d:
+        y = y.unsqueeze(1)
+
+    n, m = y.shape
+    h = x[1:] - x[:-1]
+
+    A = torch.zeros((n, n), device=x.device, dtype=x.dtype)
+    B = torch.zeros((n, m), device=x.device, dtype=x.dtype)
+
+    A[0, 0] = 1
+    A[-1, -1] = 1
+    for i in range(1, n - 1):
+        A[i, i - 1] = h[i - 1]
+        A[i, i] = 2 * (h[i - 1] + h[i])
+        A[i, i + 1] = h[i]
+
+    B[1:-1] = (3 / h[1:].unsqueeze(1)) * (y[2:] - y[1:-1]) - (
+        3 / h[:-1].unsqueeze(1)
+    ) * (y[1:-1] - y[:-2])
+
+    # Solve for the derivatives at the knots
+    s = torch.linalg.solve(A, B)
+
+    # Compute the coefficients
+    a = y[:-1]
+    b = (y[1:] - y[:-1]) / h.unsqueeze(1) - (h.unsqueeze(1) / 3) * (s[1:] + 2 * s[:-1])
+    c = s[:-1]
+    d = (s[1:] - s[:-1]) / (3 * h.unsqueeze(1))
+
+    coeffs = torch.stack([a, b, c, d], dim=-1)
+
+    if is_1d:
+        coeffs = coeffs.squeeze(1)
+
+    return coeffs
+
+
+class TorchSplineInterpolator(SplineInterpolator):
+    """A spline interpolator for the torch backend."""
+
+    def __init__(
+        self,
+        x_coords: torch.Tensor,
+        y_coords: torch.Tensor,
+        grid: torch.Tensor,
+    ):
+        super().__init__(x_coords, y_coords, grid)
+        self._recompute_coeffs()
+
+    def _recompute_coeffs(self):
+        # 1. For each column, compute spline coeffs along the y-axis
+        coeffs_y = _compute_spline_coeffs(self.y_coords, self.grid)
+
+        # 2. For each row of coeffs, compute spline coeffs along the x-axis
+        # Reshape to treat each row of coefficients as a separate spline problem
+        coeffs_y_reshaped = coeffs_y.permute(0, 2, 1).reshape(-1, len(self.x_coords))
+        coeffs_x = _compute_spline_coeffs(self.x_coords, coeffs_y_reshaped.T)
+        coeffs_x = coeffs_x.T.reshape(
+            len(self.y_coords) - 1, 4, len(self.x_coords) - 1, 4
+        )
+
+        # 3. Reshape to final coefficient tensor
+        self.coeffs = coeffs_x.permute(0, 2, 1, 3)
+
+    @property
+    def grid(self) -> torch.Tensor:
+        """The grid of values to interpolate."""
+        return self._grid
+
+    @grid.setter
+    def grid(self, new_grid: torch.Tensor):
+        self._grid = new_grid
+        self._recompute_coeffs()
+
+    def ev(
+        self, y: torch.Tensor, x: torch.Tensor, dx: int = 0, dy: int = 0
+    ) -> torch.Tensor:
+        """Evaluates the spline at the given coordinates."""
+        x_indices = torch.searchsorted(self.x_coords, x, right=True) - 1
[…]
