### 2025-10-28 a497233
feat: Add five new apodization types
Adds five new apodization functions to the Optiland library:
- SuperGaussianApodization
- CosineSquaredApodization
- TukeyApodization
- PolynomialApodization
- HannApodization

Each apodization is implemented in its own file within the `optiland/apodization` directory.

The `Optic.set_apodization` method has been updated to allow instantiation of these apodizations via string identifiers or dictionaries, using the existing registry pattern in `BaseApodization`.

Includes comprehensive unit tests for all new classes, ensuring correctness and compatibility wi
[…]
