+++ optiland/apodization/__init__.py
@@ -1,5 +1,10 @@
 # flake8: noqa
 
 from .base import BaseApodization
+from .cosine_squared import CosineSquaredApodization
 from .gaussian import GaussianApodization
+from .hann import HannApodization
+from .polynomial import PolynomialApodization
+from .super_gaussian import SuperGaussianApodization
+from .tukey import TukeyApodization
 from .uniform import UniformApodization
