+++ tests/test_grid_sag_geometry.py
@@ -1,0 +1,105 @@
+"""Unit tests for the GridSagGeometry class."""
+
+import pytest
+
+import optiland.backend as be
+from optiland.coordinate_system import CoordinateSystem
+from optiland.geometries.grid_sag import GridSagGeometry
+from optiland.rays.real_rays import RealRays
+
+
+@pytest.fixture
+def sample_grid_geometry(set_test_backend):
+    """Returns a sample GridSagGeometry instance for testing."""
+    cs = CoordinateSystem()
+    x_coords = [-1.0, 0.0, 1.0]
+    y_coords = [-1.0, 0.0, 1.0]
+    sag_values = [[0.1, 0.2, 0.1], [0.2, 0.4, 0.2], [0.1, 0.2, 0.1]]  # A simple hill
+    return GridSagGeometry(cs, x_coords, y_coords, sag_values)
+
+
+def test_grid_sag_geometry_init(sample_grid_geometry):
+    """Test the initialization of the GridSagGeometry class."""
+    assert sample_grid_geometry is not None
+    assert be.array_equal(sample_grid_geometry.x_grid, be.asarray([-1.0, 0.0, 1.0]))
+    assert be.array_equal(sample_grid_geometry.y_grid, be.asarray([-1.0, 0.0, 1.0]))
+    assert be.array_equal(
+        sample_grid_geometry.sag_grid,
+        be.asarray([[0.1, 0.2, 0.1], [0.2, 0.4, 0.2], [0.1, 0.2, 0.1]]),
+    )
+
+
+@pytest.mark.parametrize("backend", be.list_available_backends())
+def test_grid_sag_sag_calculation(backend):
+    """Test the sag calculation for the GridSagGeometry."""
+    be.set_backend(backend)
+    cs = CoordinateSystem()
+    x_coords = [0.0, 1.0]
+    y_coords = [0.0, 1.0]
+    sag_values = [[0.0, 0.1], [0.2, 0.3]]
+    geometry = GridSagGeometry(cs, x_coords, y_coords, sag_values)
+
+    # Test interpolation
+    assert be.isclose(geometry.sag(0.5, 0.5), be.asarray(0.15))
+    # Test on grid points
+    assert be.isclose(geometry.sag(0.0, 0.0), be.asarray(0.0))
+    assert be.isclose(geometry.sag(1.0, 1.0), be.asarray(0.3))
+    # Test out of bounds
+    assert be.isnan(geometry.sag(-1.0, 0.5))
+
+
+@pytest.mark.parametrize("backend", be.list_available_backends())
+def test_grid_sag_distance_and_normal(backend):
+    """Test the distance and surface normal calculations."""
+    be.set_backend(backend)
+    cs = CoordinateSystem()
+    x_coords = [-1.0, 1.0]
+    y_coords = [-1.0, 1.0]
+    # A simple plane tilted in x
+    sag_values = [[-0.1, 0.1], [-0.1, 0.1]]
+    geometry = GridSagGeometry(cs, x_coords, y_coords, sag_values)
+
+    # Ray starting at (0, 0, -1) and traveling along +z
+    rays = RealRays(
+        x=[0.0], y=[0.0], z=[-1.0], L=[0.0], M=[0.0], N=[1.0], intensity=1.0, wavelength=0.55
+    )
+
+    distance = geometry.distance(rays)
+    dtype = rays.x.dtype
+    assert be.allclose(distance, be.asarray([1.0], dtype=dtype), atol=1e-5)
+
+    # Move ray to the surface and check normal
+    rays.z = rays.z + distance
+    nx, ny, nz = geometry.surface_normal(rays)
+
[…]
