### 2025-10-29 8dfdbee
feat: Add optimizable variable for GridSagGeometry
This commit introduces the `GridSagVariable`, a new optimization variable that allows the sag values of a `GridSagGeometry` surface to be optimized.

Key changes:
- Created `GridSagVariable` in `optiland/optimization/variable/grid_sag.py` to manage the `sag_grid` parameter.
- Updated `GridSagGeometry` to accept backend arrays, enabling efficient updates during optimization.
- Added a comprehensive test suite for `GridSagVariable`, including a test to ensure compatibility with the PyTorch backend and its autograd functionality.
