### 2025-10-29 093fee4
Update optiland/geometries/grid_sag.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
