### 2025-10-29 909e9ee
refactor: Simplify test functions by removing backend parameterization
