### 2025-10-29 94a25d2
Update optiland/geometries/grid_sag.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
