### 2025-10-29 83e83b0
Update optiland/geometries/grid_sag.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
