### 2025-10-29 9177b60
Feature grid sag geometry (#387)
* feat: Add GridSagGeometry for freeform surfaces

This commit introduces a new `GridSagGeometry` class to `optiland`. This geometry allows users to define a surface using a grid of sag values, which is useful for representing complex, freeform surfaces from measured data or other simulations.

The implementation includes:
- A `GridSagGeometry` class with methods for sag calculation (using bilinear interpolation), ray-surface intersection (using a Newton-Raphson solver), and surface normal calculation.
- Integration into the existing `GeometryFactory` for easy 
[…]
