+++ optiland/geometries/grid_sag.py
@@ -1,0 +1,180 @@
+"""Grid Sag Geometry
+
+This module defines a geometry based on a grid of sag values, which are
+interpolated to find the surface shape.
+
+Kramer Harrison, 2025
+"""
+
+from __future__ import annotations
+
+from typing import Any
+
+import optiland.backend as be
+from optiland.coordinate_system import CoordinateSystem
+from optiland.geometries.base import BaseGeometry
+
+
+class GridSagGeometry(BaseGeometry):
+    """Represents a geometry defined by a grid of sag values.
+
+    The sag of the surface at any (x, y) point is determined by bilinearly
+    interpolating the surrounding points in the sag grid.
+
+    Args:
+        coordinate_system (CoordinateSystem): The coordinate system of the geometry.
+        x_coordinates (list[float]): 1D list of x-coordinates for the grid.
+        y_coordinates (list[float]): 1D list of y-coordinates for the grid.
+        sag_values (ArrayLike): 2D array-like (list of lists, numpy array, or
+            torch tensor) of sag values, with shape (len(y_coordinates),
+            len(x_coordinates)).
+        tol (float, optional): Tolerance for the iterative ray intersection solver.
+            Defaults to 1e-6.
+        max_iter (int, optional): Maximum number of iterations for the solver.
+            Defaults to 100.
+    """
+
+    def __init__(
+        self,
+        coordinate_system: CoordinateSystem,
+        x_coordinates: list[float],
+        y_coordinates: list[float],
+        sag_values: Any,
+        tol: float = 1e-6,
+        max_iter: int = 100,
+    ):
+        super().__init__(coordinate_system)
+        self.x_grid = be.asarray(x_coordinates)
+        self.y_grid = be.asarray(y_coordinates)
+        self.sag_grid = be.asarray(sag_values)
+        self.tol = tol
+        self.max_iter = max_iter
+        self.is_symmetric = False
+
+        if self.sag_grid.shape != (len(self.y_grid), len(self.x_grid)):
+            raise ValueError(
+                f"Shape of sag_values {self.sag_grid.shape} must match "
+                f"(len(y_coordinates), len(x_coordinates)) = "
+                f"({len(self.y_grid)}, {len(self.x_grid)})."
+            )
+
+    def _interpolate(self, x, y):
+        """Performs bilinear interpolation and calculates derivatives."""
+        # Find indices for lower-left corner of the interpolation cell
+        # be.searchsorted with side="right" returns the index after elements equal to x/y,
+        # so we subtract 1 to get the lower bound index for interpolation.
+        i = be.searchsorted(self.x_grid, x, side="right") - 1
+        j = be.searchsorted(self.y_grid, y, side="right") - 1
+
+        # Handle out-of-bounds cases
+        nan_mask = (
+            (x < self.x_grid[0])
+            | (x > self.x_grid[-1])
+            | (y < self.y_grid[0])
+            | (y > self.y_grid[-1])
+        )
+        i = be.where(i < 0, 0, i)
+        j = be.where(j < 0, 0, j)
+        i = be.where(i >= len(self.x_grid) - 1, len(self.x_grid) - 2, i)
+        j = be.where(j >= len(self.y_grid) - 1, len(self.y_grid) - 2, j)
+
+        # Get corner coordinates and sag values
+        x1, x2 = self.x_grid[i], self.x_grid[i + 1]
+        y1, y2 = self.y_grid[j], self.y_grid[j + 1]
+        z11, z12 = self.sag_grid[j, i], self.sag_grid[j, i + 1]
[…]
