### 2025-10-29 f18c73b
feat(visualization): Add interactive features to 2D visualizations
This commit introduces a suite of interactive features to the 2D
visualizations in the `optiland/visualization` subpackage.

A new `InteractionManager` class has been added to handle user
interactions with the Matplotlib plots. This class is responsible for
managing hover and click events, as well as displaying tooltips and
information panels.

The `OpticViewer` has been updated to use the `InteractionManager` and
the plotting methods for `Surface2D`, `Lens2D`, and `Rays2D` have been
refactored to return the Matplotlib artists 
[…]
