### 2025-10-29 399e9d1
fix: Update comments for clarity in _interpolate method
