### 2025-10-29 ae0779b
feat(visualization): Update 2D plots and fix interaction bugs
This commit addresses several issues with the 2D visualization module.

- Fixes a "RuntimeError: Another Axes already grabs mouse input" by properly disconnecting the event handler for the pop-up close button.
- Updates the pop-up info box for surfaces to display the surface number, radius, and conic constant.
- Implements color cycling for rays based on fields and wavelengths, with fields taking precedence.
- Removes the legend from all 2D plots to improve clarity.
