### 2025-10-29 37c1b33
fix(rays): Simplify line plotting syntax in Rays2D class
