### 2025-10-29 e648ad5
feat(visualization): Overhaul 2D interaction and info display
This commit addresses several issues with the 2D visualization module.

- Fixes a "RuntimeError: Another Axes already grabs mouse input" by manually releasing the mouse grabber before closing the info panel.
- Updates and modularizes the pop-up info box content generation by implementing a strategy pattern with info providers.
- Implements ray bundle highlighting on hover, making the line artists thicker.
- All changes are SOLID compliant and pass ruff checks.
