### 2025-10-29 6ccc645
fix(visualization): Overhaul 2D interaction, info display, and highlighting
This commit provides a comprehensive overhaul of the 2D visualization module, addressing several critical bugs and implementing new features.

- **Fix `RuntimeError`:** Resolves the "Another Axes already grabs mouse input" error by temporarily disconnecting the main plot's event listeners when the info panel is active, preventing event conflicts.
- **Update Pop-up Content:** Corrects the info provider logic to ensure the correct information (surface number, radius, conic) is displayed for all elements, including surfac
[…]
