### 2025-10-29 190bc8b
Return array for Abbe material extinction coefficient
Returning a float breaks ray tracing, as that relies on `be.any` which requires arrays / tensors
