### 2025-10-29 0385f66
fix(visualization): Final fixes for interaction, content, and highlighting
This commit provides a final, robust set of fixes for the 2D visualization module.

- **Fix `RuntimeError`:** Replaces the problematic Matplotlib `Button` with a custom click handler for the info panel's close button, completely eliminating the event handling conflict.
- **Implement Correct Highlighting:** Implements correct per-bundle highlighting for rays by assigning a `bundle_id` to each ray and using it to highlight all rays in the same bundle.
- **Correct Pop-up Content:** Fixes the logic for displaying pop-up con
[…]
