### 2025-10-29 e7b396e
Use zeros_like to return an array with the same shape as wavelength
