### 2025-10-29 7a2f374
fix(docs): Update author name in module docstrings
