### 2025-10-29 ababcf1
refactor(interaction): Enhance docstrings and type hints for clarity and consistency
