+++ optiland/visualization/system/interaction.py
@@ -1,4 +1,4 @@
-"""Interaction Manager Module
+"""Interaction Manager Module.
 
 This module provides the InteractionManager class for handling user interactions
 with Matplotlib-based visualizations of optical systems.
@@ -8,9 +8,17 @@ Kramer Harrison, 2025
 
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
+import matplotlib.axes
+
 from optiland.visualization.info.providers import INFO_PROVIDER_REGISTRY
 from optiland.visualization.system.lens import Lens2D
 from optiland.visualization.system.surface import Surface2D
+
+if TYPE_CHECKING:
+    import matplotlib.figure
+    from matplotlib.backend_bases import Event
 
 
 class InteractionManager:
@@ -19,13 +27,38 @@ class InteractionManager:
     This class connects to a Matplotlib figure's event loop to handle
     mouse events, such as hovering and clicking, on plotted artists.
 
-    Args:
+    Attributes:
         fig (matplotlib.figure.Figure): The Matplotlib figure to connect to.
         ax (matplotlib.axes.Axes): The Matplotlib axes containing the artists.
-
+        artist_registry (dict): Maps Matplotlib artists to Optiland objects.
+        active_artist (matplotlib.artist.Artist | None): The artist currently
+            under the mouse cursor.
+        original_props (dict): Stores original properties of artists before
+            highlighting, to allow restoration.
+        tooltip (matplotlib.text.Annotation): The annotation object used for
+            displaying tooltips.
+        tooltip_format (callable): A function that takes an Optiland object
+            and returns a string for its tooltip.
+        info_panel (matplotlib.axes.Axes | None): The axes object for the
+            detailed information panel, if currently displayed.
+        cids (list[int]): A list of connection IDs for Matplotlib events.
+        info_panel_cid (int): The connection ID for the info panel's click event.
     """
 
-    def __init__(self, fig, ax, tooltip_format=None):
+    def __init__(
+        self,
+        fig: matplotlib.figure.Figure,
+        ax: matplotlib.axes.Axes,
+        tooltip_format: callable | None = None,
+    ):
+        """Initializes the InteractionManager.
+
+        Args:
+            fig (matplotlib.figure.Figure): The Matplotlib figure to connect to.
+            ax (matplotlib.axes.Axes): The Matplotlib axes containing the artists.
+            tooltip_format (callable | None, optional): A function to format
+                tooltip text. If None, `default_tooltip_format` is used.
+        """
         self.fig = fig
         self.ax = ax
         self.artist_registry = {}
@@ -38,22 +71,36 @@ class InteractionManager:
             xy=(0, 0),
             xytext=(20, 20),
             textcoords="offset points",
-            bbox=dict(boxstyle="round", fc="w"),
-            arrowprops=dict(arrowstyle="->"),
+            bbox={"boxstyle": "round", "fc": "w"},
+            arrowprops={"arrowstyle": "->"},
         )
         self.tooltip.set_visible(False)
         self.tooltip_format = tooltip_format or self.default_tooltip_format
 
         self.info_panel = None
+        self.info_panel_cid = None
         self.cids = []
         self.connect()
 
-    def register_artist(self, artist, optiland_object):
-        """Registers a Matplotlib artist with its corresponding Optiland object."""
+    def register_artist(
+        self, artist: matplotlib.artist.Artist, optiland_object: object
+    ):
+        """Registers a Matplotlib artist with its corresponding Optiland object.
+
+        Args:
+            artist (matplotlib.artist.Artist): The Matplotlib artist (e.g.,
+                Line2D, Patch) to be tracked.
+            optiland_object (object): The Optiland object (e.g., Surface2D,
[…]
