### 2025-10-29 9de6200
fix(visualization): Comprehensive fixes for 2D/3D interaction, content, and highlighting
This commit provides a comprehensive set of fixes for the visualization module, addressing critical bugs in both 2D and 3D plotting, and implementing all requested features.

- **Fix `RuntimeError`:** Replaces the problematic Matplotlib `Button` with a custom click handler for the info panel's close button, completely eliminating the event handling conflict.
- **Implement Correct Highlighting:** Implements correct per-bundle highlighting for rays by assigning a `bundle_id` to each ray and using it to highl
[…]
