### 2025-10-29 5f634fd
fix(info): Correct number of rays calculation in RayBundleInfoProvider
