### 2025-10-29 5cba5be
feat(visualization): Overhaul tooltip functionality and disable pop-ups
This commit provides a comprehensive overhaul of the tooltip functionality in the 2D visualization module and disables the pop-up info box.

- **Disable Pop-up Box:** Removes the pop-up info box and its associated event handlers to prevent `RuntimeError`s.
- **Implement Tooltip Hover Delay:** Adds a timer to the `InteractionManager` to ensure that tooltips only appear after a short delay, improving the user experience.
- **Update Tooltip Style and Content:** Removes the arrow from the tooltip and repurposes the `INFO_PROVI
[…]
