+++ optiland/visualization/system/interaction.py
@@ -7,6 +7,8 @@ with Matplotlib-based visualizations of optical systems.
 
 from __future__ import annotations
 
+from threading import Timer
+
 from optiland.visualization.info.providers import INFO_PROVIDER_REGISTRY
 from optiland.visualization.system.lens import Lens2D
 from optiland.visualization.system.surface import Surface2D
@@ -24,9 +26,10 @@ class InteractionManager:
 
     """
 
-    def __init__(self, fig, ax, tooltip_format=None):
+    def __init__(self, fig, ax, optic, tooltip_format=None):
         self.fig = fig
         self.ax = ax
+        self.optic = optic
         self.artist_registry = {}
         self.active_artist = None
         self.original_props = {}
@@ -38,13 +41,15 @@ class InteractionManager:
             xytext=(20, 20),
             textcoords="offset points",
             bbox=dict(boxstyle="round", fc="w"),
-            arrowprops=dict(arrowstyle="->"),
         )
         self.tooltip.set_visible(False)
         self.tooltip_format = tooltip_format or self.default_tooltip_format
 
         self.info_panel = None
         self.cids = []
+        self.hover_timer = None
+        self.last_hover_time = 0
+        self.hover_delay = 0.7  # seconds
         self.connect()
 
     def register_artist(self, artist, optiland_object):
@@ -57,8 +62,7 @@ class InteractionManager:
             cid_hover = self.fig.canvas.mpl_connect(
                 "motion_notify_event", self.on_hover
             )
-            cid_click = self.fig.canvas.mpl_connect("button_press_event", self.on_click)
-            self.cids.extend([cid_hover, cid_click])
+            self.cids.extend([cid_hover])
 
     def disconnect(self):
         """Disconnects from the Matplotlib event loop."""
@@ -89,22 +93,22 @@ class InteractionManager:
         if self.active_artist != found_artist:
             if self.active_artist:
                 self.clear_hover_effects()
+            if self.hover_timer:
+                self.hover_timer.cancel()
 
             if found_artist:
                 self.active_artist = found_artist
-                self.highlight_artist(found_artist)
-                self.show_tooltip(found_artist, event)
+                self.hover_timer = Timer(
+                    self.hover_delay, self.show_tooltip, args=[found_artist, event]
+                )
+                self.hover_timer.start()
 
-    def on_click(self, event):
-        """Handles click events to show detailed information."""
-        if event.inaxes != self.ax:
-            return
-
-        for artist, optiland_object in self.artist_registry.items():
-            contains, _ = artist.contains(event)
-            if contains:
-                self.show_info_panel(optiland_object)
-                break
+    # TODO: Re-enable pop-up box functionality in a future update.
+    # The following methods are temporarily disabled.
+    # def on_click(self, event):
+    # def show_info_panel(self, optiland_object):
+    # def on_info_panel_click(self, event):
+    # def close_info_panel(self, event=None):
 
     def highlight_artist(self, artist):
         """Highlights the given artist."""
@@ -121,6 +125,7 @@ class InteractionManager:
 
     def show_tooltip(self, artist, event):
         """Shows a tooltip for the given artist."""
+        self.highlight_artist(artist)
         optiland_object = self.artist_registry[artist]
         tooltip_text = self.tooltip_format(optiland_object)
 
@@ -131,6 +136,8 @@ class InteractionManager:
 
     def clear_hover_effects(self):
         """Clears any active hover effects."""
+        if self.hover_timer:
+            self.hover_timer.cancel()
         if self.active_artist:
             obj = self.artist_registry[self.active_artist]
             if hasattr(obj, "bundle_id"):
@@ -156,12 +163,20 @@ class InteractionManager:
 
[…]
