### 2025-10-29 3afa41d
fix(surfaces): Add clear() method to SurfaceGroup to prevent GUI crash
This commit fixes a critical bug that caused the GUI to crash upon initialization.

The `surfaces` attribute of the `SurfaceGroup` class is a `cached_property` that returns a `tuple`, which is immutable. The GUI was attempting to call `.clear()` on this tuple, resulting in an `AttributeError`.

This commit resolves the issue by:
- Adding a `clear()` method to the `SurfaceGroup` class that correctly clears the internal `_surfaces` list.
- Modifying the GUI connector to call this new `clear()` method instead of attempting to 
[…]
