### 2025-10-29 105702a
fix(info): Update LensInfoProvider to return material instead of surface count
