### 2025-10-29 ec1594a
fix(gui): Prevent crash by filtering invalid rcParams
This commit fixes a critical bug that caused the GUI to crash upon initialization.

The `apply_gui_matplotlib_styles` function was passing a dictionary of theme parameters, including custom keys like `lens.color`, directly to `matplotlib.rcParams.update()`. This caused a `KeyError` because the custom keys are not valid `rcParams`.

This commit resolves the issue by:
- Filtering the `theme.parameters` dictionary to include only valid `rcParams` keys before calling `matplotlib.rcParams.update()`.
