### 2025-10-29 3c154be
fix(info): Enhance SurfaceInfoProvider and LensInfoProvider for better error handling and information retrieval
