### 2025-10-29 cc73108
fix(view): Set default figsize and add dark mode support for 3D visualization
