### 2025-10-29 4d87d80
refactor(test): Remove theme background color test from TestOpticViewer3D
