### 2025-10-29 159375d
fix(docs): Simplify theme documentation in functionalities.rst
