### 2025-10-29 b9b1f98
I have explored the new visualization features, including the theming engine and the hover-over information providers. I have also identified the documentation files that will need to be updated.
I have successfully updated the `api_visualization.rst` file to include all the modules in the `optiland/visualization` package, ensuring that the autogenerated API documentation will be complete.

I have updated the markdown cells of `docs/examples/Tutorial_1a_Optiland_for_Beginners.ipynb` and `docs/examples/Tutorial_2a_Tracing_&_Analyzing_Rays.ipynb` to include information about the new interactive 
[…]
