+++ optiland/visualization/system/mirror.py
@@ -4,6 +4,7 @@ This module contains the Mirror3D class, which is used to visualize a 3D mirror
 surface.
 
 Kramer Harrison, 2024
+Kramer Harrison, 2025
 """
 
 from __future__ import annotations
