+++ optiland/visualization/system/lens.py
@@ -3,6 +3,7 @@
 This module contains classes for visualizing lenses in 2D and 3D.
 
 Kramer Harrison, 2024
+Kramer Harrison, 2025
 """
 
 from __future__ import annotations
