+++ docs/gallery/miscellaneous.rst
@@ -15,3 +15,4 @@ or unique functionality of Optiland.
     miscellaneous/apodization
     miscellaneous/paraxial_to_thick_lens
     miscellaneous/paraxial_image_height_field
+    miscellaneous/themes
