### 2025-10-29 38b800b
feat(theme): Add theme registration functionality and list available themes
