### 2025-10-29 e078fdf
fix: Add missing newline after module docstring in analysis, info, and system packages
