+++ optiland/visualization/info/__init__.py
@@ -2,6 +2,7 @@
 
 Kramer Harrison, 2025
 """
+
 from __future__ import annotations
 
 from .lens_info_viewer import LensInfoViewer
