+++ optiland/visualization/analysis/__init__.py
@@ -2,6 +2,7 @@
 
 Kramer Harrison, 2025
 """
+
 from __future__ import annotations
 
 from .surface_sag import SurfaceSagViewer
