### 2025-10-29 457711d
docs: ipynb formatting
