+++ optiland/visualization/system/surface.py
@@ -3,7 +3,6 @@
 This module contains classes for visualizing optical surfaces in 2D and 3D.
 
 Kramer Harrison, 2024
-Kramer Harrison, 2025
 """
 
 from __future__ import annotations
