### 2025-10-30 e47fd0c
feat: Add initial framework for GUI test suite
This commit introduces the foundational framework for the Optiland GUI test suite.

- Adds `pytest-qt` to the development dependencies to enable testing of PySide6 applications.
- Creates the `tests/gui/` directory to house all future GUI-related tests.
- Implements a core `conftest.py` with a reusable `app` fixture to manage the `MainWindow` lifecycle.
- Includes an initial smoke test (`test_main_window_creation`) to verify that the main window can be instantiated in a headless environment.
- Fixes an `AttributeError` in the `OptilandConnector` w
[…]
