+++ pyproject.toml
@@ -35,7 +35,7 @@ Homepage = "https://github.com/HarrisonKramer/optiland"
 optiland = "optiland_gui.run_gui:main"
 
 [project.optional-dependencies]
-dev = ["codecov>=2.1.13", "pytest>=8.3.5", "torch"]
+dev = ["codecov>=2.1.13", "pytest>=8.3.5", "torch", "pytest-qt"]
 gui = ["PySide6", "qtconsole"]
 
 [tool.hatch.metadata]
