### 2025-10-30 a6870c5
feat(gui): Add comprehensive GUI tests
This commit introduces a suite of new tests for the `optiland-gui` package, including atomic tests for the `ActionManager` and `PanelManager`, as well as end-to-end tests for the "New System" and "Add Surface" workflows.

The following changes are included:
- New test files `tests/gui/test_action_manager.py` and `tests/gui/test_panel_manager.py`.
- New end-to-end tests in `tests/gui/test_main_window.py`.
- Updates to `pyproject.toml` to include necessary testing dependencies.
- A new `autouse` fixture in `tests/gui/conftest.py` to mock the `ViewerPanel` a
[…]
