+++ tests/gui/test_panel_manager.py
@@ -1,0 +1,123 @@
+import pytest
+from unittest.mock import MagicMock, call
+from PySide6.QtWidgets import QMainWindow, QDockWidget, QWidget
+from PySide6.QtCore import Qt
+
+from optiland_gui.panel_manager import PanelManager
+from optiland_gui.main_window import MainWindow
+
+@pytest.fixture
+def mocked_app(mocker, qtbot):
+    """
+    Creates a MainWindow instance with all panel classes mocked to return
+    QWidget instances with the necessary mock attributes.
+    """
+    # Mock all panel classes to return a QWidget instance with mock attributes
+    def mock_panel(*args, **kwargs):
+        widget = QWidget()
+        widget.menuSelected = MagicMock()
+        widget.commandExecuted = MagicMock()
+        widget.update_theme = MagicMock()
+        widget.set_theme = MagicMock()
+        widget.update_icons = MagicMock()
+        widget.shutdown_kernel = MagicMock()
+        return widget
+
+    mocker.patch('optiland_gui.panel_manager.LensEditor', side_effect=mock_panel)
+    mocker.patch('optiland_gui.panel_manager.SystemPropertiesPanel', side_effect=mock_panel)
+    mocker.patch('optiland_gui.panel_manager.AnalysisPanel', side_effect=mock_panel)
+    mocker.patch('optiland_gui.panel_manager.PythonTerminalWidget', side_effect=mock_panel)
+    mocker.patch('optiland_gui.panel_manager.SidebarWidget', side_effect=mock_panel)
+
+    # Now that panels are mocked, create the MainWindow instance
+    main_window = MainWindow()
+    qtbot.addWidget(main_window)
+    yield main_window
+    main_window.close()
+
+@pytest.fixture
+def panel_manager(mocked_app):
+    """
+    Returns the PanelManager from the mocked_app.
+    """
+    return mocked_app.panel_manager
+
+
+def test_create_all_panels(panel_manager):
+    """Test that PanelManager correctly instantiates all panels and docks."""
+    assert isinstance(panel_manager.sidebar, QDockWidget)
+    assert isinstance(panel_manager.viewer_dock, QDockWidget)
+    assert isinstance(panel_manager.lens_editor_dock, QDockWidget)
+    assert isinstance(panel_manager.system_properties_dock, QDockWidget)
+    assert isinstance(panel_manager.analysis_dock, QDockWidget)
+    assert isinstance(panel_manager.terminal_dock, QDockWidget)
+    assert len(panel_manager.all_docks) == 6
+
+
+def test_setup_default_layout(mocker, mocked_app, panel_manager):
+    """Test that the default panel layout is applied correctly."""
+    mocker.patch.object(mocked_app, 'addDockWidget')
+    mocker.patch.object(mocked_app, 'splitDockWidget')
+    mocker.patch.object(mocked_app, 'tabifyDockWidget')
+
+    for dock in panel_manager.all_docks:
+        mocker.patch.object(dock, 'setFloating')
+        mocker.patch.object(dock, 'show')
+        mocker.patch.object(dock, 'raise_')
+
+    panel_manager.setup_default_layout()
+
+    for dock in panel_manager.all_docks:
+        dock.setFloating.assert_called_with(False)
+        dock.show.assert_called()
+        dock.raise_.assert_called()
+
+    assert mocked_app.addDockWidget.called
+    mocked_app.splitDockWidget.assert_has_calls([
+        call(panel_manager.lens_editor_dock, panel_manager.analysis_dock, Qt.Horizontal),
+        call(panel_manager.lens_editor_dock, panel_manager.viewer_dock, Qt.Vertical),
[…]
