+++ optiland_gui/undo_redo_manager.py
@@ -61,10 +61,6 @@ class UndoRedoManager(QObject):
         self._redo_stack.clear()  # Clear redo stack whenever a new action is performed
         self.undoStackAvailabilityChanged.emit(self.can_undo())
         self.redoStackAvailabilityChanged.emit(self.can_redo())
-        print(
-            f"UndoRedoManager: State added. Undo stack size: {len(self._undo_stack)}, "
-            f"Redo stack size: {len(self._redo_stack)}"
-        )
 
     def undo(self, current_state_for_redo):
         """
@@ -89,10 +85,6 @@ class UndoRedoManager(QObject):
 
         self.undoStackAvailabilityChanged.emit(self.can_undo())
         self.redoStackAvailabilityChanged.emit(self.can_redo())
-        print(
-            f"UndoRedoManager: Undo. Undo stack size: {len(self._undo_stack)}, "
-            f"Redo stack size: {len(self._redo_stack)}"
-        )
         return restored_state
 
     def redo(self, current_state_for_undo):
@@ -121,10 +113,6 @@ class UndoRedoManager(QObject):
 
         self.undoStackAvailabilityChanged.emit(self.can_undo())
         self.redoStackAvailabilityChanged.emit(self.can_redo())
-        print(
-            f"UndoRedoManager: Redo. Undo stack size: {len(self._undo_stack)}, "
-            f"Redo stack size: {len(self._redo_stack)}"
-        )
         return restored_state
 
     def can_undo(self):
