### 2025-10-30 f2705fd
feat(gui): Add atomic tests for LensEditor
This commit introduces a new test module for the `LensEditor` widget, adding atomic tests for its core functionalities.

The following changes are included:
- New test file `tests/gui/test_lens_editor.py`.
- Atomic tests for adding and removing surfaces.
- A test to verify that cell data changes are correctly propagated to the connector.
- A test for the properties widget toggle functionality.

This contributes to the user's goal of building a comprehensive and modular test suite for the GUI to catch regressions.
