### 2025-10-30 4998fc7
Add new color palettes and update theme configurations
- Introduced a new 'midnight' color palette with vibrant colors for enhanced visibility.
- Added 'edges' color to existing palettes: light, dark, solarized light, solarized dark.
- Updated the 'axes.edgecolor' in the Theme class to use the 'edges' color from the palette.
- Updated the dark_palette with new colors for background, axis, text, grid, lens, and ray_cycle.
