### 2025-10-30 c442278
fix: Remove redundant theme usage examples from functionalities documentation
