+++ docs/functionalities.rst
@@ -62,20 +62,6 @@ Visualization
   Plot optical layouts, surface properties, and ray traces using Matplotlib (2D) and VTK (3D). The 2D plots are fully interactive, providing tooltips with detailed information when you hover over optical components.
 - **Customizable Visualization Themes**:
   Optiland includes a flexible theme system to control the appearance of 2D plots. You can easily switch between predefined themes (e.g., ``light``, ``dark``) or create your own. For a hands-on example, see the :doc:`gallery/miscellaneous/themes` notebook.
-
-  To set a theme globally, use the ``set_theme`` function:
-
-  .. code-block:: python
-
-      import optiland.visualization as vis
-      vis.set_theme('dark')
-
-  You can also use the ``theme_context`` context manager to apply a theme temporarily:
-
-  .. code-block:: python
-
-        with vis.theme_context('solarized_light'):
-            optic.draw()
 - **Optiland GUI**:
   A fully functional graphical user interface (GUI) for interactive design, analysis, and visualization of optical systems. See the :ref:`gui_quickstart` for more details on using the GUI.
 
