### 2025-10-30 a74e08a
refactor: Remove interactive 2D visualization section from quickstart documentation
