### 2025-10-30 db152a0
refactor: Simplify surface addition in CookeTriplet lens design
