### 2025-10-30 a7a2028
feat(gui): Add atomic tests for SidebarWidget
This commit introduces a new test module for the `SidebarWidget`, adding atomic tests for its core functionalities.

The following changes are included:
- New test file `tests/gui/test_sidebar_widget.py`.
- Atomic tests for sidebar button clicks, including work-in-progress buttons.
- A test to verify that the theme update correctly changes the button icons.

This contributes to the user's goal of building a comprehensive and modular test suite for the GUI to catch regressions.
