### 2025-10-30 649a4f8
Feature visualization themes (#389)
* feat: Implement visualization theme framework

Adds a new theme-based framework for controlling the appearance of visualizations.

The framework is managed through the `optiland.visualization.themes` module, which provides `set_theme()` and `theme_context()` for global and temporary theme management.

Includes four predefined themes: `light`, `dark`, `solarized_light`, and `solarized_dark`.

Integrates the theme system into the 2D and 3D plotting viewers, as well as the `optiland_gui`.

Adds comprehensive tests and documentation for the new functionality.

[…]
