+++ optiland/visualization/analysis/__init__.py
@@ -1,3 +1,8 @@
+"""Analysis visualization package for Optiland.
+
+Kramer Harrison, 2025
+"""
+
 from __future__ import annotations
 
 from .surface_sag import SurfaceSagViewer
