+++ optiland/visualization/system/lens.py
@@ -34,16 +34,18 @@ class Lens2D:
         # TODO: raise warning when lens surfaces overlap
         self.surfaces = surfaces
 
-    def plot(self, ax):
+    def plot(self, ax, theme=None):
         """Plots the lens on the given matplotlib axis.
 
         Args:
             ax (matplotlib.axes.Axes): The matplotlib axis on which the
                 lens will be plotted.
+            theme (Theme, optional): The theme to use for plotting.
+                Defaults to None.
 
         """
         sags = self._compute_sag()
-        self._plot_lenses(ax, sags)
+        return self._plot_lenses(ax, sags, theme=theme)
 
     def _compute_sag(self, apply_transform=True):
         """Computes the sag of the lens in local coordinates and handles
@@ -105,7 +107,7 @@ class Lens2D:
 
         return x, y, z
 
-    def _plot_single_lens(self, ax, x, y, z):
+    def _plot_single_lens(self, ax, x, y, z, theme=None):
         """Plot a single lens on the given matplotlib axis.
 
         Args:
@@ -114,18 +116,28 @@ class Lens2D:
             x (numpy.ndarray): The x coordinates of the lens.
             y (numpy.ndarray): The y coordinates of the lens.
             z (numpy.ndarray): The z coordinates of the lens.
+            theme (Theme, optional): The theme to use for plotting.
+                Defaults to None.
 
         """
+        facecolor = (0.8, 0.8, 0.8, 0.6)
+        edgecolor = (0.5, 0.5, 0.5)
+        if theme:
+            facecolor = theme.parameters.get("lens.color", facecolor)
+            edgecolor = theme.parameters.get("axes.edgecolor", edgecolor)
+
         vertices = be.to_numpy(be.column_stack((z, y)))
         polygon = Polygon(
             vertices,
             closed=True,
-            facecolor=(0.8, 0.8, 0.8, 0.6),
-            edgecolor=(0.5, 0.5, 0.5),
+            facecolor=facecolor,
+            edgecolor=edgecolor,
+            label="Lens",
         )
         ax.add_patch(polygon)
+        return polygon
 
-    def _plot_lenses(self, ax, sags):
+    def _plot_lenses(self, ax, sags, theme=None):
         """Plot the lenses on the given matplotlib axis.
 
         Args:
@@ -133,8 +145,11 @@ class Lens2D:
                 lenses will be plotted.
             sags (list): A list of tuples containing arrays of x, y, and z
                 coordinates for each surface.
+            theme (Theme, optional): The theme to use for plotting.
+                Defaults to None.
 
         """
+        artists = {}
         for k in range(len(sags) - 1):
             x1, y1, z1 = sags[k]
             x2, y2, z2 = sags[k + 1]
@@ -144,7 +159,9 @@ class Lens2D:
             y = be.concatenate([y1, be.flip(y2)])
             z = be.concatenate([z1, be.flip(z2)])
 
-            self._plot_single_lens(ax, x, y, z)
+            artist = self._plot_single_lens(ax, x, y, z, theme=theme)
+            artists[artist] = self
+        return artists
 
 
 class Lens3D(Lens2D):
@@ -196,21 +213,23 @@ class Lens3D(Lens2D):
                 return False
         return True
 
-    def plot(self, renderer):
+    def plot(self, renderer, theme=None):
         """Plots the lens or surfaces using the provided renderer.
 
         Args:
             renderer: The rendering engine used to plot the lens or surfaces.
+            theme (Theme, optional): The theme to use for plotting.
+                Defaults to None.
 
         """
         if self.is_symmetric:
             sags = self._compute_sag()  # sags are in global coordinates
-            self._plot_lenses(renderer, sags)
[…]
