### 2025-10-31 804f461
feat(gui): Add tests for SystemPropertiesPanel and OptimizationPanel
This commit adds a comprehensive test suite for the SystemPropertiesPanel and its sub-editors (Aperture, Fields, and Wavelengths). It also adds tests for the placeholder OptimizationPanel.

Key additions and changes:
- Created a new test module for the SystemPropertiesPanel, covering navigation, and full CRUD operations for aperture, fields, and wavelengths.
- Created a new test module for the placeholder OptimizationPanel.
- Fixed a bug where the OptimizationPanel was not being instantiated in the PanelManager.
