+++ tests/gui/test_system_properties_panel.py
@@ -1,0 +1,217 @@
+import pytest
+from PySide6.QtCore import Qt
+from PySide6.QtWidgets import QWidget
+
+from optiland_gui.system_properties_panel import (
+    ApertureEditor,
+    FieldsEditor,
+    SystemPropertiesPanel,
+    WavelengthsEditor,
+)
+
+@pytest.fixture
+def properties_panel(app):
+    """Returns the SystemPropertiesPanel instance from the main application window."""
+    return app.panel_manager.system_properties
+
+def test_system_properties_panel_creation(properties_panel):
+    """
+    Test that the SystemPropertiesPanel and its child editors are created.
+    """
+    assert isinstance(properties_panel, SystemPropertiesPanel)
+    assert isinstance(properties_panel.apertureEditor, ApertureEditor)
+    assert isinstance(properties_panel.fieldsEditor, FieldsEditor)
+    assert isinstance(properties_panel.wavelengthsEditor, WavelengthsEditor)
+
+def test_navigation_tree_switching(properties_panel):
+    """
+    Test that clicking an item in the navigation tree switches the stacked widget.
+    """
+    # Find the "Fields" item in the navigation tree
+    fields_item = properties_panel.navTree.findItems("Fields", Qt.MatchExactly)[0]
+
+    # Simulate a click on the "Fields" item
+    properties_panel.navTree.itemClicked.emit(fields_item, 0)
+
+    # Check that the stacked widget has switched to the FieldsEditor
+    assert properties_panel.stackedWidget.currentWidget() == properties_panel.fieldsEditor
+
+def test_aperture_editor_changes(mocker, properties_panel):
+    """
+    Test that clicking the 'Apply' button calls the connector with the correct values.
+    """
+    # Get the aperture editor
+    aperture_editor = properties_panel.apertureEditor
+
+    # Mock the set_aperture method on the Optic object
+    mock_set_aperture = mocker.patch.object(properties_panel.connector.get_optic(), 'set_aperture')
+
+    # Disconnect the reactive signals to isolate the button's functionality
+    aperture_editor.cmbApertureType.currentTextChanged.disconnect(aperture_editor.apply_aperture_changes)
+    aperture_editor.spnApertureValue.valueChanged.disconnect(aperture_editor.apply_aperture_changes)
+
+    # Simulate a user changing the aperture type and value
+    aperture_editor.cmbApertureType.setCurrentText("imageFNO")
+    aperture_editor.spnApertureValue.setValue(5.6)
+
+    # Ensure no calls happened yet
+    mock_set_aperture.assert_not_called()
+
+    # Simulate a click on the apply button
+    aperture_editor.btnApplyAperture.click()
+
+    # Check that the set_aperture method was called once with the correct arguments
+    mock_set_aperture.assert_called_once_with("imageFNO", 5.6)
+
+def test_fields_editor_add_field(mocker, properties_panel):
+    """
+    Test that clicking the 'Add Field' button calls the connector.
+    """
+    # Get the fields editor
+    fields_editor = properties_panel.fieldsEditor
+
+    # Mock the add_field method on the Optic object
+    mock_add_field = mocker.patch.object(properties_panel.connector.get_optic(), 'add_field')
+
+    # Simulate a click on the add field button
+    fields_editor.btnAddField.click()
+
+    # Check that the add_field method was called
+    mock_add_field.assert_called_once()
+
+def test_fields_editor_remove_field(mocker, properties_panel):
+    """
+    Test that clicking the 'Remove Field' button removes the selected field.
+    """
[…]
