### 2025-10-31 095a65b
feat(gui): Add integration and analysis contract tests
This commit introduces a new suite of high-level integration tests to provide a safety net against regressions in the GUI caused by changes in the `optiland` core library.

Key additions:
- Created a new test module, `tests/gui/test_integration_workflows.py`, to house end-to-end and contract tests.
- Implemented a full user workflow test that simulates loading a sample system, modifying a property, and running an analysis.
- Implemented a parameterized "analysis contract" test that iterates through all available analyses in the `AnalysisPa
[…]
