### 2025-10-31 c126601
feat(ci): Separate core and GUI test jobs
This commit refactors the CI/CD workflow to run the core library tests and the GUI tests in two separate, parallel jobs. This provides clearer, more immediate feedback and improves the robustness of the testing pipeline.

The `tests` job has been renamed to `test-core` and is now configured to exclude the `tests/gui/` directory.

A new `test-gui` job has been created with the following configuration:
- Runs on a matrix of Python versions.
- Installs necessary system dependencies (`xvfb`, `libgl1-mesa-glx`) for headless testing.
- Runs only the tests in
[…]
