### 2025-11-01 b55f3fe
fix(ci): update test commands to use pytest-cov for coverage reporting
