### 2025-11-01 1c310b5
fix(ci): update test commands to include torch in coverage runs
