### 2025-11-01 2c55a26
fix(ci): correct comments and update test commands for GUI tests to include PySide6
