### 2025-11-01 3adeeed
fix(ci): add libegl1 to system dependencies for GUI tests
