### 2025-11-01 a445bcc
fix(ci): update GUI test command to include qtconsole in temporary installations
