+++ .github/workflows/ci.yml
@@ -87,8 +87,8 @@ jobs:
 
       - name: Run GUI tests
         run: |
-          # Use --with to temporarily install pytest-cov, torch, and PySide6
-          QT_QPA_PLATFORM=offscreen uv run --with pytest-cov,torch,PySide6 pytest --durations=0 --cov --cov-report=xml tests/gui/
+          # Use --with to temporarily install pytest-cov, torch, qtconsole and PySide6
+          QT_QPA_PLATFORM=offscreen uv run --with pytest-cov,torch,qtconsole,PySide6 pytest --durations=0 --cov --cov-report=xml tests/gui/
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v4
