+++ tests/gui/test_sidebar_widget.py
@@ -1,7 +1,6 @@
 import pytest
 from PySide6.QtWidgets import QWidget
 from PySide6.QtGui import QIcon
-from optiland_gui.widgets.sidebar import SidebarWidget
 
 @pytest.fixture
 def sidebar(app):
