### 2025-11-01 2761b50
Update tests/gui/test_sidebar_widget.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
