+++ tests/gui/test_sidebar_widget.py
@@ -1,6 +1,5 @@
 import pytest
 from PySide6.QtWidgets import QWidget
-from PySide6.QtGui import QIcon
 
 @pytest.fixture
 def sidebar(app):
