+++ tests/gui/test_panel_manager.py
@@ -3,7 +3,6 @@ from unittest.mock import MagicMock, call
 from PySide6.QtWidgets import QMainWindow, QDockWidget, QWidget
 from PySide6.QtCore import Qt
 
-from optiland_gui.panel_manager import PanelManager
 from optiland_gui.main_window import MainWindow
 
 @pytest.fixture
