### 2025-11-01 5e7515f
Update tests/gui/test_panel_manager.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
