### 2025-11-01 ce92880
Update tests/gui/test_integration_workflows.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
