### 2025-11-01 7b42914
Update tests/gui/test_analysis_panel.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
