### 2025-11-01 e5d0477
Update optiland_gui/main_window.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
