### 2025-11-01 c3bd33e
Update optiland_gui/main_window.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
