### 2025-11-01 f5f94ce
Update tests/gui/test_optimization_panel.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
