### 2025-11-01 99b9e39
fix(ci): update GUI test command to include pytest-mock in temporary installations
