### 2025-11-01 00e336c
fix(ci): update GUI test command to use xvfb-run for virtual display
