+++ .github/workflows/ci.yml
@@ -83,12 +83,12 @@ jobs:
       - name: Install system dependencies
         run: |
           sudo apt-get update
-          sudo apt-get install -y libgl1 xvfb libegl1
+          sudo apt-get install -y xvfb libgl1 libegl1 libxkbcommon-x11-0 mesa-utils
 
-      - name: Run GUI tests
+      - name: Run GUI tests in virtual display
         run: |
-          # Use --with to temporarily install pytest-cov, torch, qtconsole and PySide6
-          QT_QPA_PLATFORM=offscreen uv run --with pytest-cov,pytest-mock,torch,qtconsole,PySide6 pytest --durations=0 --cov --cov-report=xml tests/gui/
+          xvfb-run -a uv run --with pytest-cov,pytest-mock,torch,qtconsole,PySide6 \
+            pytest --durations=0 --cov --cov-report=xml tests/gui/
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v4
