### 2025-11-01 e95d11b
fix: comment out unused test-core job and update test-gui dependencies
