### 2025-11-01 a71d285
fix: update GUI test command to run in virtual display and include additional dependencies
