### 2025-11-01 b1f9bf9
fix(ci): restore test-core job and update coverage upload steps
