+++ .github/workflows/ci.yml
@@ -36,34 +36,34 @@ jobs:
       - uses: ./.github/actions/setup
       - run: uvx ruff format --check .
 
-  # test-core:
-  #   runs-on: ubuntu-latest
-  #   needs: [lock_file]
-  #   strategy:
-  #     fail-fast: false
-  #     matrix:
-  #       python-version: ["3.10", "3.11", "3.12", "3.13"]
-  #   steps:
-  #     - uses: actions/checkout@v4
-  #     - uses: ./.github/actions/setup
-  #     - name: Set up Python ${{ matrix.python-version }}
-  #       uses: actions/setup-python@v5
-  #       with:
-  #         python-version: ${{ matrix.python-version }}
+  test-core:
+    runs-on: ubuntu-latest
+    needs: [lock_file]
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.10", "3.11", "3.12", "3.13"]
+    steps:
+      - uses: actions/checkout@v4
+      - uses: ./.github/actions/setup
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
 
-  #     - name: Run tests with coverage
-  #       run: |
-  #         # Use --with to temporarily install pytest-cov and torch
-  #         uv run --with pytest-cov,torch pytest --durations=0 --cov --cov-report=xml --ignore=tests/gui
+      - name: Run tests with coverage
+        run: |
+          # Use --with to temporarily install pytest-cov and torch
+          uv run --with pytest-cov,torch pytest --durations=0 --cov --cov-report=xml --ignore=tests/gui
 
-  #     - name: Upload coverage to Codecov
-  #       uses: codecov/codecov-action@v4
-  #       with:
-  #         token: ${{ secrets.CODECOV_TOKEN }}
-  #         files: ./coverage.xml
-  #         flags: unittests
-  #         name: codecov-umbrella
-  #         fail_ci_if_error: true
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v4
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          files: ./coverage.xml
+          flags: unittests
+          name: codecov-umbrella
+          fail_ci_if_error: true
 
   test-gui:
     runs-on: ubuntu-latest
@@ -102,7 +102,7 @@ jobs:
 
   build:
     runs-on: ubuntu-latest
-    needs: [linting, formatting, test-gui]
+    needs: [linting, formatting, test-core, test-gui]
     steps:
       - uses: actions/checkout@v4
       - uses: ./.github/actions/setup
