### 2025-11-01 e694f42
feat(materials): Add comprehensive tests for materials module
Adds a comprehensive test suite for the optiland/materials module,
bringing test coverage to nearly 100%.

- Creates `tests/test_material_utils.py` with mocked tests for all utility functions.
- Expands `tests/test_materials.py` to cover `Material` class logic, including error handling and initialization.
- Adds mock data files in `tests/material_files` to support isolated testing.
- Fixes several bugs discovered during testing, including handling of empty clusters in `downsample_glass_map` and backend type errors in `get_neighbour_
[…]
