+++ tests/material_files/formula_7.yml
@@ -1,0 +1,3 @@
+DATA:
+  - type: formula 7
+    coefficients: 1.5 0.5 0.1
