### 2025-11-01 aed3f43
feat: Increase test coverage for optiland_gui
Adds a new test file for the UndoRedoManager and new tests for the following files:
- main_window.py
- lens_editor.py
- system_properties_panel.py
