### 2025-11-02 57d7ffc
fix(ci): specify coverage source for tests in CI workflow
