+++ .github/workflows/ci.yml
@@ -54,7 +54,7 @@ jobs:
       - name: Run tests with coverage
         run: |
           # Use --with to temporarily install pytest-cov and torch
-          uv run --with pytest-cov,torch pytest --durations=0 --cov --cov-report=xml --ignore=tests/gui
+          uv run --with pytest-cov,torch pytest --durations=0 --cov=optiland --cov-report=xml --ignore=tests/gui
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v4
@@ -89,7 +89,7 @@ jobs:
         run: |
           export QT_QPA_PLATFORM=offscreen
           uv run --with pytest-cov,pytest-mock,torch,qtconsole,PySide6,pytest-qt,matplotlib \
-            pytest --durations=0 --cov --cov-report=xml tests/gui/
+            pytest --durations=0 --cov=optiland --cov-report=xml tests/gui/
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v4
