### 2025-11-02 3d720c3
fix(ci): update coverage source for GUI tests in CI workflow
