### 2025-11-02 e7c66ed
feat(tests): Improve test coverage for optiland/geometries
This commit improves the test coverage for the `optiland/geometries` subpackage.

- Adds new, atomic unit tests for the following modules:
  - `optiland/geometries/forbes/jacobi.py`
  - `optiland/geometries/nurbs/nurbs_fitting.py`
  - `optiland/geometries/nurbs/nurbs_basis_functions.py`
  - `optiland/geometries/nurbs/nurbs_geometry.py`
- Increases overall test coverage for the `optiland/geometries` subpackage from 61% to 74%.
- Adds `# pragma: no cover` to JIT-compiled functions in `nurbs_basis_functions.py` to correctly exclude them f
[…]
