+++ tests/test_nurbs_geometry.py
@@ -104,3 +104,53 @@ def test_nurbs_distance(set_test_backend):
     )
     distance = geo.distance(rays)
     assert_allclose(distance, 10, atol=1e-4)
+
+@pytest.fixture
+def sample_control_points():
+    """Provides a simple 3x3 grid of 3D control points as a Python list."""
+    dim, n, m = 3, 3, 3
+    points = [[[0.0 for _ in range(m)] for _ in range(n)] for _ in range(dim)]
+    for i in range(n):
+        for j in range(m):
+            points[0][i][j] = float(i)
+            points[1][i][j] = float(j)
+            points[2][i][j] = 0.0
+    return points
+
+@pytest.mark.parametrize('backend', be.list_available_backends())
+def test_nurbs_geometry_init_fitted(backend, set_test_backend):
+    """Tests the initialization of a fitted NurbsGeometry."""
+    be.set_backend(backend)
+    cs = CoordinateSystem()
+
+    # Initialize without control points to trigger fitting logic
+    geom = NurbsGeometry(cs, n_points_u=5, n_points_v=5)
+
+    assert geom.is_fitted
+    assert geom.P is None
+    assert geom.W is None
+    assert geom.ndim == 3
+    assert geom.P_size_u == 6  # n_points_u + 1
+    assert geom.P_size_v == 6  # n_points_v + 1
+
+@pytest.mark.parametrize('backend', be.list_available_backends())
+def test_nurbs_geometry_init_bezier(backend, set_test_backend, sample_control_points):
+    """Tests the Polynomial Bezier surface initialization of NurbsGeometry."""
+    be.set_backend(backend)
+    cs = CoordinateSystem()
+    control_points = be.asarray(sample_control_points)
+
+    geom = NurbsGeometry(cs, control_points=control_points)
+
+    assert not geom.is_fitted
+    assert geom.surface_type == "Bezier"
+    assert be.allclose(geom.P, control_points)
+
+    n = be.shape(control_points)[1] - 1
+    m = be.shape(control_points)[2] - 1
+
+    assert geom.p == n
+    assert geom.q == m
+    assert geom.U is not None
+    assert geom.V is not None
+    assert be.allclose(geom.W, be.ones((n + 1, m + 1)))
