### 2025-11-02 ab0cf27
refactor(zernike): streamline backend imports and enhance Zernike class handling
