### 2025-11-02 ac72fd1
feat(gui): Add tests and fix bugs in optiland_gui
This commit adds a comprehensive suite of smoke tests to the `optiland_gui` package, significantly increasing test coverage for several key components.

New test files were created for:
- `optiland_connector.py`
- `analysis_panel.py` (expanded)
- `main_window.py` (expanded)
- `frameless_window.py`
- `viewer_panel.py`
- `python_terminal.py`

Key achievements during this task include:
- **Bug Fixes:** Discovered and fixed critical bugs related to serialization and the undo/redo functionality. Specifically, the `Optic.name` and `Surface.comment` a
[…]
