### 2025-11-02 be4c3ab
Refactor: Generalize PhaseInteractionModel and 3D Gradient
This commit generalizes the `PhaseInteractionModel` to work with any surface geometry, not just planes.

Key changes:
- The `BasePhaseProfile.get_gradient` method now returns a full 3D gradient `(d_phi/dx, d_phi/dy, d_phi/dz)`.
- All concrete phase profile implementations have been updated to match the new 3D interface.
- The `PhaseInteractionModel` now uses the 3D gradient and local surface normal to correctly apply the Generalized Snell's Law on curved surfaces.
- The `TypeError` restricting the model to `Plane` geometries has been r
[…]
