+++ optiland/phase/base.py
@@ -50,7 +50,9 @@ class BasePhaseProfile(abc.ABC):
         raise NotImplementedError
 
     @abc.abstractmethod
-    def get_gradient(self, x: be.Array, y: be.Array) -> tuple[be.Array, be.Array]:
+    def get_gradient(
+        self, x: be.Array, y: be.Array
+    ) -> tuple[be.Array, be.Array, be.Array]:
         """Calculates the gradient of the phase at coordinates (x, y).
 
         Args:
@@ -58,8 +60,8 @@ class BasePhaseProfile(abc.ABC):
             y: The y-coordinates of the points of interest.
 
         Returns:
-            A tuple containing the x and y components of the phase gradient
-            (d_phi/dx, d_phi/dy).
+            A tuple containing the x, y, and z components of the phase gradient
+            (d_phi/dx, d_phi/dy, d_phi/dz).
         """
         raise NotImplementedError
 
