### 2025-11-03 89473d5
feat(vis): Add framework for 2D analysis plots on 3D surfaces
Introduces a new framework for rendering 2D analysis plots directly onto surfaces within the 3D VTK visualization.

Key features:
- A `SurfacePlot` dataclass for configuring plots.
- An `AnalysisPlot3D` class that handles analysis execution, rasterization of Matplotlib plots to a high-resolution texture, and creation of a textured VTK actor.
- Integration into the `Optic.draw3D` method via a new `surface_plots` argument, allowing users to programmatically specify which plots to render on which surfaces.
- The framework is generic an
[…]
