+++ optiland/visualization/system/analysis_plot.py
@@ -1,0 +1,163 @@
+"""
+AnalysisPlot3D: A class for rendering 2D analysis plots on 3D surfaces.
+
+This module provides the AnalysisPlot3D class, which is responsible for
+generating a VTK actor that visualizes a 2D analysis plot (e.g., a spot
+diagram) as a texture on a plane positioned and oriented at a specific
+surface within the 3D optical system view.
+
+"""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+import matplotlib.pyplot as plt
+import numpy as np
+import vtk
+
+import optiland.backend as be
+from vtk.util.numpy_support import numpy_to_vtk
+
+if TYPE_CHECKING:
+    from optiland.optic.optic import Optic
+    from optiland.visualization.system.surface_plot import SurfacePlot
+
+
+class AnalysisPlot3D:
+    """A class to generate a VTK actor for a 2D analysis plot projected
+    onto a 3D surface.
+
+    This class orchestrates the process of running a given analysis,
+    rasterizing its graphical output into a NumPy array, and mapping this
+    array as a VTK texture onto a plane actor that matches the position,
+    orientation, and size of a target optical surface.
+
+    Args:
+        optic (Optic): The optical system context.
+        config (SurfacePlot): The configuration object specifying which
+            analysis to run and on which surface to plot it.
+
+    """
+
+    def __init__(self, optic: "Optic", config: "SurfacePlot"):
+        self.optic = optic
+        self.config = config
+        self.actor = vtk.vtkActor()
+        self._create_plot_actor()
+
+    def _get_analysis_plot_as_array(self) -> np.ndarray:
+        """Runs the specified analysis and rasterizes the resulting plot into a
+        NumPy array.
+
+        Returns:
+            np.ndarray: A NumPy array representing the RGB image of the plot.
+
+        Raises:
+            AttributeError: If the analysis class does not have a 'view' or
+                'plot' method for visualization.
+
+        """
+        # Instantiate the analysis class with the optic and parameters
+        analysis = self.config.analysis_class(
+            self.optic, **self.config.analysis_params
+        )
+
+        # Generate the plot using either 'view' or 'plot' method
+        if hasattr(analysis, "view"):
+            fig, ax = analysis.view()
+        elif hasattr(analysis, "plot"):
+            fig, ax = analysis.plot()
+        else:
+            raise AttributeError(
+                f"Analysis class {self.config.analysis_class.__name__} does "
+                "not have a 'view' or 'plot' method."
+            )
+
+        # Configure the plot for clean rasterization
+        axes = ax if isinstance(ax, list) else [ax]
+        for axis in axes:
+            axis.axis("off")
+            axis.patch.set_alpha(1.0)  # Keep axes background opaque
+            axis.set_facecolor("white") # Set a fixed background for visibility
+
+        fig.tight_layout(pad=0)
+        fig.patch.set_alpha(0.0)  # Make figure background transparent
+
+        # Rasterize the figure to a NumPy array
+        fig.canvas.draw()
+        rgba_buffer = fig.canvas.buffer_rgba()
+        image_array = np.asarray(rgba_buffer)
+        # Keep only the RGB channels, discard alpha
+        rgb_array = image_array[:, :, :3]
+        plt.close(fig)  # Prevent the plot from being displayed interactively
+
+        return rgb_array
+
+    def _create_plot_actor(self):
+        """Creates the textured vtkActor for the analysis plot."""
+        # Get the plot as a high-resolution image array
+        image_array = self._get_analysis_plot_as_array()
[…]
