### 2025-11-04 dfb8bfd
Refactor SurfaceFactory to Stateless, Strategy-Based Architecture
Refactored the entire surface creation architecture to be stateless and SOLID-compliant, resolving architectural debt in the factory system.

- **Stateless Component Factories:**
  - `MaterialFactory` and `CoordinateSystemFactory` are now fully stateless with `@staticmethod` create methods.
  - `InteractionModelFactory` uses a registry pattern to be OCP-compliant.

- **Strategy Pattern Implementation:**
  - Introduced `BaseSurfaceStrategy` and concrete strategies (`Standard`, `Paraxial`, `Object`, `Grating`) to encapsulate creat
[…]
