### 2025-11-04 20d040c
refactor: Clean up type hints and imports in factory modules
