+++ optiland/surfaces/surface_group.py
@@ -12,9 +12,11 @@ from __future__ import annotations
 
 from contextlib import suppress
 from functools import cached_property
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 import optiland.backend as be
+from optiland.coatings import BaseCoatingPolarized
+from optiland.materials.base import BaseMaterial
 from optiland.surfaces.factories.coating_factory import CoatingFactory
 from optiland.surfaces.factories.coordinate_system_factory import (
     CoordinateSystemFactory,
@@ -29,17 +31,17 @@ from optiland.surfaces.factories.surface_factory import SurfaceFactory
 from optiland.surfaces.factories.types import SurfaceContext
 from optiland.surfaces.standard_surface import Surface
 
+if TYPE_CHECKING:
+    from optiland._types import SurfaceType
+
 
 class SurfaceGroup:
     """Represents a group of surfaces in an optical system."""
 
     def __init__(self, surfaces: list[Surface] | None = None):
         """Initializes a new instance of the SurfaceGroup class."""
-        if surfaces is None:
-            self._surfaces = []
-        else:
-            self._surfaces = surfaces
-            self._update_surface_links()
+        self._surfaces: list[Surface] = surfaces if surfaces is not None else []
+        self._update_surface_links()
 
         cs_factory = CoordinateSystemFactory()
         geom_factory = GeometryFactory()
@@ -56,26 +58,15 @@ class SurfaceGroup:
             int_factory,
             strategy_provider,
         )
+        self._has_absolute_z = False
 
     def _update_surface_links(self):
         with suppress(KeyError):
             self.__dict__.pop("surfaces")
-        surfaces = self._surfaces
-        if surfaces:
-            surfaces[0].previous_surface = None
-            for idx, surface in enumerate(surfaces[1:]):
-                surface.previous_surface = surfaces[idx]
-
-    def __add__(self, other: SurfaceGroup) -> SurfaceGroup:
-        offset = self.surfaces[-1].geometry.cs.z if self.surfaces else 0.0
-        object_distance = other.surfaces[0].geometry.cs.z
-        if be.isfinite(object_distance):
-            offset = offset - object_distance
-        for surf in other.surfaces[1:]:
-            surf.geometry.cs.z = surf.geometry.cs.z + offset
-        for surface in other.surfaces:
-            surface.is_stop = False
-        return SurfaceGroup(self._surfaces[:-1] + other._surfaces[1:])
+        if self._surfaces:
+            self._surfaces[0].previous_surface = None
+            for idx, surface in enumerate(self._surfaces[1:]):
+                surface.previous_surface = self._surfaces[idx]
 
     @cached_property
     def surfaces(self) -> tuple[Surface, ...]:
@@ -83,6 +74,7 @@ class SurfaceGroup:
 
     def clear(self):
         self._surfaces = []
+        self._has_absolute_z = False
         self._update_surface_links()
 
     @property
@@ -131,6 +123,123 @@ class SurfaceGroup:
         return positions.reshape(-1, 1)
 
     @property
+    def num_surfaces(self) -> int:
+        return len(self.surfaces)
+
+    @property
+    def total_track(self):
+        if self.num_surfaces < 2:
+            raise ValueError("Not enough surfaces to calculate total track.")
+        z = self.positions[1:]
+        return be.max(z) - be.min(z)
+
+    def add_surface(self, index: int, **kwargs: Any):
+        if "z" in kwargs:
+            if "thickness" in kwargs:
+                raise ValueError('Cannot define both "thickness" and "z".')
+            if "dx" in kwargs or "dy" in kwargs:
+                raise ValueError(
+                    'Cannot define "dx" or "dy" when using absolute "x", "y", "z".'
+                )
+            self._has_absolute_z = True
[…]
