### 2025-11-05 32e77ef
refactor: Clean up type hints and imports in factory modules
