### 2025-11-05 5614f4e
Update optiland/surfaces/factories/strategies/concrete.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
