### 2025-11-05 186ac4d
Update optiland/surfaces/factories/strategies/concrete.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
