### 2025-11-05 f6f76dd
fix: Simplify config retrieval in GratingStrategy's create_geometry method
