### 2025-11-05 bcf9fb9
Refactor SurfaceFactory to Stateless, Strategy-Based Architecture
Refactored the entire surface creation architecture to be stateless and SOLID-compliant, resolving architectural debt in the factory system.

- **Stateless Component Factories:**
  - `MaterialFactory` and `CoordinateSystemFactory` are now fully stateless.
  - `InteractionModelFactory` uses a registry pattern to be OCP-compliant.

- **Strategy Pattern Implementation:**
  - Introduced `BaseSurfaceStrategy` and concrete strategies to encapsulate creation logic, allowing any surface type for the object surface.
  - A `SurfaceStrateg
[…]
