+++ tests/test_surface_factory.py
@@ -1,5 +1,6 @@
 import optiland.backend as be
 import pytest
+from unittest.mock import MagicMock
 
 from .utils import assert_allclose
 from optiland.coatings import FresnelCoating, SimpleCoating
@@ -10,7 +11,9 @@ from optiland.surfaces.object_surface import ObjectSurface
 from optiland.interactions.thin_lens_interaction_model import ThinLensInteractionModel
 from optiland.phase.radial import RadialPhaseProfile
 from optiland.surfaces.standard_surface import Surface
-from optiland.surfaces import SurfaceFactory
+from optiland.surfaces.factories.surface_factory import SurfaceFactory
+from optiland.surfaces.factories.strategy_provider import SurfaceStrategyProvider
+from optiland.surfaces.factories.types import SurfaceContext
 from optiland.optic import Optic
 from optiland.fields import AngleField, Field
 
@@ -18,309 +21,70 @@ from optiland.fields import AngleField, Field
 class TestSurfaceFactory:
     @pytest.fixture(autouse=True)
     def setup(self):
-        self.lens = TessarLens()
-        self.surface_group = self.lens.surface_group
-        self.factory = SurfaceFactory(self.surface_group)
+        self.cs_factory = MagicMock()
+        self.geom_factory = MagicMock()
+        self.mat_factory = MagicMock()
+        self.coat_factory = MagicMock()
+        self.int_factory = MagicMock()
+        self.strategy_provider = SurfaceStrategyProvider()
+
+        self.factory = SurfaceFactory(
+            cs_factory=self.cs_factory,
+            geom_factory=self.geom_factory,
+            mat_factory=self.mat_factory,
+            coat_factory=self.coat_factory,
+            int_factory=self.int_factory,
+            strategy_provider=self.strategy_provider,
+        )
+
+        # Mock the return values for simple cases
+        self.mat_factory.create.return_value = IdealMaterial(n=1.0)
 
     def test_create_surface_standard(self, set_test_backend):
-        surface = self.factory.create_surface(
-            surface_type="standard",
-            comment="Standard",
-            index=1,
-            is_stop=False,
-            material="air",
-            thickness=5,
-            radius=10,
-            conic=0,
-        )
+        config = {
+            "surface_type": "standard",
+            "comment": "Standard",
+            "is_stop": False,
+            "material": "air",
+            "thickness": 5,
+            "radius": 10,
+            "conic": 0,
+        }
+        context = SurfaceContext(index=1, z=0.0, material_pre=IdealMaterial(n=1.0))
+        surface = self.factory.create_surface(config, context)
         assert isinstance(surface, Surface)
-        assert surface.geometry.radius == 10
-        assert surface.geometry.k == 0
-        assert isinstance(surface.material_pre, IdealMaterial)
-        assert isinstance(surface.material_post, IdealMaterial)
-
-    def test_create_surface_even_asphere(self, set_test_backend):
-        surface = self.factory.create_surface(
-            surface_type="even_asphere",
-            comment="Even Asphere",
-            index=1,
-            is_stop=False,
-            material="air",
-            thickness=5,
-            radius=10,
-            conic=0,
-            coefficients=[1, 2, 3],
-        )
-        assert isinstance(surface, Surface)
-        assert surface.geometry.radius == 10
-        assert surface.geometry.k == 0
-        assert surface.geometry.coefficients == [1, 2, 3]
-        assert isinstance(surface.material_pre, IdealMaterial)
-        assert isinstance(surface.material_post, IdealMaterial)
-
-    def test_create_surface_odd_asphere(self, set_test_backend):
-        surface = self.factory.create_surface(
-            surface_type="odd_asphere",
-            comment="Odd Asphere",
[…]
