+++ optiland/surfaces/factories/material_factory.py
@@ -10,6 +10,8 @@ Kramer Harrison, 2025
 
 from __future__ import annotations
 
+from typing import Union
+
 from optiland.materials import BaseMaterial, IdealMaterial, Material
 
 
@@ -18,8 +20,8 @@ class MaterialFactory:
 
     @staticmethod
     def create(
-        material_spec: BaseMaterial | tuple | str | None,
-    ) -> BaseMaterial | None:
+        material_spec: Union[BaseMaterial, tuple, str, None]
+    ) -> Union[BaseMaterial, None]:
         """Creates a material instance based on the given specification.
 
         This is a pure function that converts a material specification into a
