### 2025-11-05 1047bf3
feat(core): Introduce TracingContext and update base interfaces
This commit introduces the `TracingContext` dataclass, which will be used to pass contextual information during ray tracing. This is the first step in a larger refactoring effort to decouple ray tracing from a sequential-only architecture.

The abstract method signatures in `BaseInteractionModel` and `BaseCoating` have been updated to accept the `TracingContext`. This change intentionally breaks the build, which will be fixed in subsequent tasks.

- Created `optiland/raytrace/context.py` with `TracingContext` dataclass.
- Updated 
[…]
