### 2025-11-05 9cf63e6
Refactor: Implement Context-Based Interfaces for Interaction and Coating Models
This commit implements the new `..._with_context` method signatures for all `BaseInteractionModel` and `BaseCoating` subclasses, fixing the build breakage introduced in the previous task.

All interaction models and material-dependent coatings now source the pre-interaction material from the `TracingContext` instead of a `material_pre` attribute. This change affects both real and paraxial ray tracing logic.
