### 2025-11-05 02d621a
Refactor: Decouple Core Trace Loop
This commit refactors the core ray tracing loop to use a `TracingContext` object, decoupling the `Surface` class from its neighbors.

- Removed `previous_surface` and `material_pre` from `Surface`.
- Updated `Surface.trace` to accept and return a `TracingContext`.
- Simplified `SurfaceGroup.trace` to manage the context-passing loop.
- Updated `RealRayTracer` to initiate the context and the new trace.
