### 2025-11-08 c25a1a5
feat(fileio): Add support for Toroidal surfaces in Zemax reader
Adds the ability to load and correctly parse Zemax (.zmx) files that contain `TYPE TOROIDAL` surfaces.

Key changes:
- `ZemaxDataParser` now recognizes the "TOROIDAL" keyword.
- `ZemaxToOpticConverter` correctly handles the `toroidal` surface type, extracting the X-Radius from `CURV` and the Y-Radius from `PARM 1`.
- A `PARM 1` value of 0.0 is correctly interpreted as an infinite radius for cylindrical lenses.
- The logic is applied to both simple and complex (with COORDBRK) processing paths.
- The `ValueError` for unsupported sur
[…]
