+++ optiland/fileio/converters.py
@@ -99,40 +99,48 @@ class ZemaxToOpticConverter:
             translation, _ = self.current_cs.get_effective_transform()
             rx_, ry_, rz_ = self.current_cs.get_effective_rotation_euler()
             coeffs = self._configure_surface_coefficients(surf)
-            thickness = surf.get("thickness", 0.0)
 
-            # special care now, we ramify. for object surface,
-            # if DISZ == inf, then keep infinity
+            # Prepare surface parameters
+            surface_params = {
+                "index": surf_idx,
+                "surface_type": surf["type"],
+                "conic": surf.get("conic"),
+                "is_stop": surf.get("is_stop", False),
+                "material": surf.get("material"),
+                "coefficients": coeffs,
+            }
+
+            if surf["type"] == "toroidal":
+                radius_y = surf.get("param_1", 0.0)
+                if radius_y == 0.0:
+                    radius_y = be.inf
+                surface_params["radius_y"] = radius_y
+                surface_params["radius_x"] = surf.get("radius")
+            else:
+                surface_params["radius"] = surf.get("radius")
+
+            # Handle thickness and coordinate system parameters
+            thickness = surf.get("thickness", 0.0)
             if be.isinf(float(thickness)):
-                self.optic.add_surface(
-                    index=surf_idx,
-                    surface_type=surf["type"],
-                    radius=surf.get("radius"),
-                    conic=surf.get("conic"),
-                    thickness=thickness,
-                    is_stop=surf.get("is_stop", False),
-                    material=surf.get("material"),
-                    coefficients=coeffs,
-                    rx=float(rx_),
-                    ry=float(ry_),
-                    rz=float(rz_),
+                # For surfaces at infinity, set thickness and orientation
+                surface_params["thickness"] = thickness
+                surface_params.update(
+                    {"rx": float(rx_), "ry": float(ry_), "rz": float(rz_)}
                 )
-            else:  # normal surface, we pass no thickness argument
-                self.optic.add_surface(
-                    index=surf_idx,
-                    surface_type=surf["type"],
-                    radius=surf.get("radius"),
-                    conic=surf.get("conic"),
-                    is_stop=surf.get("is_stop", False),
-                    material=surf.get("material"),
-                    coefficients=coeffs,
-                    x=float(translation[0]),
-                    y=float(translation[1]),
-                    z=float(translation[2]),
-                    rx=float(rx_),
-                    ry=float(ry_),
-                    rz=float(rz_),
+            else:
+                # For normally positioned surfaces, set position and orientation
+                surface_params.update(
+                    {
+                        "x": float(translation[0]),
+                        "y": float(translation[1]),
+                        "z": float(translation[2]),
+                        "rx": float(rx_),
+                        "ry": float(ry_),
+                        "rz": float(rz_),
+                    }
                 )
+
+            self.optic.add_surface(**surface_params)
             surf_idx = surf_idx + 1
 
             # we need to advance the cs by the surface thickness
@@ -164,27 +172,39 @@ class ZemaxToOpticConverter:
             data (dict): The data for the surface.
         """
         coefficients = self._configure_surface_coefficients(data)
-        extra_params = {}
+
+        # Prepare surface parameters, starting with common ones
+        surface_params = {
+            "index": index,
+            "surface_type": data["type"],
+            "conic": data["conic"],
+            "thickness": data["thickness"],
+            "is_stop": data["is_stop"],
+            "material": data["material"],
+            "coefficients": coefficients,
+        }
+
[…]
