### 2025-11-09 5dd8503
fix: Improve handling of unsupported surface types in Zemax data parser
