+++ optiland/fileio/zemax_handler.py
@@ -309,7 +309,7 @@ class ZemaxDataParser:
             "ODDASPHE": "odd_asphere",
             "COORDBRK": "coordinate_break",
             "TOROIDAL": "toroidal",
-        }.get(data[1], "unsupported")
+        }.get(data[1], data[1].lower())  # Default to lowercased raw value
 
     def _read_surface_parameter(self, data):
         key = f"param_{int(data[1]) - 1}"
