### 2025-11-09 3a266d2
fix: Correct radius assignments for toroidal surfaces in Zemax converter
