### 2025-11-09 d129567
Update optiland/fileio/converters.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
