### 2025-11-09 42ca0d4
fix: Add toroidal coefficients for toroidal surfaces in Zemax converter
