### 2025-11-09 fdfa3f4
fix: Refactor surface parameter handling for toroidal surfaces in Zemax converter
