+++ optiland/fileio/converters.py
@@ -62,7 +62,6 @@ class ZemaxToOpticConverter:
         if not has_cb:
             for idx, surf_data in self.data["surfaces"].items():
                 self._configure_surface(idx, surf_data)
-            self.optic.add_surface(index=len(self.data["surfaces"]))
             return
 
         # in case there are "Coordinate Break" surfaces
@@ -99,40 +98,53 @@ class ZemaxToOpticConverter:
             translation, _ = self.current_cs.get_effective_transform()
             rx_, ry_, rz_ = self.current_cs.get_effective_rotation_euler()
             coeffs = self._configure_surface_coefficients(surf)
-            thickness = surf.get("thickness", 0.0)
 
-            # special care now, we ramify. for object surface,
-            # if DISZ == inf, then keep infinity
+            # Prepare surface parameters
+            surface_params = {
+                "index": surf_idx,
+                "surface_type": surf["type"],
+                "conic": surf.get("conic"),
+                "is_stop": surf.get("is_stop", False),
+                "material": surf.get("material"),
+                "coefficients": coeffs,
+            }
+
+            if surf["type"] == "toroidal":
+                # Zemax CURV (data["radius"]) is the Y-Radius
+                surface_params["radius_y"] = surf["radius"]
+                surface_params["toroidal_coeffs_poly_y"] = coeffs
+
+                # Zemax PARM 1 (data["param_1"]) is the X-Radius
+                radius_x = surf.get("param_1", 0.0)
+                if radius_x == 0.0:
+                    radius_x = be.inf
+                surface_params["radius_x"] = radius_x
+            else:
+                # For all other surfaces, use the standard radius.
+                surface_params["radius"] = surf["radius"]
+
+            # Handle thickness and coordinate system parameters
+            thickness = surf.get("thickness", 0.0)
             if be.isinf(float(thickness)):
-                self.optic.add_surface(
-                    index=surf_idx,
-                    surface_type=surf["type"],
-                    radius=surf.get("radius"),
-                    conic=surf.get("conic"),
-                    thickness=thickness,
-                    is_stop=surf.get("is_stop", False),
-                    material=surf.get("material"),
-                    coefficients=coeffs,
-                    rx=float(rx_),
-                    ry=float(ry_),
-                    rz=float(rz_),
+                # For surfaces at infinity, set thickness and orientation
+                surface_params["thickness"] = thickness
+                surface_params.update(
+                    {"rx": float(rx_), "ry": float(ry_), "rz": float(rz_)}
                 )
-            else:  # normal surface, we pass no thickness argument
-                self.optic.add_surface(
-                    index=surf_idx,
-                    surface_type=surf["type"],
-                    radius=surf.get("radius"),
-                    conic=surf.get("conic"),
-                    is_stop=surf.get("is_stop", False),
-                    material=surf.get("material"),
-                    coefficients=coeffs,
-                    x=float(translation[0]),
-                    y=float(translation[1]),
-                    z=float(translation[2]),
-                    rx=float(rx_),
-                    ry=float(ry_),
-                    rz=float(rz_),
+            else:
+                # For normally positioned surfaces, set position and orientation
+                surface_params.update(
+                    {
+                        "x": float(translation[0]),
+                        "y": float(translation[1]),
+                        "z": float(translation[2]),
+                        "rx": float(rx_),
+                        "ry": float(ry_),
+                        "rz": float(rz_),
+                    }
                 )
+
+            self.optic.add_surface(**surface_params)
             surf_idx = surf_idx + 1
 
[…]
