### 2025-11-09 e8326fd
docs: clarify coefficient mean for toroidal surfaces
