### 2025-11-09 55d3839
style: Format function arguments for better readability in lens and rays plotting methods
