### 2025-11-09 8e15407
feat(visualization): Add projection option to 2D plots
Adds a `projection` argument to the `Optic.draw` and `OpticViewer.view` methods, allowing the user to specify the projection plane for 2D plots.

The following projections are supported:
- 'XY': Top-down view, showing circular apertures.
- 'XZ': Side view with Z-axis horizontal.
- 'YZ': Side view with Z-axis horizontal (default).

The implementation propagates the projection choice through the rendering pipeline. All 2D plotting components (rays, lenses, surfaces) now render correctly for each projection. The axis labels and default ray di
[…]
