### 2025-11-09 5104cde
refactor(visualization): Update distribution parameter to allow None for default selection
