### 2025-11-09 4ea0d61
Refactor code structure for improved readability and maintainability
