+++ docs/gallery/miscellaneous.rst
@@ -16,3 +16,4 @@ or unique functionality of Optiland.
     miscellaneous/paraxial_to_thick_lens
     miscellaneous/paraxial_image_height_field
     miscellaneous/themes
+    miscellaneous/lens_draw_projection
