### 2025-11-09 c3f3754
feat(visualization): Update plot methods in Lens3D and Surface3D to accept additional arguments
