### 2025-11-09 8f704cd
Feature 2d plot projection (#400)
* feat(visualization): Add projection option to 2D plots

Adds a `projection` argument to the `OpticViewer.view` method, allowing the user to specify the projection plane for 2D plots.

The following projections are supported:
- 'XY'
- 'XZ'
- 'YZ' (default)

The implementation propagates the projection choice through the rendering pipeline, and all 2D plotting components (rays, lenses, surfaces) now render on the correct plane. The axis labels are also updated dynamically.

This change maintains backward compatibility by defaulting to the 'YZ' projection.

* s
[…]
