+++ optiland/visualization/system/surface.py
@@ -42,7 +42,7 @@ class Surface2D:
         else:
             self.extent = ray_extent
 
-    def plot(self, ax, theme=None):
+    def plot(self, ax, theme=None, projection="YZ"):
         """Plots the surface on the given matplotlib axis.
 
         Args:
@@ -50,13 +50,16 @@ class Surface2D:
                 surface will be plotted.
             theme (Theme, optional): The theme to use for plotting.
                 Defaults to None.
+            projection (str, optional): The projection plane. Must be 'XY',
+                'XZ', or 'YZ'. Defaults to 'YZ'.
 
         """
-        x, y, z = self._compute_sag()
+        x, y, z = self._compute_sag(projection)
 
         # convert to global coordinates and return
-        _, y, z = transform(x, y, z, self.surf, is_global=False)
+        x, y, z = transform(x, y, z, self.surf, is_global=False)
 
+        x = be.to_numpy(x)
         y = be.to_numpy(y)
         z = be.to_numpy(z)
 
@@ -64,10 +67,15 @@ class Surface2D:
         if theme:
             color = theme.parameters.get("axes.edgecolor", color)
 
-        (line,) = ax.plot(z, y, color=color, label=f"Surface {self.surf.comment}")
+        if projection == "XY":
+            (line,) = ax.plot(x, y, color=color, label=f"Surface {self.surf.comment}")
+        elif projection == "XZ":
+            (line,) = ax.plot(z, x, color=color, label=f"Surface {self.surf.comment}")
+        else:  # YZ
+            (line,) = ax.plot(z, y, color=color, label=f"Surface {self.surf.comment}")
         return {line: self}
 
-    def _compute_sag(self):
+    def _compute_sag(self, projection="YZ"):
         """Computes the sag of the surface in local coordinates and handles
         clipping due to physical apertures.
 
@@ -75,18 +83,37 @@ class Surface2D:
             tuple: A tuple containing arrays of x, y, and z coordinates.
 
         """
-        # local coordinates
-        x = be.zeros(128)
-        y = be.linspace(-self.extent, self.extent, 128)
+        if projection == "XY":
+            # local coordinates for XY circular aperture view
+            theta = be.linspace(0, 2 * be.pi, 128)
+            x = self.extent * be.cos(theta)
+            y = self.extent * be.sin(theta)
+            z = self.surf.geometry.sag(x, y)
+            # No aperture clipping needed here as we are plotting the boundary
+            return x, y, z
+
+        # local coordinates for XZ or YZ cross-section
+        if projection == "XZ":
+            y = be.zeros(128)
+            x = be.linspace(-self.extent, self.extent, 128)
+        else:  # YZ
+            x = be.zeros(128)
+            y = be.linspace(-self.extent, self.extent, 128)
         z = self.surf.geometry.sag(x, y)
 
-        # handle physical apertures
+        # handle physical apertures for line cross-sections
         if self.surf.aperture:
-            y = be.copy(y)  # required to maintain gradient for torch backend
-            intensity = be.ones_like(x)
+            if projection == "XZ":
+                x = be.copy(x)
+            else:  # YZ
+                y = be.copy(y)  # required to maintain gradient for torch backend
+            intensity = be.ones_like(x)  # works for both cases
             rays = RealRays(x, y, x, x, x, x, intensity, x)
             self.surf.aperture.clip(rays)
-            y[rays.i == 0] = be.nan
+            if projection == "XZ":
+                x[rays.i == 0] = be.nan
+            else:  # YZ
+                y[rays.i == 0] = be.nan
 
         return x, y, z
 
@@ -111,7 +138,7 @@ class Surface3D(Surface2D):
     def __init__(self, surface, extent):
[…]
