### 2025-11-21 1f83652
chore: migrate to PEP 735 dependency groups
Consolidate dependency definitions from [project.optional-dependencies]
to [dependency-groups] following PEP 735 standard.
