+++ pyproject.toml
@@ -34,10 +34,6 @@ Homepage = "https://github.com/HarrisonKramer/optiland"
 [project.scripts]
 optiland = "optiland_gui.run_gui:main"
 
-[project.optional-dependencies]
-dev = ["codecov>=2.1.13", "pytest>=8.3.5", "torch"]
-gui = ["PySide6", "qtconsole"]
-
 [tool.hatch.metadata]
 allow-direct-references = true
 
@@ -68,7 +64,5 @@ exclude = ["*.ipynb"]
 required-imports = ["from __future__ import annotations"]
 
 [dependency-groups]
-dev = [
-    "pytest",
-    "torch",
-]
+dev = ["pytest>=8.3.5", "codecov>=2.1.13", "torch"]
+gui = ["PySide6", "qtconsole"]
