### 2025-11-21 decc188
chore: add torch dependency group mentioned in README.md
