+++ pyproject.toml
@@ -66,3 +66,4 @@ required-imports = ["from __future__ import annotations"]
 [dependency-groups]
 dev = ["pytest>=8.3.5", "codecov>=2.1.13", "torch"]
 gui = ["PySide6", "qtconsole"]
+torch = ["torch"]
