### 2025-11-21 f378fb7
chore: modernize testing workflow and fix pytest-cov dependency
CONTRIBUTING.md previously instructed contributors to run `pytest --cov=optiland --cov-report=xml` before submitting PRs, but pytest-cov was not included in the dev dependency group, causing errors.

Since coverage is automatically evaluated by GitHub Actions, the manual coverage requirement is redundant. This commit:
- Adds pytest-cov to dev dependencies for reproducibility
- Updates CI workflow to use `uv sync --dev` instead of manually installing dependencies
- Removes redundant coverage instructions from CONTRIBUTING.md, clari
[…]
