+++ pyproject.toml
@@ -64,6 +64,6 @@ exclude = ["*.ipynb"]
 required-imports = ["from __future__ import annotations"]
 
 [dependency-groups]
-dev = ["pytest>=8.3.5", "codecov>=2.1.13", "torch"]
+dev = ["pytest>=8.3.5", "codecov>=2.1.13", "torch", "pytest-cov>=7.0.0"]
 gui = ["PySide6", "qtconsole"]
 torch = ["torch"]
