### 2025-11-22 666f7ca
bump to 0.5.8
