+++ docs/conf.py
@@ -9,7 +9,7 @@ project = "Optiland"
 current_year = datetime.now().year
 copyright = f"2024-{current_year}, Kramer Harrison & contributors"
 author = "Kramer Harrison"
-release = "0.5.7"
+release = "0.5.8"
 
 extensions = [
     "jupyterlite_sphinx",
