### 2025-11-26 a93c880
Refactor #246 ThinFilmOptimizer and ThinFilmReport for improved flexibility and maintainability
- Removed backward compatibility parameters from ThinFilmOptimizer's optimize method.
- Added support for additional keyword arguments in the optimizer.
- Cleaned up performance_table and convergence_plot methods in ThinFilmReport by removing commented-out code.
- Removed unused imports and unnecessary complexity in report generation methods.
- Enhanced documentation for clarity on method parameters and return types.
