+++ optiland/thin_film/optimization/report.py
@@ -12,10 +12,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any
 
-import matplotlib.pyplot as plt
 import pandas as pd
-
-import optiland.backend as be
 
 if TYPE_CHECKING:
     from .optimizer import ThinFilmOptimizer
@@ -81,317 +78,3 @@ class ThinFilmReport:
             )
 
         return pd.DataFrame(data)
-
-    def performance_table(self) -> pd.DataFrame:
-        """Generate a performance table showing target achievement.
-
-        Returns:
-            DataFrame with columns: Target, Initial, Final, Target_Value,
-            Target_Type, Weight, Achieved
-        """
-        data = []
-
-        # Reset to initial state to compute initial values
-        for _i, thickness_um in enumerate(self._initial_thicknesses):
-            if _i < len(self.stack.layers):
-                self.stack.layers[_i].update_thickness(thickness_um)
-
-        for _i, target in enumerate(self.optimizer.targets):
-            # Get target info
-            prop = target.property
-            wl_nm = target.wavelength_nm
-            aoi_deg = target.aoi_deg
-            polarization = target.polarization
-            target_value = target.value
-            target_type = target.target_type
-            weight = target.weight
-
-            # Compute initial value
-            if prop == "R":
-                initial_val = float(
-                    self.stack.reflectance_nm_deg(wl_nm, aoi_deg, polarization)
-                )
-            elif prop == "T":
-                initial_val = float(
-                    self.stack.transmittance_nm_deg(wl_nm, aoi_deg, polarization)
-                )
-            elif prop == "A":
-                initial_val = float(
-                    self.stack.absorptance_nm_deg(wl_nm, aoi_deg, polarization)
-                )
-
-            # Restore final state
-            for j, thickness_um in enumerate(self._final_thicknesses):
-                if j < len(self.stack.layers):
-                    self.stack.layers[j].update_thickness(thickness_um)
-
-            # Compute final value
-            if prop == "R":
-                final_val = float(
-                    self.stack.reflectance_nm_deg(wl_nm, aoi_deg, polarization)
-                )
-            elif prop == "T":
-                final_val = float(
-                    self.stack.transmittance_nm_deg(wl_nm, aoi_deg, polarization)
-                )
-            elif prop == "A":
-                final_val = float(
-                    self.stack.absorptance_nm_deg(wl_nm, aoi_deg, polarization)
-                )
-
-            # Check if target achieved
-            if target_type == "equal":
-                achieved = abs(final_val - target_value) < target.tolerance
-            elif target_type == "below":
-                achieved = final_val <= target_value
-            elif target_type == "over":
-                achieved = final_val >= target_value
-            else:
-                achieved = False
-
-            # Format wavelength string
-            if isinstance(wl_nm, list):
-                wl_str = f"{min(wl_nm):.0f}-{max(wl_nm):.0f} nm"
-            else:
-                wl_str = f"{wl_nm:.0f} nm"
-
-            data.append(
-                {
-                    "Target": f"{prop} at {wl_str} ({polarization}, {aoi_deg:.1f}°)",
-                    "Initial": f"{initial_val:.4f}",
-                    "Final": f"{final_val:.4f}",
-                    "Target_Value": f"{target_value:.4f}",
-                    "Target_Type": target_type,
-                    "Weight": f"{weight:.2f}",
-                    "Achieved": "✓" if achieved else "✗",
-                }
-            )
-
-        return pd.DataFrame(data)
-
-    def convergence_plot(self, ax: plt.Axes = None) -> tuple[plt.Figure, plt.Axes]:
[…]
