### 2025-11-26 88e73a9
#246 Add colorimetry module with constants, core calculations, and plotting utilities
- Implemented standard colorimetric constants including CIE observers and illuminants in `constants.py`.
- Developed core functions for spectral interpolation, XYZ conversion, and sRGB conversion in `core.py`.
- Created plotting utilities for the CIE 1931 chromaticity diagram in `plotting.py`.
- Ensured compatibility with optiland backend for mathematical operations.
- Added extensive documentation and type hints for better code clarity and usability.
