+++ optiland/colorimetry/plotting.py
@@ -1,0 +1,236 @@
+"""
+Colorimetry plotting utilities.
+"""
+
+from __future__ import annotations
+
+from typing import Literal
+
+import matplotlib.pyplot as plt
+from matplotlib.collections import LineCollection
+from matplotlib.path import Path
+from scipy.interpolate import interp1d
+
+import optiland.backend as be
+
+from .constants import CIE_1931_2DEG, WAVELENGTHS_STD
+
+
+def plot_cie_1931_chromaticity_diagram(
+    ax: plt.Axes | None = None,
+    title: str = "CIE 1931 Chromaticity Diagram",
+    color: Literal["no", "contour", "fill"] = "contour",
+) -> tuple[plt.Figure, plt.Axes]:
+    """
+    Plots the CIE 1931 chromaticity diagram (spectral locus).
+
+    Generates the spectral locus by converting the standard observer CMFs
+    to xy coordinates for each wavelength.
+
+    Args:
+        ax: Optional matplotlib Axes. If None, a new figure is created.
+        title: Title of the plot.
+        color: Coloring mode.
+               - "no": Black and white outline.
+               - "contour": Spectral locus colored by wavelength (default).
+               - "fill": Gamut filled with approximate sRGB colors.
+
+    Returns:
+        Tuple (Figure, Axes).
+    """
+    if ax is None:
+        fig, ax = plt.subplots(figsize=(6, 6))
+    else:
+        fig = ax.figure
+
+    # 1. Upsample Data for smoothness
+    # Extract raw data
+    wls_raw = be.asarray(WAVELENGTHS_STD)
+    cmfs_raw = be.asarray(CIE_1931_2DEG)
+    x_bar_raw, y_bar_raw, z_bar_raw = cmfs_raw[:, 0], cmfs_raw[:, 1], cmfs_raw[:, 2]
+
+    # Create fine wavelength grid (0.5nm steps) for smooth locus
+    wl_fine = be.linspace(390, 700, 621)
+
+    # Interpolate CMFs (Cubic)
+    # We use scipy for cubic interpolation as requested
+    def cubic_interp(x, y, x_new):
+        f = interp1d(x, y, kind="cubic", fill_value="extrapolate")
+        return be.asarray(f(x_new))
+
+    x_bar = cubic_interp(wls_raw, x_bar_raw, wl_fine)
+    y_bar = cubic_interp(wls_raw, y_bar_raw, wl_fine)
+    z_bar = cubic_interp(wls_raw, z_bar_raw, wl_fine)
+
+    # Calculate xy coordinates for the spectral locus
+    sum_xyz = x_bar + y_bar + z_bar
+    # Avoid division by zero
+    mask_valid = sum_xyz > 0
+
+    x_locus = x_bar[mask_valid] / sum_xyz[mask_valid]
+    y_locus = y_bar[mask_valid] / sum_xyz[mask_valid]
+
+    # Calculate RGB colors for the locus (normalized 0-1)
+    # We use a local float conversion to avoid integer rounding artifacts
+    def _get_rgb_float(X, Y, Z):
+        # Normalize
+        Xn, Yn, Zn = X / 100.0, Y / 100.0, Z / 100.0
+        # sRGB Matrix
+        r_l = 3.2404542 * Xn - 1.5371385 * Yn - 0.4985314 * Zn
+        g_l = -0.9692660 * Xn + 1.8760108 * Yn + 0.0415560 * Zn
+        b_l = 0.0556434 * Xn - 0.2040259 * Yn + 1.0572252 * Zn
+
+        # Gamma
+        def gamma(v):
+            return be.where(
+                v <= 0.0031308,
+                12.92 * v,
[…]
