### 2025-12-20 64916a2
feat: add real image height field
