+++ optiland/fields/field_types.py
@@ -129,6 +129,8 @@ class BaseFieldDefinition(ABC):
             return ObjectHeightField()
         elif field_type == "ParaxialImageHeightField":
             return ParaxialImageHeightField()
+        elif field_type == "RealImageHeightField":
+            return RealImageHeightField()
         else:
             raise ValueError(f"Unknown field definition: {field_type}")
 
@@ -477,3 +479,232 @@ class ParaxialImageHeightField(BaseFieldDefinition):
                 y=0, u=1, z=z_start, wavelength=wavelength, reverse=True, skip=skip
             )
             return y[-1], u[-1]
+
+
+class RealImageHeightField(BaseFieldDefinition):
+    """Defines fields by the chief ray's real height at the image plane."""
+
+    def get_ray_origins(self, optic, Hx, Hy, Px, Py, vx, vy):
+        """Calculate the initial positions for rays originating at the object.
+
+        Args:
+            Hx (float): Normalized x field coordinate.
+            Hy (float): Normalized y field coordinate.
+            Px (float or be.ndarray): x-coordinate of the pupil point.
+            Py (float or be.ndarray): y-coordinate of the pupil point.
+            vx (float): Vignetting factor in the x-direction.
+            vy (float): Vignetting factor in the y-direction.
+
+        Returns:
+            tuple: A tuple containing the x, y, and z coordinates of the
+                object position.
+        """
+        # Initial guess using ParaxialImageHeightField logic
+        paraxial_field = ParaxialImageHeightField()
+        paraxial_field.scale_chief_ray_for_field(
+            optic,
+            *paraxial_field._trace_unit_chief_ray(optic, plane="object"),
+            paraxial_field._trace_unit_chief_ray(optic, plane="image")[0],
+        )
+
+        max_field = optic.fields.max_field
+        target_x = max_field * Hx
+        target_y = max_field * Hy
+
+        # Initial guess for the field parameters (angle or height)
+        # We solve for the field parameters that result in the target image height
+        if optic.object_surface.is_infinite:
+            y_img_unit, _ = paraxial_field._trace_unit_chief_ray(optic, plane="image")
+            _, u_obj_unit = paraxial_field._trace_unit_chief_ray(optic, plane="object")
+
+            val_x = u_obj_unit * (target_x / y_img_unit)
+            val_y = u_obj_unit * (target_y / y_img_unit)
+        else:
+            # For finite object, variables are object heights
+            y_img_unit, _ = paraxial_field._trace_unit_chief_ray(optic, plane="image")
+            y_obj_unit, _ = paraxial_field._trace_unit_chief_ray(optic, plane="object")
+
+            val_x = y_obj_unit * (target_x / y_img_unit)
+            val_y = y_obj_unit * (target_y / y_img_unit)
+
+        # Iterative solver (Newton's method)
+        num_iterations = 10
+        tol = 1e-12
+
+        # We need the paraxial magnification/scaling for the Jacobian
+        # d(ImageHeight)/d(FieldParam)
+        if optic.object_surface.is_infinite:
+            jacobian = y_img_unit / u_obj_unit
+        else:
+            jacobian = y_img_unit / y_obj_unit
+
+        prev_val_x = None
+        prev_curr_x = None
+        prev_val_y = None
+        prev_curr_y = None
+
+        for _ in range(num_iterations):
+            # 1. Generate chief rays for current object guess
+            rays = self._generate_chief_rays(optic, val_x, val_y)
+
+            # 2. Trace rays
+            optic.surface_group.trace(rays)
+
+            # Propagate to image surface
[…]
