### 2025-12-20 62dfdd2
docs: update the real image height docs
