### 2025-12-20 48f4f90
feat: introduce abstract base class for optical field definitions with concrete implementations for object height and paraxial image height
