+++ optiland/fields/field_types/real_image_height.py
@@ -1,0 +1,240 @@
+"""Real Image Height Field Module
+
+Kramer Harrison, 2025
+"""
+
+from __future__ import annotations
+
+import optiland.backend as be
+
+from .base import BaseFieldDefinition
+from .paraxial_image_height import ParaxialImageHeightField
+
+
+class RealImageHeightField(BaseFieldDefinition):
+    """Defines fields by the chief ray's real height at the image plane."""
+
+    def get_ray_origins(self, optic, Hx, Hy, Px, Py, vx, vy):
+        """Calculate the initial positions for rays originating at the object.
+
+        Args:
+            Hx (float): Normalized x field coordinate.
+            Hy (float): Normalized y field coordinate.
+            Px (float or be.ndarray): x-coordinate of the pupil point.
+            Py (float or be.ndarray): y-coordinate of the pupil point.
+            vx (float): Vignetting factor in the x-direction.
+            vy (float): Vignetting factor in the y-direction.
+
+        Returns:
+            tuple: A tuple containing the x, y, and z coordinates of the
+                object position.
+        """
+        # Initial guess using ParaxialImageHeightField logic
+        paraxial_field = ParaxialImageHeightField()
+        paraxial_field.scale_chief_ray_for_field(
+            optic,
+            *paraxial_field._trace_unit_chief_ray(optic, plane="object"),
+            paraxial_field._trace_unit_chief_ray(optic, plane="image")[0],
+        )
+
+        max_field = optic.fields.max_field
+        target_x = max_field * Hx
+        target_y = max_field * Hy
+
+        # Initial guess for the field parameters (angle or height)
+        # We solve for the field parameters that result in the target image height
+        if optic.object_surface.is_infinite:
+            y_img_unit, _ = paraxial_field._trace_unit_chief_ray(optic, plane="image")
+            _, u_obj_unit = paraxial_field._trace_unit_chief_ray(optic, plane="object")
+
+            val_x = u_obj_unit * (target_x / y_img_unit)
+            val_y = u_obj_unit * (target_y / y_img_unit)
+        else:
+            # For finite object, variables are object heights
+            y_img_unit, _ = paraxial_field._trace_unit_chief_ray(optic, plane="image")
+            y_obj_unit, _ = paraxial_field._trace_unit_chief_ray(optic, plane="object")
+
+            val_x = y_obj_unit * (target_x / y_img_unit)
+            val_y = y_obj_unit * (target_y / y_img_unit)
+
+        # Iterative solver (Newton's method)
+        num_iterations = 10
+        tol = 1e-12
+
+        # We need the paraxial magnification/scaling for the Jacobian
+        # d(ImageHeight)/d(FieldParam)
+        if optic.object_surface.is_infinite:
+            jacobian = y_img_unit / u_obj_unit
+        else:
+            jacobian = y_img_unit / y_obj_unit
+
+        prev_val_x = None
+        prev_curr_x = None
+        prev_val_y = None
+        prev_curr_y = None
+
+        for _ in range(num_iterations):
+            # 1. Generate chief rays for current object guess
+            rays = self._generate_chief_rays(optic, val_x, val_y)
+
+            # 2. Trace rays
+            optic.surface_group.trace(rays)
+
+            # Propagate to image surface
+            last_surface = optic.surface_group.surfaces[-1]
+            last_surface.material_post.propagation_model.propagate(
+                rays, last_surface.thickness
+            )
+
+            # 3. Get intersection at image plane
+            curr_x = rays.x
+            curr_y = rays.y
+
+            err_x = curr_x - target_x
[…]
