+++ optiland/fields/field_types.py
@@ -1,710 +1,0 @@
-"""Field Types Module
-
-This module defines different field types for optical systems, including angle,
-object height, and paraxial image height fields. Each field type implements
-methods to calculate ray origins and paraxial object positions based on the
-field definition.
-
-Kramer Harrison, 2025
-"""
-
-from __future__ import annotations
-
-from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING
-
-import optiland.backend as be
-
-if TYPE_CHECKING:
-    from optiland import Optic
-    from optiland._types import BEArray, ScalarOrArray
-
-
-class BaseFieldDefinition(ABC):
-    """Abstract base class for defining how fields map to ray properties."""
-
-    @abstractmethod
-    def get_ray_origins(
-        self,
-        optic: Optic,
-        Hx: ScalarOrArray,
-        Hy: ScalarOrArray,
-        Px: ScalarOrArray,
-        Py: ScalarOrArray,
-        vx: ScalarOrArray,
-        vy: ScalarOrArray,
-    ) -> tuple[ScalarOrArray, ScalarOrArray, ScalarOrArray]:
-        """Calculate the initial positions for rays originating at the object.
-
-        Args:
-            Hx: Normalized x field coordinate.
-            Hy: Normalized y field coordinate.
-            Px: x-coordinate of the pupil point.
-            Py: y-coordinate of the pupil point.
-            vx: Vignetting factor in the x-direction.
-            vy: Vignetting factor in the y-direction.
-
-        Returns:
-            A tuple containing the x, y, and z coordinates of the
-                object position.
-
-        """
-        pass  # pragma: no cover
-
-    @abstractmethod
-    def get_paraxial_object_position(
-        self, optic: Optic, Hy: ScalarOrArray, y1: ScalarOrArray, EPL: ScalarOrArray
-    ) -> tuple[BEArray, BEArray]:
-        """Calculate the position of the object in the paraxial optical system.
-
-        Args:
-            Hy: The normalized field height.
-            y1: The initial y-coordinate of the ray.
-            EPL: The entrance pupil location.
-
-        Returns:
-            A tuple containing the y and z coordinates of the object
-                position.
-
-        """
-        pass  # pragma: no cover
-
-    @abstractmethod
-    def scale_chief_ray_for_field(
-        self,
-        optic: Optic,
-        y_obj_unit: ScalarOrArray,
-        u_obj_unit: ScalarOrArray,
-        y_img_unit: ScalarOrArray,
-    ) -> ScalarOrArray:
-        """Calculates the scaling factor for a unit chief ray based on the field
-        definition.
-
-        This is used in the paraxial chief_ray calculation. It uses the results
-        of a forward and backward "unit" trace from the stop to determine the
-        final scaling factor.
-
-        Args:
-            optic: The optical system.
-            y_obj_unit: The object-space height of the unit ray.
-            u_obj_unit: The object-space angle of the unit ray.
-            y_img_unit: The image-space height of the unit ray.
-
-        Returns:
-            The scaling factor.
-
-        """
-        pass  # pragma: no cover
-
-    def to_dict(self) -> dict:
-        """Convert the field definition to a dictionary.
-
-        Returns:
-            dict: A dictionary representation of the field definition.
-
-        """
-        return {"field_type": self.__class__.__name__}
-
-    @classmethod
-    def from_dict(cls, field_def_dict: dict) -> BaseFieldDefinition:
-        """Create a field definition from a dictionary.
-
-        Args:
-            field_def_dict (dict): A dictionary representation of the field
-                definition.
-
-        Returns:
-            BaseFieldDefinition: A field definition object created from the
-                dictionary.
-
-        """
[…]
