### 2025-12-20 551c027
Feat/real image height field (#417)
* feat: add real image height field

* docs: update the real image height docs

* feat: introduce abstract base class for optical field definitions with concrete implementations for object height and paraxial image height
