+++ optiland/optic/optic.py
@@ -26,6 +26,7 @@ from optiland.fields import (
     FieldGroup,
     ObjectHeightField,
     ParaxialImageHeightField,
+    RealImageHeightField,
 )
 from optiland.optic.optic_updater import OpticUpdater
 from optiland.paraxial import Paraxial
@@ -311,6 +312,8 @@ class Optic:
             self.field_definition = ObjectHeightField()
         elif field_type == "paraxial_image_height":
             self.field_definition = ParaxialImageHeightField()
+        elif field_type == "real_image_height":
+            self.field_definition = RealImageHeightField()
         else:
             raise ValueError(f"Invalid field type: {field_type}.")
 

[…]
