### 2025-12-21 fa7fad8
feat: add multi-config functionality
