### 2025-12-21 9f35ba1
feat: add draft image simulation analysis
