+++ optiland/backend/numpy_backend.py
@@ -12,6 +12,8 @@ import numpy as np
 from matplotlib.path import Path
 from numpy.typing import ArrayLike, NDArray
 from scipy.interpolate import NearestNDInterpolator
+from scipy.ndimage import map_coordinates
+from scipy.signal import fftconvolve as _fftconvolve
 from scipy.spatial.transform import Rotation as R
 from scipy.special import gamma
 
@@ -22,12 +24,16 @@ if TYPE_CHECKING:
 _lib = np
 
 
+def get_precision():
+    return 64
+
+
 ScalarOrArray = TypeVar("ScalarOrArray", float, NDArray)
 
 
 def array(x: ArrayLike) -> NDArray:
     """Create an array/tensor"""
-    return np.array(x, dtype=float)
+    return np.array(x)
 
 
 def is_array_like(x: NDArray | list | tuple) -> bool:
@@ -140,3 +146,87 @@ def path_contains_points(vertices: NDArray, points: NDArray) -> NDArray:
     path = Path(vertices)
     mask = path.contains_points(points)
     return np.asarray(mask, dtype=bool)
+
+
+def lstsq(a: ArrayLike, b: ArrayLike) -> NDArray:
+    """
+    Computes the least-squares solution to a linear matrix equation.
+
+    Args:
+        a (ArrayLike): Left-hand side matrix (M, N).
+        b (ArrayLike): Right-hand side matrix (M,) or (M, K).
+
+    Returns:
+        NDArray: Least-squares solution (N,) or (N, K).
+    """
+    return np.linalg.lstsq(a, b, rcond=None)[0]
+
+
+def fftconvolve(
+    in1: ArrayLike, in2: ArrayLike, mode: Literal["full", "valid", "same"] = "full"
+) -> NDArray:
+    """Numpy/Scipy implementation of FFT-based convolution."""
+    a = array(in1)
+    b = array(in2)
+    return _fftconvolve(a, b, mode=mode)
+
+
+def grid_sample(
+    input, grid, mode="bilinear", padding_mode="zeros", align_corners=False
+):
+    """
+    Numpy implementation of torch.nn.functional.grid_sample.
+
+    Args:
+        input (np.ndarray): Input tensor of shape (N, C, H_in, W_in).
+        grid (np.ndarray): Grid of shape (N, H_out, W_out, 2).
+                           Last dim is (x, y). Coordinates in range [-1, 1].
+        mode (str): Interpolation mode ('bilinear', 'nearest').
+        padding_mode (str): Padding mode ('zeros', 'border', 'reflection').
+        align_corners (bool): Whether to align corners.
+
+    Returns:
+        np.ndarray: Output tensor of shape (N, C, H_out, W_out).
+    """
+    N, C, H_in, W_in = input.shape
+    _N, H_out, W_out, _ = grid.shape
+    if N != _N:
+        raise ValueError("Input and grid must have same batch size")
+
+    # Map grid coordinates [-1, 1] to pixel coordinates
+    # grid[..., 0] is x (columns), grid[..., 1] is y (rows)
+    x = grid[..., 0]
+    y = grid[..., 1]
+
+    if align_corners:
+        # -1 -> 0, 1 -> W-1
+        x_pix = ((x + 1) / 2) * (W_in - 1)
+        y_pix = ((y + 1) / 2) * (H_in - 1)
+    else:
+        # Standard pytorch align_corners=False:
+        # x = (x_norm + 1) * W / 2 - 0.5
+        x_pix = ((x + 1) * W_in / 2) - 0.5
+        y_pix = ((y + 1) * H_in / 2) - 0.5
+
+    output = _lib.zeros((N, C, H_out, W_out), dtype=input.dtype)
+
+    # Scipy order: 0=nearest, 1=bilinear
[…]

[…]
