### 2025-12-22 36d318d
feat: Implement full image simulation with spatially variable blur, distortion, and lateral color, along with supporting documentation and examples.
