### 2025-12-22 d4f55c2
feat: introduce modular ray aiming system with new RayGenerator and ParaxialRayAimer.
