+++ optiland/rays/ray_aiming/__init__.py
@@ -1,0 +1,21 @@
+"""
+Ray Aiming Package
+
+This package contains infrastructure and algorithms for aiming rays at the
+stop surface in optical systems.
+
+Kramer Harrison, 2025
+"""
+
+from __future__ import annotations
+
+from optiland.rays.ray_aiming.base import BaseRayAimer
+from optiland.rays.ray_aiming.paraxial import ParaxialRayAimer
+from optiland.rays.ray_aiming.registry import create_ray_aimer, register_aimer
+
+__all__ = [
+    "BaseRayAimer",
+    "ParaxialRayAimer",
+    "create_ray_aimer",
+    "register_aimer",
+]
