### 2025-12-22 f3166d1
feat: Add iterative ray aiming algorithm and associated tests.
