### 2025-12-22 a096122
feat: add iterative and robust ray aiming strategies and integrate them into the ray generator.
