+++ optiland/optic/optic.py
@@ -83,6 +83,7 @@ class Optic:
 
     Attributes:
         name (str | None): An optional name for the optical system.
+        ray_aiming_config (dict): Configuration for ray aiming.
         aperture (Aperture | None): The aperture of the optical system.
         field_definition (BaseFieldDefinition | None): The definition of the field used
             in the optical system, e.g., AngleField or ObjectHeightField.
@@ -118,6 +119,11 @@ class Optic:
                 Defaults to None.
         """
         self.name = name
+        self.ray_aiming_config = {
+            "mode": "paraxial",
+            "max_iter": 10,
+            "tol": 1e-6,
+        }
         self.reset()
 
     def _initialize_attributes(self):
@@ -437,6 +443,24 @@ class Optic:
     def update_normalization(self, surface: Surface) -> None:
         """Update the normalization radius of surfaces."""
         self._updater.update_normalization(surface)
+
+    def set_ray_aiming(
+        self, mode: str, max_iter: int = 10, tol: float = 1e-6, **kwargs
+    ):
+        """Configure the ray aiming strategy.
+
+        Args:
+            mode: The aiming mode ("paraxial", "iterative", "robust").
+            max_iter: Maximum iterations for iterative solvers.
+            tol: Convergence tolerance for iterative solvers.
+            **kwargs: Additional configuration parameters.
+        """
+        self.ray_aiming_config = {
+            "mode": mode,
+            "max_iter": max_iter,
+            "tol": tol,
+            **kwargs,
+        }
 
     def update(self) -> None:
         """Update the surface properties (pickups, solves, paraxial properties)."""
