+++ optiland/rays/ray_generator.py
@@ -8,7 +8,7 @@ Kramer Harrison, 2024
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 import optiland.backend as be
 from optiland.rays.polarized_rays import PolarizedRays
@@ -26,7 +26,9 @@ class RayGenerator:
         self.optic = optic
         self.aimer = create_ray_aimer("paraxial", optic)
 
-    def set_ray_aiming(self, mode: str, **kwargs: Any) -> None:
+    def set_ray_aiming(
+        self, mode: str, max_iter: int = 10, tol: float = 1e-6, **kwargs
+    ):
         """
         Set the ray aiming strategy.
 
@@ -57,6 +59,20 @@ class RayGenerator:
             RealRays object containing the generated rays.
 
         """
+        if hasattr(self.optic, "ray_aiming_config"):
+            config = self.optic.ray_aiming_config
+            # Basic check: if mode changed, re-create.
+            # Ideally we check all params, but for now mode is primary.
+            # We can also check if aimer type matches config mode.
+            getattr(self.aimer, "name", "unknown")
+            # Note: Aimer instances don't store "name" by default unless we added it.
+            # Safer to recreate or store _current_config locally.
+
+            # Let's check against a stored config shadow
+            if not hasattr(self, "_current_config") or self._current_config != config:
+                self.set_ray_aiming(**config)
+                self._current_config = config.copy()
+
         # Aim rays using the configured strategy
         x0, y0, z0, L, M, N = self.aimer.aim_rays(
             (Hx, Hy),
@@ -78,5 +94,8 @@ class RayGenerator:
                     "Polarization must be set when surfaces have "
                     "polarization-dependent coatings.",
                 )
-            return RealRays(x0, y0, z0, L, M, N, intensity, wavelength)
+            rays = RealRays(
+                x0, y0, z0, L, M, N, intensity=be.ones_like(x0), wavelength=wavelength
+            )
+            return rays
         return PolarizedRays(x0, y0, z0, L, M, N, intensity, wavelength)
