+++ optiland/rays/ray_aiming/__init__.py
@@ -13,11 +13,13 @@ from optiland.rays.ray_aiming.base import BaseRayAimer
 from optiland.rays.ray_aiming.iterative import IterativeRayAimer
 from optiland.rays.ray_aiming.paraxial import ParaxialRayAimer
 from optiland.rays.ray_aiming.registry import create_ray_aimer, register_aimer
+from optiland.rays.ray_aiming.robust import RobustRayAimer
 
 __all__ = [
     "BaseRayAimer",
-    "IterativeRayAimer",
     "ParaxialRayAimer",
+    "IterativeRayAimer",
+    "RobustRayAimer",
     "create_ray_aimer",
     "register_aimer",
 ]
