### 2025-12-22 7447789
feat: Introduce core Optic class and robust ray aiming strategies with accompanying documentation and tests.
