### 2025-12-22 4fe41d2
feat: add ray aim setting in Optic
