### 2025-12-22 d4c094a
feat: implement iterative and robust methods for ray aiming
