+++ optiland/rays/ray_aiming/iterative.py
@@ -1,13 +1,14 @@
 """Iterative Ray Aiming Module
 
-This module implements the iterative ray aiming algorithm, which uses a
-Newton-Raphson-like method to aim rays at the stop surface for aberrated
-systems.
+This module implements the iterative ray aiming algorithm with robust
+derivative calculation for wide-angle systems.
 
 Kramer Harrison, 2025
 """
 
 from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any
 
 import optiland.backend as be
 from optiland.rays import RealRays
@@ -15,122 +16,156 @@ from optiland.rays.ray_aiming.base import BaseRayAimer
 from optiland.rays.ray_aiming.paraxial import ParaxialRayAimer
 from optiland.rays.ray_aiming.registry import register_aimer
 
+if TYPE_CHECKING:
+    from optiland.optic import Optic
+
 
 @register_aimer("iterative")
 class IterativeRayAimer(BaseRayAimer):
-    """Iterative ray aiming strategy.
+    """Iterative ray aiming strategy using Newton-Raphson."""
 
-    This aimer uses the paraxial result as a starting guess and then iteratively
-    adjusts the launch coordinates/direction to ensure the rays intersect the
-    stop surface at the desired normalized pupil coordinates.
-
-    Args:
-        optic: The optical system to aim rays for.
-        max_iter: Maximum number of iterations. Defaults to 10.
-        tol: Convergence tolerance for the pupil error. Defaults to 1e-6.
-        **kwargs: Additional keyword arguments.
-    """
-
-    def __init__(self, optic, max_iter: int = 20, tol: float = 1e-8, **kwargs):
+    def __init__(
+        self,
+        optic: Optic,
+        max_iter: int = 20,
+        tol: float = 1e-8,
+        **kwargs: Any,
+    ) -> None:
         super().__init__(optic, **kwargs)
         self.max_iter = max_iter
         self.tol = tol
-        # Internal paraxial aimer for initial guess
         self._paraxial_aimer = ParaxialRayAimer(optic)
 
     def aim_rays(
         self,
         fields: tuple,
-        wavelengths: float,
+        wavelengths: Any,
         pupil_coords: tuple,
         initial_guess: tuple | None = None,
     ) -> tuple:
         """Calculate ray starting coordinates using iterative aiming."""
-        if initial_guess is not None:
-            # Use provided guess
+        if initial_guess:
             x, y, z, L, M, N = initial_guess
         else:
-            # Get initial guess from paraxial aimer
             x, y, z, L, M, N = self._paraxial_aimer.aim_rays(
                 fields, wavelengths, pupil_coords
             )
 
         Px, Py = pupil_coords
-        is_infinite = (
-            self.optic.object_surface.is_infinite
-            if self.optic.object_surface
-            else False
-        )
-        stop_index = self.optic.surface_group.stop_index
-        stop_surface = self.optic.surface_group.surfaces[stop_index]
+        stop_idx = self.optic.surface_group.stop_index
+        stop_surf = self.optic.surface_group.surfaces[stop_idx]
+        is_inf = getattr(self.optic.object_surface, "is_infinite", False)
 
-        # Determine the physical target radius on the stop surface.
-        # Ideally, we use the explicit aperture dimensions if available.
-
-        if stop_surface.aperture:
-            aperture = stop_surface.aperture
-            if hasattr(aperture, "r_max"):
-                rx = ry = aperture.r_max
-            elif hasattr(aperture, "x_max") and hasattr(aperture, "y_max"):
[…]
