+++ optiland/rays/ray_aiming/robust.py
@@ -1,75 +1,104 @@
 """Robust Ray Aiming Module
 
-This module implements the robust ray aiming algorithm, which uses pupil
-expansion (continuation method) to solve for ray aiming coordinates in systems
-where standard iterative aiming might fail.
+This module implements a recursive robust ray aiming algorithm.
 
 Kramer Harrison, 2025
 """
 
 from __future__ import annotations
 
+from typing import TYPE_CHECKING, Any
+
+import optiland.backend as be
 from optiland.rays.ray_aiming.base import BaseRayAimer
 from optiland.rays.ray_aiming.iterative import IterativeRayAimer
 from optiland.rays.ray_aiming.registry import register_aimer
 
+if TYPE_CHECKING:
+    from optiland.optic import Optic
+
 
 @register_aimer("robust")
 class RobustRayAimer(BaseRayAimer):
-    """Robust ray aiming strategy using pupil expansion.
-
-    This aimer slowly expands the pupil size from a small fraction to the full
-    size, using the solution from the previous step as the starting guess for
-    the next. This improves convergence for highly aberrated systems.
-
-    Args:
-        optic: The optical system to aim rays for.
-        fractions: List of pupil fractions to solve for. Defaults to
-            [0.1, 0.5, 1.0].
-        max_iter: Maximum iterations per step. Defaults to 20.
-        tol: Convergence tolerance. Defaults to 1e-8.
-        **kwargs: Additional keyword arguments.
-    """
+    """Robust ray aiming strategy using recursive subdivision."""
 
     def __init__(
         self,
-        optic,
-        fractions: list[float] | None = None,
+        optic: Optic,
         max_iter: int = 20,
         tol: float = 1e-8,
-        **kwargs,
-    ):
+        scale_fields: bool = True,
+        **kwargs: Any,
+    ) -> None:
         super().__init__(optic, **kwargs)
-        self.fractions = fractions if fractions else [0.1, 0.5, 1.0]
-        self.max_iter = max_iter
-        self.tol = tol
-        # Internal iterative aimer
-        self._iterative_aimer = IterativeRayAimer(optic, max_iter=max_iter, tol=tol)
+        self.scale_fields = scale_fields
+        self._iterative = IterativeRayAimer(optic, max_iter=max_iter, tol=tol)
+        self._paraxial = self._iterative._paraxial_aimer
 
     def aim_rays(
         self,
         fields: tuple,
-        wavelengths: float,
+        wavelengths: Any,
         pupil_coords: tuple,
     ) -> tuple:
-        """Calculate ray starting coordinates using robust pupil expansion."""
-        Px, Py = pupil_coords
+        """Calculate ray starting coordinates using recursive robust expansion."""
+        # Anchor at t=0 (Paraxial limit)
+        p0 = (pupil_coords[0] * 0.0, pupil_coords[1] * 0.0)
+        f0 = (fields[0] * 0.0, fields[1] * 0.0) if self.scale_fields else fields
 
-        # Ensure inputs are arrays for consistent scaling
-        # (Though scalars might work if multiplied by scalar fraction)
+        sol_0 = self._paraxial.aim_rays(f0, wavelengths, p0)
 
-        current_guess = None
+        # Paraxial and Real are identical at t=0
+        return self._solve(0.0, 1.0, sol_0, sol_0, fields, wavelengths, pupil_coords)
 
-        for fraction in self.fractions:
-            # Scale target pupil coordinates
-            current_pupil = (Px * fraction, Py * fraction)
[…]
