### 2025-12-22 4eebca5
chore: fix error handling
