### 2025-12-22 804ad88
feat: Implement iterative and robust ray aiming strategies with a dedicated generator and comprehensive tests.
