+++ optiland/rays/ray_aiming/iterative.py
@@ -22,7 +22,20 @@ if TYPE_CHECKING:
 
 @register_aimer("iterative")
 class IterativeRayAimer(BaseRayAimer):
-    """Iterative ray aiming strategy using Newton-Raphson."""
+    """Iterative ray aiming strategy using Modified Newton-Raphson.
+
+    This class implements an iterative ray aiming algorithm that solves for the
+    initial ray coordinates (x, y) or directions (L, M) required to hit a specific
+    target on the stop surface. It uses a Modified Newton-Raphson method with
+    a paraxial Jacobian estimate and Broyden rank-1 updates to achieve fast
+    super-linear convergence without expensive finite-difference recalculations.
+
+    Attributes:
+        optic (Optic): The optical system being traced.
+        max_iter (int): Maximum number of iterations allowed.
+        tol (float): Convergence tolerance for ray aiming error.
+        _paraxial_aimer (ParaxialRayAimer): Helper to generate initial guesses.
+    """
 
     def __init__(
         self,
@@ -31,6 +44,14 @@ class IterativeRayAimer(BaseRayAimer):
         tol: float = 1e-8,
         **kwargs: Any,
     ) -> None:
+        """Initialize the IterativeRayAimer.
+
+        Args:
+            optic (Optic): The optical system to aim rays for.
+            max_iter (int, optional): Maximum number of iterations. Defaults to 20.
+            tol (float, optional): Error tolerance for convergence. Defaults to 1e-8.
+            **kwargs: Additional keyword arguments passed to BaseRayAimer.
+        """
         super().__init__(optic, **kwargs)
         self.max_iter = max_iter
         self.tol = tol
@@ -43,13 +64,41 @@ class IterativeRayAimer(BaseRayAimer):
         pupil_coords: tuple,
         initial_guess: tuple | None = None,
     ) -> tuple:
-        """Calculate ray starting coordinates using iterative aiming."""
+        """Calculate ray starting coordinates using iterative aiming.
+
+        This method solves the inverse ray tracing problem to find the starting
+        coordinates (on the object surface) or directions (for finite objects)
+        such that the ray passes through the specified pupil coordinates on the
+        stop surface.
+
+        Args:
+            fields (tuple): Field coordinates (Hy, Hx) or (angle_x, angle_y).
+            wavelengths (Any): Wavelengths of the rays in microns.
+            pupil_coords (tuple): Normalized pupil coordinates (Px, Py).
+            initial_guess (tuple | None, optional): Optional starting guess
+                (x, y, z, L, M, N). If None, a paraxial guess is used.
+
+        Returns:
+            tuple: A tuple containing the solved ray parameters (x, y, z, L, M, N).
+
+        Raises:
+            ValueError: If initial guess produces NaNs or if the solver fails
+                to converge within max_iter.
+        """
         if initial_guess:
             x, y, z, L, M, N = initial_guess
         else:
             x, y, z, L, M, N = self._paraxial_aimer.aim_rays(
                 fields, wavelengths, pupil_coords
             )
+
+        # Ensure arrays
+        x = be.as_array_1d(x)
+        y = be.as_array_1d(y)
+        z = be.as_array_1d(z)
+        L = be.as_array_1d(L)
+        M = be.as_array_1d(M)
+        N = be.as_array_1d(N)
 
         Px, Py = pupil_coords
         stop_idx = self.optic.surface_group.stop_index
@@ -71,11 +120,14 @@ class IterativeRayAimer(BaseRayAimer):
[…]
