+++ optiland/rays/ray_aiming/robust.py
@@ -20,7 +20,20 @@ if TYPE_CHECKING:
 
 @register_aimer("robust")
 class RobustRayAimer(BaseRayAimer):
-    """Robust ray aiming strategy using recursive subdivision."""
+    """
+    This class implements a robust ray aiming algorithm designed to handle
+    challenging optical systems where standard iterative methods might fail.
+    It employs a recursive subdivision approach to explore the pupil space,
+    ensuring that rays are correctly mapped from the object or entrance
+    pupil to the stop surface.
+
+    Attributes:
+        optic (Optic): The optical system instance.
+        max_iter (int): Maximum number of iterations for the internal solver.
+        tol (float): Numerical tolerance for convergence.
+        scale_fields (bool): If True, scales field coordinates during the
+            aiming process to improve numerical stability.
+    """
 
     def __init__(
         self,
@@ -30,6 +43,16 @@ class RobustRayAimer(BaseRayAimer):
         scale_fields: bool = True,
         **kwargs: Any,
     ) -> None:
+        """Initialize the RobustRayAimer.
+
+        Args:
+            optic (Optic): The optical system to aim rays for.
+            max_iter (int, optional): Maximum number of iterations. Defaults to 20.
+            tol (float, optional): Error tolerance for convergence. Defaults to 1e-8.
+            scale_fields (bool, optional): If True, scales field coordinates.
+                Defaults to True.
+            **kwargs: Additional keyword arguments passed to BaseRayAimer.
+        """
         super().__init__(optic, **kwargs)
         self.scale_fields = scale_fields
         self._iterative = IterativeRayAimer(optic, max_iter=max_iter, tol=tol)
@@ -41,7 +64,19 @@ class RobustRayAimer(BaseRayAimer):
         wavelengths: Any,
         pupil_coords: tuple,
     ) -> tuple:
-        """Calculate ray starting coordinates using recursive robust expansion."""
+        """Calculate ray starting coordinates using recursive robust expansion.
+
+        This method continually deforms the paraxial solution (t=0) into the
+        real solution (t=1).
+
+        Args:
+            fields (tuple): Field coordinates.
+            wavelengths (Any): Wavelengths in microns.
+            pupil_coords (tuple): Normalized pupil coordinates.
+
+        Returns:
+            tuple: Solved ray parameters (x, y, z, L, M, N).
+        """
         # Anchor at t=0 (Paraxial limit)
         p0 = (pupil_coords[0] * 0.0, pupil_coords[1] * 0.0)
         f0 = (fields[0] * 0.0, fields[1] * 0.0) if self.scale_fields else fields
@@ -52,7 +87,19 @@ class RobustRayAimer(BaseRayAimer):
         return self._solve(0.0, 1.0, sol_0, sol_0, fields, wavelengths, pupil_coords)
 
     def _solve(self, t0, t1, sol0, par0, fields, wl, pup):
-        """Recursively solve interval [t0, t1]."""
+        """Recursively solve the interval [t0, t1].
+
+        Args:
+            t0, t1: Start and end scalar parameters.
+            sol0: Solution at t0.
+            par0: Paraxial solution at t0.
+            fields: Target field coordinates.
+            wl: Wavelengths.
+            pup: Target pupil coordinates.
+
+        Returns:
+            tuple: Solution at is t1.
+        """
         # Base case: interval too small
         if (t1 - t0) < 1e-3:
             return sol0

[…]
