### 2025-12-22 ae2fa39
feat: Implement core ray generation with various aiming methods, numerical backends, and angle field definition.
