+++ optiland/rays/ray_aiming/iterative.py
@@ -88,6 +88,17 @@ class IterativeRayAimer(BaseRayAimer):
         if initial_guess:
             x, y, z, L, M, N = initial_guess
         else:
+            # Helper to ensure fields and pupil coords are backend arrays
+            Hx, Hy = fields
+            Hx = be.as_array_1d(Hx)
+            Hy = be.as_array_1d(Hy)
+            fields = (Hx, Hy)
+
+            Px, Py = pupil_coords
+            Px = be.as_array_1d(Px)
+            Py = be.as_array_1d(Py)
+            pupil_coords = (Px, Py)
+
             x, y, z, L, M, N = self._paraxial_aimer.aim_rays(
                 fields, wavelengths, pupil_coords
             )
@@ -101,6 +112,8 @@ class IterativeRayAimer(BaseRayAimer):
         N = be.as_array_1d(N)
 
         Px, Py = pupil_coords
+        Px = be.as_array_1d(Px)
+        Py = be.as_array_1d(Py)
         stop_idx = self.optic.surface_group.stop_index
         stop_surf = self.optic.surface_group.surfaces[stop_idx]
         is_inf = getattr(self.optic.object_surface, "is_infinite", False)
@@ -137,7 +150,7 @@ class IterativeRayAimer(BaseRayAimer):
             )
 
         num_rays = len(x)
-        full_indices = be.arange(num_rays)
+        full_indices = be.arange_indices(num_rays)
 
         # Precompute Paraxial Jacobian Factor
         wl_mean = (
@@ -156,15 +169,23 @@ class IterativeRayAimer(BaseRayAimer):
 
         # Store previous state for Broyden
         if is_inf:
-            p1_prev = be.copy(x)
-            p2_prev = be.copy(y)
+            be.copy(x)
+            be.copy(y)
         else:
-            p1_prev = be.copy(L)
-            p2_prev = be.copy(M)
-        ex_prev = be.copy(ex)
-        ey_prev = be.copy(ey)
+            be.copy(L)
+            be.copy(M)
+        be.copy(ex)
+        be.copy(ey)
 
-        for iter_idx in range(self.max_iter):
+        # Ensure we are not modifying leaf variables in-place
+        x = be.copy(x)
+        y = be.copy(y)
+        z = be.copy(z)
+        L = be.copy(L)
+        M = be.copy(M)
+        N = be.copy(N)
+
+        for _iter_idx in range(self.max_iter):
             # Check convergence
             error_sq = ex**2 + ey**2
             converged = error_sq < tol_sq
@@ -174,6 +195,7 @@ class IterativeRayAimer(BaseRayAimer):
 
             # Active Set Strategy: only process non-converged rays
             active_mask = ~converged
+            # Ensure indices are integers
             idx = full_indices[active_mask]
 
             # Extract active data
@@ -181,132 +203,81 @@ class IterativeRayAimer(BaseRayAimer):
             ey_curr = ey[idx]
             # Handle wavelengths (could be scalar or array)
             if hasattr(wavelengths, "__len__"):
-                wl_active = wavelengths[idx]
+                wavelengths[idx]
             else:
-                wl_active = wavelengths
+                pass
 
-            # Extract current params
-            if is_inf:
-                p1_curr = x[idx]
-                p2_curr = y[idx]
-            else:
-                p1_curr = L[idx]
-                p2_curr = M[idx]
+            # Update solution (Newton Step)
+            # [dx] = - [J]^-1 * [error]
+            # Determinant for active rays
+            det = J11[idx] * J22[idx] - J12[idx] * J21[idx]
 
-            # Broyden Update (skip first iter)
-            if iter_idx > 0:
-                # Get deltas from previous step
[…]
