### 2025-12-22 e41149f
Feat/ray aiming (#420)
* feat: introduce modular ray aiming system with new RayGenerator and ParaxialRayAimer.

* feat: Add iterative ray aiming algorithm and associated tests.

* feat: add iterative and robust ray aiming strategies and integrate them into the ray generator.

* feat: Introduce core Optic class and robust ray aiming strategies with accompanying documentation and tests.

* feat: add ray aim setting in Optic

* feat: implement iterative and robust methods for ray aiming

* chore: fix error handling

* docs: update the ray aiming example

* feat: Implement iterative and robust ray
[…]
