### 2025-12-22 90631d0
Fix Optic.plot_surface_sag kwargs forwarding
Accept and forward fig_to_plot_on, max_extent, num_points_grid, and buffer_factor to SurfaceSagViewer.

This restores compatibility with gallery notebooks that call plot_surface_sag(..., max_extent=..., num_points_grid=...).
