### 2025-12-23 d34cf08
feat: add draft cached ray aimer
