### 2025-12-23 6b13f40
feat: implement robust ray aiming algorithm and add tests for cached and robust aimers
