+++ tests/test_cached_aimer.py
@@ -1,113 +1,156 @@
 
-import unittest
-from unittest.mock import MagicMock, call
+import pytest
+from unittest.mock import MagicMock
 import optiland.backend as be
 from optiland.rays.ray_aiming.cached import CachedRayAimer
 from optiland.rays.ray_aiming.base import BaseRayAimer
+from optiland.rays.ray_aiming.robust import RobustRayAimer
 from optiland.optic import Optic
 
-class TestCachedRayAimer(unittest.TestCase):
-    def setUp(self):
-        # Create a mock optic
-        self.optic = MagicMock(spec=Optic)
-        surface_group_mock = MagicMock()
-        surface_group_mock.to_dict.return_value = {"surfaces": []}
-        self.optic.surface_group = surface_group_mock
-        
-        # Create a mock wrapped aimer
-        self.wrapped_aimer = MagicMock(spec=BaseRayAimer)
-        
-        # Instantiate CachedRayAimer
-        self.cached_aimer = CachedRayAimer(self.optic, self.wrapped_aimer)
+@pytest.fixture
+def mock_dependencies():
+    optic = MagicMock(spec=Optic)
+    surface_group_mock = MagicMock()
+    surface_group_mock.to_dict.return_value = {"surfaces": []}
+    optic.surface_group = surface_group_mock
+    
+    wrapped_aimer = MagicMock(spec=BaseRayAimer)
+    return optic, wrapped_aimer
 
-    def test_cache_miss_calls_wrapped(self):
-        """Test that the wrapped aimer is called on a cache miss."""
-        inputs = ((0.0,), 0.55, (0.0, 0.0))
-        expected_result = (1, 2, 3, 4, 5, 6)
-        self.wrapped_aimer.aim_rays.return_value = expected_result
-        
-        result = self.cached_aimer.aim_rays(*inputs)
-        
-        self.assertEqual(result, expected_result)
-        self.wrapped_aimer.aim_rays.assert_called_once()
+def test_cache_miss_calls_wrapped(set_test_backend, mock_dependencies):
+    """Test that the wrapped aimer is called on a cache miss."""
+    optic, wrapped_aimer = mock_dependencies
+    cached_aimer = CachedRayAimer(optic, wrapped_aimer)
+    
+    inputs = ((0.0,), 0.55, (0.0, 0.0))
+    expected_result = (1, 2, 3, 4, 5, 6)
+    wrapped_aimer.aim_rays.return_value = expected_result
+    
+    result = cached_aimer.aim_rays(*inputs)
+    
+    assert result == expected_result
+    wrapped_aimer.aim_rays.assert_called_once()
 
-    def test_cache_hit_skips_wrapped(self):
-        """Test that the wrapped aimer is NOT called on a cache hit."""
-        inputs = ((0.0,), 0.55, (0.0, 0.0))
-        expected_result = (1, 2, 3, 4, 5, 6)
-        self.wrapped_aimer.aim_rays.return_value = expected_result
-        
-        # First call (miss)
-        self.cached_aimer.aim_rays(*inputs)
-        self.wrapped_aimer.aim_rays.reset_mock()
-        
-        # Second call (hit)
-        result = self.cached_aimer.aim_rays(*inputs)
-        
-        self.assertEqual(result, expected_result)
-        self.wrapped_aimer.aim_rays.assert_not_called()
+def test_cache_hit_skips_wrapped(set_test_backend, mock_dependencies):
+    """Test that the wrapped aimer is NOT called on a cache hit."""
+    optic, wrapped_aimer = mock_dependencies
+    cached_aimer = CachedRayAimer(optic, wrapped_aimer)
+    
+    inputs = ((0.0,), 0.55, (0.0, 0.0))
[…]
