### 2025-12-23 71092a0
feat: hash all Optic components for robustness
