+++ optiland/rays/ray_aiming/cached.py
@@ -137,5 +137,11 @@ class CachedRayAimer(BaseRayAimer):
 
     def _get_system_hash(self) -> str:
         """Generate a hash for the current state of the optical system."""
-        sys_dict = self.optic.surface_group.to_dict()
-        return hashlib.md5(str(sys_dict).encode("utf-8")).hexdigest()
+        data = (
+            self.optic.surface_group.to_dict(),
+            self.optic.fields.to_dict(),
+            self.optic.wavelengths.to_dict(),
+            self.optic.aperture.to_dict() if self.optic.aperture else None,
+            self.optic.ray_aiming_config,
+        )
+        return hashlib.md5(str(data).encode("utf-8")).hexdigest()
