### 2025-12-23 6160817
feat: Implement robust ray aiming with recursive subdivision and add corresponding tests.
