### 2025-12-23 b0a4919
feat: add initial stop radius calculation methods
