+++ optiland/rays/ray_aiming/initialization.py
@@ -1,0 +1,173 @@
+"""Ray Aiming Initialization Module
+
+This module implements the initialization logic for determining the physical
+aperture stop size (Stop Radius) before the main Ray Aiming iteration begins.
+It uses a Strategy Pattern to handle different ways of calculating this radius.
+
+Kramer Harrison, 2025
+"""
+
+from __future__ import annotations
+
+import abc
+import warnings
+from typing import TYPE_CHECKING
+
+import optiland.backend as be
+from optiland.rays import RealRays
+
+if TYPE_CHECKING:
+    from optiland.optic import Optic
+
+
+class StopSizeStrategy(abc.ABC):
+    """Abstract base class for stop size determination strategies."""
+
+    def __init__(self, optic: Optic):
+        self.optic = optic
+
+    @abc.abstractmethod
+    def calculate_stop_radius(self) -> float:
+        """Calculate the radius of the stop surface."""
+        pass
+
+
+class FloatByStopStrategy(StopSizeStrategy):
+    """Strategy for 'Float By Stop Size' aperture type.
+
+    Simply returns the user-defined semi-diameter of the Stop Surface.
+    """
+
+    def calculate_stop_radius(self) -> float:
+        stop_index = self.optic.surface_group.stop_index
+        surface = self.optic.surface_group.surfaces[stop_index]
+
+        # Check for explicit aperture object first
+        if surface.aperture and hasattr(surface.aperture, "r_max"):
+            return surface.aperture.r_max
+        elif surface.aperture and hasattr(surface.aperture, "x_max"):
+            return (surface.aperture.x_max + surface.aperture.y_max) / 2.0
+
+        # Fallback to semi_aperture attribute
+        return float(surface.semi_aperture)
+
+
+class ParaxialReferenceStrategy(StopSizeStrategy):
+    """Strategy using paraxial ray trace to determine stop radius.
+
+    Traces a Paraxial Marginal Ray from the center of the object to the
+    Stop Surface.
+    """
+
+    def calculate_stop_radius(self) -> float:
+        stop_index = self.optic.surface_group.stop_index
+        para = self.optic.paraxial
+
+        # Determine marginal ray height at the stop surface
+        y_marginal, _ = para.marginal_ray()
+        return be.abs(float(y_marginal[stop_index]))
+
+
+class RealReferenceStrategy(StopSizeStrategy):
+    """Strategy using real ray trace to determine stop radius.
+
+    Traces a Real Ray from the center of the object toward the edge of the
+    Paraxial Entrance Pupil. Fallbacks to ParaxialReferenceStrategy on failure.
+    """
+
+    def calculate_stop_radius(self) -> float:
+        try:
+            return self._trace_real_marginal_ray()
+        except Exception as e:
+            warnings.warn(
+                f"RealReferenceStrategy failed: {e}. "
+                "Falling back to ParaxialReferenceStrategy.",
+                stacklevel=2,
+            )
+            fallback = ParaxialReferenceStrategy(self.optic)
+            return fallback.calculate_stop_radius()
+
+    def _trace_real_marginal_ray(self) -> float:
+        wavelength = self.optic.primary_wavelength
+        EPL = float(self.optic.paraxial.EPL())
+        EPD = float(self.optic.paraxial.EPD())
+
+        stop_index = self.optic.surface_group.stop_index
+
+        # Determine launch ray parameters (x, y, z, L, M, N)
+        obj_surf = self.optic.object_surface
+        if obj_surf and obj_surf.is_infinite:
+            # For infinite objects, we launch a ray parallel to the axis
+            # starting slightly before surface 1 to ensure robust intersection.
[…]
