### 2025-12-23 46c69d2
fix: update image sim for backend compatibility
