### 2025-12-23 5586f01
Feat/image simulation (#422)
* feat: add draft image simulation analysis

* feat: Implement full image simulation with spatially variable blur, distortion, and lateral color, along with supporting documentation and examples.

* chore: update markdown labels

* fix: update image sim for backend compatibility
