### 2025-12-24 ed47197
docs: Add API, developer guide, and gallery example for ray tracing and aiming capabilities.
