### 2025-12-24 6d05c8b
feat: add new ray aiming tests.
