+++ tests/rays/ray_aiming/test_robust_aimer.py
@@ -15,7 +15,7 @@ class TestRobustRayAimer(unittest.TestCase):
         
         # Wide angle or highly aberrated system to justify robust aiming
         # Simple setup for now: Aperture stop at surface 2
-        self.optic.add_surface(index=0, thickness=100)
+        self.optic.add_surface(index=0, thickness=np.inf)
         self.optic.add_surface(
             index=1, radius=20, thickness=10, material=IdealMaterial(n=1.5)
         )
@@ -25,7 +25,7 @@ class TestRobustRayAimer(unittest.TestCase):
         self.optic.set_field_type("angle")
         self.optic.add_field(y=0)
         self.optic.add_field(y=20) # 20 degrees
-        self.optic.set_aperture("EPD", 10.0)
+        self.optic.set_aperture("float_by_stop_size", 1.0)
         self.optic.add_wavelength(0.55)
 
     def test_robust_aimer_initialization(self):
