### 2025-12-24 f8b1bfd
fix: update backend call for material tests
