### 2025-12-24 164a52d
Feat/cached ray aiming (#428)
* feat: add draft cached ray aimer

* feat: implement robust ray aiming algorithm and add tests for cached and robust aimers

* feat: hash all Optic components for robustness

* feat: Implement robust ray aiming with recursive subdivision and add corresponding tests.

* feat: add initial stop radius calculation methods

* docs: Add API, developer guide, and gallery example for ray tracing and aiming capabilities.

* feat: add new ray aiming tests.

* fix: update backend.where call

* fix: update backend call for material tests
