### 2025-12-24 198989d
feat: Introduce `PickupManager` and `Pickup` classes for linking optical parameters and lay groundwork for multi-configuration.
