+++ optiland/multiconfig/multi_configuration.py
@@ -8,10 +8,13 @@ Kramer Harrison, 2025
 
 from __future__ import annotations
 
-import copy  # Use standard copy module
-from typing import TYPE_CHECKING
+import copy
+from typing import TYPE_CHECKING, Any, Literal
+
+from optiland.utils import set_attr_by_path
 
 if TYPE_CHECKING:
+    from optiland.materials.base import BaseMaterial
     from optiland.optic import Optic
 
 
@@ -93,6 +96,248 @@ class MultiConfiguration:
                     target_surface_idx=i,
                     source_optic=source,
                 )
+
+    def set_property(
+        self,
+        value: Any,
+        configurations: list[int] | Literal["all"] = "all",
+        surface_index: int | None = None,
+        attribute_path: str = None,
+    ):
+        """Set a property value across one or more configurations.
+
+        Args:
+            value: The value to set.
+            configurations: A list of configuration indices to update, or "all"
+                to update the base configuration and ensure links (pickups)
+                exist (or are created) for other configurations.
+            surface_index: The index of the surface to modify. If None, the
+                property is assumed to be on the Optic itself.
+            attribute_path: The dot-separated path to the attribute, or a
+                known alias ('radius', 'thickness', 'conic', 'material').
+        """
+        if configurations == "all":
+            configs_to_update = list(range(len(self.configurations)))
+        else:
+            configs_to_update = configurations
+
+        # Standardize aliases
+        if attribute_path == "radius":
+            self.set_radius(surface_index, value, configurations)
+            return
+        elif attribute_path == "thickness":
+            self.set_thickness(surface_index, value, configurations)
+            return
+        elif attribute_path == "conic":
+            self.set_conic(surface_index, value, configurations)
+            return
+        elif attribute_path == "material":
+            self.set_material(surface_index, value, configurations)
+            return
+
+        # Generic handling
+        for config_idx in configs_to_update:
+            if config_idx == 0:
+                # Set on base optic
+                self._set_generic_value(0, surface_index, attribute_path, value)
+            else:
+                if configurations == "all":
+                    # If setting "all", we want to ensure it is linked to base
+                    self._ensure_generic_pickup(
+                        config_idx, 0, surface_index, attribute_path
+                    )
+                else:
+                    # If setting specific config (not 0), assume unique val & break link
+                    self._remove_generic_pickup(
+                        config_idx, surface_index, attribute_path
+                    )
+                    self._set_generic_value(
+                        config_idx, surface_index, attribute_path, value
+                    )
+
+    def set_radius(
+        self,
+        surface_index: int,
+        value: float,
+        configurations: list[int] | Literal["all"] = "all",
+    ):
+        """Set the radius of a surface."""
+        self._set_standard_property("radius", surface_index, value, configurations)
+
+    def set_thickness(
+        self,
+        surface_index: int,
+        value: float,
+        configurations: list[int] | Literal["all"] = "all",
+    ):
+        """Set the thickness of a surface."""
+        self._set_standard_property("thickness", surface_index, value, configurations)
+
+    def set_conic(
+        self,
+        surface_index: int,
+        value: float,
+        configurations: list[int] | Literal["all"] = "all",
+    ):
+        """Set the conic constant of a surface."""
+        self._set_standard_property("conic", surface_index, value, configurations)
+
+    def set_material(
+        self,
+        surface_index: int,
+        value: str | BaseMaterial,
+        configurations: list[int] | Literal["all"] = "all",
+    ):
+        """Set the material of a surface."""
+        self._set_standard_property("material", surface_index, value, configurations)
+
+    def set_surface_property(
[…]
