+++ tests/test_multiconfig.py
@@ -16,52 +16,119 @@ def base_optic():
     return optic
 
 
-def test_multiconfig_initial_linking(base_optic):
-    """Verify that a new configuration is correctly linked to the base."""
+
+def test_set_global_radius(base_optic):
+    """Test setting a global radius updates all linked configs."""
     mc = MultiConfiguration(base_optic)
-    config1 = mc.add_configuration()
+    mc.add_configuration()
 
-    # Verify initial link by modifying base
-    base_optic.set_radius(200, 1)
-    base_optic.update()
-    config1.update()
+    mc.set_radius(1, 200, configurations="all")
+    mc.current_config(0).update()
+    mc.current_config(1).update()
 
-    r1_base = base_optic.surface_group.surfaces[1].geometry.radius
-    r1_conf = config1.surface_group.surfaces[1].geometry.radius
-    
-    assert r1_base == 200
-    assert r1_conf == 200
+    assert mc.current_config(0).surface_group.surfaces[1].geometry.radius == 200
+    assert mc.current_config(1).surface_group.surfaces[1].geometry.radius == 200
 
 
-def test_multiconfig_break_link_logic(base_optic):
-    """Verify that we can manually break the link (simulating optimization variable addition)."""
+def test_set_unique_radius(base_optic):
+    """Test setting a unique radius breaks the link."""
     mc = MultiConfiguration(base_optic)
-    config1 = mc.add_configuration()
+    mc.add_configuration()
 
-    # 1. Identify and remove the pickup for Surface 1, Radius
-    surface_number = 1
-    variable_type = 'radius'
+    # Set unique on config 1
+    mc.set_radius(1, 300, configurations=[1])
+    mc.current_config(0).update()
+    mc.current_config(1).update()
+
+    assert mc.current_config(0).surface_group.surfaces[1].geometry.radius == -100 # Original
+    assert mc.current_config(1).surface_group.surfaces[1].geometry.radius == 300
+
+    # Modify global, config 1 should NOT change
+    mc.set_radius(1, 400, configurations=[0])
+    mc.current_config(0).update()
+    mc.current_config(1).update()
+
+    assert mc.current_config(0).surface_group.surfaces[1].geometry.radius == 400
+    assert mc.current_config(1).surface_group.surfaces[1].geometry.radius == 300
+
+
+def test_relink_radius(base_optic):
+    """Test that setting configurations='all' relinks unique configs."""
+    mc = MultiConfiguration(base_optic)
+    mc.add_configuration()
+
+    # Make unique
+    mc.set_radius(1, 300, configurations=[1])
+    assert mc.current_config(1).surface_group.surfaces[1].geometry.radius == 300
+
+    # Set global 'all', should overwrite unique
+    mc.set_radius(1, 500, configurations="all")
+    mc.current_config(0).update()
+    mc.current_config(1).update()
+
+    assert mc.current_config(0).surface_group.surfaces[1].geometry.radius == 500
+    assert mc.current_config(1).surface_group.surfaces[1].geometry.radius == 500
+
+    # Update global again, check link persists
+    mc.set_radius(1, 600, configurations=[0]) # Only set 0, but 1 is linked now
+    mc.current_config(0).update()
+    mc.current_config(1).update()
+
+    assert mc.current_config(0).surface_group.surfaces[1].geometry.radius == 600
+    assert mc.current_config(1).surface_group.surfaces[1].geometry.radius == 600
+
+
+def test_set_generic_property(base_optic):
+    """Test setting a generic property (e.g. asphere coefficient)."""
+    # Use asphere coefficient on surface 1 (need to change type first?)
+    # Or just use an existing property like 'material_post.n_p'? No, read-only mostly.
[…]
