### 2025-12-24 1dc11f7
feat: add draw method to multiconfig
