### 2025-12-24 dd1d642
feat: update multiconfig plotting to use themes
