### 2025-12-24 74bba04
docs: update documentation for multiconfig
