+++ docs/learning_guide.rst
@@ -82,6 +82,7 @@ This learning guide breaks down Optiland into a series of tutorials that cover t
    examples/Tutorial_7c_Freeform_Surfaces
    examples/Tutorial_7d_Three_Mirror_Anastigmat
    examples/Tutorial_7e_Glass_Expert
+   examples/Tutorial_7f_Multi_Configuration_Zoom_Lenses
 
 
 8. Tolerancing
