+++ docs/gallery/specialized_lenses.rst
@@ -18,3 +18,4 @@ While these lenses are perhaps more complex, they are built using the same basic
     specialized_lenses/infrared
     specialized_lenses/infrared_triplet
     specialized_lenses/lithography
+    specialized_lenses/zoom_lens
