+++ README.md
@@ -134,7 +134,7 @@ Optiland is continually evolving to provide new functionalities for optical desi
 
 - [x] **GUI (PySide6-based)** - *Initial version available, ongoing enhancements.*
 - [ ] **Multi-Path Sequential Ray Tracing**
-- [ ] **Multiple Configurations (Zoom Lenses)**
+- [x] **Multiple Configurations (Zoom Lenses)**
 - [ ] **Thin Film Design and Optimization** 
 - [ ] **Jones Pupils**
 - [ ] **Additional Freeforms (Superconic, etc.)**
