+++ docs/api/api_multiconfig.rst
@@ -1,0 +1,10 @@
+Multi-Configuration
+===================
+
+The multi-configuration module allows for the definition of optical systems with multiple configurations, such as zoom lenses.
+
+.. autosummary::
+   :toctree: multiconfig/
+   :caption: Multi-Configuration Modules
+
+   multiconfig.multi_configuration
