### 2025-12-25 b1cbdd4
Feat/multiconfig (#429)
* feat: add multi-config functionality

* feat: Introduce `PickupManager` and `Pickup` classes for linking optical parameters and lay groundwork for multi-configuration.

* feat: add draw method to multiconfig

* feat: update multiconfig plotting to use themes

* docs: update documentation for multiconfig

* docs: update zoom lens examples
