### 2025-12-25 be25915
feat: Add specialized projection and wide-angle lens examples to gallery and include new objectives sample.
