+++ optiland/samples/objectives.py
@@ -533,3 +533,186 @@ class Telephoto(optic.Optic):
 
         # scale from inches to mm
         self.scale_system(25.4)
+
+
+class WideAngle100FOV(optic.Optic):
+    """A wide-angle lens design with a 100° field of view.
+
+    Reference: Milton Laikin, Lens Design, 4th ed., CRC Press, 2007, p. 108.
+    """
+
+    def __init__(self):
+        super().__init__()
+
+        self.add_surface(index=0, radius=be.inf, thickness=be.inf)
+        self.add_surface(index=1, radius=8.0107, thickness=0.2500, material="N-SK4")
+        self.add_surface(index=2, radius=1.1856, thickness=0.9613)
+        self.add_surface(index=3, radius=1.6747, thickness=0.3578, material="SF1")
+        self.add_surface(index=4, radius=-7.5157, thickness=0.2136, material="N-SK4")
+        self.add_surface(index=5, radius=0.9411, thickness=0.4146)
+        self.add_surface(index=6, radius=-1.7688, thickness=0.3333, material="SF1")
+        self.add_surface(index=7, radius=-1.5531, thickness=0.3863, material="N-SK4")
+        self.add_surface(index=8, radius=-2.2281, thickness=0.9842)
+        self.add_surface(index=9, radius=be.inf, thickness=0.1000, is_stop=True)
+        self.add_surface(index=10, radius=13.6803, thickness=0.1400, material="N-SK4")
+        self.add_surface(index=11, radius=-3.4279, thickness=0.0605)
+        self.add_surface(index=12, radius=20.0257, thickness=0.3332, material="N-SK4")
+        self.add_surface(index=13, radius=-0.9258, thickness=0.4374, material="SF1")
+        self.add_surface(index=14, radius=-3.2233, thickness=0.2679)
+        self.add_surface(index=15, radius=10.3847, thickness=0.1400, material="SF1")
+        self.add_surface(index=16, radius=2.4272, thickness=0.2609, material="N-SK4")
+        self.add_surface(index=17, radius=-3.8828, thickness=0.0150)
+        self.add_surface(index=18, radius=3.3650, thickness=0.3167, material="N-SK4")
+        self.add_surface(index=19, radius=-14.7547, thickness=1.5604)
+        self.add_surface(index=20, radius=be.inf, thickness=0.0)
+
+        self.set_aperture("imageFNO", 4.0)
+
+        self.set_field_type("angle")
+        self.add_field(y=0)
+        self.add_field(y=35.0)
+        self.add_field(y=50.0)
+
+        self.add_wavelength(value=0.4861)
+        self.add_wavelength(value=0.5876, is_primary=True)
+        self.add_wavelength(value=0.6563)
+
+        self.set_ray_aiming("iterative", cache=True)
+
+
+class ProjectionLens120FOV(optic.Optic):
+    """A 120-degree FOV projection lens design.
+
+    Reference: Milton Laikin, Lens Design, 4th ed., CRC Press, 2007, p. 110.
+    """
+
+    def __init__(self):
+        super().__init__()
+
+        self.add_surface(index=0, radius=be.inf, thickness=be.inf)
+        self.add_surface(index=1, radius=28.2482, thickness=0.5000, material="N-SK4")
+        self.add_surface(index=2, radius=2.2027, thickness=1.4379)
+        self.add_surface(index=3, radius=2.7819, thickness=1.1174, material="SF1")
+        self.add_surface(index=4, radius=-10.4886, thickness=0.3022, material="N-SK4")
+        self.add_surface(index=5, radius=1.2417, thickness=0.7247)
[…]
