### 2025-12-25 069f39b
test: add wide angle lens creation checks
