+++ optiland/analysis/jones_pupil.py
@@ -1,0 +1,271 @@
+"""Jones Pupil Analysis
+
+This module provides a Jones pupil analysis for optical systems.
+
+Kramer Harrison, 2025
+"""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Literal
+
+import matplotlib.pyplot as plt
+import numpy as np
+
+import optiland.backend as be
+from optiland.analysis.base import BaseAnalysis
+from optiland.rays import PolarizationState
+from optiland.utils import resolve_fields
+
+if TYPE_CHECKING:
+    from matplotlib.axes import Axes
+    from matplotlib.figure import Figure
+    from optiland.optic import Optic
+
+
+class JonesPupil(BaseAnalysis):
+    """Generates and plots Jones pupil maps.
+
+    This class computes the spatially resolved Jones matrix at the exit pupil
+    (or image plane) as a function of normalized pupil coordinates. It visualizes
+    the real and imaginary parts of the Jones matrix elements (Jxx, Jxy, Jyx, Jyy).
+
+    Attributes:
+        optic: Instance of the optic object to be assessed.
+        fields: Fields at which data is generated.
+        wavelengths: Wavelengths at which data is generated.
+        grid_size: The side length of the square grid of rays (NxN).
+        data: Contains Jones matrix data in a nested list, ordered by field, then
+            wavelength.
+    """
+
+    def __init__(
+        self,
+        optic: Optic,
+        fields: str | list = "all",
+        wavelengths: str | list = "all",
+        grid_size: int = 33,
+    ):
+        """Initializes the JonesPupil analysis.
+
+        Args:
+            optic: An instance of the optic object to be assessed.
+            fields: Fields at which to generate data. If 'all', all defined
+                field points are used. Defaults to "all".
+            wavelengths: Wavelengths at which to generate data. If 'all', all
+                defined wavelengths are used. Defaults to "all".
+            grid_size: The number of points along one dimension of the pupil grid.
+                Defaults to 33.
+        """
+        self.fields = resolve_fields(optic, fields)
+        self.grid_size = grid_size
+        super().__init__(optic, wavelengths)
+
+    def view(
+        self,
+        fig_to_plot_on: Figure | None = None,
+        figsize: tuple[float, float] = (16, 8),
+    ) -> tuple[Figure, list[Axes]]:
+        """Displays the Jones pupil plots.
+
+        Args:
+            fig_to_plot_on: An existing Matplotlib figure to plot on. If None,
+                a new figure is created. Defaults to None.
+            figsize: The figure size for the output window. Defaults to (16, 8).
+
+        Returns:
+            A tuple containing the Matplotlib figure and a list of its axes.
+        """
+        if not self.fields or self.fields == [(0, 0)]:
+            # resolve_fields returns [(0,0)] if max_field is 0 which happens when
+            # fields are empty or explicitly (0,0).
+            # We should check if optic has fields.
+            if not self.optic.fields.fields:
+                return None, None
+
+        # Determine layout
+        num_fields = len(self.fields)
+        num_wavelengths = len(self.wavelengths)
+
+        # For simplicity, we plot the first field and first wavelength by default
+        # if multiple are present, or we could tile them.
+        # Given standard usage, usually one looks at one field/wavelength or
[…]
