### 2025-12-28 414bc76
Implement polychromatic MTF calculation in GeometricMTF
- Updated `GeometricMTF` to support `wavelength="all"` and list of wavelengths.
- Implemented weighted average logic for polychromatic MTF calculation.
- Fixed backend compatibility issue (Torch in-place operations) in MTF generation.
- Updated `resolve_wavelengths` in `optiland/utils.py` to handle scalar inputs.
- Verified with regression tests for both polychromatic and single-wavelength cases.
