### 2025-12-28 61760de
Fix formatting and linting issues in JonesPupil analysis.
- Removed unused `Literal` import.
- Sorted imports in `TYPE_CHECKING` block.
- Combined nested `if` statements for clarity.
- Removed unused variables: `num_fields`, `num_wavelengths`, `x_in`, `y_in`.
- Resolved long line length warning.
- Applied `ruff` auto-fixes.
