### 2025-12-28 6c1264b
Implement warning for overlapping lens surfaces in Lens2D
- Updated `Lens2D.__init__` to accept `check_overlap` argument (default: False).
- Implemented `_check_overlaps` method to detect surface overlaps in the YZ plane.
- Added `tests/visualization/system/test_lens_overlap.py` to verify warning behavior.
