### 2025-12-28 e088e68
Fix Ruff formatting and linting errors
- Resolved `zip()` strict mode issue in `optiland/mtf/geometric.py` (B905).
- Applied Ruff formatting to `optiland/mtf/geometric.py` and `optiland/utils.py`.
- Verified all tests pass.
