### 2025-12-28 3daf017
feat: Add Jones Pupil analysis module, example notebook, and documentation.
