+++ optiland/analysis/jones_pupil.py
@@ -15,7 +15,6 @@ import numpy as np
 import optiland.backend as be
 from optiland.analysis.base import BaseAnalysis
 from optiland.rays import PolarizationState
-from optiland.utils import resolve_fields
 
 if TYPE_CHECKING:
     from matplotlib.axes import Axes
@@ -33,32 +32,33 @@ class JonesPupil(BaseAnalysis):
 
     Attributes:
         optic: Instance of the optic object to be assessed.
-        fields: Fields at which data is generated.
+    Attributes:
+        optic: Instance of the optic object to be assessed.
+        field: Field at which data is generated (Hx, Hy).
         wavelengths: Wavelengths at which data is generated.
         grid_size: The side length of the square grid of rays (NxN).
-        data: Contains Jones matrix data in a nested list, ordered by field, then
-            wavelength.
+        data: Contains Jones matrix data in a list, ordered by wavelength.
     """
 
     def __init__(
         self,
         optic: Optic,
-        fields: str | list = "all",
+        field: tuple[float, float] = (0, 0),
         wavelengths: str | list = "all",
-        grid_size: int = 33,
+        grid_size: int = 65,
     ):
         """Initializes the JonesPupil analysis.
 
         Args:
             optic: An instance of the optic object to be assessed.
-            fields: Fields at which to generate data. If 'all', all defined
-                field points are used. Defaults to "all".
+            field: The normalized field coordinates (Hx, Hy) at which to
+                generate data. Defaults to (0, 0).
             wavelengths: Wavelengths at which to generate data. If 'all', all
                 defined wavelengths are used. Defaults to "all".
             grid_size: The number of points along one dimension of the pupil grid.
-                Defaults to 33.
+                Defaults to 65.
         """
-        self.fields = resolve_fields(optic, fields)
+        self.field = field
         self.grid_size = grid_size
         super().__init__(optic, wavelengths)
 
@@ -77,29 +77,12 @@ class JonesPupil(BaseAnalysis):
         Returns:
             A tuple containing the Matplotlib figure and a list of its axes.
         """
-        if (
-            not self.fields or self.fields == [(0, 0)]
-        ) and not self.optic.fields.fields:
-            return None, None
-
-        # Determine layout
-
-        # For simplicity, we plot the first field and first wavelength by default
-        # if multiple are present, or we could tile them.
-        # Given standard usage, usually one looks at one field/wavelength or
-        # tabs through them. But BaseAnalysis `view` usually static.
-        # Let's plot the first field and first wavelength for now,
-        # or iterate? SpotDiagram plots all fields.
-        # Jones Pupil is 8 plots per field/wavelength. That's too many to tile.
-        # We will plot the PRIMARY field (or first) and PRIMARY wavelength (or first).
-
-        # Select first field and primary wavelength index
-        field_idx = 0
+        # Select primary wavelength index
         wl_idx = 0
         if self.optic.primary_wavelength in self.wavelengths:
             wl_idx = self.wavelengths.index(self.optic.primary_wavelength)
 
-        data_fw = self.data[field_idx][wl_idx]
+        data_fw = self.data[wl_idx]
 
         if fig_to_plot_on:
             fig = fig_to_plot_on
@@ -121,12 +104,6 @@ class JonesPupil(BaseAnalysis):
         px = be.to_numpy(data_fw["Px"]).reshape(self.grid_size, self.grid_size)
[…]
