### 2025-12-28 938070f
fix: update tests after Jones pupil update
