### 2025-12-28 4ea9e32
Add Jones Pupil Analysis (#432)
* feat: Add JonesPupil analysis class

- Added `JonesPupil` class to `optiland/analysis/jones_pupil.py`.
- Exposed `JonesPupil` in `optiland/analysis/__init__.py`.
- Added `JonesPupil` to `optiland/analysis/jones_pupil.py`.
- Added unit tests in `tests/analysis/test_jones_pupil.py`.
- Jones pupil analysis generates a grid of rays and computes the Jones matrix elements (Jxx, Jxy, Jyx, Jyy) by projecting the 3x3 global polarization matrix onto a local basis defined by the ray direction.
- Visualization displays the real and imaginary parts of the Jones matrix elem
[…]
