### 2026-01-01 5dcee06
fixes to the cache problem: now faster for higher number of coeffs. autograd remains intact
