+++ optiland/geometries/forbes/qpoly.py
@@ -11,30 +11,11 @@ Copyright (c) 2017 Brandon Dube
 from __future__ import annotations
 
 from collections import defaultdict
-from functools import lru_cache, wraps
+from functools import cache
 
 from scipy import special
 
 import optiland.backend as be
-
-
-def autograd_aware_cache(func):
-    """
-    A decorator that provides LRU caching but automatically bypasses the
-    cache if the PyTorch backend is active with gradient tracking enabled
-    """
-    cached_func = lru_cache(2000)(func)
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        if be.get_backend() == "torch" and be.grad_mode.requires_grad:
-            # If gradients are on, bypass the cache and run the original function
-            return func(*args, **kwargs)
-        else:
-            # Otherwise (Numpy backend or PyTorch with no_grad), use the cache
-            return cached_func(*args, **kwargs)
-
-    return wrapper
 
 
 def kronecker(i: int, j: int) -> int:
@@ -42,7 +23,7 @@ def kronecker(i: int, j: int) -> int:
     return 1 if i == j else 0
 
 
-@autograd_aware_cache
+@cache
 def gamma_func(n: int, m: int) -> float:
     """Recursive gamma function for Q2D polynomials."""
     if n == 1 and m == 2:
@@ -62,7 +43,7 @@ def gamma_func(n: int, m: int) -> float:
 # bfs polynomials logic
 
 
-@autograd_aware_cache
+@cache
 def g_qbfs(n_minus_1: int) -> float:
     """Recurrence coefficient g for Q-BFS polynomials."""
     if n_minus_1 == 0:
@@ -71,24 +52,26 @@ def g_qbfs(n_minus_1: int) -> float:
     return -(1 + g_qbfs(n_minus_2) * h_qbfs(n_minus_2)) / f_qbfs(n_minus_1)
 
 
-@autograd_aware_cache
+@cache
 def h_qbfs(n_minus_2: int) -> float:
     """Recurrence coefficient h for Q-BFS polynomials."""
     n = n_minus_2 + 2
     return -n * (n - 1) / (2 * f_qbfs(n_minus_2))
 
 
-@autograd_aware_cache
+@cache
 def f_qbfs(n: int) -> float:
     """Recurrence coefficient f for Q-BFS polynomials."""
     if n == 0:
         return 2.0
     if n == 1:
-        return be.sqrt(19) / 2
+        return 19**0.5 / 2
+
     term1 = float(n * (n + 1) + 3)
     term2 = g_qbfs(n - 1) ** 2
     term3 = h_qbfs(n - 2) ** 2
-    return be.sqrt(term1 - term2 - term3)
+
+    return (term1 - term2 - term3) ** 0.5
 
 
 def change_basis_qbfs_to_pn(cs: list[float]) -> be.array:
@@ -293,7 +276,7 @@ def compute_z_zprime_qbfs(
 # q2d polynomials logic
 
 
-@autograd_aware_cache
+@cache
 def _g_q2d_raw(n: int, m: int) -> float:
     """Raw G coefficient for Q2D polynomials."""
     if n == 0:
@@ -317,7 +300,7 @@ def _g_q2d_raw(n: int, m: int) -> float:
     return term1 * gamma_func(n, m)
 
 
-@autograd_aware_cache
+@cache
 def _f_q2d_raw(n: int, m: int) -> float:
     """Raw F coefficient for Q2D polynomials."""
     if n == 0 and m == 1:
[…]
