### 2026-01-07 0b54771
feat: Implement GRIN (Gradient-Index) material and propagation
Add comprehensive support for gradient-index optical materials:

- Add GradientMaterial class with Zemax Gradient3-like profile:
  n(r, z) = n0 + nr2*r² + nr4*r⁴ + nr6*r⁶ + nz1*z + nz2*z² + nz3*z³
  Supports both radial and axial refractive index gradients

- Implement GRINPropagation using RK4 numerical integration:
  Solves ray equation: d/ds(n·dr/ds) = ∇n
  Adaptive step size control for accuracy
  Handles boundary conditions and normalization

- Add comprehensive test suite (9 tests):
  Material initialization and refractive in
[…]
