+++ optiland/propagation/grin.py
@@ -1,9 +1,16 @@
-"""Placeholder for Graded-Index (GRIN) propagation model."""
+"""Gradient-Index (GRIN) propagation model.
+
+This module implements ray propagation through gradient-index media using
+the Runge-Kutta 4th order (RK4) numerical integration method.
+
+Kramer Harrison, 2025
+"""
 
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+import optiland.backend as be
 from optiland.propagation.base import BasePropagationModel
 
 if TYPE_CHECKING:
@@ -12,33 +19,197 @@ if TYPE_CHECKING:
 
 
 class GRINPropagation(BasePropagationModel):
-    """Placeholder for GRIN propagation.
+    """Gradient-Index (GRIN) ray propagation model.
 
-    This model is not yet implemented and will raise an error if used.
+    This model implements ray propagation through gradient-index media
+    using numerical integration (RK4 method) to solve the ray equation.
+
     """
 
+    def __init__(self, material: BaseMaterial):
+        """Initializes the GRINPropagation model.
+
+        Args:
+            material: A reference to the parent material instance, used to
+                query for refractive index and gradient.
+
+        """
+        self.material = material
+        self.step_size = 0.01  # Default step size for RK4 integration (in mm)
+
     def propagate(self, rays: RealRays, t: float) -> None:
-        """Raises NotImplementedError.
+        """Propagates rays through the GRIN medium to the exit surface.
+
+        This method uses RK4 numerical integration to solve the ray equation
+        in a gradient-index medium. The propagation continues until rays
+        reach the exit surface defined by the distance t.
 
         Args:
             rays: The rays object to be propagated.
-            t: The distance to propagate.
+            t: The approximate distance to the exit surface along the z-axis.
+                This is used as a stopping criterion for the integration.
+
         """
-        raise NotImplementedError("GRIN propagation is not yet implemented.")
+        # For GRIN media, we need to use numerical integration
+        # For now, use the provided t as a target distance and integrate
+        # through the GRIN medium.
+
+        num_rays = len(rays.x)
+
+        # Store initial state for each ray
+        x = be.copy(rays.x)
+        y = be.copy(rays.y)
+        z = be.copy(rays.z)
+        L = be.copy(rays.L)
+        M = be.copy(rays.M)
+        N = be.copy(rays.N)
+
+        # Target z position (exit surface)
+        z_target = z + t
+
+        # Get wavelength
+        w = rays.w
+
+        # Maximum iterations to prevent infinite loops
+        max_iterations = 10000
+
+        for iteration in range(max_iterations):
+            # Check which rays still need to propagate
+            # Rays that have reached or passed the target z are done
+            active = (rays.N > 0) & (rays.z < z_target - 1e-6)
+
+            if not be.any(active):
+                break
+
+            # RK4 integration step for active rays
+            if be.any(active):
+                self._rk4_step(rays, active, z_target)
+
+        # Update OPD (optical path difference)
+        # For GRIN, this should be accumulated during propagation
+        # For simplicity, we approximate it here
+        # Use _calculate_n to avoid caching issues with array kwargs
[…]
