+++ optiland/materials/__init__.py
@@ -6,6 +6,7 @@ from __future__ import annotations
 
 from .abbe import AbbeMaterial
 from .base import BaseMaterial
+from .gradient_material import GradientMaterial
 from .ideal import IdealMaterial
 from .material import Material
 from .material_file import MaterialFile
@@ -24,6 +25,8 @@ __all__ = [
     "AbbeMaterial",
     # From base.py
     "BaseMaterial",
+    # From gradient_material.py
+    "GradientMaterial",
     # From ideal.py
     "IdealMaterial",
     # From material.py
