### 2026-01-07 38851b8
test: Add axial gradient reflection effect tests and example
Add comprehensive testing for axial GRIN reflection effect where
refractive index increases with z, causing rays to bend toward
the normal (gradual reflection).

New test cases:
- test_grin_axial_gradient_reflection_effect: Validates that obliquely
  incident rays bend toward z-axis in axial GRIN medium
- test_grin_axial_gradient_strong_reflection: Tests strong gradient
  that creates significant ray bending

Key findings:
- n(z) = 1.5 + 0.05*z over 20mm creates 40% direction change
- Higher incident angles produce more bending
- Ray
[…]
