+++ tests/propagation/test_grin.py
@@ -166,3 +166,104 @@ def test_grin_propagation_serialization():
     # Check that it's a valid GRINPropagation instance
     assert isinstance(model2, GRINPropagation)
     assert model2.material is material
+
+
+def test_grin_axial_gradient_reflection_effect():
+    """Test GRIN propagation with axial gradient creating gradual reflection.
+
+    When the refractive index increases in the z-direction, obliquely incident
+    rays bend toward the normal (z-axis), creating a gradual "reflection"
+    or turning effect.
+    """
+    # Create a material with strong axial gradient
+    # n(z) = n0 + nz1*z
+    # nz1 > 0 means refractive index increases with z
+    n0 = 1.5
+    nz1 = 0.05  # Strong gradient: n increases from 1.5 at z=0 to 2.5 at z=20
+    material = GradientMaterial(n0=n0, nz1=nz1)
+
+    # Create a ray starting at origin with oblique angle
+    # Ray direction: L=0.3, M=0, N=sqrt(1-0.3^2) ≈ 0.954
+    # This means the ray is at ~16.7° from the z-axis
+    initial_angle = 0.3  # sin(angle) ≈ 0.3
+    L_init = initial_angle
+    N_init = be.sqrt(1 - initial_angle**2)
+
+    rays = RealRays(
+        x=[0.0], y=[0.0], z=[0.0],
+        L=[L_init], M=[0.0], N=[N_init],
+        intensity=[1.0], wavelength=[0.5]
+    )
+
+    # Store initial direction
+    L_initial = rays.L[0]
+    N_initial = rays.N[0]
+
+    # Propagate through the axial GRIN medium
+    thickness = 20.0
+    model = GRINPropagation(material)
+    model.propagate(rays, t=thickness)
+
+    # As n increases with z, the ray should bend toward the normal (z-axis)
+    # This means L should decrease and N should increase
+    assert rays.L[0] < L_initial, "Ray should bend toward z-axis (L decreases)"
+    assert rays.N[0] > N_initial, "Ray should bend toward z-axis (N increases)"
+
+    # The ray should still be propagating forward (z should increase)
+    assert rays.z[0] > 0
+
+    # Verify refractive index at different positions
+    n_start = material.n(0.5, x=0.0, y=0.0, z=0.0)
+    n_end = material.n(0.5, x=0.0, y=0.0, z=thickness)
+
+    assert be.allclose(n_start, n0, atol=1e-6)
+    assert be.allclose(n_end, n0 + nz1 * thickness, atol=1e-6)
+
+    # The stronger the gradient, the more the ray bends
+    # Calculate the bending angle
+    final_L = rays.L[0]
+    final_N = rays.N[0]
+    bending_amount = L_initial - final_L
+
+    assert bending_amount > 0, "Ray should have bent toward z-axis"
+
+    # With this strong gradient, we expect significant bending
+    # (at least 10% reduction in L component)
+    assert bending_amount > 0.1 * L_initial, \
+        f"Expected significant bending, got {bending_amount:.4f}"
+
+
+def test_grin_axial_gradient_strong_reflection():
+    """Test strong axial gradient that can almost reflect the ray.
+
+    With a very strong gradient in n, the ray can bend so much that it
+    approaches a turning point or even reflects back.
+    """
+    # Create a material with very strong axial gradient
+    n0 = 1.3
+    nz1 = 0.1  # Very strong gradient
+
+    material = GradientMaterial(n0=n0, nz1=nz1)
+
+    # Create a ray with significant initial angle
+    L_init = 0.5  # ~30° from z-axis
+    N_init = be.sqrt(1 - L_init**2)
+
+    rays = RealRays(
[…]
