### 2026-01-07 a8025ab
feat: Support bidirectional GRIN propagation with reflection
Enable GRIN media to handle ray reflection when refractive index
gradient is strong enough to reverse ray direction (N < 0).

Changes:
- Remove N > 0 constraint in GRINPropagation
- Implement adaptive step size for bidirectional propagation
- Add boundary detection for both entrance and exit surfaces
- Allow rays to exit from either surface after reflection

This enables proper modeling of gradual reflection effects in
gradient-index media where rays bend toward higher refractive
index regions until completely reversing direction.
