+++ examples/axial_grin_reflection.py
@@ -1,252 +1,0 @@
-"""
-Example: Axial GRIN Reflection Effect
-
-This example demonstrates how an axial gradient in refractive index
-can create a gradual "reflection" effect, where obliquely incident rays
-bend toward the normal as they propagate through the medium.
-
-Physical principle:
-When n increases with z, rays propagating at an angle gradually bend
-toward the z-axis (the normal), similar to how light bends when entering
-a higher-index medium, but continuously throughout the GRIN medium.
-"""
-
-import matplotlib.pyplot as plt
-import numpy as np
-
-from optiland.materials import GradientMaterial
-from optiland.propagation.grin import GRINPropagation
-from optiland.rays import RealRays
-
-
-def visualize_axial_gradient_reflection():
-    """Visualize the reflection effect of axial GRIN medium."""
-    print("\n" + "="*70)
-    print("Axial GRIN Medium - Gradual Reflection Effect")
-    print("="*70)
-
-    # Create a material with axial gradient
-    # n(z) = n0 + nz1*z
-    # nz1 > 0 means refractive index increases with z
-    n0 = 1.5
-    nz1 = 0.05  # n increases from 1.5 at z=0 to 2.5 at z=20
-
-    material = GradientMaterial(n0=n0, nz1=nz1)
-
-    print(f"\nMaterial properties:")
-    print(f"  n0  = {n0}")
-    print(f"  nz1 = {nz1}")
-    print(f"  Formula: n(z) = {n0} + {nz1}*z")
-
-    # Calculate refractive index at different z positions
-    z_positions = [0, 5, 10, 15, 20]
-    print(f"\nRefractive index distribution:")
-    for z in z_positions:
-        n = material.n(0.55, x=0.0, y=0.0, z=z)
-        print(f"  z = {z:2d} mm: n = {n:.3f}")
-
-    # Test rays at different incident angles
-    angles = [0.1, 0.2, 0.3, 0.4, 0.5]  # sin(angle)
-    thickness = 20.0
-
-    ray_data = []
-
-    for angle in angles:
-        # Create ray with oblique angle
-        L_init = angle
-        N_init = np.sqrt(1 - angle**2)
-
-        rays = RealRays(
-            x=[0.0], y=[0.0], z=[0.0],
-            L=[L_init], M=[0.0], N=[N_init],
-            intensity=[1.0], wavelength=[0.55]
-        )
-
-        L_initial = rays.L[0]
-        N_initial = rays.N[0]
-
-        # Propagate
-        prop_model = GRINPropagation(material)
-        prop_model.propagate(rays, t=thickness)
-
-        # Calculate results
-        L_final = rays.L[0]
-        N_final = rays.N[0]
-        bending = L_initial - L_final
-        angle_deg = np.arcsin(angle) * 180 / np.pi
-
-        ray_data.append({
-            'angle_deg': angle_deg,
-            'L_init': L_initial,
-            'L_final': L_final,
-            'bending': bending,
-            'z_final': rays.z[0]
-        })
-
-        print(f"\nRay at {angle_deg:.1f}° incidence:")
-        print(f"  Initial direction: L={L_init:.3f}, N={N_initial:.3f}")
-        print(f"  Final direction:   L={L_final:.3f}, N={N_final:.3f}")
-        print(f"  Bending amount:    {bending:.4f}")
-        print(f"  Final z position:  {rays.z[0]:.2f} mm")
-
-    # Visualization
[…]
