+++ examples/grin_lens_example.py
@@ -1,225 +1,0 @@
-"""
-Example: GRIN Lens Ray Tracing
-
-This example demonstrates how to create and use a Gradient-Index (GRIN) lens
-in Optiland. GRIN lenses have a refractive index that varies with position,
-allowing for unique optical properties.
-
-In this example, we create a simple GRIN rod lens with a radial gradient
-and trace rays through it to observe the focusing/defocusing behavior.
-"""
-
-import matplotlib.pyplot as plt
-import numpy as np
-
-from optiland import optic
-from optiland.materials import GradientMaterial
-from optiland.geometries import Plane
-
-
-def create_grin_lens_system():
-    """Create an optical system with a GRIN lens.
-
-    Returns:
-        An Optiland OpticalSystem instance.
-
-    """
-    # Create a new optical system
-    lens = optic.OpticalSystem()
-
-    # Add object surface (point source at infinity)
-    lens.add_surface(
-        geometry=Plane(),
-        material=None,  # Air before first surface
-        is_stop=True
-    )
-
-    # Create a GRIN material with radial gradient
-    # n(r) = n0 + nr2*r^2 + nr4*r^4
-    # Negative nr2 creates a focusing lens (higher index at center)
-    grin_material = GradientMaterial(
-        n0=1.6,      # Base refractive index
-        nr2=-0.05,   # Negative coefficient for focusing (n decreases with r²)
-        nr4=0.0,     # No r^4 term
-        nr6=0.0,     # No r^6 term
-    )
-
-    # Add GRIN lens entrance surface (plane at z=0)
-    lens.add_surface(
-        geometry=Plane(),
-        material=grin_material,
-        thickness=10.0  # 10 mm thick GRIN rod
-    )
-
-    # Add GRIN lens exit surface (plane at z=10)
-    lens.add_surface(
-        geometry=Plane(),
-        material=None  # Air after lens
-    )
-
-    # Add image surface
-    lens.add_surface(
-        geometry=Plane()
-    )
-
-    return lens
-
-
-def trace_rays_through_grin_lens():
-    """Trace rays through the GRIN lens and visualize the results."""
-    # Create the GRIN lens system
-    system = create_grin_lens_system()
-
-    # Define ray source parameters
-    wavelength = 0.55  # 550 nm (green light)
-    num_rays = 11       # Number of rays
-    ray_height = 2.0    # Maximum ray height (mm)
-
-    # Create a fan of rays starting at different heights
-    y_positions = np.linspace(-ray_height, ray_height, num_rays)
-
-    # Trace rays
-    for y in y_positions:
-        # Add ray starting at x=0, y=current height, z=-infinity
-        # Direction: along z-axis (0, 0, 1)
-        system.add_rays(
-            x=[0.0],
-            y=[y],
-            z=[np.inf],  # Object at infinity
-            L=[0.0],
-            M=[0.0],
-            N=[1.0],
-            wavelength=[wavelength]
-        )
-
-    # Trace through system
-    system.trace()
-
-    # Extract ray paths for visualization
-    # Plot the ray trace
-    fig, ax = plt.subplots(figsize=(12, 6))
-
-    # Draw lens boundaries
-    lens_length = 10.0
-    lens_radius = 3.0
-
-    # Draw GRIN lens
-    lens_rect = plt.Rectangle((0, -lens_radius), lens_length, 2*lens_radius,
-                              fill=False, edgecolor='blue',
-                              linewidth=2, label='GRIN Lens')
-    ax.add_patch(lens_rect)
-
[…]
