### 2026-01-08 e83d422
Delete GRIN_IMPLEMENTATION_SUMMARY.md
move the description to PR
