+++ GRIN_IMPLEMENTATION_SUMMARY.md
@@ -1,220 +1,0 @@
-# GRIN (Gradient-Index) 功能实现总结
-
-## 概述
-
-成功实现了 Optiland 的 GRIN（梯度折射率）功能，允许用户模拟和分析具有空间变化折射率的光学元件。
-
-## 实现内容
-
-### 1. GradientMaterial 类
-**文件**: [optiland/materials/gradient_material.py](optiland/materials/gradient_material.py)
-
-新的材料类，支持类似 Zemax Gradient3 格式的折射率分布：
-
-```python
-n(r, z) = n0 + nr2*r² + nr4*r⁴ + nr6*r⁶ + nz1*z + nz2*z² + nz3*z³
-```
-
-其中 r² = x² + y²
-
-**主要功能**：
-- `_calculate_n()`: 计算给定位置的折射率
-- `get_index_and_gradient()`: 计算折射率和梯度（用于光线追踪）
-- 序列化/反序列化支持
-
-### 2. GRINPropagation 类
-**文件**: [optiland/propagation/grin.py](optiland/propagation/grin.py)
-
-实现了使用 RK4（Runge-Kutta 4阶）数值积分的光线传播模型。
-
-**核心算法**：
-- 求解光线方程：d/ds(n·dr/ds) = ∇n
-- 支持自适应步长控制
-- 处理边界条件和归一化
-
-**主要方法**：
-- `propagate()`: 主传播方法
-- `_rk4_step()`: RK4 数值积分步骤
-- `_ray_derivative()`: 计算光线状态导数
-
-### 3. 测试套件
-**文件**: [tests/propagation/test_grin.py](tests/propagation/test_grin.py)
-
-包含 9 个测试用例：
-- 材料初始化测试
-- 折射率计算测试
-- 梯度计算测试
-- 直线传播测试（均匀介质）
-- 径向梯度传播测试
-- 多光线传播测试
-- 序列化测试
-
-**测试结果**: ✅ 全部通过（25/25）
-
-### 4. 使用示例
-**文件**: [examples/simple_grin_test.py](examples/simple_grin_test.py)
-
-演示了：
-- 径向 GRIN 透镜（会聚/发散）
-- 轴向 GRIN 介质
-- 光线路径可视化
-
-## 使用方法
-
-### 基本用法
-
-```python
-from optiland.materials import GradientMaterial
-from optiland.propagation.grin import GRINPropagation
-from optiland.rays import RealRays
-
-# 1. 创建 GRIN 材料
-# 负 nr2 = 会聚透镜（中心折射率高）
-material = GradientMaterial(
-    n0=1.6,      # 基础折射率
-    nr2=-0.02,   # 径向梯度系数
-    nr4=0.0,
-    nr6=0.0,
-    nz1=0.0,     # 轴向梯度系数
-    nz2=0.0,
-    nz3=0.0
-)
-
-# 2. 创建光线
-rays = RealRays(
-    x=[0.0], y=[2.0], z=[0.0],  # 位置
-    L=[0.0], M=[0.0], N=[1.0],  # 方向（沿 z 轴）
-    intensity=[1.0],
-    wavelength=[0.55]  # 波长（微米）
-)
-
-# 3. 创建传播模型并传播
-prop_model = GRINPropagation(material)
-prop_model.propagate(rays, t=10.0)  # 传播 10 mm
-
-# 4. 查看结果
-print(f"最终位置: x={rays.x[0]:.3f}, y={rays.y[0]:.3f}, z={rays.z[0]:.3f}")
-```
-
-### 常见 GRIN 配置
-
-#### 1. 会聚 GRIN 透镜（径向）
-```python
-# 折射率从中心向外递减
-material = GradientMaterial(n0=1.6, nr2=-0.02)
-```
-
[…]
