### 2026-01-14 e563f4e
docs: ruff format & update author
