+++ optiland/propagation/grin.py
@@ -3,7 +3,7 @@
 This module implements ray propagation through gradient-index media using
 the Runge-Kutta 4th order (RK4) numerical integration method.
 
-Kramer Harrison, 2025
+Jie Cai (zhazhajust), 2025
 """
 
 from __future__ import annotations
@@ -56,7 +56,7 @@ class GRINPropagation(BasePropagationModel):
         # For now, use the provided t as a target distance and integrate
         # through the GRIN medium.
 
-        num_rays = len(rays.x)
+        len(rays.x)
 
         # Store initial state for each ray
         z_initial = be.copy(rays.z)
@@ -65,15 +65,14 @@ class GRINPropagation(BasePropagationModel):
         z_target = z_initial + t
 
         # Get wavelength
-        w = rays.w
 
         # Maximum iterations to prevent infinite loops
         max_iterations = 10000
 
-        for iteration in range(max_iterations):
+        for _iteration in range(max_iterations):
             # Check which rays still need to propagate
             # Rays are active if they are within the medium bounds
-            # Allow rays to propagate in either direction (N can be positive or negative)
+            # Allow rays to propagate in both directions (N can be positive or negative)
             z_min = be.minimum(z_initial, z_target)
             z_max = be.maximum(z_initial, z_target)
             active = (rays.z > z_min - 1e-6) & (rays.z < z_max - 1e-6)
@@ -119,14 +118,18 @@ class GRINPropagation(BasePropagationModel):
         if be.any(N > 0):
             # Rays propagating toward z_max
             remaining_forward = (z_max - z) * (N > 0)
-            step_size_forward = be.minimum(self.step_size, be.abs(remaining_forward * 0.9))
+            step_size_forward = be.minimum(
+                self.step_size, be.abs(remaining_forward * 0.9)
+            )
         else:
             step_size_forward = be.full_like(z, self.step_size)
 
         if be.any(N < 0):
             # Rays propagating toward z_min
             remaining_backward = (z - z_min) * (N < 0)
-            step_size_backward = be.minimum(self.step_size, be.abs(remaining_backward * 0.9))
+            step_size_backward = be.minimum(
+                self.step_size, be.abs(remaining_backward * 0.9)
+            )
         else:
             step_size_backward = be.full_like(z, self.step_size)
 
@@ -168,12 +171,12 @@ class GRINPropagation(BasePropagationModel):
         k4 = self._ray_derivative(x4, y4, z4, L4, M4, N4, w)
 
         # Update state
-        rays.x[active] = x + (step_size / 6.0) * (k1[0] + 2*k2[0] + 2*k3[0] + k4[0])
-        rays.y[active] = y + (step_size / 6.0) * (k1[1] + 2*k2[1] + 2*k3[1] + k4[1])
-        rays.z[active] = z + (step_size / 6.0) * (k1[2] + 2*k2[2] + 2*k3[2] + k4[2])
-        rays.L[active] = L + (step_size / 6.0) * (k1[3] + 2*k2[3] + 2*k3[3] + k4[3])
-        rays.M[active] = M + (step_size / 6.0) * (k1[4] + 2*k2[4] + 2*k3[4] + k4[4])
-        rays.N[active] = N + (step_size / 6.0) * (k1[5] + 2*k2[5] + 2*k3[5] + k4[5])
+        rays.x[active] = x + (step_size / 6.0) * (k1[0] + 2 * k2[0] + 2 * k3[0] + k4[0])
+        rays.y[active] = y + (step_size / 6.0) * (k1[1] + 2 * k2[1] + 2 * k3[1] + k4[1])
[…]
