+++ examples/simple_grin_test.py
@@ -5,6 +5,8 @@ This example demonstrates basic usage of GRIN material and propagation
 without requiring a full optical system.
 """
 
+from __future__ import annotations
+
 import matplotlib.pyplot as plt
 import numpy as np
 
@@ -15,30 +17,30 @@ from optiland.rays import RealRays
 
 def test_radial_grin_lens():
     """Test ray propagation through a radial GRIN medium."""
-    print("\n" + "="*60)
+    print("\n" + "=" * 60)
     print("Testing Radial GRIN Lens")
-    print("="*60)
+    print("=" * 60)
 
     # Create a GRIN material with negative nr2 (focusing lens)
     # n(r) = n0 + nr2*r^2
     # Negative nr2 means higher index at center → focusing
     material = GradientMaterial(
-        n0=1.6,      # Base refractive index at center
-        nr2=-0.2,   # Negative coefficient (focusing)
+        n0=1.6,  # Base refractive index at center
+        nr2=-0.2,  # Negative coefficient (focusing)
         nr4=0.0,
         nr6=0.0,
         nz1=0.0,
         nz2=0.0,
-        nz3=0.0
+        nz3=0.0,
     )
 
-    print(f"\nGRIN Material Properties:")
+    print("\nGRIN Material Properties:")
     print(f"  n0  = {material.n0}")
     print(f"  nr2 = {material.nr2}")
-    print(f"  Formula: n(r) = n0 + nr2*r²")
+    print("  Formula: n(r) = n0 + nr2*r²")
 
     # Test refractive index at different positions
-    print(f"\nRefractive Index at different positions:")
+    print("\nRefractive Index at different positions:")
     for r in [0.0, 0.5, 1.0, 1.5, 2.0]:
         n = material.n(0.55, x=r, y=0.0, z=0.0)
         print(f"  r = {r:4.1f} mm: n = {n:.4f}")
@@ -58,7 +60,7 @@ def test_radial_grin_lens():
             M=[0.0],
             N=[1.0],  # Propagating in +z direction
             intensity=[1.0],
-            wavelength=[0.55]  # 550 nm
+            wavelength=[0.55],  # 550 nm
         )
         rays_list.append(ray)
 
@@ -84,57 +86,81 @@ def test_radial_grin_lens():
         final_y.append(rays.y[0])
         final_L.append(rays.L[0])  # Should be ~0 for symmetric lens
 
-        print(f"  Ray {i:2d}: y_start = {y_start:+6.2f} mm → "
-              f"y_final = {rays.y[0]:+6.3f} mm, "
-              f"z = {rays.z[0]:.2f} mm")
+        print(
+            f"  Ray {i:2d}: y_start = {y_start:+6.2f} mm → "
+            f"y_final = {rays.y[0]:+6.3f} mm, "
+            f"z = {rays.z[0]:.2f} mm"
+        )
 
     # Visualization
     fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(14, 5))
 
     # Plot 1: Ray paths
-    for i, (y_start, y_end) in enumerate(zip(initial_y, final_y)):
+    for i, (y_start, y_end) in enumerate(zip(initial_y, final_y, strict=False)):
         color = plt.cm.viridis(i / len(initial_y))
         # Draw simplified ray path (straight line approximation for visualization)
-        ax1.plot([0, thickness], [y_start, y_end], 'o-',
-                color=color, alpha=0.7, linewidth=1.5,
-                label=f'y={y_start:.1f}' if i % 2 == 0 else '')
+        ax1.plot(
+            [0, thickness],
[…]
