### 2026-01-16 #9152 @kyleniemeyer
:wave: Hello @HarrisonKramer! Thank you for your patience while your submission has been waiting in our pre-review queue.

## What this means for your submission

1. Software Design (~200 words): Explain the trade-offs you weighed, the design/architecture you settled on, and why it matters. If other packages exist in this space, you must include an explicit "build vs. contribute" justification explaining why you created new software rather than contributing to existing projects.
2. Research Impact Statement (~200 words): Provide concise evidence of either realized impact (e.g., external use, integrations, enabled results) or credible near-term significance (novel capability with benchmarks, reproducible materials, and community-readiness signals such as documentation, tests, license, releases, and contribution process).
3. AI Usage Disclosure: Disclose any use of generative AI in your submission, including software creation, documentation generation, or paper authoring. If you did not use AI tools, please state this explicitly.

For full details, please see:

- Our updated scope definition: https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance
- An example paper in the new format: https://joss.readthedocs.io/en/latest/example_paper.html

We understand this may be frustrating, particularly if you've been waiting for some time. We apologise for that. These changes reflect the rapidly evolving landscape of software development and our commitment to ensuring JOSS publications recognize genuine scholarly contributions.
