### 2026-01-18 0d04672
docs: update following new JOSS guidelines
