### 2026-01-18 e0c727f
docs: format fix
