### 2026-01-18 #9152 @HarrisonKramer
Hi @kyleniemeyer,

I have updated the paper based on the new guidelines. It is ready for review.

Let me know if anything else is needed.
