### 2026-01-20 36c4fef
fix: units in tests related to phase interaction model
