### 2026-01-22 bcf60b5
feat: add new Buchdahl AbbeMaterial model
