+++ optiland/materials/abbe.py
@@ -1,79 +1,61 @@
 """Abbe Material
 
 This module defines a material based on the refractive index at the Fraunhofer
-d-line (587.56 nm) and the Abbe number. The refractive index is based on a
-polynomial fit to glass data from the Schott catalog. The extinction
-coefficient is ignored in this model and is always set to zero.
+d-line (587.56 nm) and the Abbe number. It provides two models for calculating
+the refractive index:
+1.  "polynomial": A polynomial fit to glass data (Legacy).
+2.  "buchdahl": A Buchdahl 3-term model with LASSO-derived coefficients (New).
 
 Kramer Harrison, 2024
 """
 
 from __future__ import annotations
 
+import warnings
+from abc import ABC, abstractmethod
 from importlib import resources
 
 import optiland.backend as be
 from optiland.materials.base import BaseMaterial
 
 
-class AbbeMaterial(BaseMaterial):
-    """Represents a material based on the refractive index at the Fraunhofer
-    d-line (587.56 nm) and the Abbe number. The refractive index is based on
-    a polynomial fit to glass data from the Schott catalog. The extinction
-    coefficient is ignored in this model and is always set to zero.
+class AbbeModel(ABC):
+    """Abstract base class for Abbe number based material models."""
 
-    Attributes:
-        index (float): The refractive index of the material at 587.56 nm.
-        abbe (float): The Abbey number of the material.
+    @abstractmethod
+    def predict_n(self, wavelength: float | be.ndarray) -> float | be.ndarray:
+        """Predicts the refractive index at a given wavelength."""
+        pass
 
+    @abstractmethod
+    def predict_k(self, wavelength: float | be.ndarray) -> float | be.ndarray:
+        """Predicts the extinction coefficient at a given wavelength."""
+        pass
+
+
+class AbbePolynomialModel(AbbeModel):
+    """Legacy polynomial model for Abbe materials.
+
+    This model uses a polynomial fit to glass data from the Schott catalog.
+    It corresponds to the original implementation of AbbeMaterial.
     """
 
-    def __init__(self, n, abbe):
-        super().__init__()
-        self.index = be.array([n])
+    def __init__(self, index: float, abbe: float):
+        self.index = be.array([index])
         self.abbe = be.array([abbe])
         self._p = self._get_coefficients()
 
-    def _calculate_n(self, wavelength, **kwargs):
-        """Returns the refractive index of the material.
-
-        Args:
-            wavelength (float or be.ndarray): The wavelength(s) of light in microns.
-
-        Returns:
-            be.ndarray: The refractive index of the material at the given
-            wavelength(s).
-
-        """
+    def predict_n(self, wavelength: float | be.ndarray) -> float | be.ndarray:
         wavelength = be.array(wavelength)
         if be.any(wavelength < 0.380) or be.any(wavelength > 0.750):
+            # This legacy check is preserved
             raise ValueError("Wavelength out of range for this model.")
         return be.atleast_1d(be.polyval(self._p, wavelength))
 
-    def _calculate_k(self, wavelength, **kwargs):
-        """Returns the extinction coefficient of the material.
-
-        Args:
-            wavelength (float or be.ndarray): The wavelength(s) of light in microns.
-
-        Returns:
-            float or be.ndarray: The extinction coefficient of the material, which
-            is always 0 for this model. Returns a scalar 0 if wavelength is scalar,
[…]
