### 2026-01-22 c479710
feat: add abbe material for e-line def
