+++ tests/test_abbe_new.py
@@ -1,8 +1,8 @@
 
 import pytest
 import optiland.backend as be
-from optiland.materials import AbbeMaterial
-from optiland.materials.abbe import AbbePolynomialModel, AbbeBuchdahlModel
+from optiland.materials import AbbeMaterial, AbbeMaterialE
+from optiland.materials.abbe import AbbePolynomialModel, BuchdahlDModel, BuchdahlEModel
 from .utils import assert_allclose
 
 class TestAbbePolynomialModel:
@@ -22,29 +22,48 @@ class TestAbbePolynomialModel:
         model = AbbePolynomialModel(1.5, 64.17)
         assert model.predict_k(0.5) == 0
 
-class TestAbbeBuchdahlModel:
+class TestBuchdahlDModel:
     def test_initialization(self, set_test_backend):
-        model = AbbeBuchdahlModel(index=1.5, abbe=50.0)
-        assert model.nd == 1.5
-        assert model.vd == 50.0
+        model = BuchdahlDModel(index=1.5, abbe=50.0)
+        assert model.index == 1.5
+        assert model.abbe == 50.0
         # Coefficients should be calculated
         assert model.v1 is not None
         assert model.v2 is not None
         assert model.v3 is not None
+        assert model.WAVE_REF == 0.5875618
 
     def test_predict_n_at_d_line(self, set_test_backend):
         # By definition, n(lambda_d) should be close to nd
-        # omega at lambda_d is 0
-        # n = nd + v1*0 + ... = nd
         nd_input = 1.5168
         vd_input = 64.17
-        model = AbbeBuchdahlModel(nd_input, vd_input)
+        model = BuchdahlDModel(nd_input, vd_input)
         n_pred = model.predict_n(0.5875618) # lambda_d
         assert_allclose(n_pred, nd_input, atol=1e-6)
 
     def test_predict_k(self, set_test_backend):
-        model = AbbeBuchdahlModel(1.5, 64.17)
+        model = BuchdahlDModel(1.5, 64.17)
         assert model.predict_k(0.5) == 0
+
+class TestBuchdahlEModel:
+    def test_initialization(self, set_test_backend):
+        # Test values roughly corresponding to N-BK7 e-line
+        # N-BK7: nd=1.5168, vd=64.17
+        # ne ~ 1.5187, ve ~ 63.96 (approximate conversion for testing init)
+        ne = 1.51872
+        ve = 63.96
+        model = BuchdahlEModel(index=ne, abbe=ve)
+        assert model.index == ne
+        assert model.abbe == ve
+        assert model.WAVE_REF == 0.546074
+        assert model.v1 is not None
+
+    def test_predict_n_at_e_line(self, set_test_backend):
+        ne = 1.51872
+        ve = 63.96
+        model = BuchdahlEModel(ne, ve)
+        n_pred = model.predict_n(0.546074) # lambda_e
+        assert_allclose(n_pred, ne, atol=1e-6)
 
 class TestAbbeMaterialWrapper:
     def test_default_legacy_model(self, set_test_backend):
@@ -65,7 +84,6 @@ class TestAbbeMaterialWrapper:
             assert len(future_warnings) == 0
         assert isinstance(mat.model, AbbePolynomialModel)
 
-
     def test_buchdahl_model(self, set_test_backend):
         # Should NOT raise warning
         import warnings
@@ -75,7 +93,7 @@ class TestAbbeMaterialWrapper:
             # Check that no FutureWarning was raised related to AbbeMaterial
             future_warnings = [w for w in record if issubclass(w.category, FutureWarning) and "AbbeMaterial" in str(w.message)]
[…]
