+++ optiland/materials/abbe.py
@@ -1,10 +1,12 @@
 """Abbe Material
 
-This module defines a material based on the refractive index at the Fraunhofer
-d-line (587.56 nm) and the Abbe number. It provides two models for calculating
-the refractive index:
-1.  "polynomial": A polynomial fit to glass data (Legacy).
-2.  "buchdahl": A Buchdahl 3-term model with LASSO-derived coefficients (New).
+This module defines a material based on the refractive index and Abbe number.
+It provides support for both d-line (587.56 nm) and e-line (546.07 nm)
+definitions.
+
+Models:
+1.  "polynomial": A polynomial fit to glass data (Legacy, D-line only).
+2.  "buchdahl": A Buchdahl 3-term model with LASSO-derived coefficients.
 
 Kramer Harrison, 2024
 """
@@ -34,7 +36,7 @@ class AbbeModel(ABC):
 
 
 class AbbePolynomialModel(AbbeModel):
-    """Legacy polynomial model for Abbe materials.
+    """Legacy polynomial model for Abbe materials (d-line).
 
     This model uses a polynomial fit to glass data from the Schott catalog.
     It corresponds to the original implementation of AbbeMaterial.
@@ -79,25 +81,52 @@ class AbbePolynomialModel(AbbeModel):
         return be.matmul(X_poly, coefficients)
 
 
-class AbbeBuchdahlModel(AbbeModel):
-    """Buchdahl 3-term model with LASSO-derived coefficients.
+class BuchdahlModel(AbbeModel):
+    """Base class for Buchdahl 3-term models."""
 
-    This model predicts refractive index using the Buchdahl coordinate $\omega$:
-    $$ \omega = \frac{\lambda - \lambda_D}{1 + \alpha(\lambda - \lambda_D)} $$
-    $$ n(\lambda) = n_d + v_1 \omega + v_2 \omega^2 + v_3 \omega^3 $$
-
-    where $\lambda_D = 0.5875618$ $\mu m$ and $\alpha = 2.5$.
-    The coefficients $v_1, v_2, v_3$ are predicted from $n_d$ and $V_d$ using a
-    linear regression model derived via LASSO.
-    """
-
-    # Buchdahl constants
-    WAVE_D = 0.5875618
     ALPHA = 2.5
 
+    def __init__(self, index: float, abbe: float):
+        self.index = float(index)
+        self.abbe = float(abbe)
+        self.v1, self.v2, self.v3 = self._calculate_buchdahl_coefficients()
+
+    @property
+    @abstractmethod
+    def WAVE_REF(self) -> float:
+        """Reference wavelength in microns."""
+        pass
+
+    @abstractmethod
+    def _calculate_buchdahl_coefficients(self) -> tuple[float, float, float]:
+        """Calculates v1, v2, v3 based on index and abbe."""
+        pass
+
+    def predict_n(self, wavelength: float | be.ndarray) -> float | be.ndarray:
+        wavelength = be.array(wavelength)
+
+        # Calculate Buchdahl coordinate omega
+        # omega = (lambda - lambda_d) / (1 + alpha * (lambda - lambda_d))
+        d_lambda = wavelength - self.WAVE_REF
+        omega = d_lambda / (1 + self.ALPHA * d_lambda)
+
+        # Buchdahl polynomial: n = nd + v1*w + v2*w^2 + v3*w^3
+        n_pred = (
+            self.index + self.v1 * omega + self.v2 * (omega**2) + self.v3 * (omega**3)
+        )
+
+        return be.atleast_1d(n_pred)
+
+    def predict_k(self, wavelength: float | be.ndarray) -> float | be.ndarray:
+        return be.zeros_like(0)
+
+
+class BuchdahlDModel(BuchdahlModel):
[…]
