### 2026-01-22 7190c5c
fix: update inputs to be be arrays
