### 2026-01-22 fdf28d5
Feat/abbe material (#439)
* feat: add new Buchdahl AbbeMaterial model

* docs: update Abbe material docs

* feat: add abbe material for e-line def

* docs: update notebook to demonstrate the new abbe models

* fix: update inputs to be be arrays
