+++ tests/visualization/test_visualization.py
@@ -492,7 +492,7 @@ class TestLensInfoViewer:
 
     def test_view_abbe_material(self, set_test_backend):
         lens = ReverseTelephoto()
-        lens.surface_group.surfaces[2].material_post = AbbeMaterial(1.5, 60)
+        lens.surface_group.surfaces[2].material_post = AbbeMaterial(1.5, 60, model="polynomial")
         viewer = LensInfoViewer(lens)
         viewer.view()
 
