### 2026-01-25 5dcd516
docs: add Jie Cai to authors list
