+++ docs/authors.rst
@@ -28,7 +28,7 @@ We sincerely appreciate the contributions of the following individuals, whose ef
 - **Robert Moerland** (`GitHub <https://github.com/rjmoerland>`_)
 - **Littie28** (`GitHub <https://github.com/Littie28>`_)
 - **Hsoj** (`GitHub <https://github.com/Hhsoj>`_)
-
+- **Jie Cai** (`GitHub <https://github.com/zhazhajust>`_)
 
 Your contributions, whether in the form of code, documentation, feedback, or discussions, are what make **Optiland** better for everyone.
 
