### 2026-01-25 6b3fd2f
feat: update OPD calculation and tests
