+++ optiland/propagation/grin.py
@@ -56,40 +56,30 @@ class GRINPropagation(BasePropagationModel):
         # For now, use the provided t as a target distance and integrate
         # through the GRIN medium.
 
-        len(rays.x)
-
         # Store initial state for each ray
         z_initial = be.copy(rays.z)
 
         # Target z position (exit surface)
         z_target = z_initial + t
 
-        # Get wavelength
-
         # Maximum iterations to prevent infinite loops
         max_iterations = 10000
+
+        # Pre-calculate bounds for all rays
+        z_min_all = be.minimum(z_initial, z_target)
+        z_max_all = be.maximum(z_initial, z_target)
 
         for _iteration in range(max_iterations):
             # Check which rays still need to propagate
             # Rays are active if they are within the medium bounds
             # Allow rays to propagate in both directions (N can be positive or negative)
-            z_min = be.minimum(z_initial, z_target)
-            z_max = be.maximum(z_initial, z_target)
-            active = (rays.z > z_min - 1e-6) & (rays.z < z_max - 1e-6)
+            active = (rays.z > z_min_all - 1e-6) & (rays.z < z_max_all + 1e-6)
 
             if not be.any(active):
                 break
 
             # RK4 integration step for active rays
-            if be.any(active):
-                self._rk4_step(rays, active, z_min, z_max)
-
-        # Update OPD (optical path difference)
-        # For GRIN, this should be accumulated during propagation
-        # For simplicity, we approximate it here
-        # Use _calculate_n to avoid caching issues with array kwargs
-        n_final = self.material._calculate_n(rays.w, x=rays.x, y=rays.y, z=rays.z)
-        rays.opd = rays.opd + be.abs(t * n_final)
+            self._rk4_step(rays, active, z_min_all[active], z_max_all[active])
 
         # Normalize direction vectors
         rays.normalize()
@@ -100,8 +90,8 @@ class GRINPropagation(BasePropagationModel):
         Args:
             rays: The rays object (modified in-place).
             active: Boolean array indicating which rays are active.
-            z_min: Minimum z boundary (entrance surface).
-            z_max: Maximum z boundary (exit surface).
+            z_min: Minimum z boundary (entrance surface) for active rays.
+            z_max: Maximum z boundary (exit surface) for active rays.
 
         """
         # Get active ray positions and directions
@@ -114,34 +104,36 @@ class GRINPropagation(BasePropagationModel):
         w = rays.w[active]
 
         # Adaptive step size based on distance to boundaries
-        # Limit step size to prevent overshooting boundaries
-        if be.any(N > 0):
-            # Rays propagating toward z_max
-            remaining_forward = (z_max - z) * (N > 0)
-            step_size_forward = be.minimum(
-                self.step_size, be.abs(remaining_forward * 0.9)
-            )
-        else:
-            step_size_forward = be.full_like(z, self.step_size)
+        # Initialize with default step size
+        step_size = be.full_like(z, self.step_size)
 
-        if be.any(N < 0):
-            # Rays propagating toward z_min
-            remaining_backward = (z - z_min) * (N < 0)
-            step_size_backward = be.minimum(
-                self.step_size, be.abs(remaining_backward * 0.9)
[…]
