+++ .history/optiland/optic/optic_20251222222552.py
@@ -1,729 +1,0 @@
-"""Optic Module
-
-This is the core module of Optiland, which provides the class to define
-generic optical systems. The Optic class encapsulates the core properties
-of an optical system, such as the aperture, fields, surfaces, and
-wavelengths. It also provides methods to draw the optical system, trace rays,
-and perform paraxial and aberration analyses. Instances of the Optic class
-are used as arguments to various analysis, optimization, and visualization
-functions in Optiland.
-
-Kramer Harrison, 2024
-"""
-
-from __future__ import annotations
-
-from copy import deepcopy
-from typing import TYPE_CHECKING, Any, Literal
-
-from optiland.aberrations import Aberrations
-from optiland.aperture import Aperture
-from optiland.apodization import BaseApodization
-from optiland.fields import (
-    AngleField,
-    BaseFieldDefinition,
-    Field,
-    FieldGroup,
-    ObjectHeightField,
-    ParaxialImageHeightField,
-    RealImageHeightField,
-)
-from optiland.optic.optic_updater import OpticUpdater
-from optiland.paraxial import Paraxial
-from optiland.pickup import PickupManager
-from optiland.rays import PolarizationState
-from optiland.raytrace.real_ray_tracer import RealRayTracer
-from optiland.solves import SolveManager
-from optiland.surfaces import ObjectSurface, SurfaceGroup
-from optiland.visualization import (
-    LensInfoViewer,
-    OpticViewer,
-    OpticViewer3D,
-    SurfaceSagViewer,
-)
-from optiland.wavelength import WavelengthGroup
-
-if TYPE_CHECKING:
-    from matplotlib.axes import Axes
-    from matplotlib.figure import Figure
-
-    from optiland._types import (
-        ApertureType,
-        BEArray,
-        DistributionType,
-        FieldType,
-        ReferenceRay,
-        ScalarOrArray,
-        SurfaceParameters,
-        SurfaceType,
-        Unpack,
-        WavelengthUnit,
-    )
-    from optiland.distribution import BaseDistribution
-    from optiland.materials.base import BaseMaterial
-    from optiland.rays import RealRays
-    from optiland.surfaces.standard_surface import Surface
-
-
-class Optic:
-    """A class for defining and manipulating optical systems.
-
-    The Optic class is central to the Optiland library, providing a comprehensive
-    framework for representing optical systems. It encapsulates essential
-    properties such as aperture, fields of view, constituent surfaces, and
-    operating wavelengths. This class also offers a suite of methods for
-    visualizing the optical system, performing ray tracing, and conducting
-    paraxial and aberration analyses. Instances of Optic are fundamental objects
-    passed to various analysis, optimization, and visualization functions
-    throughout Optiland.
-
-    Args:
-        name (str, optional): An optional name for the optical system.
-            Defaults to None.
-
-    Attributes:
-        name (str | None): An optional name for the optical system.
-        aperture (Aperture | None): The aperture of the optical system.
-        field_definition (BaseFieldDefinition | None): The definition of the field used
-            in the optical system, e.g., AngleField or ObjectHeightField.
-        surface_group (SurfaceGroup): The group of surfaces that constitute
-            the optical system.
-        fields (FieldGroup): The group of fields defined for the system.
-        wavelengths (WavelengthGroup): The group of wavelengths used for
-            analysis.
-        paraxial (Paraxial): A helper class for paraxial analysis of the
-            optical system.
-        aberrations (Aberrations): A helper class for analyzing aberrations.
-        ray_tracer (RealRayTracer): The ray tracer for performing real ray
-            traces.
-        polarization (PolarizationState | Literal['ignore']): The polarization
-            state of the light. Defaults to 'ignore'.
[…]
