### 2026-01-28 cbd404c
test: update tests with new method signature
