### 2026-01-28 6644bf4
Fix Optic.plot_surface_sag kwargs forwarding (#424)
* Fix Optic.plot_surface_sag kwargs forwarding

Accept and forward fig_to_plot_on, max_extent, num_points_grid, and buffer_factor to SurfaceSagViewer.

This restores compatibility with gallery notebooks that call plot_surface_sag(..., max_extent=..., num_points_grid=...).

* chore: delete obsolete files

* test: update tests with new method signature

---------

Co-authored-by: Kramer <kdanielharrison@gmail.com>
