### 2026-01-30 2f73719
feat(optic): add a method remove_surface in optic.Optic which calls surface_group.remove_surface
