### 2026-01-30 ffc1df2
correct the type check of index
