### 2026-01-30 3ab4baf
test(optic): add test for remove_surface
