+++ tests/test_optic.py
@@ -522,7 +522,14 @@ class TestOptic:
         lens.flip()
         assert lens.pickups.pickups[0].source_surface_idx == 2
         assert lens.pickups.pickups[0].target_surface_idx == 1
-
+    
+    def test_remove_surface(self, set_test_backend):
+        lens = singlet_infinite_object()
+        
+        num_surfaces_before = len(lens.surface_group.surfaces)
+        lens.remove_surface(index=2)
+        assert len(lens.surface_group.surfaces) == num_surfaces_before - 1
+    
     @patch("optiland.optic.optic.SurfaceSagViewer")
     def test_plot_surface_sag(self, mock_viewer, set_test_backend):
         lens = singlet_infinite_object()
