### 2026-01-30 c1c84f1
Add circular aperture zemax handler
Some surfaces in Zemax have circular apertures. This commit add a read
circular aperture method for the Zemax file parser.
