+++ optiland/fileio/zemax_handler.py
@@ -16,11 +16,11 @@ import tempfile
 from dataclasses import dataclass, field
 from typing import Any
 
-import requests
-
 import optiland.backend as be
+import requests
 from optiland.fileio.converters import ZemaxToOpticConverter
 from optiland.materials import AbbeMaterial, BaseMaterial, Material
+from optiland.physical_apertures import RadialAperture
 
 
 def load_zemax_file(source: str):
@@ -147,6 +147,7 @@ class ZemaxDataParser:
             "VCXN": self._read_vignette_compress_x,
             "VCYN": self._read_vignette_compress_y,
             "VANN": self._read_vignette_tangent_angle,
+            "CLAP": self._read_circular_aperture,
         }
 
     def parse(self) -> ZemaxDataModel:
@@ -239,6 +240,7 @@ class ZemaxDataParser:
             "is_stop": False,
             "conic": 0.0,
             "material": "air",
+            "aperture": None,
         }
 
     def _read_radius(self, data):
@@ -309,7 +311,9 @@ class ZemaxDataParser:
             "ODDASPHE": "odd_asphere",
             "COORDBRK": "coordinate_break",
             "TOROIDAL": "toroidal",
-        }.get(data[1], data[1].lower())  # Default to lowercased raw value
+        }.get(
+            data[1], data[1].lower()
+        )  # Default to lowercased raw value
 
     def _read_surface_parameter(self, data):
         key = f"param_{int(data[1]) - 1}"
@@ -348,6 +352,11 @@ class ZemaxDataParser:
         self.data_model.fields["vignette_tangent_angle"] = [
             float(v) for v in data[1 : n + 1]
         ]
+
+    def _read_circular_aperture(self, data):
+        self._current_surf_data["aperture"] = RadialAperture(
+            r_min=float(data[1]), r_max=float(data[2])
+        )
 
     # ------------------ Helpers ------------------
     def _finalize_fields(self):
