+++ tests/test_analysis_extended.py
@@ -1,0 +1,108 @@
+
+import pytest
+from unittest.mock import MagicMock, patch
+import numpy as np
+import optiland.backend as be
+from optiland.analysis import angle_vs_height, image_simulation, jones_pupil, through_focus_mtf
+from optiland.samples.objectives import CookeTriplet
+from optiland.optic import Optic
+
+@pytest.fixture
+def cooke_triplet():
+    return CookeTriplet()
+
+# -----------------------------------------------------------------------------
+# Image Simulation Tests (Mocked)
+# -----------------------------------------------------------------------------
+
+def test_simulator_init(cooke_triplet, set_test_backend):
+    # ImageSimulator likely takes source_image_path or similar
+    # Constructor signature: (optic, source_image_path, ...)
+    # Let's mock the path or pass a dummy string
+    sim = image_simulation.ImageSimulator(cooke_triplet, "dummy_path.png")
+    assert sim.optic == cooke_triplet
+
+def test_distortion_warper_init(cooke_triplet, set_test_backend):
+    warper = image_simulation.DistortionWarper(cooke_triplet)
+    assert warper.optic == cooke_triplet
+
+def test_psf_basis_generator(cooke_triplet, set_test_backend):
+    # PSFBasisGenerator likely takes (optic, num_psfs=...) or just (optic)
+    # Check signature: (optic, kernel_size, num_psfs) perhaps?
+    # Error was: unexpected keyword 'num_psfs'.
+    # Maybe it's positional or named differently.
+    # Let's try default init.
+    gen = image_simulation.PSFBasisGenerator(cooke_triplet)
+    assert gen.optic == cooke_triplet
+
+# -----------------------------------------------------------------------------
+# Jones Pupil Tests
+# -----------------------------------------------------------------------------
+
+def test_jones_pupil_init(cooke_triplet, set_test_backend):
+    # Mocking trace_generic to avoid backend issues during init (which calls _generate_data)
+    with patch.object(cooke_triplet, 'trace_generic') as mock_trace:
+        # Mock rays return
+        mock_rays = MagicMock()
+        mock_rays.x = be.array([0.0])
+        # Need to ensure polarization matrix p is present if expected
+        mock_rays.p = be.zeros((1, 3, 3))
+        if be.get_backend() == "torch":
+             # Fix for AttributeError: 'numpy.ndarray' object has no attribute 'clone'
+             # If backend is torch, ensure we return tensors
+             import torch
+             mock_rays.p = torch.zeros((1, 3, 3), dtype=torch.float64)
+             mock_rays.x = torch.zeros(1, dtype=torch.float64)
+
+        mock_trace.return_value = mock_rays
+
+        jp = jones_pupil.JonesPupil(cooke_triplet)
+        assert jp.optic == cooke_triplet
+
+def test_jones_pupil_generate_data(cooke_triplet, set_test_backend):
+    with patch.object(cooke_triplet, 'trace_generic') as mock_trace:
+        mock_rays = MagicMock()
+        mock_rays.x = be.array([0.0])
+        mock_rays.p = be.zeros((1, 3, 3))
+        if be.get_backend() == "torch":
+             import torch
+             mock_rays.p = torch.zeros((1, 3, 3), dtype=torch.float64)
+             mock_rays.x = torch.zeros(1, dtype=torch.float64)
+        mock_trace.return_value = mock_rays
+
+        jp = jones_pupil.JonesPupil(cooke_triplet)
+        assert jp.optic == cooke_triplet
+
+# -----------------------------------------------------------------------------
+# Through Focus MTF Tests
+# -----------------------------------------------------------------------------
+
+def test_through_focus_mtf_init(cooke_triplet, set_test_backend):
[…]
