+++ tests/test_geometries_extended.py
@@ -1,0 +1,157 @@
+
+import pytest
+from unittest.mock import MagicMock
+import numpy as np
+import optiland.backend as be
+from optiland.coordinate_system import CoordinateSystem
+from optiland.geometries.standard_grating import StandardGratingGeometry
+from optiland.geometries.plane_grating import PlaneGrating
+from optiland.rays import RealRays
+from optiland.geometries.forbes import jacobi
+from optiland.geometries.nurbs import nurbs_basis_functions
+
+def test_standard_grating_init(set_test_backend):
+    cs = CoordinateSystem()
+    g = StandardGratingGeometry(cs, radius=100.0, grating_order=1, grating_period=1.0, groove_orientation_angle=0.0, conic=0.0)
+    assert g.radius == 100.0
+    assert g.grating_order == 1
+    assert g.grating_period == 1.0
+
+def test_standard_grating_sag(set_test_backend):
+    cs = CoordinateSystem()
+    g = StandardGratingGeometry(cs, radius=100.0, grating_order=1, grating_period=1.0, groove_orientation_angle=0.0)
+    assert g.sag(0, 0) == 0.0
+    s = g.sag(10, 0)
+    assert abs(s - 0.5) < 0.1
+
+def test_standard_grating_serialization(set_test_backend):
+    cs = CoordinateSystem()
+    g = StandardGratingGeometry(cs, radius=100.0, grating_order=1, grating_period=2.0, groove_orientation_angle=0.5, conic=-1.0)
+    d = g.to_dict()
+
+    g2 = StandardGratingGeometry.from_dict(d)
+    assert g2.radius == 100.0
+    assert g2.grating_order == 1
+    assert g2.grating_period == 2.0
+    assert g2.groove_orientation_angle == 0.5
+    assert g2.k == -1.0
+
+def test_plane_grating_init(set_test_backend):
+    cs = CoordinateSystem()
+    g = PlaneGrating(cs, grating_order=1, grating_period=1.0, groove_orientation_angle=0.0)
+    assert g.grating_order == 1
+
+def test_plane_grating_sag(set_test_backend):
+    cs = CoordinateSystem()
+    g = PlaneGrating(cs, grating_order=1, grating_period=1.0, groove_orientation_angle=0.0)
+    assert g.sag(10, 10) == 0.0
+
+def test_plane_grating_serialization(set_test_backend):
+    cs = CoordinateSystem()
+    g = PlaneGrating(cs, grating_order=1, grating_period=2.0, groove_orientation_angle=0.5)
+    d = g.to_dict()
+
+    g2 = PlaneGrating.from_dict(d)
+    assert g2.grating_order == 1
+    assert g2.grating_period == 2.0
+    assert g2.groove_orientation_angle == 0.5
+
+def test_standard_grating_distance(set_test_backend):
+    cs = CoordinateSystem()
+    g = StandardGratingGeometry(cs, radius=100.0, grating_order=1, grating_period=1.0, groove_orientation_angle=0.0)
+
+    rays = RealRays(x=be.array([0.0]), y=be.array([0.0]), z=be.array([0.0]),
+                    L=be.array([0.0]), M=be.array([0.0]), N=be.array([1.0]),
+                    wavelength=0.55, intensity=1.0)
+    dist = g.distance(rays)
+    val = dist[0]
[…]
