### 2026-01-31 0d229db
Improve test coverage for core modules to >90%
- Added `tests/test_backend_functions.py` and `tests/test_backend_utils.py` for comprehensive backend testing.
- Added `tests/test_geometries_extended.py` to cover gratings, Jacobis, and NURBS.
- Added `tests/test_materials_extended.py` for Abbe material models and utilities.
- Added `tests/test_optimization_extended.py` for optimization variables.
- Added `tests/test_analysis_extended.py` for analysis components (ImageSim, JonesPupil, MTF).
- Added `tests/visualization/test_visualization_extended.py` for visualization helpers.
- Fixed bugs in `St
[…]
