### 2026-01-31 50a6ab3
Add optiland.reporting module for publication-quality reports
Implemented a comprehensive reporting module with:
- Metrics calculation (first-order, image quality, wavefront, aberrations).
- Visualization engine using Matplotlib with publication-quality styling.
- PDF report generation with dashboard and detailed analysis sections.
- Engine to orchestrate report generation.
- Unit tests for the new module.

Co-authored-by: HarrisonKramer <29144587+HarrisonKramer@users.noreply.github.com>
