+++ optiland/reporting/engine.py
@@ -1,0 +1,248 @@
+"""
+Reporting Engine Module.
+
+This module contains the core logic that assembles a report.
+"""
+
+from __future__ import annotations
+
+import abc
+from typing import TYPE_CHECKING
+
+import matplotlib.pyplot as plt
+import numpy as np
+
+import optiland.backend as be
+from optiland.physical_apertures.radial import RadialAperture
+from optiland.reporting.metrics import (
+    FirstOrderMetrics,
+    MaterialMetrics,
+    SeidelMetrics,
+)
+from optiland.reporting.pdf_export import PDFReportRenderer
+from optiland.reporting.visualizations import (
+    FieldCurvatureDistortionPlot,
+    LensLayoutPlot,
+    MTFPlot,
+    OPDPlot,
+    PlotStyler,
+    RayFanPlot,
+    SpotDiagramPlot,
+)
+
+if TYPE_CHECKING:
+    from optiland.optic.optic import Optic
+
+
+class ReportBuilder(abc.ABC):
+    """Abstract base class for report builders."""
+
+    def __init__(self, optic: Optic):
+        self.optic = optic
+
+    @abc.abstractmethod
+    def save(self, filename: str):
+        """Builds and saves the report to a file."""
+        pass
+
+
+class StandardPerformanceReport(ReportBuilder):
+    """Generates a standard performance report."""
+
+    def save(self, filename: str):
+        PlotStyler.apply_style()
+
+        with PDFReportRenderer(filename) as pdf:
+            # 1. Dashboard
+            self._create_dashboard(pdf)
+
+            # 2. System Prescription
+            self._create_prescription_table(pdf)
+
+            # 3. First Order Properties (Detailed)
+            self._create_first_order_table(pdf)
+
+            # 4. Image Quality Assessment
+            self._create_image_quality_section(pdf)
+
+            # 5. Aberration Analysis
+            self._create_aberration_analysis(pdf)
+
+    def _create_dashboard(self, pdf: PDFReportRenderer):
+        """Creates the dashboard page."""
+        # 1. Get Lens Layout
+        layout_plot = LensLayoutPlot(self.optic)
+        fig = layout_plot.plot()
+        fig.set_size_inches(8.5, 11)  # Portrait
+
+        # Adjust existing axes to make room for table at bottom
+        if fig.axes:
+            ax_layout = fig.axes[0]
+            pos = ax_layout.get_position()
+            # Move up and shrink height: [left, bottom, width, height]
+            # Set bottom to 0.4 (leave 40% for table)
+            ax_layout.set_position([pos.x0, 0.4, pos.width, 0.55])
+            ax_layout.set_title("System Layout", fontsize=14, weight="bold")
+
+        # 2. Calculate Key Metrics for Summary Table
+        fo_metrics = FirstOrderMetrics.calculate(self.optic)
+
+        summary_data = [
+            ["EFL", f"{fo_metrics['EFL']:.4f} mm"],
+            ["F/#", f"{fo_metrics['F/#']:.4f}"],
+            ["EPD", f"{fo_metrics['EPD']:.4f} mm"],
+            ["Field of View", self._get_fov_string()],
+            ["Total Track", f"{fo_metrics['Total Track']:.4f} mm"],
+            [
+                "Wavelengths",
+                f"{self.optic.wavelengths.primary_wavelength.value:.4f} µm (Primary)",
+            ],
+        ]
+
+        # Add table
+        ax_table = fig.add_axes([0.1, 0.05, 0.8, 0.3])  # Bottom 30%
+        ax_table.axis("off")
+        ax_table.set_title("Performance Summary", fontsize=12, weight="bold")
+
+        table = ax_table.table(
+            cellText=summary_data,
+            colLabels=["Metric", "Value"],
+            loc="center",
+            cellLoc="center",
[…]
