+++ optiland/reporting/metrics.py
@@ -1,0 +1,257 @@
+"""
+Metrics Module for Optiland Reporting.
+
+This module provides stateless service classes responsible for calculating
+numerical performance data for optical systems. It adheres to SOLID principles,
+separating calculation logic from visualization.
+"""
+
+from __future__ import annotations
+
+import contextlib
+from typing import TYPE_CHECKING, Any
+
+import optiland.backend as be
+from optiland.analysis.spot_diagram import SpotDiagram
+from optiland.wavefront.wavefront import Wavefront
+
+if TYPE_CHECKING:
+    from optiland.optic import Optic
+
+
+class FirstOrderMetrics:
+    """Calculates first-order properties of the optical system."""
+
+    @staticmethod
+    def calculate(optic: Optic) -> dict[str, Any]:
+        """Calculates all first-order metrics.
+
+        Args:
+            optic: The optical system to analyze.
+
+        Returns:
+            A dictionary containing EFL, F/#, EPD, Magnification, etc.
+        """
+        paraxial = optic.paraxial
+        metrics = {}
+
+        try:
+            metrics["EFL"] = float(be.to_numpy(paraxial.f2()))
+        except Exception:
+            metrics["EFL"] = float("nan")
+
+        try:
+            metrics["F/#"] = float(be.to_numpy(paraxial.FNO()))
+        except Exception:
+            metrics["F/#"] = float("nan")
+
+        try:
+            metrics["EPD"] = float(be.to_numpy(paraxial.EPD()))
+        except Exception:
+            metrics["EPD"] = float("nan")
+
+        try:
+            metrics["EPL"] = float(be.to_numpy(paraxial.EPL()))
+        except Exception:
+            metrics["EPL"] = float("nan")
+
+        try:
+            metrics["XPD"] = float(be.to_numpy(paraxial.XPD()))
+        except Exception:
+            metrics["XPD"] = float("nan")
+
+        try:
+            metrics["XPL"] = float(be.to_numpy(paraxial.XPL()))
+        except Exception:
+            metrics["XPL"] = float("nan")
+
+        try:
+            metrics["Magnification"] = float(
+                be.to_numpy(paraxial.magnification())
+            )
+        except Exception:
+            metrics["Magnification"] = float("nan")
+
+        metrics["Total Track"] = float(optic.total_track)
+
+        return metrics
+
+
+class ImageQualityMetrics:
+    """Calculates image quality metrics based on spot diagrams."""
+
+    @staticmethod
+    def calculate(
+        optic: Optic, fields: str | list = "all", wavelength="primary"
+    ) -> dict[str, Any]:
+        """Calculates spot diagram metrics.
+
+        Args:
+            optic: The optical system.
+            fields: Fields to analyze.
+            wavelength: Wavelength to analyze.
+
+        Returns:
+            Dictionary with RMS Radius, Geometric Radius, Centroid for each field.
+        """
+        wls = "primary" if wavelength == "primary" else [wavelength]
+        spot = SpotDiagram(optic, fields=fields, wavelengths=wls)
+        results = {}
+
+        rms_radii = spot.rms_spot_radius()
+        geo_radii = spot.geometric_spot_radius()
+        centroids = spot.centroid()
+
+        for i, field in enumerate(spot.fields):
+            # rms_radii is list of lists [field][wavelength]
+            # We assume single wavelength analysis for metrics report per call
+            # or we take the first one if multiple are implicitly present
+            rms = rms_radii[i][0]
+            geo = geo_radii[i][0]
+            cx, cy = centroids[i]
+
+            results[f"Field {i+1}"] = {
+                "Field Coordinate": field,
+                "RMS Radius": float(be.to_numpy(rms)),
+                "Geo Radius": float(be.to_numpy(geo)),
+                "Centroid X": float(be.to_numpy(cx)),
[…]
