### 2026-01-31 8157a23
Feat/add remove_surface method in optic.optic (#443)
* feat(optic): add a method remove_surface in optic.Optic which calls surface_group.remove_surface

* correct the type check of index

* test(optic): add test for remove_surface
