### 2026-01-31 786cf13
docs: remove obsolete comments
