+++ tests/test_analysis_extended.py
@@ -16,32 +16,8 @@ def cooke_triplet(set_test_backend):
 # -----------------------------------------------------------------------------
 
 def test_simulator_init(cooke_triplet, set_test_backend):
-    # ImageSimulator might be named differently or not exposed directly.
-    # Check imports. If module has 'simulator' submodule, use that.
-    # From file listing: optiland/analysis/image_simulation/simulator.py
-    # So it should be image_simulation.simulator.ImageSimulator if not exposed in __init__.
-    # Or maybe it's just 'Simulator'.
-
-    # Try instantiation via submodule if main package access fails,
-    # but based on previous error 'module ... has no attribute ImageSimulator',
-    # it seems it's not exported in __init__.py of image_simulation.
-
-    # Actually, check optiland/analysis/image_simulation/__init__.py or just use the module path.
-    # The file listing showed optiland/analysis/image_simulation/simulator.py
-    # But previous import error said "cannot import name 'ImageSimulator'".
-    # This implies the class name in the file might be different.
-    # Let's inspect optiland/analysis/image_simulation/simulator.py content first if possible, but I can't interactively.
-    # Assuming standard naming convention failed, maybe it's just 'Simulator'?
-    # Or maybe it's exposed in image_simulation module itself if imported correctly.
-
-    # Let's try to inspect the module dynamically or just guess based on standard practice.
-    # If file is simulator.py, class is likely ImageSimulator.
-    # If ImportError, maybe circular import or class not there?
-
-    # Let's try importing the module and checking attributes
     from optiland.analysis.image_simulation import simulator
     sim = simulator.SpatiallyVariableSimulator()
-    # SpatiallyVariableSimulator.__init__ currently takes no arguments and has no optic attribute
     assert isinstance(sim, simulator.SpatiallyVariableSimulator)
 
 def test_distortion_warper_init(cooke_triplet, set_test_backend):
@@ -49,8 +25,6 @@ def test_distortion_warper_init(cooke_triplet, set_test_backend):
     assert warper.optic == cooke_triplet
 
 def test_psf_basis_generator(cooke_triplet, set_test_backend):
-    # Error: missing 'wavelength' argument.
-    # So signature is likely (optic, wavelength, ...)
     gen = image_simulation.PSFBasisGenerator(cooke_triplet, wavelength=0.55)
     assert gen.optic == cooke_triplet
 
@@ -59,16 +33,11 @@ def test_psf_basis_generator(cooke_triplet, set_test_backend):
 # -----------------------------------------------------------------------------
 
 def test_jones_pupil_init(cooke_triplet, set_test_backend):
-    # Mocking trace_generic to avoid backend issues during init (which calls _generate_data)
     with patch.object(cooke_triplet, 'trace_generic') as mock_trace:
-        # Mock rays return
         mock_rays = MagicMock()
         mock_rays.x = be.array([0.0])
-        # Need to ensure polarization matrix p is present if expected
         mock_rays.p = be.zeros((1, 3, 3))
         if be.get_backend() == "torch":
-             # Fix for AttributeError: 'numpy.ndarray' object has no attribute 'clone'
-             # If backend is torch, ensure we return tensors
              import torch
              mock_rays.p = torch.zeros((1, 3, 3), dtype=torch.float64)
              mock_rays.x = torch.zeros(1, dtype=torch.float64)
@@ -97,15 +66,6 @@ def test_jones_pupil_generate_data(cooke_triplet, set_test_backend):
 # -----------------------------------------------------------------------------
 
 def test_through_focus_mtf_init(cooke_triplet, set_test_backend):
[…]
