+++ tests/test_materials_extended.py
@@ -1,0 +1,187 @@
+
+import pytest
+from unittest.mock import MagicMock, patch, mock_open
+import numpy as np
+import optiland.backend as be
+from optiland.materials import abbe, material_utils
+from optiland.materials.abbe import AbbeMaterial, AbbeMaterialE
+
+# -----------------------------------------------------------------------------
+# AbbeMaterial Tests (Buchdahl & E-line)
+# -----------------------------------------------------------------------------
+
+def test_abbe_buchdahl_init(set_test_backend):
+    # Test initialization with 'buchdahl' model
+    mat = AbbeMaterial(1.5, 64.0, model="buchdahl")
+    assert mat.model_name == "buchdahl"
+    assert isinstance(mat.model, abbe.BuchdahlDModel)
+
+    # Test calculations
+    n = mat.n(0.5876)
+    assert abs(n - 1.5) < 1e-3
+
+def test_abbe_invalid_model(set_test_backend):
+    with pytest.raises(ValueError, match="Unknown model"):
+        AbbeMaterial(1.5, 64.0, model="invalid")
+
+def test_abbe_e_init(set_test_backend):
+    # Test AbbeMaterialE
+    mat = AbbeMaterialE(1.5, 64.0)
+    assert isinstance(mat.model, abbe.BuchdahlEModel)
+
+    # Test calculations (check near reference wavelength)
+    n = mat.n(0.5461)
+    assert abs(n - 1.5) < 1e-3
+
+def test_abbe_to_from_dict(set_test_backend):
+    mat = AbbeMaterial(1.5, 64.0, model="buchdahl")
+    d = mat.to_dict()
+    assert d["model"] == "buchdahl"
+    assert d["index"] == 1.5
+
+    mat2 = AbbeMaterial.from_dict(d)
+    assert mat2.model_name == "buchdahl"
+    assert mat2.index.item() == 1.5
+
+def test_abbe_e_to_from_dict(set_test_backend):
+    mat = AbbeMaterialE(1.5, 64.0)
+    d = mat.to_dict()
+    assert d["index"] == 1.5
+
+    mat2 = AbbeMaterialE.from_dict(d)
+    assert mat2.index.item() == 1.5
+
+# -----------------------------------------------------------------------------
+# Material Utils Tests
+# -----------------------------------------------------------------------------
+
+@patch("optiland.materials.material_utils.resources.files")
+def test_glasses_selection(mock_files, set_test_backend):
+    # Mock CSV content
+    csv_content = (
+        "group,filename,filename_no_ext,min_wavelength,max_wavelength\n"
+        "glass,catalog/glass1,glass1,0.3,1.0\n"
+        "glass,catalog/glass2,glass2,0.4,0.8\n"
+        "plastic,catalog/plastic1,plastic1,0.3,1.0\n"
+    )
+
+    mock_path = MagicMock()
+    # Ensure the path behaves like a string or path-like object
+    mock_file_path = MagicMock()
+    mock_file_path.__str__.return_value = "dummy.csv"
+    mock_path.joinpath.return_value = mock_file_path
+    mock_files.return_value = mock_path
+
+    with patch("builtins.open", mock_open(read_data=csv_content)):
+        # Case 1: Broad range, only glass1 fits
+        selection = material_utils.glasses_selection(0.35, 0.95)
+        assert "glass1" in selection
+        assert "glass2" not in selection # 0.8 < 0.95
+
+        # Case 2: Narrow range
+        selection = material_utils.glasses_selection(0.5, 0.6)
+        assert "glass1" in selection
+        assert "glass2" in selection
+
[…]
