### 2026-02-02 915910f
Add ax parmeter to Optic.draw method.
Adding the option to pass a matplolib axes as argument. Just passing the
argument that is already on the view method.
