### 2026-02-02 1fbbef6
feat: Introduce reporting module with visualization, engine, metrics, and PDF export capabilities.
