+++ optiland/reporting/engine.py
@@ -22,7 +22,6 @@ from optiland.reporting.metrics import (
 from optiland.reporting.pdf_export import PDFReportRenderer
 from optiland.reporting.visualizations import (
     FieldCurvatureDistortionPlot,
-    LensLayoutPlot,
     MTFPlot,
     OPDPlot,
     PlotStyler,
@@ -53,7 +52,7 @@ class StandardPerformanceReport(ReportBuilder):
         PlotStyler.apply_style()
 
         with PDFReportRenderer(filename) as pdf:
-            # 1. Dashboard
+            # 1. Dashboard / Title Page
             self._create_dashboard(pdf)
 
             # 2. System Prescription
@@ -70,21 +69,66 @@ class StandardPerformanceReport(ReportBuilder):
 
     def _create_dashboard(self, pdf: PDFReportRenderer):
         """Creates the dashboard page."""
-        # 1. Get Lens Layout
-        layout_plot = LensLayoutPlot(self.optic)
-        fig = layout_plot.plot()
-        fig.set_size_inches(8.5, 11)  # Portrait
+        import os
 
-        # Adjust existing axes to make room for table at bottom
-        if fig.axes:
-            ax_layout = fig.axes[0]
-            pos = ax_layout.get_position()
-            # Move up and shrink height: [left, bottom, width, height]
-            # Set bottom to 0.4 (leave 40% for table)
-            ax_layout.set_position([pos.x0, 0.4, pos.width, 0.55])
-            ax_layout.set_title("System Layout", fontsize=14, weight="bold")
+        from matplotlib.offsetbox import AnnotationBbox, OffsetImage
 
-        # 2. Calculate Key Metrics for Summary Table
+        from optiland.visualization.system.optic_viewer import OpticViewer
+
+        fig = plt.figure(figsize=(8.5, 11))
+
+        # 1. Logo and Title
+        ax_header = fig.add_axes([0.05, 0.85, 0.9, 0.1])
+        ax_header.axis("off")
+
+        # Try to find logo
+        logo_path = None
+        possible_paths = [
+            "optiland_gui/resources/logo.png",
+            "../optiland_gui/resources/logo.png",
+            "../../optiland_gui/resources/logo.png",
+            os.path.join(
+                os.path.dirname(__file__), "../../optiland_gui/resources/logo.png"
+            ),
+        ]
+
+        for p in possible_paths:
+            if os.path.exists(p):
+                logo_path = p
+                break
+
+        if logo_path:
+            try:
+                img = plt.imread(logo_path)
+                imagebox = OffsetImage(img, zoom=0.15)
+                ab = AnnotationBbox(
+                    imagebox, (0.05, 0.5), frameon=False, boxcoords="axes fraction"
+                )
+                ax_header.add_artist(ab)
+            except Exception:
+                pass  # Fail silently on logo
+
+        # Title
+        ax_header.text(
+            0.5,
+            0.5,
+            "Optical System Performance Report",
+            ha="center",
+            va="center",
+            fontsize=24,
+            weight="bold",
+            color="#2c3e50",
+        )
+
+        # 2. Lens Layout (Center)
+        ax_layout = fig.add_axes([0.1, 0.45, 0.8, 0.35])
+
+        # Use OpticViewer directly to plot on specific axes
+        viewer = OpticViewer(self.optic)
+        viewer.view(projection="YZ", ax=ax_layout)
+        ax_layout.set_title("System Layout", fontsize=14, weight="bold")
+
+        # 3. Summary Table (Bottom)
         fo_metrics = FirstOrderMetrics.calculate(self.optic)
 
         summary_data = [
[…]
