### 2026-02-02 0199b00
feat: Implement scale method for all geometries and update optic updater to use this
