### 2026-02-03 84228b8
feat: add material formatting for lens info viewer (#448)
