### 2026-02-03 ca913f7
feat: implement scale for all subclasses
