+++ tests/test_scale_system.py
@@ -11,8 +11,15 @@ from optiland.geometries import (
     ZernikePolynomialGeometry,
     GridSagGeometry,
     BiconicGeometry,
-    ToroidalGeometry
+    ToroidalGeometry,
+    Plane,
+    PlaneGrating,
+    StandardGratingGeometry,
+    ForbesQbfsGeometry,
+    ForbesQ2dGeometry,
+    NurbsGeometry
 )
+from optiland.geometries.forbes import ForbesSurfaceConfig
 import optiland.backend as be
 
 def test_scale_standard():
@@ -126,3 +133,98 @@ def test_scale_toroidal():
     # s^(1 - 2) = s^-1.
     expected = coeffs[0] * s**(1 - 2*(1))
     assert np.isclose(surface.geometry.coeffs_poly_y[0], expected)
+
+def test_scale_plane():
+    lens = optic.Optic()
+    lens.add_surface(index=0, radius=np.inf, thickness=5)
+    lens.add_surface(index=1)
+    lens.scale_system(2.0)
+    
+    surface = lens.surface_group.surfaces[0]
+    assert isinstance(surface.geometry, Plane)
+    assert np.isinf(surface.geometry.radius)
+    assert surface.thickness == 10
+
+def test_scale_plane_grating():
+    lens = optic.Optic()
+    # PlaneGrating
+    lens.add_surface(index=0, radius=np.inf, surface_type="grating", 
+                     grating_order=1, grating_period=1.0, groove_orientation_angle=0.0)
+    lens.scale_system(2.0)
+    
+    surface = lens.surface_group.surfaces[0]
+    assert isinstance(surface.geometry, PlaneGrating)
+    assert surface.geometry.grating_period == 2.0
+
+def test_scale_standard_grating():
+    lens = optic.Optic()
+    lens.add_surface(index=0, radius=10, surface_type="grating", 
+                     grating_order=1, grating_period=1.0, groove_orientation_angle=0.0)
+    lens.scale_system(2.0)
+    
+    surface = lens.surface_group.surfaces[0]
+    assert isinstance(surface.geometry, StandardGratingGeometry)
+    assert surface.geometry.radius == 20
+    assert surface.geometry.grating_period == 2.0
+
+def test_scale_forbes_qbfs():
+    lens = optic.Optic()
+    # ForbesQbfs
+    # terms: {m: val}
+    terms = {0: 1e-3, 1: 1e-4}
+    lens.add_surface(index=0, radius=10, surface_type="forbes_qbfs", 
+                     norm_radius=1.0, radial_terms=terms)
+    lens.scale_system(2.0)
+    
+    surface = lens.surface_group.surfaces[0]
+    assert isinstance(surface.geometry, ForbesQbfsGeometry)
+    assert surface.geometry.radius == 20
+    assert surface.geometry.norm_radius == 2.0
+    
+    # Coefficients scale by s (linear with sag)
+    s = 2.0
+    assert np.allclose(surface.geometry.radial_terms[0], terms[0] * s)
+    assert np.allclose(surface.geometry.radial_terms[1], terms[1] * s)
+
+def test_scale_forbes_q2d():
+    lens = optic.Optic()
+    # ForbesQ2d
+    # terms: {('a', m, n): val}
+    terms = {('a', 0, 0): 1e-3}
+    lens.add_surface(index=0, radius=10, surface_type="forbes_q2d", 
+                     norm_radius=1.0, freeform_coeffs=terms)
+    lens.scale_system(2.0)
+    
+    surface = lens.surface_group.surfaces[0]
[…]
