### 2026-02-03 e874854
fix: add required method
