### 2026-02-03 8a91c02
feat: Implement scale method for all geometries (#449)
* feat: Implement scale method for all geometries and update optic updater to use this

* feat: implement scale for all subclasses

* fix: add required method
