+++ optiland/geometries/standard_grating.py
@@ -74,6 +74,15 @@ class StandardGratingGeometry(BaseGeometry):
         """
         self.radius = -self.radius
 
+    def scale(self, scale_factor: float):
+        """Scale the geometry parameters.
+
+        Args:
+            scale_factor (float): The factor by which to scale the geometry.
+        """
+        self.radius = self.radius * scale_factor
+        self.grating_period = self.grating_period * scale_factor
+
     def sag(self, x=0, y=0):
         """Calculate the surface sag of the geometry at the given coordinates.
 
