### 2026-02-03 700b617
chore: ruff
