+++ optiland/fileio/zemax_handler.py
@@ -16,8 +16,9 @@ import tempfile
 from dataclasses import dataclass, field
 from typing import Any
 
-import optiland.backend as be
 import requests
+
+import optiland.backend as be
 from optiland.fileio.converters import ZemaxToOpticConverter
 from optiland.materials import AbbeMaterial, BaseMaterial, Material
 from optiland.physical_apertures import RadialAperture
@@ -311,9 +312,7 @@ class ZemaxDataParser:
             "ODDASPHE": "odd_asphere",
             "COORDBRK": "coordinate_break",
             "TOROIDAL": "toroidal",
-        }.get(
-            data[1], data[1].lower()
-        )  # Default to lowercased raw value
+        }.get(data[1], data[1].lower())  # Default to lowercased raw value
 
     def _read_surface_parameter(self, data):
         key = f"param_{int(data[1]) - 1}"
