### 2026-02-03 9067990
Add circular aperture to zemax import (#450)
* Add circular aperture zemax handler

Some surfaces in Zemax have circular apertures. This commit add a read
circular aperture method for the Zemax file parser.

* Add ax parmeter to Optic.draw method.

Adding the option to pass a matplolib axes as argument. Just passing the
argument that is already on the view method.

* Add a demo notebook to show that the implementation is working.

The notebook uses two files that were added to the tests/zemax_files.

* chore: ruff

---------

Co-authored-by: Kramer <kdanielharrison@gmail.com>
