### 2026-02-03 a113397
feat: Add reference geometries for wavefront calcs, including for afocal option
