+++ tests/wavefront/test_afocal_strategies.py
@@ -1,0 +1,116 @@
+
+import pytest
+import numpy as np
+from optiland.optic import Optic
+from optiland.surfaces import ObjectSurface
+from optiland.wavefront import Wavefront, PlanarReference, SphericalReference
+from optiland.geometries import Plane
+from optiland.materials import IdealMaterial
+from optiland.coordinate_system import CoordinateSystem
+
+def create_object_surface():
+    cs = CoordinateSystem()
+    geometry = Plane(cs)
+    material = IdealMaterial(1.0, 0)
+    return ObjectSurface(geometry, material)
+
+def test_afocal_chief_ray_strategy():
+    """Test ChiefRayStrategy in afocal mode with a perfect collimator."""
+    optic = Optic()
+    # Object at infinity
+    optic.surface_group.add_surface(new_surface=create_object_surface(), index=0)
+    optic.surface_group.surfaces[0].thickness = 1000 # Put object far away or just some distance
+    optic.surface_group.surfaces[0].geometry.cs.z = float('inf')
+    
+    # Simple paraxial lens that collimates light
+    # Focal length f = 100mm. Place stop at lens.
+    optic.surface_group.add_surface(radius=100, thickness=0, material='air', index=1, is_stop=True)
+    # Let's make an afocal telescope: Keplerian
+    # f1 = 100, f2 = 50. Separation = 150.
+    # Obj -> Lens 1 (f=100) -> Lens 2 (f=50) -> Image (inf)
+    
+    # Reset optic for telescope
+    optic = Optic()
+    optic.surface_group.add_surface(new_surface=create_object_surface(), index=0)
+    optic.surface_group.add_surface(radius=50.0, thickness=5.0, material='N-BK7', index=1)
+    optic.surface_group.add_surface(radius=-1e9, thickness=150.0, material='air', index=2)
+
+    optic = Optic()
+    optic.surface_group.add_surface(new_surface=create_object_surface(), index=0) # Obj
+    optic.set_thickness(1e12, 0)
+    
+    cs_stop = CoordinateSystem()
+    stop_surf = ObjectSurface(Plane(cs_stop), IdealMaterial(1.0, 0))
+    from optiland.surfaces import Surface
+    def create_plane_surface(is_stop=False):
+        cs = CoordinateSystem()
+        return Surface(previous_surface=None, geometry=Plane(cs), material_post=IdealMaterial(1.0, 0), is_stop=is_stop)
+
+    optic.surface_group.add_surface(new_surface=create_plane_surface(is_stop=True), index=1, thickness=10)
+    
+    optic.surface_group.add_surface(new_surface=create_plane_surface(is_stop=False), index=2, thickness=20)
+    
+    optic.set_aperture('EPD', 10)
+    optic.set_field_type('angle')
+    optic.add_field(0, 0)
+    optic.add_wavelength(0.55)
+    
+    optic.update_paraxial()
+    wf = Wavefront(optic, fields=[(0,0)], wavelengths=[0.55], strategy='chief_ray', afocal=True)
+    data = wf.get_data((0,0), 0.55)
+    
+    # Check rays
+    valid = data.intensity > 0
+    rms = np.std(data.opd[valid])
+    print(f"RMS: {rms}")
+    assert rms < 1e-10
+
+def test_afocal_best_fit_strategy():
+    """Test BestFitStrategy in afocal mode removes tilt from tilted plane wave."""
+    optic = Optic()
+    optic.surface_group.add_surface(new_surface=create_object_surface(), index=0)
+    optic.set_thickness(1e9, 0)
+    optic.surface_group.add_surface(radius=float('inf'), thickness=10, material='air', index=1, is_stop=True)
[…]
