+++ optiland/wavefront/strategy.py
@@ -16,16 +16,17 @@ from typing import TYPE_CHECKING, Literal
 import optiland.backend as be
 
 from ..fields.field_types import AngleField
+from .reference_geometry import PlanarReference, ReferenceGeometry, SphericalReference
 from .wavefront_data import WavefrontData
 
 if TYPE_CHECKING:
     from optiland._types import BEArrayT
     from optiland.distribution import BaseDistribution
     from optiland.optic.optic import Optic
-    from optiland.rays.base import BaseRays
     from optiland.rays.real_rays import RealRays
 
-WavefrontStrategyType = Literal["chief_ray", "centroid_sphere", "best_fit_sphere"]
+WavefrontStrategyType = Literal["chief_ray", "centroid", "best_fit"]
+ReferenceType = Literal["sphere", "plane"]
 
 
 class ReferenceStrategy(ABC):
@@ -39,11 +40,19 @@ class ReferenceStrategy(ABC):
     Args:
         optic (Optic): The optical system to analyze.
         distribution (Distribution): The pupil sampling distribution.
+        reference_type (str): The type of reference geometry ("sphere" or "plane").
     """
 
-    def __init__(self, optic: Optic, distribution: BaseDistribution, **kwargs) -> None:
+    def __init__(
+        self,
+        optic: Optic,
+        distribution: BaseDistribution,
+        reference_type: ReferenceType = "sphere",
+        **kwargs,
+    ) -> None:
         self.optic = optic
         self.distribution = distribution
+        self.reference_type = reference_type
         self.n_image = optic.n()[-1]
 
     @abstractmethod
@@ -65,55 +74,17 @@ class ReferenceStrategy(ABC):
         """
         pass
 
-    def _opd_image_to_xp(
-        self,
-        rays_at_image: RealRays,
-        xc: float,
-        yc: float,
-        zc: float,
-        R: float,
-        wavelength: float,
-    ) -> float:
-        """Computes propagation distance from image plane to exit pupil sphere.
+    @abstractmethod
+    def _create_reference_geometry(self, rays: RealRays) -> ReferenceGeometry:
+        """Creates the reference geometry based on the traced rays.
 
         Args:
-            rays_at_image (object): An object with ray data attributes (x, y, z,
-                L, M, N) at the final image surface.
-            xc (float): The x-coordinate of the reference sphere center.
-            yc (float): The y-coordinate of the reference sphere center.
-            zc (float): The z-coordinate of the reference sphere center.
-            R (float): The radius of the reference sphere.
-            wavelength (float): The wavelength of the light.
+            rays: The traced rays at the image surface.
 
         Returns:
-            float: The optical path length from the image surface to the
-                   reference sphere.
+            ReferenceGeometry: The computed reference geometry.
         """
-        xr, yr, zr = rays_at_image.x, rays_at_image.y, rays_at_image.z
-        L, M, N = -rays_at_image.L, -rays_at_image.M, -rays_at_image.N
-
-        a = L**2 + M**2 + N**2
-        b = 2 * (L * (xr - xc) + M * (yr - yc) + N * (zr - zc))
-        c = (
-            xr**2
-            + yr**2
-            + zr**2
-            - 2 * (xr * xc + yr * yc + zr * zc)
-            + xc**2
-            + yc**2
-            + zc**2
-            - R**2
-        )
-        d = b**2 - 4 * a * c
-        d = be.where(d < 0, 0, d)  # Ensure non-negative for sqrt
-        t = (-b - be.sqrt(d)) / (2 * a)
-
-        # If the first solution for t is negative, the ray is pointing away
-        # from the sphere, so we take the other root.
-        mask = t < 0
-        t = be.where(mask, (-b + be.sqrt(d)) / (2 * a), t)
-
[…]
