### 2026-02-03 aec0aef
chore: clean up strategies and correct tests
