+++ optiland/wavefront/strategy.py
@@ -201,15 +201,32 @@ class ChiefRayStrategy(ReferenceStrategy):
             raise ValueError("Chief ray cannot be determined. It must be traced alone.")
 
         if self.reference_type == "sphere":
-            R = be.sqrt(x**2 + y**2 + (z - self.pupil_z) ** 2)
-            return SphericalReference((float(x), float(y), float(z)), R.item())
+            return self._create_spherical_ref(x, y, z)
         elif self.reference_type == "plane":
-            L, M, N = rays.L, rays.M, rays.N
-            return PlanarReference(
-                (float(x), float(y), float(z)), (float(L), float(M), float(N))
-            )
+            return self._create_planar_ref(x, y, z, rays.L, rays.M, rays.N)
         else:
             raise ValueError(f"Unknown reference type: {self.reference_type}")
+
+    def _create_spherical_ref(
+        self, x: BEArrayT, y: BEArrayT, z: BEArrayT
+    ) -> SphericalReference:
+        """Create a spherical reference geometry."""
+        R = be.sqrt(x**2 + y**2 + (z - self.pupil_z) ** 2)
+        return SphericalReference((float(x), float(y), float(z)), R.item())
+
+    def _create_planar_ref(
+        self,
+        x: BEArrayT,
+        y: BEArrayT,
+        z: BEArrayT,
+        L: BEArrayT,
+        M: BEArrayT,
+        N: BEArrayT,
+    ) -> PlanarReference:
+        """Create a planar reference geometry."""
+        return PlanarReference(
+            (float(x), float(y), float(z)), (float(L), float(M), float(N))
+        )
 
 
 class CentroidStrategy(ReferenceStrategy):
@@ -348,32 +365,43 @@ class CentroidStrategy(ReferenceStrategy):
         centroid = be.sum(image_points * weights[:, None], axis=0) / total_weight
 
         if self.reference_type == "sphere":
-            distances_wf = be.linalg.norm(wavefront_points - centroid, axis=1)
-            radius = float(be.sum(weights * distances_wf) / be.sum(weights))
-            return SphericalReference(
-                (float(centroid[0]), float(centroid[1]), float(centroid[2])), radius
-            )
-
+            return self._create_spherical_ref(wavefront_points, centroid, weights)
         elif self.reference_type == "plane":
-            L, M, N = rays.L[valid_mask], rays.M[valid_mask], rays.N[valid_mask]
-            directions = be.stack((L, M, N), axis=1)
-            mean_direction = (
-                be.sum(directions * weights[:, None], axis=0) / total_weight
-            )
-            norm = be.linalg.norm(mean_direction)
-            if norm > 0:
-                mean_direction = mean_direction / norm
-
-            return PlanarReference(
-                (float(centroid[0]), float(centroid[1]), float(centroid[2])),
-                (
-                    float(mean_direction[0]),
-                    float(mean_direction[1]),
-                    float(mean_direction[2]),
-                ),
-            )
+            return self._create_planar_ref(centroid, rays, valid_mask, weights)
         else:
             raise ValueError(f"Unknown reference type: {self.reference_type}")
+
+    def _create_spherical_ref(
+        self, wavefront_points: be.ndarray, centroid: be.ndarray, weights: be.ndarray
+    ) -> SphericalReference:
+        distances_wf = be.linalg.norm(wavefront_points - centroid, axis=1)
+        radius = float(be.sum(weights * distances_wf) / be.sum(weights))
+        return SphericalReference(
+            (float(centroid[0]), float(centroid[1]), float(centroid[2])), radius
+        )
+
[…]
