### 2026-02-03 6f40cd0
docs: add doc strings to methods
