### 2026-02-03 2c6eb64
docs: update for afocal mode
