### 2026-02-04 7dca6f1
Add aperture visualization to optic draw
- Draw aperture line segments on 2D projections (XZ/YZ).
- Add stop arrows on the aperture stop surface.
- Derive stop size from physical aperture, semi-aperture, floating stop,
or traced ray extent.\n- Skip XY projection to avoid clutter.
