### 2026-02-04 339624d
updated code resolves issues with optimization using autograd
