+++ optiland/geometries/forbes/geometry.py
@@ -219,24 +219,10 @@ class ForbesQbfsGeometry(ForbesGeometryBase):
                 k: be.array(v) for k, v in (self.surface_config.terms or {}).items()
             }
         else:
-            self.radial_terms = (self.surface_config.terms or {}).copy()
+            self.radial_terms = self.surface_config.terms or {}
 
         self.norm_radius = be.array(self.surface_config.norm_radius)
         self.is_symmetric = True
-
-    def scale(self, scale_factor: float):
-        """Scale the geometry parameters.
-
-        Args:
-            scale_factor (float): The factor by which to scale the geometry.
-        """
-        super().scale(scale_factor)
-        self.surface_config.radius = self.radius
-        self.surface_config.norm_radius *= scale_factor
-        self.norm_radius = be.array(self.surface_config.norm_radius)
-
-        for k in self.radial_terms:
-            self.radial_terms[k] = self.radial_terms[k] * scale_factor
 
     def _prepare_coeffs(self):
         """Prepares the internal coefficient lists from the radial_terms dictionary."""
@@ -283,9 +269,8 @@ class ForbesQbfsGeometry(ForbesGeometryBase):
     def _surface_normal(self, x, y):
         """Calculates the unit vector normal to the surface.
 
-        Dispatches to an autograd-based method for the torch backend and an
-        analytical method for the numpy backend. It also patches the `NaN`
-        gradient that autograd produces at the vertex.
+        Uses analytical method for both torch and numpy backends, with proper
+        handling of the vertex case to avoid NaN gradients.
 
         Args:
             x (float or array_like): X coordinate(s).
@@ -298,36 +283,17 @@ class ForbesQbfsGeometry(ForbesGeometryBase):
         self._prepare_coeffs()
         x_in, y_in = be.array(x), be.array(y)
 
-        if be.get_backend() == "torch":
-            # ensure inputs require gradients for autograd to work
-            x_grad = x_in.clone().detach().requires_grad_(True)
-            y_grad = y_in.clone().detach().requires_grad_(True)
-            z0 = self.sag(x_grad, y_grad)
-
-            grad_outputs = be.ones_like(z0)
-            gradients = be.autograd.grad(
-                outputs=z0,
-                inputs=(x_grad, y_grad),
-                grad_outputs=grad_outputs,
-                create_graph=True,
-                allow_unused=True,
-            )
-            df_dx, df_dy = gradients[0], gradients[1]
-
-            # replace possible NaNs at the vertex with the analytical value
-            # (which is 0 for Qbfs)
-            is_vertex = be.logical_and(be.abs(x_in) < _EPSILON, be.abs(y_in) < _EPSILON)
-            df_dx = be.where(is_vertex, 0.0, df_dx)
-            df_dy = be.where(is_vertex, 0.0, df_dy)
-        else:
-            df_dx, df_dy = self._surface_normal_analytical(x_in, y_in)
+        df_dx, df_dy = self._surface_normal_analytical(x_in, y_in)
 
         mag = be.sqrt(df_dx**2 + df_dy**2 + 1)
         safe_mag = be.where(mag < _EPSILON, 1.0, mag)
         return df_dx / safe_mag, df_dy / safe_mag, -1 / safe_mag
 
     def _surface_normal_analytical(self, x, y):
-        """Computes the analytical surface derivatives for the numpy backend.
+        """Computes the analytical surface derivatives.
+
+        This method handles the vertex case by ensuring numerically stable
[…]
