### 2026-02-04 26269af
Fix GeometricMTF diffraction scale when max_freq is used
