### 2026-02-04 ca64e09
Fix formatting using ruff
I was using black with isort
