+++ optiland/geometries/forbes/geometry.py
@@ -3,30 +3,29 @@
 This module provides implementations for optical surfaces described by
 Forbes polynomials, which are superimposed on a base conic section.
 Forbes polynomials offer a modern alternative to standard power-series
-aspheres, providing better control over the surface shape and its derivatives.
-Their orthogonality helps to decouple the effects of different coefficients,
-which is advantageous for optimization, tolerancing, and assessing
-manufacturability.
+aspheres.
 
 The implementation is based on the work of G. W. Forbes. See, for example,
 G. W. Forbes, "Manufacturability estimates for optical aspheres," Opt. Express (2011).
 
 This module implements two types of Forbes surfaces:
-    1.  **ForbesQbfsGeometry (Q-type, Best Fit Sphere)**:
-        This class represents rotationally symmetric aspheric surfaces. The "Qbfs"
-        polynomials are orthogonal over a circular aperture and describe the
-        departure from a base conic.
+    1.  **ForbesQNormalSlopeGeometry** (slope-orthogonal Q polynomials):
+        This class represents rotationally symmetric aspheric surfaces using
+        Forbes Q polynomials (historically called Q^bfs in Forbes 2007). These
+        polynomials are orthogonal with respect to the normal-departure
+        slope metric, and the implementation supports a general conic
+        reference surface (Forbes 2011 generalization). The deprecated alias
+        ``ForbesQbfsGeometry`` is retained for backward compatibility.
     2.  **ForbesQ2dGeometry (Q-type, 2D)**:
         This class represents non-rotationally symmetric, or "freeform," surfaces.
         The Q2D polynomials extend the Q-type formalism to two dimensions, allowing
         for the description of complex, freeform optical surfaces that lack
         rotational symmetry.
-
-Manuel Fragata Mendes, 2025
 """
 
 from __future__ import annotations
 
+import warnings
 from dataclasses import asdict, dataclass
 from typing import Any
 
@@ -180,11 +179,17 @@ class ForbesGeometryBase(NewtonRaphsonGeometry):
         return factor, derivative
 
 
-class ForbesQbfsGeometry(ForbesGeometryBase):
-    r"""Represents a Forbes Q-bfs surface (rotationally symmetric Q-type).
+class ForbesQNormalSlopeGeometry(ForbesGeometryBase):
+    r"""Represents a Forbes Q surface using slope-orthogonal polynomials.
 
-    The Q-bfs surface is defined by the sag equation:
-    $z(\rho) = z_{base}(\rho) + \frac{1}{\sigma(\rho)}
+    This surface uses the Forbes Q polynomials that are orthogonal with respect
+    to the normal-departure slope metric. These were historically called
+    Q^bfs ("best-fit sphere") in Forbes 2007, but the implementation here
+    supports a general conic reference surface following the Forbes 2011
+    generalization.
+
+    The surface sag is defined as:
+    $z(\rho) = z_{base}(\rho) + \phi(\rho)
     \left[ u^2(1-u^2) \sum_{m=0}^{M} a_m Q_m(u^2) \right]$
 
     where:
@@ -192,10 +197,16 @@ class ForbesQbfsGeometry(ForbesGeometryBase):
           is the base conic.
         - $c = 1/R$ is the curvature, $k$ is the conic constant.
         - $u = \rho/\rho_{max}$ is the normalized radial coordinate.
-        - $Q_m(u^2)$ are the Forbes orthogonal polynomials.
+        - $Q_m(u^2)$ are the Forbes slope-orthogonal polynomials.
         - $a_m$ are the polynomial coefficients.
-        - $\sigma(\rho) = \sqrt{\frac{1 - kc^2\rho^2}{1 - (1+k)c^2\rho^2}}$ is a
-          conic scaling factor.
[…]
