### 2026-02-04 61794bf
Do not draw if the surface is a lens
It will always draw if it is the aperture stop.
