### 2026-02-04 1e9a0a0
docs: add Renato Ferracini Alves to authors list
