### 2026-02-04 1d37a74
Make aperture drawing optional an add a guard only for 2D visualization
