### 2026-02-04 fe7b98a
Fix coordinates from tensor to numpy for matplolib plotting
