### 2026-02-04 bfa0d49
Wavefront Analysis for Afocal Systems (#456)
* feat: Add reference geometries for wavefront calcs, including for afocal option

* chore: clean up strategies and correct tests

* docs: add doc strings to methods

* feat: code cleanup and impacted MTF tests fix

* docs: update for afocal mode

* docs: fix num rays argument for opd calc

* docs: reformat OPD plots

* docs: update rst references

* docs: mention afocal mode in ZernikeOPD
