### 2026-02-04 96663af
Add aperture visualization to optic draw method (#455)
* Add aperture visualization to optic draw

- Draw aperture line segments on 2D projections (XZ/YZ).
- Add stop arrows on the aperture stop surface.
- Derive stop size from physical aperture, semi-aperture, floating stop,
or traced ray extent.\n- Skip XY projection to avoid clutter.

* Do not draw if the surface is a lens

It will always draw if it is the aperture stop.

* Make aperture drawing optional an add a guard only for 2D visualization

* Fix coordinates from tensor to numpy for matplolib plotting
