### 2026-02-05 65a78e9
Fix/forbes issues (#460)
* fixes to the cache problem: now faster for higher number of coeffs. autograd remains intact

* updated code resolves issues with optimization using autograd

* fixes to the ambigous naming in the forbes geometry (associated calls were updated too)

* fix: issues with the example in the Q2D surface. Fixes (#419)
