+++ optiland/geometries/forbes/__init__.py
@@ -6,7 +6,8 @@ from __future__ import annotations
 
 from .geometry import (
     ForbesQ2dGeometry,
-    ForbesQbfsGeometry,
+    ForbesQbfsGeometry,  # Deprecated alias
+    ForbesQNormalSlopeGeometry,
     ForbesSolverConfig,
     ForbesSurfaceConfig,
 )
@@ -17,7 +18,8 @@ from .qpoly import (
 
 __all__ = [
     "ForbesQ2dGeometry",
-    "ForbesQbfsGeometry",
+    "ForbesQNormalSlopeGeometry",
+    "ForbesQbfsGeometry",  # Deprecated alias for backward compatibility
     "compute_z_zprime_q2d",
     "q2d_nm_coeffs_to_ams_bms",
     "ForbesSurfaceConfig",
