### 2026-02-06 f18c03d
Add option to remove vignetted rays on 2D visualization
