### 2026-02-06 a32ad98
Add option to hide vignetted rays in the 3D layout drawing
