+++ optiland/visualization/system/rays.py
@@ -8,9 +8,8 @@ Kramer Harrison, 2024
 from __future__ import annotations
 
 import numpy as np
-import vtk
-
 import optiland.backend as be
+import vtk
 from optiland.utils import resolve_fields, resolve_wavelengths
 from optiland.visualization.system.ray_bundle import RayBundle
 from optiland.visualization.system.utils import transform
@@ -298,6 +297,7 @@ class Rays3D(Rays2D):
         distribution="line_y",
         reference=None,
         theme=None,
+        hide_vignetted=False,
     ):
         """Plots the rays for the given fields and wavelengths.
 
@@ -312,6 +312,8 @@ class Rays3D(Rays2D):
             reference (str, optional): The reference rays to plot. Options
                 include "chief" and "marginal". Defaults to None.
             theme (Theme, optional): The theme to apply. Defaults to None.
+            hide_vignetted (bool, optional): If True, rays that vignette at any
+                surface are not shown. Defaults to False.
 
         """
         fields = resolve_fields(self.optic, fields)
@@ -328,12 +330,25 @@ class Rays3D(Rays2D):
                 else:
                     # trace rays and plot lines
                     self._trace(field, wavelength, num_rays, distribution)
-                    self._plot_lines(ax, color_idx, field, theme=theme)
+                    self._plot_lines(
+                        ax,
+                        color_idx,
+                        field,
+                        theme=theme,
+                        hide_vignetted=hide_vignetted,
+                    )
 
                 # trace reference rays and plot lines
                 if reference is not None:
                     self._trace_reference(field, wavelength, reference)
-                    self._plot_lines(ax, color_idx, field, linewidth=1.5, theme=theme)
+                    self._plot_lines(
+                        ax,
+                        color_idx,
+                        field,
+                        linewidth=1.5,
+                        theme=theme,
+                        hide_vignetted=hide_vignetted,
+                    )
 
     def __init__(self, optic):
         super().__init__(optic)
@@ -352,13 +367,18 @@ class Rays3D(Rays2D):
             (0.090, 0.745, 0.812),
         ]
 
-    def _plot_lines(self, ax, color_idx, field, linewidth=1, theme=None):
+    def _plot_lines(
+        self, ax, color_idx, field, linewidth=1, theme=None, hide_vignetted=False
+    ):
         # loop through rays
         for k in range(self.z.shape[1]):
             xk = be.to_numpy(self.x[:, k])
             yk = be.to_numpy(self.y[:, k])
             zk = be.to_numpy(self.z[:, k])
             ik = be.to_numpy(self.i[:, k])
+
+            if hide_vignetted and np.any(ik == 0):
+                continue
 
             # remove rays outside aperture
             xk[ik == 0] = np.nan
