### 2026-02-07 d6ba01f
Feat/add option to remove vignetted rays from optic draw (#462)
* Add option to remove vignetted rays on 2D visualization

* Add hide_vignetted option example

* Add option to hide vignetted rays in the 3D layout drawing

* Add hide_vignetted test for 3D in the example notebook

* Fix linting
