+++ optiland/visualization/system/rays.py
@@ -58,6 +58,7 @@ class Rays2D:
         reference=None,
         theme=None,
         projection="YZ",
+        hide_vignetted=False,
     ):
         """Plots the rays for the given fields and wavelengths.
 
@@ -72,6 +73,8 @@ class Rays2D:
             reference (str, optional): The reference rays to plot. Options
                 include "chief" and "marginal". Defaults to None.
             theme (Theme, optional): The theme to apply. Defaults to None.
+            hide_vignetted (bool, optional): If True, rays that vignette at any
+                surface are not shown. Defaults to False.
 
         """
         fields = resolve_fields(self.optic, fields)
@@ -91,7 +94,12 @@ class Rays2D:
                     self._trace(field, wavelength, num_rays, distribution)
                     artists.update(
                         self._plot_lines(
-                            ax, color_idx, field, theme=theme, projection=projection
+                            ax,
+                            color_idx,
+                            field,
+                            theme=theme,
+                            projection=projection,
+                            hide_vignetted=hide_vignetted,
                         )
                     )
 
@@ -106,6 +114,7 @@ class Rays2D:
                             linewidth=1.5,
                             theme=theme,
                             projection=projection,
+                            hide_vignetted=hide_vignetted,
                         )
                     )
         return artists
@@ -168,7 +177,14 @@ class Rays2D:
         self.r_extent = be.fmax(self.r_extent, r_extent_new)
 
     def _plot_lines(
-        self, ax, color_idx, field, linewidth=1, theme=None, projection="YZ"
+        self,
+        ax,
+        color_idx,
+        field,
+        linewidth=1,
+        theme=None,
+        projection="YZ",
+        hide_vignetted=False,
     ):
         """Plots multiple lines on the given axis.
 
@@ -200,6 +216,9 @@ class Rays2D:
             yk = be.to_numpy(self.y[:, k])
             zk = be.to_numpy(self.z[:, k])
             ik = be.to_numpy(self.i[:, k])
+
+            if hide_vignetted and np.any(ik == 0):
+                continue
 
             # remove rays outside aperture
             xk[ik == 0] = np.nan
@@ -279,6 +298,7 @@ class Rays3D(Rays2D):
         distribution="line_y",
         reference=None,
         theme=None,
+        hide_vignetted=False,
     ):
         """Plots the rays for the given fields and wavelengths.
 
@@ -293,6 +313,8 @@ class Rays3D(Rays2D):
             reference (str, optional): The reference rays to plot. Options
                 include "chief" and "marginal". Defaults to None.
             theme (Theme, optional): The theme to apply. Defaults to None.
+            hide_vignetted (bool, optional): If True, rays that vignette at any
+                surface are not shown. Defaults to False.
 
         """
         fields = resolve_fields(self.optic, fields)
@@ -309,12 +331,25 @@ class Rays3D(Rays2D):
                 else:
                     # trace rays and plot lines
                     self._trace(field, wavelength, num_rays, distribution)
-                    self._plot_lines(ax, color_idx, field, theme=theme)
+                    self._plot_lines(
+                        ax,
+                        color_idx,
+                        field,
+                        theme=theme,
+                        hide_vignetted=hide_vignetted,
+                    )
 
                 # trace reference rays and plot lines
                 if reference is not None:
                     self._trace_reference(field, wavelength, reference)
-                    self._plot_lines(ax, color_idx, field, linewidth=1.5, theme=theme)
+                    self._plot_lines(
+                        ax,
+                        color_idx,
+                        field,
+                        linewidth=1.5,
[…]
