### 2026-02-07 b5ee851
feat: implement orthogonal descent optimizer & add docs
