### 2026-02-07 41af27a
feat: implement orthogonal descent optimizer & add docs (#464)
* feat: implement orthogonal descent optimizer & add docs

* chore: transfer notebook from master
