+++ optiland/__init__.py
@@ -2,4 +2,4 @@
 
 from __future__ import annotations
 
-__version__ = "0.5.8"
+__version__ = "0.5.9"
