### 2026-02-08 68df009
feat: add curvature solves and refactor subpackage
