+++ tests/test_solves.py
@@ -1,11 +1,16 @@
 import pytest
+import numpy as np
 
 import optiland.backend as be
 from optiland.samples.objectives import CookeTriplet
 from optiland.solves import (
     BaseSolve,
-    ChiefRayHeightSolve,
+    ThicknessSolve,
     MarginalRayHeightSolve,
+    ChiefRayHeightSolve,
+    CurvatureSolve,
+    MarginalRayAngleSolve,
+    ChiefRayAngleSolve,
     QuickFocusSolve,
     SolveFactory,
     SolveManager,
@@ -25,14 +30,21 @@ class TestMarginalRayHeightSolve:
         assert solve.optic == optic
         assert solve.surface_idx == surface_idx
         assert solve.height == height
+        assert isinstance(solve, ThicknessSolve)
 
     def test_marginal_ray_height_solve_apply(self, set_test_backend):
         optic = CookeTriplet()
         surface_idx = 7
         height = 0.5
 
+        # Paraxial trace
         ya, ua = optic.paraxial.marginal_ray()
-        offset = (height - ya[surface_idx]) / ua[surface_idx]
+        
+        # Incident slope is ua[surface_idx-1] (or ua[0] if surface_idx=0)
+        u_incident = ua[0] if surface_idx == 0 else ua[surface_idx - 1]
+        
+        offset = (height - ya[surface_idx]) / u_incident
+        
         surf = optic.surface_group.surfaces[surface_idx]
         z_orig = be.copy(surf.geometry.cs.z)
 
@@ -62,39 +74,25 @@ class TestMarginalRayHeightSolve:
         optic = CookeTriplet()
         data = {"type": "MarginalRayHeightSolve", "surface_idx": 7, "height": 0.5}
 
-        solve = BaseSolve.from_dict(optic, data)  # Use imported BaseSolve
+        solve = BaseSolve.from_dict(optic, data)
 
         assert solve.surface_idx == data["surface_idx"]
         assert solve.height == data["height"]
         assert isinstance(solve, MarginalRayHeightSolve)
 
-    def test_from_dict_invalid_type(self, set_test_backend):
-        optic = CookeTriplet()
-        data = {"type": "Invalid", "surface_idx": 7, "height": 0.5}
-
-        with pytest.raises(ValueError):
-            BaseSolve.from_dict(optic, data)  # Changed to use imported BaseSolve
-
 
 class TestChiefRayHeightSolve:
     def test_chief_ray_height_solve_constructor(self, set_test_backend):
         optic = CookeTriplet()
-        surface_idx = 7  # Example surface index
-        height = 0.2  # Example target chief ray height
+        surface_idx = 7
+        height = 0.2
 
         solve = ChiefRayHeightSolve(optic, surface_idx, height)
 
         assert solve.optic == optic
         assert solve.surface_idx == surface_idx
         assert solve.height == height
-
-    def test_invalid_surface_idx(self, set_test_backend):
-        with pytest.raises(ValueError):
-            optic = CookeTriplet()
-            surface_idx = None
-            height = 0.2
-
-            solve = ChiefRayHeightSolve(optic, surface_idx, height)
+        assert isinstance(solve, ThicknessSolve)
 
     def test_chief_ray_height_solve_apply(self, set_test_backend):
         optic = CookeTriplet()
@@ -102,13 +100,16 @@ class TestChiefRayHeightSolve:
         height = 0.2
 
         yc, uc = optic.paraxial.chief_ray()
-        # Ensure uc[surface_idx] is not zero to avoid division by zero
-        if be.abs(uc[surface_idx] - 0.0) < 1e-6:
+        
+        # Incident slope
+        u_incident = uc[0] if surface_idx == 0 else uc[surface_idx - 1]
+
+        if be.abs(u_incident - 0.0) < 1e-6:
[…]
