+++ optiland/solves/curvature.py
@@ -1,0 +1,279 @@
+"""Curvature Solve Module
+
+This module defines `CurvatureSolve`, an abstract base class for solves
+that adjust surface curvature (radius) to satisfy a condition.
+
+Kramer Harrison, 2026
+"""
+
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
+
+from optiland.solves.base import BaseSolve
+
+
+class CurvatureSolve(BaseSolve, ABC):
+    """Abstract base class for curvature solves.
+
+    This class provides the common structure for solves that aim to satisfy
+    a condition by adjusting the curvature of a specific surface.
+
+    Attributes:
+        optic (Optic): The optic object.
+        surface_idx (int): The index of the surface where the curvature
+            is to be controlled.
+    """
+
+    def __init__(self, optic, surface_idx: int):
+        """Initializes a CurvatureSolve object.
+
+        Args:
+            optic (Optic): The optic object.
+            surface_idx (int): The index of the surface.
+        """
+        if surface_idx is None:
+            raise ValueError("'surface_idx' argument must be provided.")
+        super().__init__()
+        self.optic = optic
+        self.surface_idx = surface_idx
+
+    @abstractmethod
+    def apply(self):
+        """Applies the curvature solve to the optic."""
+        pass  # pragma: no cover
+
+    def to_dict(self):
+        """Returns a dictionary representation of the solve."""
+        solve_dict = super().to_dict()
+        solve_dict.update(
+            {
+                "surface_idx": self.surface_idx,
+            }
+        )
+        return solve_dict
+
+    @classmethod
+    def from_dict(cls, optic, data):
+        """Creates a solve instance from a dictionary representation."""
+        if cls is CurvatureSolve:
+            raise TypeError(
+                "CurvatureSolve is an abstract class and cannot be "
+                "instantiated directly."
+            )
+
+        if cls is MarginalRayAngleSolve:
+            return cls(optic, data["surface_idx"], data["angle"])
+
+        if cls is ChiefRayAngleSolve:
+            return cls(optic, data["surface_idx"], data["angle"])
+
+        raise NotImplementedError(f"from_dict not implemented for {cls.__name__}")
+
+
+class MarginalRayAngleSolve(CurvatureSolve):
+    """Adjusts surface curvature to achieve a target marginal ray exit angle.
+
+    This solve uses the paraxial refraction equation:
+        n'u' - nu = -y(n' - n)c
+
+    to solve for curvature c:
+        c = (nu - n'u') / (y(n' - n))
+
+    Attributes:
+        optic (Optic): The optic object.
+        surface_idx (int): The index of the surface.
+        angle (float): The target exit angle (u') of the marginal ray.
+    """
+
+    def __init__(self, optic, surface_idx: int, angle: float):
+        """Initializes a MarginalRayAngleSolve object.
+
+        Args:
+            optic (Optic): The optic object.
+            surface_idx (int): The index of the surface.
+            angle (float): The target exit angle.
+        """
+        super().__init__(optic, surface_idx)
+        self.angle = angle
+
+    def apply(self):
+        """Applies the marginal ray angle solve."""
+        y, u = self.optic.paraxial.marginal_ray()
+
+        u_in = u[0] if self.surface_idx == 0 else u[self.surface_idx - 1]
+
+        y_surf = y[self.surface_idx]
+        u_out_target = self.angle
+
+        # Indices of refraction
+        if self.surface_idx == 0:
+            n_pre = self.optic.surface_group.surfaces[
+                self.surface_idx - 1
+            ].material_post.n(self.optic.primary_wavelength)
+        else:
+            n_pre = self.optic.surface_group.surfaces[
+                self.surface_idx - 1
+            ].material_post.n(self.optic.primary_wavelength)
+
[…]
