### 2026-02-08 6b1fdae
refactor: update naming conventions and clean up
