+++ tests/test_solves.py
@@ -6,11 +6,11 @@ from optiland.samples.objectives import CookeTriplet
 from optiland.solves import (
     BaseSolve,
     ThicknessSolve,
-    MarginalRayHeightSolve,
-    ChiefRayHeightSolve,
+    MarginalRayHeightThicknessSolve,
+    ChiefRayHeightThicknessSolve,
     CurvatureSolve,
-    MarginalRayAngleSolve,
-    ChiefRayAngleSolve,
+    MarginalRayAngleCurvatureSolve,
+    ChiefRayAngleCurvatureSolve,
     QuickFocusSolve,
     SolveFactory,
     SolveManager,
@@ -19,13 +19,13 @@ from optiland.solves import (
 from .utils import assert_allclose
 
 
-class TestMarginalRayHeightSolve:
+class TestMarginalRayHeightThicknessSolve:
     def test_marginal_ray_height_solve_constructor(self, set_test_backend):
         optic = CookeTriplet()
         surface_idx = 7
         height = 0.5
 
-        solve = MarginalRayHeightSolve(optic, surface_idx, height)
+        solve = MarginalRayHeightThicknessSolve(optic, surface_idx, height)
 
         assert solve.optic == optic
         assert solve.surface_idx == surface_idx
@@ -48,7 +48,7 @@ class TestMarginalRayHeightSolve:
         surf = optic.surface_group.surfaces[surface_idx]
         z_orig = be.copy(surf.geometry.cs.z)
 
-        solve = MarginalRayHeightSolve(optic, surface_idx, height)
+        solve = MarginalRayHeightThicknessSolve(optic, surface_idx, height)
         solve.apply()
 
         # Check that surface has been shifted
@@ -63,31 +63,31 @@ class TestMarginalRayHeightSolve:
         surface_idx = 7
         height = 0.5
 
-        solve = MarginalRayHeightSolve(optic, surface_idx, height)
+        solve = MarginalRayHeightThicknessSolve(optic, surface_idx, height)
         data = solve.to_dict()
 
-        assert data["type"] == "MarginalRayHeightSolve"
+        assert data["type"] == "MarginalRayHeightThicknessSolve"
         assert data["surface_idx"] == surface_idx
         assert data["height"] == height
 
     def test_from_dict(self, set_test_backend):
         optic = CookeTriplet()
-        data = {"type": "MarginalRayHeightSolve", "surface_idx": 7, "height": 0.5}
+        data = {"type": "MarginalRayHeightThicknessSolve", "surface_idx": 7, "height": 0.5}
 
         solve = BaseSolve.from_dict(optic, data)
 
         assert solve.surface_idx == data["surface_idx"]
         assert solve.height == data["height"]
-        assert isinstance(solve, MarginalRayHeightSolve)
+        assert isinstance(solve, MarginalRayHeightThicknessSolve)
 
 
-class TestChiefRayHeightSolve:
+class TestChiefRayHeightThicknessSolve:
     def test_chief_ray_height_solve_constructor(self, set_test_backend):
         optic = CookeTriplet()
         surface_idx = 7
         height = 0.2
 
-        solve = ChiefRayHeightSolve(optic, surface_idx, height)
+        solve = ChiefRayHeightThicknessSolve(optic, surface_idx, height)
 
         assert solve.optic == optic
         assert solve.surface_idx == surface_idx
@@ -113,7 +113,7 @@ class TestChiefRayHeightSolve:
         surf = optic.surface_group.surfaces[surface_idx]
         z_orig = be.copy(surf.geometry.cs.z)
 
-        solve = ChiefRayHeightSolve(optic, surface_idx, height)
+        solve = ChiefRayHeightThicknessSolve(optic, surface_idx, height)
[…]
