### 2026-02-08 354961f
test: fix solve test assertion for torch backend
