### 2026-02-08 96f44a2
Solve Refactor & Curvature Solve Addition (#468)
* feat: add curvature solves and refactor subpackage

* refactor: update naming conventions and clean up

* docs: update solve docs

* test: fix solve test assertion for torch backend
