+++ tests/test_solves.py
@@ -1,11 +1,16 @@
 import pytest
+import numpy as np
 
 import optiland.backend as be
 from optiland.samples.objectives import CookeTriplet
 from optiland.solves import (
     BaseSolve,
-    ChiefRayHeightSolve,
-    MarginalRayHeightSolve,
+    ThicknessSolve,
+    MarginalRayHeightThicknessSolve,
+    ChiefRayHeightThicknessSolve,
+    CurvatureSolve,
+    MarginalRayAngleCurvatureSolve,
+    ChiefRayAngleCurvatureSolve,
     QuickFocusSolve,
     SolveFactory,
     SolveManager,
@@ -14,29 +19,36 @@ from optiland.solves import (
 from .utils import assert_allclose
 
 
-class TestMarginalRayHeightSolve:
+class TestMarginalRayHeightThicknessSolve:
     def test_marginal_ray_height_solve_constructor(self, set_test_backend):
         optic = CookeTriplet()
         surface_idx = 7
         height = 0.5
 
-        solve = MarginalRayHeightSolve(optic, surface_idx, height)
+        solve = MarginalRayHeightThicknessSolve(optic, surface_idx, height)
 
         assert solve.optic == optic
         assert solve.surface_idx == surface_idx
         assert solve.height == height
+        assert isinstance(solve, ThicknessSolve)
 
     def test_marginal_ray_height_solve_apply(self, set_test_backend):
         optic = CookeTriplet()
         surface_idx = 7
         height = 0.5
 
+        # Paraxial trace
         ya, ua = optic.paraxial.marginal_ray()
-        offset = (height - ya[surface_idx]) / ua[surface_idx]
+        
+        # Incident slope is ua[surface_idx-1] (or ua[0] if surface_idx=0)
+        u_incident = ua[0] if surface_idx == 0 else ua[surface_idx - 1]
+        
+        offset = (height - ya[surface_idx]) / u_incident
+        
         surf = optic.surface_group.surfaces[surface_idx]
         z_orig = be.copy(surf.geometry.cs.z)
 
-        solve = MarginalRayHeightSolve(optic, surface_idx, height)
+        solve = MarginalRayHeightThicknessSolve(optic, surface_idx, height)
         solve.apply()
 
         # Check that surface has been shifted
@@ -51,50 +63,36 @@ class TestMarginalRayHeightSolve:
         surface_idx = 7
         height = 0.5
 
-        solve = MarginalRayHeightSolve(optic, surface_idx, height)
+        solve = MarginalRayHeightThicknessSolve(optic, surface_idx, height)
         data = solve.to_dict()
 
-        assert data["type"] == "MarginalRayHeightSolve"
+        assert data["type"] == "MarginalRayHeightThicknessSolve"
         assert data["surface_idx"] == surface_idx
         assert data["height"] == height
 
     def test_from_dict(self, set_test_backend):
         optic = CookeTriplet()
-        data = {"type": "MarginalRayHeightSolve", "surface_idx": 7, "height": 0.5}
+        data = {"type": "MarginalRayHeightThicknessSolve", "surface_idx": 7, "height": 0.5}
 
-        solve = BaseSolve.from_dict(optic, data)  # Use imported BaseSolve
+        solve = BaseSolve.from_dict(optic, data)
 
         assert solve.surface_idx == data["surface_idx"]
         assert solve.height == data["height"]
-        assert isinstance(solve, MarginalRayHeightSolve)
-
-    def test_from_dict_invalid_type(self, set_test_backend):
-        optic = CookeTriplet()
-        data = {"type": "Invalid", "surface_idx": 7, "height": 0.5}
-
-        with pytest.raises(ValueError):
-            BaseSolve.from_dict(optic, data)  # Changed to use imported BaseSolve
[…]
