### 2026-02-09 166177d
tests: add tests for fileio, image sim, & variables
