### 2026-02-09 55332f4
fix: get device for operations in torch backend (#469)
