### 2026-02-10 00016fa
pytest: negative and positive thin/thick lenses
