### 2026-02-11 2173b57
refactor source logic: Add SMF Source model, ExtendedSourceOptic wrapper, and a correction term added for radiant intensity computation (#224)
