+++ optiland/sources/smf.py
@@ -1,0 +1,207 @@
+"""Single-Mode Fiber (SMF) Source Module
+
+This module implements an extended Gaussian source that generates a ray bundle
+correctly representing the output of a single-mode fiber (SMF). The source
+produces rays with Gaussian distributions in both the spatial and angular
+domains, using quasi-random Sobol sequences for high-quality sampling.
+
+Key features:
+    - Spatial profile defined by the Mode Field Diameter (MFD).
+    - Angular profile defined by the 1/e² full divergence angle.
+    - Non-paraxial direction cosines computed via tangent mapping.
+    - Sobol quasi-random sampling for low-discrepancy ray sets.
+    - Optional point-source mode (zero spatial extent).
+
+Note on ``num_rays``:
+    Sobol sequences require the sample count to be a power of two.  The
+    ``generate_rays`` method automatically rounds the requested ``num_rays``
+    **up** to the nearest power of two, so the returned ``RealRays`` object
+    may contain more rays than requested.
+"""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from optiland import backend as be
+from optiland.sources.base import BaseSource
+
+if TYPE_CHECKING:
+    from optiland.rays import RealRays
+
+
+class SMFSource(BaseSource):
+    """Extended Gaussian source representing a single-mode fiber output.
+
+    This source generates rays with independent Gaussian distributions in
+    both spatial (x, y) and angular (θ_x, θ_y) phase-space dimensions.
+    It is designed to faithfully represent the far-field emission profile
+    of a single-mode optical fiber.
+
+    Sampling uses 4-dimensional Sobol sequences (two spatial, two angular)
+    transformed to Gaussian distributions via the inverse error function.
+    Direction cosines are computed non-paraxially from the sampled angles
+    using the tangent mapping, ensuring physical accuracy at large
+    divergence angles.
+
+    Args:
+        mfd_um (float): Mode Field Diameter in micrometers (µm).  The 1/e²
+            beam diameter at the fiber facet.  The beam waist w₀ is ``mfd / 2``.
+        divergence_deg_1e2 (float): Full-angle 1/e² far-field divergence in
+            degrees.
+        wavelength_um (float): Operating wavelength in micrometers (µm).
+        total_power (float): Total optical power in Watts.  Defaults to 1.0.
+        position (tuple[float, float, float]): Source position (x, y, z) in
+            millimeters.  Defaults to ``(0, 0, 0)``.
+        is_point_source (bool): If ``True``, spatial coordinates are set to
+            zero (ideal point source).  Defaults to ``False``.
+
+    Attributes:
+        wavelength (float): Wavelength in µm.
+        total_power (float): Total power in W.
+        mfd_um (float): Mode Field Diameter in µm.
+        divergence_deg_1e2 (float): Full 1/e² divergence in degrees.
+        sigma_spatial_mm (float): Spatial sampling sigma (mm), equal to w₀/2
+            converted to mm.
+        sigma_angular_rad (float): Angular sampling sigma (rad), equal to
+            half-angle / 2.
+        is_point_source (bool): Whether to emit from a single point.
+
+    Note:
+        The actual number of rays returned by ``generate_rays`` is rounded
+        up to the nearest power of two because Sobol sampling requires this.
+        For example, requesting 1000 rays will generate 1024.
+    """
+
+    def __init__(
+        self,
+        mfd_um: float,
+        wavelength_um: float,
+        divergence_deg_1e2: float | None = None,
[…]
