+++ optiland/sources/__init__.py
@@ -1,5 +1,4 @@
 # flake8: noqa
 
 from .base import BaseSource
-from .gaussian import GaussianSource
-from .collimated_gaussian import CollimatedGaussianSource
+from .smf import SMFSource
