### 2026-02-11 fccff58
docs: update documentation to include ExtendedSource approach. Includes an example Notebook Tutorial and for the gallery as well
