+++ docs/index.rst
@@ -108,6 +108,7 @@ Not sure what to type in the shell? Here are a few ideas to explore Optiland rig
    gallery/differentiable_ray_tracing
    gallery/real_world_projects
    gallery/external_tools
+   gallery/extended_sources
    gallery/miscellaneous
 
 
@@ -168,6 +169,7 @@ Not sure what to type in the shell? Here are a few ideas to explore Optiland rig
    api/api_materials
    api/api_ml
    api/api_optic
+   api/api_sources
    api/api_optimization
    api/api_paraxial
    api/api_phase
