+++ docs/api/sources/sources.visualization.rst
@@ -1,0 +1,12 @@
+sources.visualization
+=====================
+
+.. automodule:: sources.visualization
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      SourceViewer
+   
