+++ docs/api/api_optic.rst
@@ -3,6 +3,7 @@ Optic
 
 This section includes the core optic module and the optic updater module. The Optic class is the primary container for an optical system in Optiland.
 The OpticUpdater class is used to update various properties of an Optic object, such as surface properties or polarization.
+The ExtendedSourceOptic class is a wrapper around the Optic class that enables extended source ray tracing.
 
 .. autosummary::
    :toctree: optic/
@@ -10,3 +11,4 @@ The OpticUpdater class is used to update various properties of an Optic object, 
 
    optic.optic
    optic.optic_updater
+   optic.extended_source_optic
