+++ docs/gallery/extended_sources.rst
@@ -1,0 +1,12 @@
+.. _gallery_extended_sources:
+
+Extended Sources
+================
+
+This section contains examples of extended source modeling in **Optiland**. Extended sources allow
+users to define spatially and angularly extended light sources and trace them through optical systems,
+enabling physically accurate simulations of fiber-coupled and other non-point-source illumination.
+
+.. nbgallery::
+
+    extended_sources/beam_shaping_singlet
