### 2026-02-11 459b882
Add comprehensive tests for extended sources and fix torch XY projection
- Add tests/test_extended_sources.py with 160 tests covering:
  - SMFSource: ray generation, parameter computation, edge cases, visualization
  - BaseSource: abstract interface contract verification
  - ExtendedSourceOptic: delegation, tracing, drawing, projections, repr
  - IncoherentIrradiance: source integration and conflict handling
  - SourceViewer: 6-panel diagnostic visualization
  - Standard analysis delegation via .optic for field-based analyses

- Fix torch backend compatibility in Lens2D XY projection
  (optila
[…]
