+++ tests/test_extended_sources.py
@@ -1,0 +1,789 @@
+"""Comprehensive tests for the extended sources feature.
+
+This module tests:
+  - SMFSource ray generation, parameters, and edge cases
+  - BaseSource interface contract
+  - ExtendedSourceOptic delegation, tracing, drawing, and restrictions
+  - Source visualization (SourceViewer)
+  - Integration with IncoherentIrradiance analysis
+"""
+
+from __future__ import annotations
+
+import math
+
+import matplotlib
+import matplotlib.pyplot as plt
+import numpy as np
+import pytest
+
+import optiland.backend as be
+from optiland.coordinate_system import CoordinateSystem
+from optiland.optic import ExtendedSourceOptic, Optic
+from optiland.physical_apertures import RectangularAperture
+from optiland.rays import RealRays
+from optiland.sources import BaseSource, SMFSource
+
+
+# ---------------------------------------------------------------------------
+# Helper fixtures
+# ---------------------------------------------------------------------------
+
+
+def _simple_singlet():
+    """Create a minimal singlet for testing."""
+    lens = Optic()
+    lens.add_surface(index=0, thickness=be.inf)
+    lens.add_surface(
+        index=1,
+        thickness=7,
+        radius=43.7354,
+        is_stop=True,
+        material="N-SF11",
+    )
+    lens.add_surface(index=2, radius=-46.2795, thickness=50)
+    lens.add_surface(index=3)
+    lens.set_aperture(aperture_type="EPD", value=25)
+    lens.set_field_type(field_type="angle")
+    lens.add_field(y=0)
+    lens.add_wavelength(value=0.55, is_primary=True)
+    return lens
+
+
+def _simple_singlet_with_detector():
+    """Create a singlet that has a rectangular aperture on the image surface."""
+    lens = Optic()
+    lens.add_surface(index=0, thickness=be.inf)
+    lens.add_surface(
+        index=1,
+        thickness=7,
+        radius=43.7354,
+        is_stop=True,
+        material="N-SF11",
+    )
+    lens.add_surface(index=2, radius=-46.2795, thickness=50)
+    lens.add_surface(
+        index=3,
+        aperture=RectangularAperture(
+            x_min=-15, x_max=15, y_min=-15, y_max=15
+        ),
+    )
+    lens.set_aperture(aperture_type="EPD", value=25)
+    lens.set_field_type(field_type="angle")
+    lens.add_field(y=0)
+    lens.add_wavelength(value=0.55, is_primary=True)
+    return lens
+
+
+def _default_source():
+    """Create a default SMFSource for testing."""
+    return SMFSource(
+        mfd_um=10.4,
+        wavelength_um=0.55,
+        total_power=1.0,
+    )
+
+
+# ===========================================================================
+# SMFSource Tests
+# ===========================================================================
+
+
+class TestSMFSource:
+    """Tests for the Single-Mode Fiber source."""
+
+    @pytest.fixture(autouse=True)
+    def _backend(self, set_test_backend):
+        pass
+
+    # --- Initialization & parameter computation ---
+
+    def test_default_divergence_calculation(self):
+        """Divergence should be computed from the Gaussian beam formula."""
+        src = SMFSource(mfd_um=10.4, wavelength_um=1.55)
+        w0 = 10.4 / 2.0
+        expected_half_rad = 1.55 / (math.pi * w0)
+        expected_deg = 2 * math.degrees(expected_half_rad)
+        assert abs(src.divergence_deg_1e2 - expected_deg) < 1e-10
+
+    def test_explicit_divergence_overrides(self):
+        """When divergence is explicitly provided it should be used as-is."""
[…]
