### 2026-02-15 69580d5
Add HeightProfile with wavelength-dependent material support
