+++ optiland/phase/base.py
@@ -37,7 +37,7 @@ class BasePhaseProfile(abc.ABC):
         return 1.0
 
     @abc.abstractmethod
-    def get_phase(self, x: be.Array, y: be.Array) -> be.Array:
+    def get_phase(self, x: be.Array, y: be.Array, wavelength: be.Array) -> be.Array:
         """Calculates the phase added by the profile at coordinates (x, y).
 
         Args:
@@ -51,7 +51,7 @@ class BasePhaseProfile(abc.ABC):
 
     @abc.abstractmethod
     def get_gradient(
-        self, x: be.Array, y: be.Array
+        self, x: be.Array, y: be.Array, wavelength: be.Array
     ) -> tuple[be.Array, be.Array, be.Array]:
         """Calculates the gradient of the phase at coordinates (x, y).
 
@@ -66,7 +66,7 @@ class BasePhaseProfile(abc.ABC):
         raise NotImplementedError
 
     @abc.abstractmethod
-    def get_paraxial_gradient(self, y: be.Array) -> be.Array:
+    def get_paraxial_gradient(self, y: be.Array, wavelength: be.Array) -> be.Array:
         """Calculates the paraxial phase gradient at y-coordinate.
 
         This is the gradient d_phi/dy evaluated at x=0.
