+++ optiland/phase/radial.py
@@ -23,7 +23,9 @@ class RadialPhaseProfile(BasePhaseProfile):
     def __init__(self, coefficients: list[float]):
         self.coefficients = coefficients
 
-    def get_phase(self, x: be.Array, y: be.Array, wavelength: be.Array = None) -> be.Array:
+    def get_phase(
+        self, x: be.Array, y: be.Array, wavelength: be.Array = None
+    ) -> be.Array:
         """Calculates the phase added by the profile at coordinates (x, y).
 
         Args:
@@ -74,7 +76,9 @@ class RadialPhaseProfile(BasePhaseProfile):
 
         return d_phi_dx, d_phi_dy, d_phi_dz
 
-    def get_paraxial_gradient(self, y: be.Array, wavelength: be.Array = None) -> be.Array:
+    def get_paraxial_gradient(
+        self, y: be.Array, wavelength: be.Array = None
+    ) -> be.Array:
         """Calculates the paraxial phase gradient at y-coordinate.
 
         This is the gradient d_phi/dy evaluated at x=0.
