+++ optiland/phase/constant.py
@@ -24,7 +24,9 @@ class ConstantPhaseProfile(BasePhaseProfile):
     def __init__(self, phase: float = 0.0):
         self.phase = phase
 
-    def get_phase(self, x: be.Array, y: be.Array, wavelength: be.Array = None) -> be.Array:
+    def get_phase(
+        self, x: be.Array, y: be.Array, wavelength: be.Array = None
+    ) -> be.Array:
         """Calculates the phase added by the profile at coordinates (x, y).
 
         Args:
@@ -51,7 +53,9 @@ class ConstantPhaseProfile(BasePhaseProfile):
         """
         return be.zeros_like(x), be.zeros_like(y), be.zeros_like(x)
 
-    def get_paraxial_gradient(self, y: be.Array, wavelength: be.Array = None) -> be.Array:
+    def get_paraxial_gradient(
+        self, y: be.Array, wavelength: be.Array = None
+    ) -> be.Array:
         """Calculates the paraxial phase gradient at y-coordinate.
 
         This is the gradient d_phi/dy evaluated at x=0.
