+++ optiland/phase/grid.py
@@ -56,7 +56,9 @@ class GridPhaseProfile(BasePhaseProfile):
             self.y_coords, self.x_coords, self.phase_grid
         )
 
-    def get_phase(self, x: be.Array, y: be.Array, wavelength: be.Array = None) -> be.Array:
+    def get_phase(
+        self, x: be.Array, y: be.Array, wavelength: be.Array = None
+    ) -> be.Array:
         """Calculates the phase added by the profile at coordinates (x, y).
 
         Args:
@@ -88,7 +90,9 @@ class GridPhaseProfile(BasePhaseProfile):
         d_phi_dz = be.zeros_like(x)
         return d_phi_dx, d_phi_dy, d_phi_dz
 
-    def get_paraxial_gradient(self, y: be.Array, wavelength: be.Array = None) -> be.Array:
+    def get_paraxial_gradient(
+        self, y: be.Array, wavelength: be.Array = None
+    ) -> be.Array:
         """Calculates the paraxial phase gradient at y-coordinate.
 
         This is the gradient d_phi/dy evaluated at x=0.
