### 2026-02-16 eb1f068
chore: ruff & add Gustavo to authors
