### 2026-02-16 fd572a2
Add HeightProfile with wavelength-dependent material support (#475)
* Add HeightProfile with wavelength-dependent material support

* formating using ruff

* chore: ruff & add Gustavo to authors

---------

Co-authored-by: Kramer <kdanielharrison@gmail.com>
