### 2026-02-16 6cd1e2b
feat: #246 Update colorimetry constants to English and enhance thin film analysis with color handling methods
