### 2026-02-16 5c2458e
#246 add an example to the gallery
