### 2026-02-16 21b139c
#246 add test for colorimetry module
