### 2026-02-16 110ab97
feat: #246 update tests for thin_film
