+++ tests/test_thin_film.py
@@ -824,7 +824,8 @@ class TestSpectralAnalyzer:
         wavelengths_um = be.linspace(0.4, 0.8, 3)
         angles_deg = be.linspace(0, 60, 3)
 
-        # Test multiple quantities with single polarization - should return array of axes
+        # Test multiple quantities with single
+        # polarization - should return array of axes
         fig, axs = analyzer.map_view(
             wavelengths_um, aoi_values=angles_deg, to_plot=["R", "T", "A"]
         )
@@ -855,7 +856,8 @@ class TestSpectralAnalyzer:
         plt.close(fig)
 
     def test_multiple_polarizations_comprehensive(self, set_test_backend, simple_stack):
-        """Test comprehensive support for multiple polarizations across all view methods."""
+        """Test comprehensive support for multiple polarizations across
+        all view methods."""
         analyzer = SpectralAnalyzer(simple_stack)
         wavelengths_um = be.linspace(0.4, 0.8, 5)
         angles_deg = be.linspace(0, 60, 5)
@@ -1000,7 +1002,8 @@ class TestSpectralAnalyzer:
         result_um = analyzer.stack.compute_rtRTA(wl_um, be.deg2rad(angles_deg), "s")
 
         # Test consistency - all should convert to the same wavelength internally
-        # We can't easily test the plotting results directly, but we can test the conversion methods
+        # We can't easily test the plotting results directly, but we can test the
+        # conversion methods
         assert_allclose(
             analyzer._convert_to_wavelength_um(wl_nm, "nm"), wl_um, rtol=1e-10
         )
@@ -1047,3 +1050,211 @@ class TestThinFilmStackComplex:
             values = result[quantity]
             assert be.all(values >= 0), f"{quantity} has negative values"
             assert be.all(values <= 1), f"{quantity} has values > 1"
+
+
+class TestEdgeCasesAndErrors:
+    """Test edge cases and error handling."""
+
+    def test_wavelength_unit_unknown(self, set_test_backend, simple_stack):
+        """Test error handling for unknown wavelength unit."""
+        analyzer = SpectralAnalyzer(simple_stack)
+        with pytest.raises(ValueError, match="Unknown wavelength unit"):
+            analyzer._convert_to_wavelength_um(0.5, "invalid_unit")
+
+    def test_relative_wavenumber_without_reference(self, set_test_backend, air, glass):
+        """Test error handling for relative_wavenumber without reference_wl_um."""
+        # Create stack without reference_wl_um
+        stack = ThinFilmStack(incident_material=air, substrate_material=glass)
+        analyzer = SpectralAnalyzer(stack)
+
+        with pytest.raises(ValueError, match="reference_wl_um must be set"):
+            analyzer._convert_to_wavelength_um(0.5, "relative_wavenumber")
+
+    def test_angle_unit_conversion_deg(self, set_test_backend, simple_stack):
+        """Test angle conversion in degrees."""
+        analyzer = SpectralAnalyzer(simple_stack)
+        angles_deg = 45.0
+        angles_rad = analyzer._convert_angle_to_radians(angles_deg, "deg")
+        assert_allclose(angles_rad, be.deg2rad(45.0), rtol=1e-10)
+
+    def test_angle_unit_conversion_rad(self, set_test_backend, simple_stack):
+        """Test angle conversion in radians (identity)."""
+        analyzer = SpectralAnalyzer(simple_stack)
+        angles_rad_input = 0.7854  # 45 degrees in radians
+        angles_rad_output = analyzer._convert_angle_to_radians(angles_rad_input, "rad")
[…]
