### 2026-02-16 4c75bc7
Update test for thin film optimization #246
