### 2026-02-17 6bf627a
#246 Update doc tutorial for thin films
