+++ docs/learning_guide.rst
@@ -69,7 +69,9 @@ This learning guide breaks down Optiland into a series of tutorials that cover t
 
    examples/Tutorial_6a_Introduction_to_Coatings
    examples/Tutorial_6b_Introduction_to_Polarization
-
+   examples/Tutorial_6c_Multilayer_Stack
+   examples/Tutorial_6d_Thin_Film_Optimization
+   examples/Tutorial_6e_Color_Analysis_For_Thin_Film
 
 7. Advanced Optical Design
 --------------------------
