### 2026-02-21 36a45d9
Solves #483 and #484
