### 2026-02-21 a431fac
Solves #483 and #484 (#485)
