+++ docs/api/sources/sources.base.rst
@@ -1,12 +1,0 @@
-sources.base
-============
-
-.. automodule:: sources.base
-
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      BaseSource
-   
