+++ docs/api/sources/sources.smf.rst
@@ -1,12 +1,0 @@
-sources.smf
-===========
-
-.. automodule:: sources.smf
-
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      SMFSource
-   
