+++ docs/api/sources/sources.visualization.rst
@@ -1,12 +1,0 @@
-sources.visualization
-=====================
-
-.. automodule:: sources.visualization
-
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      SourceViewer
-   
