+++ docs/api/optic/optic.extended_source_optic.rst
@@ -1,12 +1,0 @@
-optic.extended\_source\_optic
-============================
-
-.. automodule:: optic.extended_source_optic
-
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      ExtendedSourceOptic
-   
