### 2026-02-22 d3433b9
docs: add extended sources to MD learning guide
