### 2026-02-22 24c076b
fix: update aperture extent plotting logic. Remove update_paraxial calls from examples.
