### 2026-02-22 8d54c95
feat: add draft MTF vs. field analysis type
