+++ optiland/analysis/mtf_vs_field.py
@@ -1,0 +1,216 @@
+"""MTF versus Field Analysis
+
+This module enables the calculation of the Modulation Transfer Function (MTF)
+versus field coordinate of an optical system.
+
+Kramer Harrison, 2026
+"""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+import matplotlib.pyplot as plt
+
+import optiland.backend as be
+from optiland.analysis.base import BaseAnalysis
+from optiland.mtf.sampled import SampledMTF
+
+if TYPE_CHECKING:
+    from matplotlib.axes import Axes
+    from matplotlib.figure import Figure
+
+    from optiland.optic import Optic
+
+
+class MTFvsField(BaseAnalysis):
+    """MTF versus Field Coordinate.
+
+    This class is used to analyze the Modulation Transfer Function (MTF) versus
+    the field coordinate of an optical system for specified spatial frequencies.
+
+    Args:
+        optic (Optic): the optical system.
+        frequencies (list[float]): the spatial frequencies (in cycles/mm) to analyze.
+        num_fields (int): the number of fields in the Y direction. Default is 32.
+        wavelengths (str or list): the wavelengths to be analyzed. Default is 'all'.
+        num_rays (int): the number of rays across the pupil in 1D for the SampledMTF
+            calculation. Default is 128.
+        override_limits (bool): If True, bypasses the limit on the number of frequencies
+            and wavelengths to prevent cluttered plots. Default is False.
+
+    """
+
+    MAX_FREQUENCIES = 5
+    MAX_WAVELENGTHS = 3
+
+    def __init__(
+        self,
+        optic: Optic,
+        frequencies: list[float],
+        num_fields: int = 32,
+        wavelengths: str | list[float] = "all",
+        num_rays: int = 128,
+        override_limits: bool = False,
+    ):
+        self.frequencies = frequencies
+        self.num_fields = num_fields
+        self.num_rays = num_rays
+
+        self._check_limits(override_limits, wavelengths, optic)
+
+        # Base Analysis will set self.wavelengths
+        super().__init__(optic, wavelengths)
+
+    def _check_limits(self, override_limits: bool, wavelengths, optic):
+        """Check to ensure inputs won't produce an overly cluttered plot."""
+        if override_limits:
+            return
+
+        if len(self.frequencies) > self.MAX_FREQUENCIES:
+            raise ValueError(
+                f"Number of frequencies ({len(self.frequencies)}) exceeds the "
+                f"recommended limit of {self.MAX_FREQUENCIES} for clean plots. "
+                "Set override_limits=True to bypass this check."
+            )
+
+        from optiland.utils import resolve_wavelengths
+
+        resolved_wls = resolve_wavelengths(optic, wavelengths)
+        num_wl = len(resolved_wls)
+
+        if num_wl > self.MAX_WAVELENGTHS:
+            raise ValueError(
+                f"Number of wavelengths ({num_wl}) exceeds the recommended "
+                f"limit of {self.MAX_WAVELENGTHS} for clean plots. "
+                "Set override_limits=True to bypass this check."
+            )
+
+    def _generate_data(self):
+        """Generate the MTF data across fields, wavelengths, and frequencies."""
+        fields = [(0.0, float(Hy)) for Hy in be.linspace(0.0, 1.0, self.num_fields)]
+        self._field_coords = be.array(fields)
+
+        results = []
+        for wl in self.wavelengths:
+            wl_results = []
+            for freq in self.frequencies:
+                freq_tan = (freq, 0.0)
+                freq_sag = (0.0, freq)
+
+                tan_mtf = []
+                sag_mtf = []
+
+                for field in fields:
+                    sampled_mtf = SampledMTF(
[…]
