+++ tests/test_mtf_vs_field.py
@@ -1,0 +1,104 @@
+import pytest
+import matplotlib.pyplot as plt
+from matplotlib.axes import Axes
+from matplotlib.figure import Figure
+
+from optiland.analysis import MTFvsField
+from optiland.samples.objectives import CookeTriplet
+import optiland.backend as be
+import numpy as np
+
+
+class TestMTFvsField:
+    """Test suite for the MTFvsField class."""
+
+    def test_init_defaults(self, set_test_backend):
+        """Test initialization with default parameters."""
+        optic = CookeTriplet()
+        mtf_vf = MTFvsField(optic, frequencies=[10.0, 20.0], num_fields=3, num_rays=12)
+
+        assert mtf_vf.frequencies == [10.0, 20.0]
+        assert mtf_vf.num_fields == 3
+        assert mtf_vf.num_rays == 12
+        assert mtf_vf.wavelengths == [w.value for w in optic.wavelengths.wavelengths]
+
+        # Check results structure
+        assert len(mtf_vf.data) == len(mtf_vf.wavelengths)
+        assert len(mtf_vf.data[0]) == len(mtf_vf.frequencies)
+        
+        # Check tangential and sagittal arrays
+        res = mtf_vf.data[0][0]
+        assert "tangential" in res
+        assert "sagittal" in res
+        assert len(res["tangential"]) == mtf_vf.num_fields
+        assert len(res["sagittal"]) == mtf_vf.num_fields
+
+    def test_init_limits_frequencies(self, set_test_backend):
+        """Test the frequency limit checks."""
+        optic = CookeTriplet()
+        freqs = [10, 20, 30, 40, 50, 60]  # length 6 > MAX=5
+        
+        with pytest.raises(ValueError, match="Number of frequencies"):
+            MTFvsField(optic, frequencies=freqs)
+            
+        # Try with override
+        mtf_vf = MTFvsField(optic, frequencies=freqs, num_rays=32, num_fields=5, override_limits=True)
+        assert len(mtf_vf.frequencies) == 6
+
+    def test_init_limits_wavelengths(self, set_test_backend):
+        """Test the frequency limit checks."""
+        optic = CookeTriplet()
+        optic.add_wavelength(0.5, weight=1.0)
+        optic.add_wavelength(0.6, weight=1.0)
+        optic.add_wavelength(0.7, weight=1.0)
+        # Optic now has 4 wavelengths (>3 MAX)
+        
+        with pytest.raises(ValueError, match="Number of wavelengths"):
+            MTFvsField(optic, frequencies=[10], wavelengths="all")
+            
+        # Try with override
+        mtf_vf = MTFvsField(optic, frequencies=[10], wavelengths="all", num_rays=32, num_fields=5, override_limits=True)
+        assert len(mtf_vf.wavelengths) == 6
+
+    def test_custom_params(self, set_test_backend):
+        """Test initialization with custom parameters."""
+        optic = CookeTriplet()
+        mtf_vf = MTFvsField(
+            optic, 
+            frequencies=[15.0],
+            num_fields=16,
+            wavelengths=[0.55],
+            num_rays=64
+        )
+
+        assert mtf_vf.frequencies == [15.0]
+        assert mtf_vf.num_fields == 16
+        assert mtf_vf.num_rays == 64
+        assert mtf_vf.wavelengths == [0.55]
+
+        # Check data
+        res = mtf_vf.data[0][0]
+        assert len(res["tangential"]) == 16
+        
+        # Check values
+        assert be.all(res["tangential"] >= 0.0)
+        assert be.all(res["tangential"] <= 1.0)
+        assert be.all(res["sagittal"] >= 0.0)
+        assert be.all(res["sagittal"] <= 1.0)
+
[…]
