+++ docs/api/api_analysis.rst
@@ -19,6 +19,7 @@ encircled energy, field curvature, distortion, etc.
    analysis.image_simulation
    analysis.irradiance
    analysis.jones_pupil
+   analysis.mtf_vs_field
    analysis.pupil_aberration
    analysis.ray_fan
    analysis.rms_vs_field
