### 2026-02-22 ee80e20
feat: optimize calculation for speed. Patch in tests for speed.
