### 2026-02-22 14ab028
feat: Add pickup functionality to define relationships between lens components, supporting generic attribute paths with surface index `[i]`.
