### 2026-02-24 f6cd48d
Parse NAME zemax file command and save it in Optiland .json; Open dialog supports an additional file type Zemax .zmx and be able to open zmx directly from the GUI; Surface diameter DIAM in zmx is now imported (for now, only circular surfaces are recongized).
