+++ optiland/optic/optic.py
@@ -727,6 +727,7 @@ class Optic:
         """
         data = {
             "version": 1.0,
+            "name": self.name,
             "aperture": self.aperture.to_dict() if self.aperture else None,
             "fields": self.fields.to_dict(),
             "wavelengths": self.wavelengths.to_dict(),
@@ -755,6 +756,7 @@ class Optic:
 
         """
         optic = cls()
+        optic.name = data["name"]
         optic.aperture = Aperture.from_dict(data["aperture"])
         optic.surface_group = SurfaceGroup.from_dict(data["surface_group"])
         optic.fields = FieldGroup.from_dict(data["fields"])
