### 2026-03-01 01e2deb
feat: Add pickup functionality to define relationships between lens components, supporting generic attribute paths with surface index `[i]`. (#489)
