### 2026-03-01 afc8ce4
feat: Add OptilandEncoder to handle tensor/array serialization across backends. Add tests
