### 2026-03-01 4d82ae9
feat: implement state for norm radius update configuration. Update docs
