+++ optiland/geometries/base.py
@@ -1,3 +1,4 @@
+# ruff: noqa: B027
 """Base Geometry
 
 The base geometry class for all geometries.
@@ -95,6 +96,16 @@ class BaseGeometry(ABC):
         curvature is inverted. For example, the radius of curvature should
         change sign. Conic constants and polynomial coefficients should not
         change. The coordinate system of the geometry should not be modified.
+        """
+        pass  # pragma: no cover
+
+    def update_normalization(self, semi_aperture: float) -> None:
+        """Update the normalization attributes of the geometry based on
+        its defined normalization_mode ('auto' or 'manual').
+        Base geometry generally does not maintain a normalization radius.
+
+        Args:
+            semi_aperture (float): The current semi-aperture of the surface.
         """
         pass  # pragma: no cover
 
