### 2026-03-01 3ee4a5c
fix: update paraxial checks for surface aperture before changing limit
