### 2026-03-01 234cad1
fix: make extent a be array before checking finite
