### 2026-03-01 94018b4
chore: remove print statement
