### 2026-03-01 4fcb646
Parse NAME zemax file command and save it in Optiland .json; Open dia… (#495)
* Parse NAME zemax file command and save it in Optiland .json; Open dialog supports an additional file type Zemax .zmx and be able to open zmx directly from the GUI; Surface diameter DIAM in zmx is now imported (for now, only circular surfaces are recongized).

* feat: Add pickup functionality to define relationships between lens components, supporting generic attribute paths with surface index `[i]`. (#489)

* fix: add comment to to_dict (#499)

* feat: Add OptilandEncoder to handle tensor/array serialization across
[…]
