### 2026-03-01 27b65d4
docs: add cyberstudio10 to authors list in docs
