### 2026-03-01 ad9673c
fix: fix dict retrieval in Optic
