### 2026-03-01 ee6fa4c
Add Torch grid interpolation using grid_sample with autograd support (#493)
* Add Torch grid interpolation using grid_sample with autograd support

* Fix ruff TC003: move Callable import to TYPE_CHECKING block
