### 2026-03-01 54bf089
fix: use global coords instead of local for stop surface
