### 2026-03-03 cb375d3
Remove implicit array to float conversions
