### 2026-03-03 4af2240
Add explicit float conversion in calculate_stop_radius
