### 2026-03-06 46831fb
Remove implicit array to scalar conversions (#503)
* Remove implicit array to float conversions

* Add explicit float conversion in calculate_stop_radius

* Fix incorrect name of ScalarOrArrayT TypeVar
