### 2026-03-06 58eec4a
Refactoring Normalization Radius Logic (#501)
* feat: implement state for norm radius update configuration. Update docs

* fix: update paraxial checks for surface aperture before changing limit

* fix: make extent a be array before checking finite
