### 2026-03-06 e3aa002
simplify `ParaxialRayTracer.trace_generic`
