+++ optiland/raytrace/paraxial_ray_tracer.py
@@ -101,7 +101,6 @@ class ParaxialRayTracer:
 
         heights = []
         slopes = []
-        backward_propagating = False
 
         for k in range(skip, len(R)):
             if isinstance(surfs[k], ObjectSurface):
@@ -118,15 +117,12 @@ class ParaxialRayTracer:
             if surfs[k].interaction_model.is_reflective:
                 if surfs[k].surface_type == "paraxial":
                     f = surfs[k].interaction_model.f * (-1 if reverse else 1)
-                    u_ = -u_ - y_ / f
+                    u_ = -u_ + y_ / f
                 else:
                     u_ = -u_ - 2 * y_ / R[k]
-                backward_propagating = not backward_propagating
             else:
                 if surfs[k].surface_type == "paraxial":
-                    f = surfs[k].interaction_model.f * (
-                        -1 if backward_propagating else 1
-                    )
+                    f = surfs[k].interaction_model.f * (-1 if t < 0 else 1)
                     u_ = (n[k - 1] * u_ - y_ / f) / n[k]
                 else:
                     u_ = (n[k - 1] * u_ - y_ * power[k]) / n[k]
