### 2026-03-06 9fbebd3
Fix interaction of real rays with paraxial surfaces
