### 2026-03-07 5afb464
feat: refactor spot diagram and add reference option (chief ray or centroid)
