### 2026-03-07 a59b142
chore: remove StrEnum which is not supported in 3.10
