### 2026-03-07 e8e66b8
Spot Diagram Refactor & Reference Strategies (#505)
* feat: refactor spot diagram and add reference option (chief ray or centroid)

* chore: remove StrEnum which is not supported in 3.10
