+++ tests/analysis/test_spot_reference.py
@@ -1,0 +1,97 @@
+import pytest
+import numpy as np
+
+import optiland.backend as be
+from optiland.optic import Optic
+from optiland.analysis.spot_diagram import SpotDiagram, SpotReferenceType
+from optiland.analysis.spot_diagram.reference import ChiefRayReference, CentroidReference
+from optiland.samples.objectives import CookeTriplet
+
+
+@pytest.fixture
+def cooke_triplet():
+    return CookeTriplet()
+
+
+class TestSpotReference:
+    def test_default_reference(self, set_test_backend, cooke_triplet):
+        """SpotDiagram should default to ChiefRayReference."""
+        spot = SpotDiagram(cooke_triplet)
+        assert isinstance(spot._reference_strategy, ChiefRayReference)
+
+    def test_centroid_reference(self, set_test_backend, cooke_triplet):
+        """SpotDiagram with centroid reference should use CentroidReference."""
+        spot = SpotDiagram(cooke_triplet, reference="centroid")
+        assert isinstance(spot._reference_strategy, CentroidReference)
+
+    def test_invalid_reference(self, cooke_triplet):
+        """SpotDiagram with invalid reference should raise ValueError."""
+        with pytest.raises(ValueError):
+            SpotDiagram(cooke_triplet, reference="invalid_ref")
+
+    def test_centroid_centering_logic(self, set_test_backend, cooke_triplet):
+        """When using centroid reference, the mean of the centered rays at the ref wavelength should be ~zero."""
+        spot = SpotDiagram(cooke_triplet, reference="centroid")
+        centered_data = spot._center_spots(spot.data)
+        
+        ref_idx = spot._analysis_ref_wavelength_index
+        for field_data in centered_data:
+            ref_spot = field_data[ref_idx]
+            mean_x = be.to_numpy(be.mean(ref_spot.x)).item()
+            mean_y = be.to_numpy(be.mean(ref_spot.y)).item()
+            
+            # The centered spots should have a centroid of (0, 0)
+            assert np.isclose(mean_x, 0.0, atol=1e-7)
+            assert np.isclose(mean_y, 0.0, atol=1e-7)
+
+    def test_chief_ray_centering_logic(self, set_test_backend, cooke_triplet):
+        """When using chief ray reference, the chief ray coordinates should be (0, 0) in the centered data."""
+        spot = SpotDiagram(cooke_triplet, reference="chief_ray")
+        
+        ref_idx = spot._analysis_ref_wavelength_index
+        ref_wl = spot.wavelengths[ref_idx]
+        
+        # Center the data using the chief ray
+        centered_data = spot._center_spots(spot.data)
+        
+        # Also compute the uncentered chief ray centers
+        chief_centers = spot.generate_chief_rays_centers(ref_wl)
+        chief_centers_np = be.to_numpy(chief_centers)
+        
+        # Original centroid of the data (not centered on chief ray yet)
+        raw_centroid = spot.centroid()
+        
+        for i, field_data in enumerate(centered_data):
+            # The chief ray should be at (0, 0) in the centered data plot.
+            # But the spots are just dots. The difference between the uncentered chief ray
+            # and the uncentered centroid should match the centered centroid.
+            
+            ref_spot = field_data[ref_idx]
+            centered_mean_x = be.to_numpy(be.mean(ref_spot.x)).item()
+            centered_mean_y = be.to_numpy(be.mean(ref_spot.y)).item()
+            
+            raw_cx, raw_cy = raw_centroid[i]
+            chief_x, chief_y = chief_centers_np[i]
+            
+            # The centered data centroid should equal (raw centroid - chief ray)
[…]
