+++ optiland/analysis/spot_diagram/plotting.py
@@ -1,0 +1,228 @@
+"""Spot Diagram Plotting Utilities
+
+This module provides standalone plotting helper functions for spot diagrams.
+These are extracted from the SpotDiagram class to keep the core analysis
+class focused on data generation and metrics.
+
+Kramer Harrison, 2025
+"""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+import matplotlib.pyplot as plt
+import numpy as np
+from matplotlib import patches
+
+import optiland.backend as be
+
+if TYPE_CHECKING:
+    from matplotlib.axes import Axes
+    from matplotlib.figure import Figure
+    from numpy.typing import NDArray
+
+    from .core import SpotData
+
+
+def handle_no_fields(fig: Figure | None) -> tuple[None, None]:
+    """Handles the case where there are no fields to plot.
+
+    Args:
+        fig: An optional existing figure to draw a message on.
+
+    Returns:
+        A tuple of (None, None).
+    """
+    print("Warning (SpotDiagram.view): No fields to plot.")
+    if fig and hasattr(fig, "canvas") and fig.canvas:
+        fig.text(0.5, 0.5, "No fields to plot Spot Diagram", ha="center", va="center")
+        fig.canvas.draw_idle()
+    return None, None
+
+
+def setup_plot_layout(
+    num_fields: int,
+    fig_to_plot_on: Figure | None,
+    figsize: tuple[float, float],
+) -> tuple[Figure, NDArray[np.object_]]:
+    """Sets up the Matplotlib figure and axes grid.
+
+    Args:
+        num_fields: Number of field points to plot.
+        fig_to_plot_on: An existing figure to use, or None to create one.
+        figsize: The size for the figure.
+
+    Returns:
+        A tuple of the figure and a flattened array of its axes.
+    """
+    num_cols = 3
+    num_rows = (num_fields + num_cols - 1) // num_cols
+
+    if fig_to_plot_on:
+        fig = fig_to_plot_on
+        fig.clear()
+    else:
+        fig = plt.figure(figsize=(figsize[0], num_rows * figsize[1]))
+
+    axs = fig.subplots(num_rows, num_cols, sharex=True, sharey=True).flatten()
+    return fig, axs
+
+
+def calculate_axis_limits(
+    centered_data: list[list[SpotData]],
+    fields: list[tuple[float, float]],
+    airy_disk_data: dict | None = None,
+    buffer: float = 1.05,
+) -> float:
+    """Calculates the axis limits to encompass all spots and Airy disks.
+
+    Args:
+        centered_data: The centered spot data.
+        fields: List of field coordinates.
+        airy_disk_data: The prepared Airy disk data, if any.
+        buffer: A multiplicative buffer to apply to the final limit.
+
+    Returns:
+        A float representing the symmetric axis limit (+/- limit).
+    """
+    max_radii = [
+        be.to_numpy(be.max(be.sqrt(sd.x**2 + sd.y**2)))
+        for field in centered_data
+        for sd in field
+    ]
+    max_geom_radius = max(max_radii) if max_radii else 0.01
+
+    if not airy_disk_data:
+        return max_geom_radius * buffer
+
+    rad_x = airy_disk_data["radii_x"]
+    rad_y = airy_disk_data["radii_y"]
+    centroids = airy_disk_data["airy_centers"]
+
+    max_extent = 0
+    for i in range(len(fields)):
+        offset = np.max(np.abs(centroids[i]))
+        radius = max(rad_x[i], rad_y[i], max_geom_radius)
+        max_extent = max(max_extent, offset + radius)
+
[…]

[…]
