### 2026-03-08 9a1342c
chore: add workflow to close inactive issues/PRs
