+++ .github/workflows/stale.yml
@@ -1,0 +1,26 @@
+name: Close inactive issues
+
+on:
+  schedule:
+    - cron: '30 1 * * *' # Runs at 01:30 UTC every day
+
+permissions:
+  issues: write
+  pull-requests: write
+
+jobs:
+  stale:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/stale@v9
+        with:
+          repo-token: ${{ secrets.GITHUB_TOKEN }}
+          stale-issue-message: 'Hi there! 👋 This issue has been automatically marked as stale because it has not had recent activity. It will be closed in 7 days if no further activity occurs. Thank you for your contributions. If you believe this is still relevant, please leave a comment to keep it open!'
+          stale-pr-message: 'Hi there! 👋 This pull request has been automatically marked as stale because it has not had recent activity. It will be closed in 7 days if no further activity occurs. Thank you for your contributions. If you are still working on this, please let us know or push a new commit!'
+          stale-issue-label: 'stale'
+          stale-pr-label: 'stale'
+          days-before-stale: 60
+          days-before-close: 7
+          exempt-issue-labels: 'bug,enhancement,security,pinned,help wanted,good first issue'
+          exempt-pr-labels: 'dependencies,security'
+          operations-per-run: 50
