### 2026-03-08 bf6868f
refactor: clean up Jones functionalities (#507)
