+++ tests/test_materials_extended.py
@@ -1,4 +1,3 @@
-
 import pytest
 from unittest.mock import MagicMock, patch, mock_open
 import numpy as np
@@ -10,6 +9,7 @@ from optiland.materials.abbe import AbbeMaterial, AbbeMaterialE
 # AbbeMaterial Tests (Buchdahl & E-line)
 # -----------------------------------------------------------------------------
 
+
 def test_abbe_buchdahl_init(set_test_backend):
     # Test initialization with 'buchdahl' model
     mat = AbbeMaterial(1.5, 64.0, model="buchdahl")
@@ -20,9 +20,11 @@ def test_abbe_buchdahl_init(set_test_backend):
     n = mat.n(0.5876)
     assert abs(n - 1.5) < 1e-3
 
+
 def test_abbe_invalid_model(set_test_backend):
     with pytest.raises(ValueError, match="Unknown model"):
         AbbeMaterial(1.5, 64.0, model="invalid")
+
 
 def test_abbe_e_init(set_test_backend):
     # Test AbbeMaterialE
@@ -32,6 +34,7 @@ def test_abbe_e_init(set_test_backend):
     # Test calculations (check near reference wavelength)
     n = mat.n(0.5461)
     assert abs(n - 1.5) < 1e-3
+
 
 def test_abbe_to_from_dict(set_test_backend):
     mat = AbbeMaterial(1.5, 64.0, model="buchdahl")
@@ -43,6 +46,7 @@ def test_abbe_to_from_dict(set_test_backend):
     assert mat2.model_name == "buchdahl"
     assert mat2.index.item() == 1.5
 
+
 def test_abbe_e_to_from_dict(set_test_backend):
     mat = AbbeMaterialE(1.5, 64.0)
     d = mat.to_dict()
@@ -51,9 +55,11 @@ def test_abbe_e_to_from_dict(set_test_backend):
     mat2 = AbbeMaterialE.from_dict(d)
     assert mat2.index.item() == 1.5
 
+
 # -----------------------------------------------------------------------------
 # Material Utils Tests
 # -----------------------------------------------------------------------------
+
 
 @patch("optiland.materials.material_utils.resources.files")
 def test_glasses_selection(mock_files, set_test_backend):
@@ -76,12 +82,13 @@ def test_glasses_selection(mock_files, set_test_backend):
         # Case 1: Broad range, only glass1 fits
         selection = material_utils.glasses_selection(0.35, 0.95)
         assert "glass1" in selection
-        assert "glass2" not in selection # 0.8 < 0.95
+        assert "glass2" not in selection  # 0.8 < 0.95
 
         # Case 2: Narrow range
         selection = material_utils.glasses_selection(0.5, 0.6)
         assert "glass1" in selection
         assert "glass2" in selection
+
 
 @patch("optiland.materials.material_utils.Material")
 def test_get_nd_vd(mock_material_cls, set_test_backend):
@@ -102,13 +109,14 @@ def test_get_nd_vd(mock_material_cls, set_test_backend):
             assert nd == 1.5
             assert vd == 64.0
 
+
 def test_downsample_glass_map(set_test_backend):
     # Create dummy glass dict
     glass_dict = {
         "G1": (1.5, 60.0),
         "G2": (1.51, 61.0),
         "G3": (1.8, 30.0),
-        "G4": (1.81, 31.0)
+        "G4": (1.81, 31.0),
     }
 
     # We expect 2 clusters essentially (low index/high abbe, high index/low abbe)
@@ -123,15 +131,15 @@ def test_downsample_glass_map(set_test_backend):
     assert min(nd_vals) < 1.6
[…]
