### 2026-03-08 2293683
feat: fix Jones calculus for 3D. Return p matrices in wavefront calcs. Create generalized polarization coatings.
