+++ optiland/jones.py
@@ -117,8 +117,103 @@ class JonesFresnel(BaseJones):
         return jones_matrix
 
 
+class JonesLinearPolarizer(BaseJones):
+    """Class representing a general linear polarizer in 3D space.
+
+    Args:
+        axis (tuple | list | be.ndarray): A 3D vector representing the transmission
+            axis in global coordinates (e.g., [1, 0, 0] for horizontal).
+    """
+
+    def __init__(self, axis):
+        self.axis = be.array(axis)
+        self.axis = self.axis / be.linalg.norm(self.axis)
+
+    def calculate_matrix(
+        self,
+        rays: RealRays,
+        reflect: bool = False,
+        aoi: be.ndarray = None,
+    ):
+        """Calculate the Jones matrix for the given rays.
+
+        Args:
+            rays (RealRays): Object representing the rays.
+            reflect (bool, optional): Indicates whether the rays are reflected.
+            aoi (be.ndarray, optional): Array representing the angle of incidence.
+
+        Returns:
+            be.ndarray: The calculated Jones matrix.
+        """
+        from optiland.rays.polarized_rays import PolarizedRays  # noqa: PLC0415
+
+        k0 = be.stack([rays.L0, rays.M0, rays.N0]).T
+        k1 = be.stack([rays.L, rays.M, rays.N]).T
+
+        s, p0, p1, o_in, o_out = PolarizedRays.get_local_basis(k0, k1)
+
+        # Broadcast axis to match rays
+        axis_b = be.broadcast_to(self.axis, k0.shape)
+
+        # Project transmission axis onto local incident and exit planes
+        ts_in = be.sum(axis_b * s, axis=1)
+        tp_in = be.sum(axis_b * p0, axis=1)
+        norm_in = be.sqrt(ts_in**2 + tp_in**2)
+        norm_in = be.where(norm_in == 0, be.ones_like(norm_in), norm_in)
+
+        ts_out = be.sum(axis_b * s, axis=1)
+        tp_out = be.sum(axis_b * p1, axis=1)
+        norm_out = be.sqrt(ts_out**2 + tp_out**2)
+        norm_out = be.where(norm_out == 0, be.ones_like(norm_out), norm_out)
+
+        us_in = ts_in / norm_in
+        up_in = tp_in / norm_in
+        us_out = ts_out / norm_out
+        up_out = tp_out / norm_out
+
+        jones_matrix = be.to_complex(be.zeros((be.size(rays.x), 3, 3)))
+        jones_matrix[:, 0, 0] = us_out * us_in
+        jones_matrix[:, 0, 1] = us_out * up_in
+        jones_matrix[:, 1, 0] = up_out * us_in
+        jones_matrix[:, 1, 1] = up_out * up_in
+        jones_matrix[:, 2, 2] = 1.0
+
+        return jones_matrix
+
+
+class JonesPolarizerH(JonesLinearPolarizer):
+    """Class representing the Jones matrix for a horizontal polarizer."""
+
+    def __init__(self):
+        super().__init__([1, 0, 0])
+
+
+class JonesPolarizerV(JonesLinearPolarizer):
+    """Class representing the Jones matrix for a vertical polarizer."""
+
+    def __init__(self):
+        super().__init__([0, 1, 0])
+
+
+class JonesPolarizerL45(JonesLinearPolarizer):
+    """Class representing the Jones matrix for a linear polarizer at 45 degrees."""
+
+    def __init__(self):
+        # 45 deg in X-Y plane
+        val = 1.0 / be.sqrt(be.array(2.0))
+        super().__init__([val, val, 0])
+
+
+class JonesPolarizerL135(JonesLinearPolarizer):
+    """Class representing the Jones matrix for a linear polarizer at 135 degrees."""
+
+    def __init__(self):
[…]
