### 2026-03-08 a7664b4
tests: update after jones refactor
