### 2026-03-09 d4fa29f
test: add jones and coatings tests for better coverage
