+++ tests/test_grid_phase.py
@@ -1,7 +1,11 @@
+from __future__ import annotations
+
 import pytest
+
 from optiland import backend as be
-from .utils import assert_allclose
 from optiland.phase.grid import GridPhaseProfile
+
+from .utils import assert_allclose
 
 
 @pytest.fixture
