### 2026-03-09 3f99071
feat: update wavefront data. Add vectorial PSF. Polarized rays cleanup.
