### 2026-03-09 cf72287
feat: add adapt method to allow coating to adapt to material on both sides. #246
