+++ docs/learning_guide.rst
@@ -72,6 +72,8 @@ This learning guide breaks down Optiland into a series of tutorials that cover t
    examples/Tutorial_6c_Multilayer_Stack
    examples/Tutorial_6d_Thin_Film_Optimization
    examples/Tutorial_6e_Color_Analysis_For_Thin_Film
+   examples/Tutorial_6f_Thin_Film_Coated_System
+   examples/Tutorial_6g_AR_Coating_System
 
 7. Advanced Optical Design
 --------------------------
