+++ optiland/colorimetry/constants.py
@@ -1,187 +1,20 @@
-"""
-Standard colorimetric data (CIE).
-Contains standard observers and illuminants.
-These values can be modified by the user at runtime if necessary.
+"""Standard colorimetric data (CIE).
+
+Data are sampled at 1 nm from 380 nm to 780 nm.
+Values are frozen in a local JSON file generated from official CIE datasets,
+so runtime does not depend on external colorimetry libraries.
 """
 
-# Standard wavelength range: 380nm to 780nm in 5nm steps
 from __future__ import annotations
 
-WAVELENGTHS_STD = list(range(380, 785, 5))
+import json
+from pathlib import Path
 
-# Illuminant D65 (Standard daylight) normalized
-# Standard approximate values in 5nm steps
-ILLUMINANT_D65 = [
-    50.0,
-    54.6,
-    60.0,
-    68.7,
-    78.2,
-    87.7,
-    95.7,
-    103.5,
-    110.3,
-    115.0,
-    117.8,
-    117.4,
-    114.4,
-    109.8,
-    104.5,
-    99.2,
-    94.8,
-    91.5,
-    89.6,
-    88.7,
-    88.7,
-    89.4,
-    90.7,
-    92.2,
-    93.7,
-    95.2,
-    96.5,
-    97.7,
-    99.1,
-    101.0,
-    103.3,
-    105.9,
-    108.7,
-    111.1,
-    112.8,
-    113.8,
-    114.1,
-    113.7,
-    112.4,
-    110.4,
-    107.9,
-    104.9,
-    101.9,
-    99.1,
-    96.3,
-    93.9,
-    91.9,
-    90.3,
-    89.3,
-    88.6,
-    88.1,
-    87.7,
-    87.1,
-    86.1,
-    84.7,
-    82.7,
-    80.3,
-    78.2,
-    76.3,
-    74.6,
-    72.9,
-    71.6,
-    70.2,
-    68.6,
-    67.1,
-    65.8,
-    64.5,
-    63.3,
-    62.1,
-    61.0,
-    60.0,
-    59.1,
-    58.2,
-    57.4,
-    56.6,
-    55.8,
-    55.0,
-    54.2,
-    53.4,
-    52.7,
-    52.0,
-]
+_DATA_PATH = Path(__file__).with_name("colorimetric_data_1nm.json")
+_DATA = json.loads(_DATA_PATH.read_text())
 
-# Standard CIE 1931 2 degree observer (x_bar, y_bar, z_bar)
-# Simplified data for example (380-780nm, 5nm)
-CIE_1931_2DEG = [
-    (0.0014, 0.0000, 0.0065),
-    (0.0022, 0.0001, 0.0105),
-    (0.0042, 0.0001, 0.0201),
-    (0.0076, 0.0002, 0.0362),
-    (0.0143, 0.0004, 0.0679),
-    (0.0232, 0.0006, 0.1102),
-    (0.0435, 0.0012, 0.2074),
-    (0.0776, 0.0022, 0.3713),
-    (0.1344, 0.0040, 0.6456),
-    (0.2148, 0.0073, 1.0391),
-    (0.2839, 0.0116, 1.3856),
-    (0.3285, 0.0168, 1.6230),
-    (0.3483, 0.0230, 1.7471),
-    (0.3481, 0.0298, 1.7826),
-    (0.3362, 0.0380, 1.7721),
-    (0.3187, 0.0480, 1.7441),
[…]
