### 2026-03-10 5750537
fix test for thin film #246
