### 2026-03-10 5e23708
feat: update normalization for PSF
