+++ optiland/thin_film/core.py
@@ -21,8 +21,13 @@ PolSP = Literal["s", "p"]
 
 
 def _complex_index(material: BaseMaterial, wavelength_um: float | Array) -> Array:
-    n = material.n(wavelength_um)
-    k = material.k(wavelength_um)
+    if be.get_backend() == "torch" and hasattr(wavelength_um, "detach"):
+        # Avoid material cache-key conversion that may require tensor->numpy bridge.
+        n = material._calculate_n(wavelength_um)
+        k = material._calculate_k(wavelength_um)
+    else:
+        n = material.n(wavelength_um)
+        k = material.k(wavelength_um)
     n = be.atleast_1d(n)
     k = be.atleast_1d(k)
     return be.asarray(n, dtype=be.complex128) + 1j * be.asarray(k, dtype=be.complex128)
@@ -78,10 +83,10 @@ def _tmm_coh(stack: ThinFilmStack, wavelength_um, theta0_rad, pol: PolSP):
     etas = _admittance(ns, coss, pol)
 
     # Id initial matrix
-    A = be.ones_like(eta0, dtype=be.complex128)
-    B = be.zeros_like(eta0, dtype=be.complex128)
-    C = be.zeros_like(eta0, dtype=be.complex128)
-    D = be.ones_like(eta0, dtype=be.complex128)
+    A = be.asarray(be.ones_like(eta0), dtype=be.complex128)
+    B = be.asarray(be.zeros_like(eta0), dtype=be.complex128)
+    C = be.asarray(be.zeros_like(eta0), dtype=be.complex128)
+    D = be.asarray(be.ones_like(eta0), dtype=be.complex128)
 
     for layer in stack.layers:
         n_l = layer.n_complex(wavelength_um)
