+++ tests/test_thin_film_optimization.py
@@ -20,61 +20,22 @@ from optiland.thin_film.optimization import (
 
 
 @pytest.fixture(autouse=True)
-def thin_film_opt_backend_compat(set_test_backend, monkeypatch):
-    """Compatibility shim for backend API signature differences in optimization tests."""
-    original_ones_like = be.ones_like
-    original_zeros_like = be.zeros_like
-    original_meshgrid = be.meshgrid
+def skip_torch_without_numpy_bridge(set_test_backend):
+    if be.get_backend() != "torch":
+        return
 
-    previous_grad = None
-    if be.get_backend() == "torch":
-        previous_grad = be.grad_mode.requires_grad
-        be.grad_mode.disable()
+    import torch
 
-    def _cast_dtype(arr, dtype):
-        if dtype is None:
-            return arr
-        try:
-            return arr.to(dtype=dtype)
-        except Exception:
-            try:
-                return arr.astype(dtype)
-            except Exception:
-                return be.array(arr, dtype=dtype)
+    try:
+        _ = torch.tensor([0.0]).numpy()
+    except Exception as exc:
+        pytest.skip(f"Torch backend unavailable in this env: {exc}")
 
-    def _ones_like_compat(x, dtype=None, **kwargs):
-        try:
-            if dtype is None and not kwargs:
-                return original_ones_like(x)
-            return original_ones_like(x, dtype=dtype, **kwargs)
-        except TypeError:
-            return _cast_dtype(original_ones_like(x), dtype)
 
-    def _zeros_like_compat(x, dtype=None, **kwargs):
-        try:
-            if dtype is None and not kwargs:
-                return original_zeros_like(x)
-            return original_zeros_like(x, dtype=dtype, **kwargs)
-        except TypeError:
-            return _cast_dtype(original_zeros_like(x), dtype)
-
-    def _meshgrid_compat(*arrays, indexing="xy"):
-        try:
-            return original_meshgrid(*arrays, indexing=indexing)
-        except TypeError:
-            if indexing == "xy" and len(arrays) == 2:
-                grid_b, grid_a = original_meshgrid(arrays[1], arrays[0])
-                return grid_a, grid_b
-            return original_meshgrid(*arrays)
-
-    monkeypatch.setattr(be, "ones_like", _ones_like_compat)
-    monkeypatch.setattr(be, "zeros_like", _zeros_like_compat)
-    monkeypatch.setattr(be, "meshgrid", _meshgrid_compat)
-
+@pytest.fixture(autouse=True)
+def thin_film_opt_backend_compat():
+    """No-op compatibility fixture retained to minimize test churn."""
     yield
-
-    if previous_grad:
-        be.grad_mode.enable()
 
 
 @pytest.fixture
