### 2026-03-10 0115549
Refactor Polarization Architecture (#508)
* fix: fix ABC in jones. fix coating update in surface. Ruff formatting.

* feat: fix Jones calculus for 3D. Return p matrices in wavefront calcs. Create generalized polarization coatings.

* tests: update after jones refactor

* test: add jones and coatings tests for better coverage

* docs: do not check tests or docs dirs for ruff

* docs: update jones and coating docs for new polarization functionality
