+++ tests/test_variable.py
@@ -1,31 +1,34 @@
-import optiland.backend as be
-import pytest
-from unittest.mock import patch
-import numpy as np
+from __future__ import annotations
 
-from optiland.optic import Optic
+from unittest.mock import patch
+
+import pytest
+
+import optiland.backend as be
 from optiland.coordinate_system import CoordinateSystem
 from optiland.geometries import (
     ChebyshevPolynomialGeometry,
-    ForbesQbfsGeometry,
     ForbesQ2dGeometry,
+    ForbesQbfsGeometry,
+    ForbesSurfaceConfig,
     PolynomialGeometry,
     ZernikePolynomialGeometry,
-    ForbesSurfaceConfig,
 )
-from optiland.optimization import variable, OptimizationProblem, OptimizerGeneric
-from optiland.samples.microscopes import Objective60x, UVReflectingMicroscope
-from optiland.samples.simple import AsphericSinglet, Edmund_49_847
 from optiland.materials.abbe import AbbeMaterial
-from optiland.optimization.variable.material import MaterialVariable
+from optiland.optic import Optic
+from optiland.optimization import OptimizationProblem, OptimizerGeneric, variable
 from optiland.optimization.scaling.identity import IdentityScaler
 from optiland.optimization.variable import (
+    AsphereCoeffVariable,
     DecenterVariable,
     TiltVariable,
-    AsphereCoeffVariable,
+    Variable,
     ZernikeCoeffVariable,
-    Variable
 )
+from optiland.optimization.variable.material import MaterialVariable
+from optiland.samples.microscopes import Objective60x, UVReflectingMicroscope
+from optiland.samples.simple import AsphericSinglet, Edmund_49_847
+
 from .utils import assert_allclose
 
 
@@ -48,7 +51,9 @@ class TestReciprocalRadiusVariable:
     def setup(self):
         self.optic = Edmund_49_847()
         self.reciprocal_radius_var = variable.ReciprocalRadiusVariable(self.optic, 1)
-        self.radius_var = variable.RadiusVariable(self.optic, 1, scaler=IdentityScaler())
+        self.radius_var = variable.RadiusVariable(
+            self.optic, 1, scaler=IdentityScaler()
+        )
         self.problem = OptimizationProblem()
         input_data = {
             "optic": self.optic,
@@ -391,7 +396,10 @@ class TestVariable:
         )
         assert_allclose(radius_var.value, 4.5326)
         captured = capsys.readouterr()
-        assert "Warning: unrecognized_attribute is not a recognized attribute" in captured.out
+        assert (
+            "Warning: unrecognized_attribute is not a recognized attribute"
+            in captured.out
+        )
 
     def test_invalid_type(self, set_test_backend):
         optic = Objective60x()
@@ -430,7 +438,9 @@ class TestTiltVariable:
 
     def test_get_value_no_scaling(self, set_test_backend):
         self.optic = Objective60x()
-        self.tilt_var_x = variable.TiltVariable(self.optic, 1, "x", scaler=IdentityScaler())
+        self.tilt_var_x = variable.TiltVariable(
+            self.optic, 1, "x", scaler=IdentityScaler()
+        )
         assert_allclose(self.tilt_var_x.get_value(), 0.0)
 
 
@@ -455,9 +465,9 @@ class TestDecenterVariable:
         self.decenter_var_y.update_value(5.0)
         assert_allclose(self.decenter_var_y.get_value(), 5.0)
 
-#    def test_invalid_axis(self, set_test_backend):
-#        with pytest.raises(ValueError):
[…]
