+++ tests/test_fields.py
@@ -1,7 +1,10 @@
-import optiland.backend as be
+from __future__ import annotations
+
 import pytest
 
+import optiland.backend as be
 from optiland import fields
+
 from .utils import assert_allclose
 
 
