### 2026-08-18 #9152 @manunicholasjacob
@matthewfeickert I saw you're still chasing reviewers here. I can take a slot if one's open.

The dual backend is what I'd dig into. Optiland exposes NumPy for CPU and PyTorch for GPU behind one interface, and "same interface, pick your engine" is easy to state and rarely tested by a reviewer. I'd want to know where the crossover sits. In my own work the answer has been counterintuitive: I've measured a workload where batch 1 loses to the CPU by 4x and batch 64 wins on the GPU by 29x. If the crossover for raytracing sits at a problem size most users never reach, the paper should say so.

I'd also test the GPU install path on a 4 GB card. That's a realistic configuration for people who'll try this, and probably not the one it was developed on.

Then the standard work: clean-environment install, test suite, docs and examples.

Background, briefly. I work on computer architecture, edge AI and LLM inference, with published roofline and energy measurements across ARM Cortex-A76 and x86. I maintain llama-roofline (10.5281/zenodo.21842493) and ml-systems-lab (10.5281/zenodo.21867055).

Scheduling, up front: I'm on the USENIX ATC 2026 Artifact Evaluation Committee from 22 September to 14 October and will be slower through it. Doesn't block me taking this, and I won't go quiet on you.

Reviewer database: @manunicholasjacob. ORCID 0009-0007-6589-6572.
