### 2026-08-18 #11170 @editorialbot
[REVIEW]: Optiland: Open-Source Optical Design Software in Python
**Submitting author:** @HarrisonKramer (<a href="http://orcid.org/0009-0000-5494-139X">Kramer Harrison</a>)
**Repository:** https://github.com/optiland/optiland
**Branch with paper.md** (empty if default branch): JOSS_paper
**Version:** v0.5.6
