### 2026-08-18 #11170 @matthewfeickert
@manunicholasjacob, @Jashcraf Thanks for agreeing to review this submission! This is the review thread for the paper. All of our communications will happen here from now on. :+1:

As you go over the submission, please check any items that you feel have been satisfied (and if you leave notes on each item that's even better). There are also links to the JOSS reviewer guidelines. I find it particularly helpful to also use the JOSS [review criteria](https://joss.readthedocs.io/en/pre-2026-submission/review_criteria.html) and [review checklist](https://joss.readthedocs.io/en/pre-2026-submission/review_checklist.html) docs as supplement/guides to the reviewer checklist @editorialbot will make for you.

> [!IMPORTANT]
> Please note that in 2026 [JOSS updated its submission scope requirements](https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future). These are reflected in the review criteria links above, but please be mindful of these revisions during your review.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, reviewers are encouraged to submit issues (I'd suggest for clarity prefixing the title with `[JOSS Review]`) and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#11170` so that a link is created to this Issue thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within about 4 weeks. Please let me know if either of you require some more time (that's perfectly okay). We can also use @editorialbot to set automatic reminders if you know you'll be away for a known period of time.
