### 2026-08-18 optiland/optiland#742 @manunicholasjacob
[JOSS Review] Ship the benchmark script behind Table 1, plus three small reference fixes
Reviewing for JOSS, openjournals/joss-reviews#11170.

**Benchmark script.** The paper reports original results in Table 1, from 2.3e6 to 2.3e8 ray-surfaces/s, with the 100x figure. I could not find the script behind those numbers. The `benchmarks/` directory on master covers the nonsequential harness, which is a different thing. I also found nothing matching the Table 1 sequential Cooke triplet measurement on either `master` or `JOSS_paper`.

That leaves the ray count, distribution, repetition count, and whether the reported number is a mean, median, or best case unstated. I reimplemented the benchmark from the prose to check the numbers. A reader should not have to do that, especially here because the numbers move quite a bit depending on those choices. I measured run-to-run spread between 4% and 60%, depending on the configuration.

For what it is worth, my reimplementation does support the absolute GPU figures. Scaled for SM count my measurement lands within 0.3% of the reported throughput, so this is a request for the method rather than a doubt about the result.

Adding the script alongside the existing `benchmarks/` tree would resolve this. It would also let readers reproduce the comparison on their own hardware, which is the interesting part. Two details should also be stated next to the table: the ray count and how many repetitions the reported figure summarizes.

**References.** `@editorialbot` flagged six entries with no DOI. I checked all six against Crossref, and only one is genuinely findable:

* `lam2015numba` has a DOI: 10.1145/2833157.2833162

The other five have none. The PyTorch NeurIPS 2019 paper has no Crossref DOI, and RayOptics, OpticStudio, CODE V, and your own documentation are not the kind of items that carry one. So the bot's list is mostly a false alarm. Only the Numba entry needs changing.

Two smaller things are worth fixing in the same area:

* `2020SciPy-NMeth` is in `paper.bib` but is never cited in the paper text. SciPy is a runtime dependency, and the paper cites NumPy, PyTorch, and Numba in the same breath. It is probably worth citing SciPy alongside them rather than dropping the entry.
* The `codev` entry has the title `CODE {V`, with an unbalanced brace that will render incorrectly.

The citation syntax itself is consistent and correct throughout. No citation in the text is missing a bib entry.
