### 2026-08-18 optiland/optiland#743 @manunicholasjacob
[JOSS Review] GPU throughput collapses before OOM on a 4 GB card
From the JOSS review, openjournals/joss-reviews#11170. This is not a bug, but it is worth a line in the documentation.

I traced a Cooke triplet on an RTX 3050 Laptop with 4 GB of VRAM. Throughput peaked at around 1.6M rays and then started falling well before the card ran out of memory:

| Rays | GPU float64 | Peak VRAM | GPU float32 | Peak VRAM |
| ---: | ---: | ---: | ---: | ---: |
| 822k | 2.77e7 | 0.53 GiB | 2.49e7 | |
| 1.64M | **3.15e7 peak** | 1.45 GiB | **5.00e7 peak** | 0.72 GiB |
| 3.29M | 2.56e7 | 2.96 GiB | 3.05e7 | 1.45 GiB |
| 6.58M | **2.82e6** | 5.84 GiB | 1.56e7 | 2.89 GiB |
| 13.2M | out of memory | | 7.79e6 | 5.79 GiB |

The float64 result drops 11x at 6.58M rays and then runs out of memory. The collapse starts once the allocator passes physical VRAM and the driver begins spilling. Users therefore hit a performance cliff before they get an actual out-of-memory error, which is the harder failure to diagnose.

Consumer GPUs with 4 to 8 GB of VRAM are likely to be common among users running this locally. A short note that throughput is not monotonic with ray count, with a rough sweet spot, would probably save some trial and error.

Method, in case it is useful: medians of repeated runs on an otherwise idle machine, with peak VRAM taken from `torch.cuda.max_memory_allocated`.
