### 2026-08-18 optiland/optiland#744 @manunicholasjacob
[JOSS Review] Paper example 1 ends on an interactive call, so it cannot be run as a script
From the JOSS review, openjournals/joss-reviews#11170. Very minor, filing for completeness.

Example 1 in the paper ends with `lens.draw3D()`, which opens an interactive VTK window. When I ran the paper's examples in order as a script, it appeared to hang rather than finish. It took me a little time to work out that the window was waiting for interaction.

`Tutorial_1a` already has a note saying, "Note this opens a new window." The paper example is the only one missing it. A one-line comment in the code block would be enough.

Examples 2 and 3 run cleanly. Example 4 is covered by the version point I raised in the review thread rather than here, since that issue is already fixed upstream.
