### 2026-08-18 #11170 @manunicholasjacob
@matthewfeickert checklist generated, and I have done a full pass on install, tests, the performance claims, the docs examples and the paper.

**The main thing I want to flag is the version.** The submission is v0.5.6 and the repo is at v0.6.1, and that gap is doing real damage to what a reader gets:

- Installing **v0.5.6 specifically**, which is on PyPI and is what the archive would point at, `to_dict()` raises on current NumPy, so an optical system cannot be saved or loaded. Nine tests fail, and `docs/examples/Tutorial_1c_Save_and_Load_Files.ipynb` fails with it. **Fixed in v0.5.9.** To be clear, a plain `pip install optiland` today gets 0.6.1 and is unaffected; this bites only someone pinning the reviewed version.
- The paper's Example 4 imports `optiland.ml`, which does not exist in v0.5.6. **Added in v0.5.7.**
- v0.5.6 has no code of conduct, which bears on the 2026 open-source practice criterion. **Present on master.**

I verified each of those by installing and running v0.5.6, v0.5.7, v0.5.8, v0.5.9, v0.6.0 and v0.6.1 rather than reading diffs, on both Windows and Linux. All three are resolved by v0.5.9 or earlier, so bumping the submitted version and re-archiving would clear them in one step. I have not filed issues for them, since there is nothing left to fix upstream.

The reason the NumPy break survived to a release at all is worth a line: `ci.yml` runs `uv lock --locked`, which pins NumPy 2.1.3, where the offending conversion is still only a DeprecationWarning. A fresh `pip install` gets 2.4.6, where it raises. A CI job that resolves dependencies fresh alongside the locked one would catch that class of problem. That is a suggestion, not a blocker.

Issues filed for the things that are still open on master: optiland/optiland#742, optiland/optiland#743, optiland/optiland#744.

**On the performance claims,** which is what I offered to check. The absolute GPU throughput holds up. I measured 7.99e7 ray-surfaces/s float32 on an RTX 3050 Laptop against the paper's 2.3e8 on an RTX 3070. My card has 16 SMs to the 3070's 46, a ratio of 0.348, and the measured throughput ratio is 0.347, so the reported figure is consistent with the hardware difference.

The 100x speedup does not reproduce, and I do not think the GPU is the reason. I get 42x. My NumPy CPU baseline on the same i7-12700H is faster than the paper's, 1.89e6 to 4.17e6 depending on ray count against the reported 2.3e6, most likely a newer NumPy. The GPU scales down with the weaker card while the CPU baseline holds, so the ratio compresses. The absolute figures look like the robust claim and the multiple like the fragile one, since it moves with the baseline, the ray count and the NumPy version. Since 100x is quoted in the Research Impact Statement, it may be worth presenting as measured on a specific pairing rather than as a property of the software. Medians of 15 repetitions on an idle machine.

I also checked cross-platform, since my main pass was Windows only: `pip install optiland` on Ubuntu under WSL2 with Python 3.12 is clean in 51 s, and the Cooke triplet paraxial focal length comes out at 49.9998 on both platforms, on different NumPy versions.

**Everything else looks good to me.** I ran the tutorials rather than reading them, six of eight pass on v0.5.6 with the two failures explained above. 11 of 12 feature-table claims verify. On the 2026 criteria you flagged, this is a strong submission: public since August 2017, 40 contributors, 453 PRs, 222 issues, 18 releases, and a named downstream user in Visisipy.

I will work through the checklist items now and leave notes as I go.
