### 2026-08-18 #11170 @manunicholasjacob
Twenty-five of the thirty-one items are checked. Six remain open. Two actions close five of those six, and the last is a question for the authors:

* A version bump closes Functionality and Community Guidelines.
* [optiland/optiland#742](https://github.com/optiland/optiland/issues/742) closes Reproducibility and References.
* Either a benchmark script or clarification of the reported pairing closes Performance.
* A one-line answer from the authors closes Contribution and Authorship.

### Contribution and authorship

@HarrisonKramer has 2,351 commits and is clearly the major contributor. I could map three of the four paper authors to GitHub accounts: Harrison Kramer (2,351 commits), `manuelFragata` (119), and `crnh` (68, with a profile matching LUMC). I could not map Grégoire Hein to a GitHub account from public metadata.

JOSS authorship also covers non-code contributions, so this may be completely fine. I am not raising it as a finding. I am just asking the authors to confirm.

### Reproducibility

The paper reports original results in Table 1, but I could not find a benchmark script on either `master` or `JOSS_paper`. I reimplemented the method from the paper's description and filed this as optiland/optiland#742.

### Functionality

I verified 11 of the 12 README feature-table claims on v0.5.6. These include Zemax import, Zernike, Monte Carlo tolerancing, sensitivity analysis, GlassExpert, polarization, coatings, the material database, PSF/MTF, freeform surfaces, and VTK visualization.

The remaining claim is `optiland.ml`, which is used by paper example 4. It is absent from v0.5.6 and arrived in v0.5.7.

As a spot check, the Cooke triplet gives a paraxial focal length of 49.9998 for the 50 mm design. I got the same result on Windows and Linux.

### Performance

The paper reports 2.3e6 for NumPy CPU, 7.1e6 for torch CPU, 5.7e7 for GPU f64, and 2.3e8 for GPU f32 on an RTX 3070, described as "100x."

I measured 1.89e6 to 4.17e6, 8.66e6, 3.45e7, and 7.99e7 respectively on an RTX 3050 Laptop 4 GB.

The absolute GPU figure holds. My card has 16 SMs versus 46 on the RTX 3070, giving a ratio of 0.348. The measured throughput ratio is 0.347.

The 100x result did not reproduce. I measured 42x. The difference comes from the CPU baseline, not the GPU. My NumPy baseline on the same i7-12700H is faster than the paper's. I used the median of 15 repetitions on an idle machine.

This could be closed by shipping the benchmark script, or by presenting 100x explicitly as the result for the specific hardware pairing used in the paper.

### Community guidelines and references

`CONTRIBUTING.md` is present, along with four issue templates. Issues and discussions are enabled. `CODE_OF_CONDUCT.md` is absent in v0.5.6 but present on master.

There are 12 bibliography entries and 11 citations. `2020SciPy-NMeth` is never cited. The DOI for `lam2015numba` exists: 10.1145/2833157.2833162. The `codev` title has an unbalanced brace. I also checked the six missing DOIs flagged by @editorialbot against Crossref. Only the Numba DOI exists, so the other five can be ignored.

### Additional checks

I installed v0.5.6 cleanly in 350 seconds with 29 packages. The `[gui]` extra also installed cleanly. The documented CUDA path resolved to working torch 2.7.1+cu118. A clean Ubuntu WSL2 install with Python 3.12 took 51 seconds.

I ran all 34 tutorials. Twenty-four passed on v0.5.6. Four require interactive `draw3D()`, two are global optimisers that exceeded my 200 second timeout, one hits the save/load break, and three are tutorials that have since been renamed or removed upstream.

The test suite has 53 files with 1,546 passing tests. The documented core-only install produces 49 failures and errors: 35 are `ValueError: Unknown backend` because torch is absent and the backend parameter does not skip those tests, and 14 are the NumPy `to_dict` break described above. Master largely fixes the torch half with `importorskip` in 11 files, compared with one in v0.5.6.

The documentation has 19 API pages, 150 `autosummary` entries, and 121 `automodule` entries. Docstrings cover 870 of 1,031 public symbols, or 84.4%.

The build-versus-contribute justification for not extending RayOptics is also worth noting. It directly addresses the criterion and is often missing from submissions.

### Limits

I did not exercise the GUI interactively. It installs, imports, and its entry point resolves, but the VTK widget needs a real OpenGL context that I could not provide headlessly. I am also not assessing the optical correctness of the results.

My environment was Windows 11 with Python 3.11.4 and Ubuntu WSL2 with Python 3.12, on an i7-12700H with an RTX 3050 Laptop 4 GB.
