### 2026-08-19 #11170 @HarrisonKramer
@matthewfeickert - following the recommendation of @manunicholasjacob, can you please bump the submitted version to v0.6.1?
