### 2026-08-19 #11170 @HarrisonKramer
@manunicholasjacob, thanks for the comprehensive review! All makes sense and I'll work through the issues you opened in the Optiland repo.

On authorship, Grégoire Hein is @drpaprika. I see him listed as "DrPaprikaa" and "drpaprika" in the _Commit count by author_ list in openjournals/joss-reviews#9152. Grégoire is a long-term contributor to the project.

Thanks,
Kramer
