### 2026-08-19 750e9d3
docs: bib fixes following #742
