+++ paper/paper.bib
@@ -112,7 +112,7 @@ url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-perf
 }
 
 @misc{codev,
-  title        = {CODE {V}},
+  title        = {CODE V},
   howpublished = {Commercial optical system design software},
   note         = {Used in optical design (see general overview)},
   year         = {2025},
@@ -124,7 +124,8 @@ url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-perf
   author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
   pages={1--6},
-  year={2015}
+  year={2015},
+  doi={10.1145/2833157.2833162}
 }
 
 @article{polyanskiy2024,
