### 2026-08-19 1e78110
feat(benchmarks): add sequential Cooke-triplet throughput script (#742)
Ships the benchmark script behind the paper's Table 1 (previously
reimplemented from prose by a JOSS reviewer). Ports the manual notebook
procedure (optiland/jupyter/JOSS_NumPy_vs_PyTorch_Speed_Check.ipynb) into
benchmarks/sequential/cooke_triplet.py, mirroring the existing
benchmarks/nonsequential/ layout. Not part of the pytest suite or CI.

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
