+++ benchmarks/sequential/cooke_triplet.py
@@ -1,0 +1,250 @@
+"""Sequential ray-tracing throughput benchmark behind the JOSS paper's Table 1.
+
+This is a direct port of the manual procedure used to produce the numbers in
+``paper/paper.md`` (originally run interactively in
+``optiland/jupyter/JOSS_NumPy_vs_PyTorch_Speed_Check.ipynb``): trace a fixed
+random distribution of rays through a Cooke triplet once per backend
+configuration and time only the ``surface_group.trace()`` call. Ray
+generation, backend setup, and any warm-up are excluded from the timed
+region.
+
+Each configuration is measured with a *single* trace call, matching how the
+published table was produced -- this script does not average repeated runs.
+Reviewers/readers have measured 4-60% run-to-run spread depending on machine
+load and configuration, so treat single-run numbers as illustrative rather
+than precise; pass ``--repeats`` to see that spread directly on your own
+hardware.
+
+Not run as part of the pytest suite or CI -- see ``pyproject.toml``
+(``testpaths = ["tests"]``). Run directly::
+
+    .venv/Scripts/python.exe -m benchmarks.sequential.cooke_triplet
+    .venv/Scripts/python.exe -m benchmarks.sequential.cooke_triplet --repeats 5
+    .venv/Scripts/python.exe -m benchmarks.sequential.cooke_triplet \
+        --output results.json
+
+Kramer Harrison, 2026
+"""
+
+from __future__ import annotations
+
+import argparse
+import dataclasses
+import json
+import statistics
+import sys
+import time
+
+import optiland.backend as be
+from optiland import distribution
+from optiland.samples.objectives import CookeTriplet
+
+# Matches the notebook: 10M rays for the float64 configurations, 20M for the
+# float32 GPU configuration (float32 traces fast enough that 10M rays no
+# longer dominates fixed per-call overhead).
+DEFAULT_NUM_RAYS = 10_000_000
+DEFAULT_NUM_RAYS_FLOAT32 = 20_000_000
+DEFAULT_SEED = 0
+
+CONFIGS = [
+    # (label, backend, precision, device)
+    ("numpy_cpu_f64", "numpy", None, None),
+    ("torch_cpu_f64", "torch", "float64", "cpu"),
+    ("torch_gpu_f64", "torch", "float64", "cuda"),
+    ("torch_gpu_f32", "torch", "float32", "cuda"),
+]
+
+
+@dataclasses.dataclass(frozen=True)
+class BenchmarkResult:
+    """One backend configuration's measured throughput.
+
+    Attributes:
+        label: Configuration name, e.g. "torch_gpu_f32".
+        num_rays: Rays traced.
+        num_surfaces: Surfaces in the Cooke triplet (including object/image).
+        trace_times_sec: One elapsed time per repeat, in seconds.
+        rays_surfaces_per_sec: ``num_rays * num_surfaces / time`` per repeat.
+    """
+
+    label: str
+    num_rays: int
+    num_surfaces: int
+    trace_times_sec: list[float]
+    rays_surfaces_per_sec: list[float]
+
+
+def _build_rays(num_rays: int, seed: int):
+    """Build a Cooke triplet and a fixed random distribution of rays for it.
+
+    Args:
+        num_rays: Number of rays to generate across the pupil.
+        seed: Seed for the pupil-sampling RNG, for reproducibility.
+
+    Returns:
+        A ``(lens, rays)`` tuple, ready for ``lens.surface_group.trace(rays)``.
+    """
+    lens = CookeTriplet()
+    d = distribution.RandomDistribution(seed=seed)
+    d.generate_points(num_rays)
[…]
