+++ benchmarks/sequential/__init__.py
@@ -1,0 +1,6 @@
+"""Sequential ray-tracing throughput benchmark (Cooke triplet, JOSS Table 1).
+
+Kramer Harrison, 2026
+"""
+
+from __future__ import annotations
