### 2026-08-19 optiland/optiland#742 @HarrisonKramer
Thanks for the thorough review. All items addressed on `JOSS_paper`:

**Benchmark script.** Added `benchmarks/sequential/cooke_triplet.py`, a direct port of the procedure used to produce Table 1 (previously only run manually in a notebook). It traces the same Cooke triplet with the same ray counts as the paper (10M rays, 20M for the float32 GPU configuration) and is not part of the pytest suite or CI, consistent with the existing `benchmarks/nonsequential/` tree. A `--repeats` flag reruns a configuration multiple times so readers can see run-to-run spread directly on their own hardware.

The text next to Table 1 now states the ray count and that each reported figure is a single measurement rather than an average. I reran all four configurations on the original benchmark machine under the current release and confirmed throughput is consistent with the published numbers, so the table itself is unchanged.

**References.**
- `lam2015numba` now has `doi = {10.1145/2833157.2833162}`.
- The `codev` entry's title is now `{CODE V}`, fixing the brace issue.
- `2020SciPy-NMeth` is cited in the text alongside NumPy, PyTorch, and Numba (Optimization and Tolerancing bullet).

Closing this issue, as I think it's now complete, but let me know if anything's missing and I can reopen.

Best,
Kramer
