### 2026-08-19 f9aa4be
docs: note that Example 1's draw3D() opens an interactive window (#744)
Matches the existing note in Tutorial_1a. Running the paper's examples
in order as a script otherwise appears to hang on the VTK window.

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
