### 2026-08-19 optiland/optiland#744 @HarrisonKramer
Fixed. Added `# Visualize in 3D (note: this opens a new window)` above `lens.draw3D()` in Example 1, matching the note already in `Tutorial_1a`.

I believe this closes it out, but flag anything I missed and I'll pick it back up.

Best,
Kramer
