### 2026-08-19 95165be
docs: note GPU throughput cliff on VRAM-limited cards (#743)
Add a tip about non-monotonic PyTorch/CUDA trace throughput as VRAM
fills up, per JOSS review feedback in #743.

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
