### 2026-08-19 #11170 @manunicholasjacob
Thanks @HarrisonKramer, that was quick. I have gone through each of the fixes.

**Confirmed on `JOSS_paper`:** `benchmarks/sequential/cooke_triplet.py` is present and runs; `lam2015numba` now carries `doi={10.1145/2833157.2833162}`; the `codev` title is `{CODE V}`; `2020SciPy-NMeth` is cited in the text. The paper now states the ray count and that the figures are single measurements. On authorship, `DrPaprikaa` and `drpaprika` both appear in the commit history, so that answers my question. optiland/optiland#742 and optiland/optiland#744 are correctly closed from my side, no need to reopen either.

**I also need to correct something I said, and it goes in your favour.** I reported that the 100x speedup did not reproduce and that I measured 42x. That was at my own choice of ray count, before your script told me yours. Running it on an RTX 3050 Laptop (4 GB):

| Rays | float32 GPU (M ray-surf/s) | Speedup vs NumPy |
|---:|---:|---:|
| 100k | 20.17 | 3.6x |
| 1M | 90.20 | 35.4x |
| 2M | 74.19 | 45.6x |
| 4M | 37.83 | 20.4x |
| 10M | does not complete | |

The speedup depends strongly on ray count, and on my card it peaks around 1 to 2M rays before VRAM pressure pulls it back down. Your table is measured at 10M, and 20M for float32, which a 4 GB card cannot reach. So I was measuring a different operating point rather than failing to reproduce yours, and the trend I can see is consistent with your figure. Apologies for the earlier framing.

It also makes optiland/optiland#743 more relevant than I first thought. On a small card the throughput cliff arrives well before the memory limit does, which is precisely why I could not get near your ray counts. It is a documentation suggestion rather than a defect, so treat it as optional.

One small heads-up from re-checking: the paper is now 2076 words, up from 1988, since the benchmark paragraph was added. @editorialbot flagged the earlier figure, so it may be worth asking @matthewfeickert whether that matters rather than having it come up later.

That leaves two unticked items, both waiting on the version bump you have already requested rather than on anything from you: `optiland.ml` for Functionality, and the code of conduct for Community guidelines. Both are in v0.6.1. Once the version is set I will tick those and post my recommendation.
