### 2026-08-19 optiland/optiland#743 @HarrisonKramer
Thanks for the detailed report and methodology. I reproduced this on a different card (RTX 3070 Laptop, 8 GB VRAM) using the sequential tracer on a Cooke triplet, and saw the same pattern: throughput peaked, then collapsed about 5-6x (float64: 4.2e6 -> 8.0e5 rays/s; float32: 1.4e7 -> 2.3e6 rays/s) as VRAM approached the card's limit, followed by a hard OOM at the next ray count. That matches your observation that the performance cliff hits before the OOM does.

Added a note about this to the docs in #745 (merged), so users hitting the same thing have a pointer to what's going on and some workarounds (checking torch.cuda.max_memory_allocated(), reducing num_rays/batching, or using float32).

Closing this as complete, thanks again for the thorough writeup.

Best,
Kramer
