### 2026-08-19 #11170 @matthewfeickert
> I worked through the review checklist and did a full pass covering installation, tests, performance, documentation, examples, and the paper.

Thanks @manunicholasjacob for your fast and detailed review so far. 👍 If there are things that you can not complete in your checklist please open JOSS Review Issues on the project repository to discuss them rather than here.

> The main issue I found is the submitted version. The paper is submitted against v0.5.6, while the repository is now at v0.6.1.

@manunicholasjacob For all JOSS reviews the version number is a final artifact. Reviewers should start with the released distributions and the published documentation and also cover the current state of the repository. In almost all cases a JOSS review finishes with a new release of the software and so the version number at the start of the review is not something to focus on.

---

> following the recommendation of [@manunicholasjacob](https://github.com/manunicholasjacob), can you please bump the submitted version to v0.6.1?

@HarrisonKramer We wait to bump the version until we are in the submitting for publication process. Don't worry about the version number during the review.

> One small heads-up from re-checking: the paper is now 2076 words, up from 1988, since the benchmark paragraph was added.

@manunicholasjacob @HarrisonKramer This will come up in editorial review as well, but the paper is too long at the moment. As mentioned in [the docs](https://joss.readthedocs.io/en/latest/paper.html), a JOSS paper has required sections and should not significantly expand beyond them. Code examples should not exist in the paper unless there is a significantly compelling reason, nor should benchmarking; these should be added to the documentation website instead, if they do not already exist.
