### 2026-08-19 #11170 @manunicholasjacob
Understood on both counts, thank you.

On the version: I have re-assessed the two items I was holding against the current repository rather than the submitted tag. `optiland.ml` and the code of conduct are both present on master, so my checklist is now complete at 31 of 31.

On paper length: my request in optiland/optiland#742 asked for the ray count and repetition detail to be stated next to Table 1, which is part of why the count went up. Given your steer, that methodology belongs in the documentation and in the `benchmarks/sequential/cooke_triplet.py` script @HarrisonKramer added rather than in the paper itself, so please treat my request as satisfied by the script alone and disregard the part about the paper text.

**Recommendation.** From my side I am happy to recommend acceptance, once the paper length and code-example points you raised are resolved. Every issue I opened has been addressed and closed, and I verified each fix myself. The software is well tested, the documentation is thorough, and the performance claims hold up when measured. I did not assess the optical correctness of the results.

Noted on keeping detail in repository issues.
