### 2026-08-19 #11170 @matthewfeickert
@Jammy2211 thank you for offering to join the review. As mentioned in https://github.com/openjournals/joss-reviews/issues/9152#issuecomment-5333816007, it is useful to have more than the minimal number of reviewers. Please read through https://github.com/openjournals/joss-reviews/issues/11170#issuecomment-5334048252 and generate your reviewer checklist. If you have any questions please feel free to ask them here. 👍
