### 2026-08-25 #11170 @Jashcraf
Hi y'all I've recovered from the hurricane and SPIE and am ready to start the review
