### 2026-08-25 optiland/optiland#754 @Jashcraf
[JOSS Review] "Your first 5 minutes" spot.view() diagram doesn't plot
This issue is part of my JOSS review at https://github.com/openjournals/joss-reviews/issues/11170#issuecomment-5334048252

I'm running through the example at https://optiland.readthedocs.io/en/latest/quickstart.html

Fresh conda environment with python 3.10 installed. `pip install optiland` to install optiland. Run the following in the conda environment.

```python
from optiland import optic
from optiland.analysis import SpotDiagram

lens = optic.Optic(name="Singlet")
lens.surfaces.add(index=0, radius=float("inf"), thickness=float("inf"))  # object at infinity
lens.surfaces.add(index=1, radius=50.0, thickness=5.0, material="N-BK7", is_stop=True)
lens.surfaces.add(index=2, radius=-50.0, thickness=0.0)
lens.surfaces.add(index=3)  # image plane
lens.set_aperture(aperture_type="EPD", value=10.0)
lens.fields.set_type("angle")
lens.fields.add(y=0.0)
lens.wavelengths.add(value=0.5876, is_primary=True)

# Works
lens.draw3D()

lens.updater.image_solve()   # moves image surface to paraxial focus

# Works
lens.draw3D()

rays = lens.trace(Hx=0, Hy=0, wavelength=0.5876, num_rays=64, distribution="hexapolar")
print("x range:", rays.x.min(), "to", rays.x.max())
print("y range:", rays.y.min(), "to", rays.y.max())

# chief ray for the on-axis field
ray = lens.trace_generic(Hx=0, Hy=0, Px=0, Py=0, wavelength=0.5876)
print("x range:", ray.x.min(), "to", ray.x.max())
print("y range:", ray.y.min(), "to", ray.y.max())

spot = SpotDiagram(lens)

# NOTE: No Plot
spot.view()

```
