### 2026-08-25 optiland/optiland#754 @crnh
Hi! Thanks for opening this issue. Python 3.10 is not supported as of Optiland 0.6.2. Can you retry this with the latest Optiland version and Python >= 3.11? I ran your code and the spot diagram is shown correctly with Optiland 0.6.2, Python 3.11.
