### 2026-09-03 optiland/optiland#766 @Jashcraf
[JOSS Review] Cheat sheet step 7 missing display
This issue is part of my JOSS review, available at https://github.com/openjournals/joss-reviews/issues/11170#issuecomment-5422104802

Review environment includes:
- python 3.11.15
- optiland 0.6.2
- torch 2.14.0

I'm going through and verifying optiland's documented functionality based on the stable docs and the latest release (0.6.2). "Your First 5 Minutes" looks good 👍 . At step 7 (shown above), the documentation instructs the user to run `lens.draw()`. Executing this after loading a lens, say
```python
from optiland.samples.objectives import CookeTriplet
lens = CookeTriplet()
lens.draw(num_rays=5, distribution="line_y")
```

The lens will not display, which is dissimilar behavior to `lens.draw3D()`. However, calling `plt.show()` will plot the lens
```python
from optiland.samples.objectives import CookeTriplet
import matplotlib.pyplot as plt
lens = CookeTriplet()
lens.draw(num_rays=5, distribution="line_y")
plt.show()
```

This could confuse new users, so I wanted to point this out to see if this is intended behavior. I'd recommend comend including instructions in element 7 of the cheat sheet so that the user can copy + paste the code and get a plot, however is preferable to do this.
