### 2026-09-03 optiland/optiland#767 @Jashcraf
[JOSS Review] `Wavefront' object has no attribute 'view'
This issue is part of my JOSS review, available at https://github.com/openjournals/joss-reviews/issues/11170#issuecomment-5422104802

Review environment includes:
- python 3.11.15
- optiland 0.6.2
- torch 2.14.0

Step 12 of the cheat sheet instructs the user to perform a wavefront analysis and Zernike polynomial decomposition. However, the call to `Wavefront.view()` appears to break with `AttributeError: 'Wavefront' object has no attribute 'view'`

```python
from optiland.samples.objective import CookeTriplet
from optiland.wavefront import Wavefront, ZernikeOPD

lens = CookeTriplet()

# NOTE: AttributeError: 'Wavefront' object has no attribute 'view'
wf = Wavefront(lens, field=(0, 0), wavelength="primary")
wf.view()
```
