### 2026-09-03 optiland/optiland#768 @Jashcraf
[JOSS Review] Cheat sheet step 12 missing display
This issue is part of my JOSS review, available at https://github.com/openjournals/joss-reviews/issues/11170#issuecomment-5422104802

Review environment includes:
- python 3.11.15
- optiland 0.6.2
- torch 2.14.0

This is identical behavior to #766 , the call to `zfit.view()` does not actually display the fit wavefront. However, a `plt.show()` after this generates the Zernike Fringe fit map.
