### 2026-09-03 optiland/optiland#766 @Jashcraf
Since this is the same behavior across a couple steps, I'll just add here. Looks like the behavior is identical for Step 13 (and step 12 #768, which I think I'll close)
