### 2026-09-03 optiland/optiland#768 @Jashcraf
Closing so that this behavior is tracked on a single issue #766
