### 2026-09-03 optiland/optiland#771 @Jashcraf
[JOSS Paper] Software Design does not include substantive discussion of trade-offs
This issue is part of my review at https://github.com/openjournals/joss-reviews/issues/11170#issuecomment-5422104802

**The JOSS requirements state that:**
Does the paper include a "Software Design" section that explains the trade-offs weighed, the design/architecture chosen, and why it matters? Is the content compelling and does it demonstrate meaningful design thinking?

The software design section of the paper states that optiland's goals are
- provide a fully-featured lens design software
- provide seamless integration with ML

To accomplish
- Use an object-oriented framework to ensure that the interface is intuitive
- Utilize a hybrid backend for swapping to differentiable backend

This is a good and concise high-level description, but I'd appreciate if the authors can speak to the trade-offs weighed in the design of the software. Specifically, it would be interesting to understand why PyTorch was chosen v.s. other GPU-accelerated and automatic differentiation libraries (e.g. JAX, Tensorflow).
