### 2026-09-03 optiland/optiland#772 @Jashcraf
[JOSS Review] Installation page does not include list of dependencies
This issue is part of my JOSS review available at https://github.com/openjournals/joss-reviews/issues/11170#issuecomment-5422104802

JOSS requests that the installation instructions include "a clearly-stated list of dependencies". This doesn't appear on the installation page https://optiland.readthedocs.io/en/stable/quickstart.html#install

Please list all dependencies on the installation page

Having torch be an optional backend is a great idea, and is well-documented with optional installation instructions 👍
