### 2026-09-03 #11170 @Jashcraf
@matthewfeickert a question for what JOSS prefers

the optiland documentation does not include a section marked "Statement of Need", but I would argue that the homepage covers the content that would make up a statement of need https://optiland.readthedocs.io/en/stable/index.html

Should I request that the authors include a specific section titled "Statement of Need"?
