### 2026-09-03 #11170 @matthewfeickert
> the optiland documentation does not include a section marked "Statement of Need", but I would argue that the homepage covers the content that would make up a statement of need https://optiland.readthedocs.io/en/stable/index.html
>
> Should I request that the authors include a specific section titled "Statement of Need"?

@Jashcraf no, that's not necessary. For the documentation we ask that

> **A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

but you don't have to explicitly put the words "statement of need". For the paper we do require this though

> **A statement of need (required section)**: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?
