### 2026-09-03 optiland/optiland#772 @crnh
Hi! Thanks for opening this review issue. I checked the JOSS review guidelines for this one and interpret them differently.

According to the review checklist:

> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

The [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation) elaborate on this:

> Software dependencies should be clearly documented and their installation handled by an automated procedure. Where possible, software installation should be managed with a package manager. However, this criterion depends upon the maturity and availability of software packaging and distribution in the programming language being used. For example, Python packages should be pip installable, and have adopted [packaging conventions](https://packaging.python.org/), while Fortran submissions with a Makefile may be sufficient.
>
>    **Good**: The software is simple to install, and follows established distribution and dependency management approaches for the language being used
>    **OK**: A list of dependencies to install, together with some kind of script to handle their installation (e.g., a Makefile)
>    **Bad** (not acceptable): Dependencies are unclear, and/or installation process lacks automation

We adhere to 'established distribution and dependency management approaches' for Python, and in my interpretation this is 'good' according to the review criteria. Could you please check with the editor if listing all dependencies in the documentation is really required? I am willing to add them, but a second list of dependencies in the documentation is likely to be missed when dependencies change.
