### 2026-09-08 134836e
fix: JOSS review issues #766, #767, #769, #770
- Optic.draw() / OpticViewer.view() now accept show: bool = True and
  call plt.show() when not embedding into an existing ax, matching the
  convention already used by every BaseAnalysis.view() (#766).
- Cheat sheet step 12 used Wavefront(...).view(), but Wavefront has no
  view() method -- only its OPD/ZernikeOPD subclasses do (#767).
- Add a concise, code-verified methodology note to the FFT Vectorial
  PSF, Huygens PSF, and Vectorial Huygens PSF gallery notebooks (#769,
  #770).

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
Claude-S
[…]
